Module smithay::backend::allocator

Buffer allocation and management.

Collection of common traits and implementations around buffer creation and handling from various sources.

Allocators provided:

Dumb Buffers through crate::backend::drm::DrmDevice

Gbm Buffers through ::gbm::Device

Buffer types supported:

DumbBuffers

GbmBuffers

DmaBufs

Helpers:

Swapchain to help with buffer management for framebuffers

Modules

dmabuf

Module for dmabuf buffers.

dumbbackend_drm

Module for DumbBuffer buffers

format

Format info tables for DRM formats.

gbmbackend_gbm

Module for Buffers created using libgbm.

vulkanbackend_vulkan

Module for Buffers created using Vulkan.

Structs

Format

Slot

Slot of a swapchain containing an allocated buffer and its userdata.

Swapchain

Swapchain handling a fixed set of re-usable buffers e.g. for scan-out.

UnrecognizedFourcc

Wraps some u32 that isn’t a DRM fourcc we recognize

UnrecognizedVendor

Wraps some u8 that isn’t a DRM vendor we recognize

Enums

Fourcc

Modifier

Vendor

Traits

Allocator

Interface to create Buffers

Buffer

Common trait describing common properties of most types of buffers.