smithay::backend::allocator::format
Function has_alpha

pub const fn has_alpha(fourcc: Fourcc) -> bool

Returns true if the format has an alpha channel.

This function may be useful to know if the alpha channel may need to be swizzled when rendering with some graphics apis.

Unknown formats will always return false.