Struct smithay::backend::allocator::gbm::GbmBuffer

pub struct GbmBuffer<T>where
    T: 'static,{ /* private fields */ }

A GBM buffer object

Implementations

impl<T> BufferObject<T>where
T: 'static,

pub fn width(&self) -> Result<u32, DeviceDestroyedError>

Get the width of the buffer object

pub fn height(&self) -> Result<u32, DeviceDestroyedError>

Get the height of the buffer object

pub fn stride(&self) -> Result<u32, DeviceDestroyedError>

Get the stride of the buffer object

pub fn stride_for_plane(&self, plane: i32) -> Result<u32, DeviceDestroyedError>

Get the stride of the buffer object

pub fn format(&self) -> Result<DrmFourcc, DeviceDestroyedError>

Get the format of the buffer object

pub fn bpp(&self) -> Result<u32, DeviceDestroyedError>

Get the bits per pixel of the buffer object

pub fn offset(&self, plane: i32) -> Result<u32, DeviceDestroyedError>

Get the offset for a plane of the buffer object

pub fn plane_count(&self) -> Result<u32, DeviceDestroyedError>

Get the plane count of the buffer object

pub fn modifier(&self) -> Result<DrmModifier, DeviceDestroyedError>

Get the modifier of the buffer object

pub fn fd(&self) -> Result<i32, DeviceDestroyedError>

Get a DMA-BUF file descriptor for the buffer object

This function creates a DMA-BUF (also known as PRIME) file descriptor handle for the buffer object. Each call to Self::fd() returns a new file descriptor and the caller is responsible for closing the file descriptor.

pub fn handle(&self) -> Result<gbm_bo_handle, DeviceDestroyedError>

Get the handle of the buffer object

This is stored in the platform generic union BufferObjectHandle type. However the format of this handle is platform specific.

pub fn fd_for_plane(&self, plane: i32) -> Result<i32, DeviceDestroyedError>

Get a DMA-BUF file descriptor for a plane of the buffer object

This function creates a DMA-BUF (also known as PRIME) file descriptor handle for a plane of the buffer object. Each call to Self::fd_for_plane() returns a new file descriptor and the caller is responsible for closing the file descriptor.

pub fn handle_for_plane(
&self,
plane: i32
) -> Result<gbm_bo_handle, DeviceDestroyedError>

Get the handle of a plane of the buffer object

This is stored in the platform generic union BufferObjectHandle type. However the format of this handle is platform specific.

pub fn map<'a, D, F, S>(
&'a self,
device: &Device<D>,
x: u32,
y: u32,
width: u32,
height: u32,
f: F
) -> Result<Result<S, Error>, WrongDeviceError>where
D: 'static + AsRawFd,
F: FnOnce(&MappedBufferObject<'a, T>) -> S,

Map a region of a GBM buffer object for cpu access

This function maps a region of a GBM bo for cpu read access.

pub fn map_mut<'a, D, F, S>(
&'a mut self,
device: &Device<D>,
x: u32,
y: u32,
width: u32,
height: u32,
f: F
) -> Result<Result<S, Error>, WrongDeviceError>where
D: 'static + AsRawFd,
F: FnOnce(&mut MappedBufferObject<'a, T>) -> S,

Map a region of a GBM buffer object for cpu access

This function maps a region of a GBM bo for cpu read/write access.

pub fn write(
&mut self,
buffer: &[u8]
) -> Result<Result<(), Error>, DeviceDestroyedError>

Write data into the buffer object

If the buffer object was created with the BufferObjectFlags::WRITE flag, this function can be used to write data into the buffer object. The data is copied directly into the object and it’s the responsibility of the caller to make sure the data represents valid pixel data, according to the width, height, stride and format of the buffer object.

pub fn set_userdata(
&mut self,
userdata: T
) -> Result<Option<T>, DeviceDestroyedError>

Sets the userdata of the buffer object.

If previously userdata was set, it is returned.

pub fn clear_userdata(&mut self) -> Result<(), DeviceDestroyedError>

Clears the set userdata of the buffer object.

pub fn userdata(&self) -> Result<Option<&T>, DeviceDestroyedError>

Returns a reference to set userdata, if any.

pub fn userdata_mut(&mut self) -> Result<Option<&mut T>, DeviceDestroyedError>

Returns a mutable reference to set userdata, if any.

pub fn take_userdata(&mut self) -> Result<Option<T>, DeviceDestroyedError>

Takes ownership of previously set userdata, if any.

This removes the userdata from the buffer object.