Struct smithay::backend::allocator::vulkan::VulkanAllocator

pub struct VulkanAllocator { /* private fields */ }

An allocator which uses Vulkan to create buffers.

Implementations

impl VulkanAllocator

pub const MAX_INSTANCE_VERSION: Version = Version::VERSION_1_3

Maximum supported version instance version that may be used with the allocator.

pub fn required_extensions(phd: &PhysicalDevice) -> Vec<&'static CStr>

Returns the list of device extensions required by the Vulkan allocator.

This function may return a different list for each PhysicalDevice, meaning each device should be filtered using it’s own call to this function.

pub fn new(
phd: &PhysicalDevice,
default_usage: ImageUsageFlags
) -> Result<VulkanAllocator, Error>

Creates a VulkanAllocator.

Panics

If the version of instance which created the PhysicalDevice is higher than VulkanAllocator::MAX_INSTANCE_VERSION.

If the default ImageUsageFlags are empty.

pub fn is_format_supported(
&self,
format: DrmFormat,
usage: ImageUsageFlags
) -> bool

Returns whether this allocator supports the specified format with the usage flags.

pub fn create_buffer_with_usage(
&mut self,
width: u32,
height: u32,
fourcc: DrmFourcc,
modifiers: &[DrmModifier],
usage: ImageUsageFlags
) -> Result<VulkanImage, Error>

Try to create a buffer with the given dimensions, pixel format and usage flags.

This may return Err for one of the following reasons:

The usage is empty.

The fourcc format is not supported.

All of the allowed modifiers are not supported.

The size of the buffer is too large for the usage, fourcc format or modifiers.

The fourcc format and modifiers do not support the specified usage.

pub fn physical_device(&self) -> &PhysicalDevice

Returns the PhysicalDevice this allocator was created with.