Struct smithay::backend::drm::atomic::AtomicDrmDevice

pub struct AtomicDrmDevice<A: AsRawFd + 'static> { /* fields omitted */ }

Open raw drm device utilizing atomic mode-setting

Methods

impl<A: AsRawFd + 'static> AtomicDrmDevice<A>

pub fn new<L>(fd: A, logger: L) -> Result<Self, Error> where
L: Into<Option<Logger>>,

Create a new [AtomicDrmDevice] from an open drm node

Returns an error if the file is no valid drm node or context creation was not successful.