Module smithay::backend::drm::common::fallback

Types to make fallback device initialization easier

Enums

EitherError
Enum uniting two kinds of possible errors.

FallbackDevice
Device Wrapper to assist fallback in case initialization of the preferred device type fails.

FallbackDeviceObserver
SessionObserver Wrapper to assist fallback in case initialization of the preferred device type fails.

FallbackSurface
Surface Wrapper to assist fallback in case initialization of the preferred device type fails.