Enum smithay::backend::drm::common::fallback::FallbackDevice

pub enum FallbackDevice<D1: Device + 'static, D2: Device + 'static> {
    Preference(D1),
    Fallback(D2),
}

Device Wrapper to assist fallback in case initialization of the preferred device type fails.

Variants

Preference(D1)

Variant for successful initialization of the preferred device

Fallback(D2)

Variant for the fallback device

Implementations

impl<A: AsRawFd + Clone + 'static> FallbackDevice<AtomicDrmDevice<A>, LegacyDrmDevice<A>>

pub fn new<L>(fd: A, disable_connectors: bool, logger: L) -> Result<Self, Error> where
L: Into<Option<Logger>>,

Try to initialize an AtomicDrmDevice and fall back to a LegacyDrmDevice if atomic-modesetting is not supported.

Arguments

fd - Open drm node (needs to be clonable to be passed to multiple initializers)

disable_connectors - Setting this to true will initialize all connectors
as disabled on device creation. smithay enables connectors, when attached
to a surface, and disables them, when detached. Setting this to false
requires usage of drm-rs to disable unused connectors to prevent them
showing garbage, but will also prevent flickering of already turned on
connectors (assuming you won't change the resolution).

logger - Optional [slog::Logger] to be used by the resulting device.

Return

Returns an error, if both devices fail to initialize due to fd being no valid drm node or the device being not accessible.

impl<D> FallbackDevice<GbmDevice<D>, EglStreamDevice<D>> where
D: RawDevice + ControlDevice + 'static,

pub fn new<L>(
dev: D,
logger: L
) -> Result<Self, EitherError<GbmError<<<D as Device>::Surface as Surface>::Error>, EglStreamError<<<D as Device>::Surface as Surface>::Error>>> where
L: Into<Option<Logger>>,

Try to initialize a GbmDevice or a EglStreamDevice depending on the used driver.

Arguments

dev - Open drm device (needs implement RawDevice)

logger - Optional [slog::Logger] to be used by the resulting device.

Return

Returns an error, if the choosen device fails to initialize.

impl<D1, D2> FallbackDevice<D1, D2> where
D1: Device + 'static,
<D1 as Device>::Surface: NativeSurface<Error = <<D1 as Device>::Surface as Surface>::Error>,
D2: Device + 'static,
<D2 as Device>::Surface: NativeSurface<Error = <<D2 as Device>::Surface as Surface>::Error>,

pub fn new_egl<B1, B2, L>(
dev: FallbackDevice<D1, D2>,
logger: L
) -> Result<FallbackDevice<EglDevice<B1, D1>, EglDevice<B2, D2>>, EitherError<EglDeviceError<<<D1 as Device>::Surface as Surface>::Error>, EglDeviceError<<<D2 as Device>::Surface as Surface>::Error>>> where
B1: Backend<Surface = <D1 as Device>::Surface, Error = <<D1 as Device>::Surface as Surface>::Error> + 'static,
D1: NativeDisplay<B1, Arguments = (Handle, Mode, Vec<Handle>)>,
B2: Backend<Surface = <D2 as Device>::Surface, Error = <<D2 as Device>::Surface as Surface>::Error> + 'static,
D2: NativeDisplay<B2, Arguments = (Handle, Mode, Vec<Handle>)>,
L: Into<Option<Logger>>,

Try to create a new EglDevice from a FallbackDevice containing two compatible device types.

This helper function is necessary as implementing NativeDevice for FallbackDevice is impossible as selecting the appropriate Backend would be impossible without knowing the underlying device type, that was selected by FallbackDevice.

Returns an error if the context creation was not successful.

pub fn new_egl_with_defaults<B1, B2, L>(
dev: FallbackDevice<D1, D2>,
default_attributes: GlAttributes,
default_requirements: PixelFormatRequirements,
logger: L
) -> Result<FallbackDevice<EglDevice<B1, D1>, EglDevice<B2, D2>>, EitherError<EglDeviceError<<<D1 as Device>::Surface as Surface>::Error>, EglDeviceError<<<D2 as Device>::Surface as Surface>::Error>>> where
B1: Backend<Surface = <D1 as Device>::Surface, Error = <<D1 as Device>::Surface as Surface>::Error> + 'static,
D1: NativeDisplay<B1, Arguments = (Handle, Mode, Vec<Handle>)>,
B2: Backend<Surface = <D2 as Device>::Surface, Error = <<D2 as Device>::Surface as Surface>::Error> + 'static,
D2: NativeDisplay<B2, Arguments = (Handle, Mode, Vec<Handle>)>,
L: Into<Option<Logger>>,

Try to create a new EglDevice from a FallbackDevice containing two compatible device types with the given attributes and requirements as defaults for new surfaces.

This helper function is necessary as implementing NativeDevice for FallbackDevice is impossible as selecting the appropriate Backend would be impossible without knowing the underlying device type, that was selected by FallbackDevice.

Returns an error if the context creation was not successful.