smithay::backend::drm::compositor

Struct DrmCompositor

pub struct DrmCompositor<A, F, U, G>
where
    A: Allocator,
    F: ExportFramebuffer<A::Buffer>,
    <F as ExportFramebuffer<A::Buffer>>::Framebuffer: Debug + Send + Sync + 'static,
    G: AsFd + 'static,
{ /* private fields */ }

Available on crate features wayland_frontend and backend_gbm and backend_drm only.

Composite an output using a combination of planes and rendering

see the module docs for more information

Implementations

impl<A, F, U, G> DrmCompositor<A, F, U, G>
where A: Allocator, <A as Allocator>::Error: Error + Send + Sync, <A as Allocator>::Buffer: AsDmabuf, <A::Buffer as AsDmabuf>::Error: Error + Send + Sync + Debug, F: ExportFramebuffer<A::Buffer>, <F as ExportFramebuffer<A::Buffer>>::Framebuffer: Debug + Send + Sync + 'static, <F as ExportFramebuffer<A::Buffer>>::Error: Error + Send + Sync, G: AsFd + Clone,

pub fn new( output_mode_source: impl Into<OutputModeSource> + Debug, surface: DrmSurface, planes: Option<Planes>, allocator: A, framebuffer_exporter: F, color_formats: impl IntoIterator<Item = DrmFourcc>, renderer_formats: impl IntoIterator<Item = DrmFormat>, cursor_size: Size<u32, BufferCoords>, gbm: Option<GbmDevice<G>>, ) -> Result<Self, FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Initialize a new DrmCompositor.

The OutputModeSource can be created from an Output, which will automatically track the output’s mode changes. An OutputModeSource::Static variant should only be used when manually updating modes using DrmCompositor::set_output_mode_source.

output_mode_source is used to determine the current mode, scale and transform

surface for the compositor to use

planes defines which planes the compositor is allowed to use for direct scan-out. None will result in the compositor to use all planes as specified by DrmSurface::planes

allocator used for the primary plane swapchain

color_formats are tested in order until a working configuration is found

renderer_formats as reported by the used renderer, used to build the intersection between the possible scan-out formats of the primary plane and the renderer

framebuffer_exporter is used to create drm framebuffers for the swapchain buffers (and if possible for element buffers) for scan-out

cursor_size as reported by the drm device, used for creating buffer for the cursor plane

gbm device used for creating buffers for the cursor plane, None will disable the cursor plane

pub fn with_format( output_mode_source: impl Into<OutputModeSource> + Debug, surface: DrmSurface, planes: Option<Planes>, allocator: A, framebuffer_exporter: F, code: DrmFourcc, modifiers: impl IntoIterator<Item = DrmModifier>, cursor_size: Size<u32, BufferCoords>, gbm: Option<GbmDevice<G>>, ) -> Result<Self, FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Initialize a new DrmCompositor with a pre-selected format.

The OutputModeSource can be created from an Output, which will automatically track the output’s mode changes. An OutputModeSource::Static variant should only be used when manually updating modes using DrmCompositor::set_output_mode_source.

output_mode_source is used to determine the current mode, scale and transform

surface for the compositor to use

planes defines which planes the compositor is allowed to use for direct scan-out. None will result in the compositor to use all planes as specified by DrmSurface::planes

allocator used for the primary plane swapchain

framebuffer_exporter is used to create drm framebuffers for the swapchain buffers (and if possible for element buffers) for scan-out

code is the fixed format to initialize the framebuffer with

modifiers is the set of modifiers allowed, when allocating buffers with the specified color format

cursor_size as reported by the drm device, used for creating buffer for the cursor plane

gbm device used for creating buffers for the cursor plane, None will disable the cursor plane

pub fn render_frame<'a, R, E>( &mut self, renderer: &mut R, elements: &'a [E], clear_color: impl Into<Color32F>, frame_flags: FrameFlags, ) -> Result<RenderFrameResult<'a, A::Buffer, F::Framebuffer, E>, RenderFrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error, <R as RendererSuper>::Error>>
where E: RenderElement<R>, R: Renderer + Bind<Dmabuf>, R::TextureId: Texture + 'static,

Render the next frame

elements for this frame in front-to-back order

frame_flags specifies techniques allowed to realize the frame

pub fn queue_frame( &mut self, user_data: U, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Queues the current frame for scan-out.

If render_frame has not been called prior to this function or returned no damage this function will return FrameError::EmptyFrame. Instead of calling queue_frame it is the callers responsibility to re-schedule the frame. A simple strategy for frame re-scheduling is to queue a one-shot timer that will trigger after approximately one retrace duration.

Note: It is your responsibility to synchronize rendering if the RenderFrameResult returned by the previous render_frame call returns true on RenderFrameResult::needs_sync.

Note: This function needs to be followed up with DrmCompositor::frame_submitted when a vblank event is received, that denotes successful scan-out of the frame. Otherwise the underlying swapchain will eventually run out of buffers.

user_data can be used to attach some data to a specific buffer and later retrieved with DrmCompositor::frame_submitted

pub fn commit_frame( &mut self, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Commits the current frame for scan-out.

If render_frame has not been called prior to this function or returned no damage this function will return FrameError::EmptyFrame. Instead of calling commit_frame it is the callers responsibility to re-schedule the frame. A simple strategy for frame re-scheduling is to queue a one-shot timer that will trigger after approximately one retrace duration.

Note: It is your responsibility to synchronize rendering if the RenderFrameResult returned by the previous render_frame call returns true on RenderFrameResult::needs_sync.

Note: This function should not be followed up with DrmCompositor::frame_submitted and will not generate a vblank event on the underlying device.

pub fn reset_state(&mut self) -> Result<(), DrmError>

Re-evaluates the current state of the crtc and forces calls to render_frame to return false for RenderFrameResult::is_empty until a frame is queued with queue_frame.

It is recommended to call this function after this used Session gets re-activated / VT switched to.

Usually you do not need to call this in other circumstances, but if the state of the crtc is modified elsewhere, you may call this function to reset it’s internal state.

pub fn frame_submitted( &mut self, ) -> Result<Option<U>, FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Marks the current frame as submitted.

Note: Needs to be called, after the vblank event of the matching DrmDevice was received after calling DrmCompositor::queue_frame on this surface. Otherwise the underlying swapchain will run out of buffers eventually.

pub fn reset_buffers(&mut self)

Reset the underlying buffers

pub fn reset_buffer_ages(&mut self)

Reset the age for all buffers.

This can be used to efficiently clear the damage history without having to modify the damage for each surface.

pub fn crtc(&self) -> Handle

Returns the underlying crtc of this surface

pub fn plane(&self) -> Handle

Returns the underlying plane of this surface

pub fn current_connectors(&self) -> impl IntoIterator<Item = Handle>

Currently used connectors of this Surface

pub fn pending_connectors(&self) -> impl IntoIterator<Item = Handle>

Returns the pending connectors used for the next frame queued via queue_frame.

pub fn add_connector( &self, connector: Handle, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Tries to add a new connector to be used after the next commit.

Warning: You need to make sure, that the connector is not used with another surface or was properly removed via remove_connector + commit before adding it to another surface. Behavior if failing to do so is undefined, but might result in rendering errors or the connector getting removed from the other surface without updating it’s internal state.

Fails if the connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn remove_connector( &self, connector: Handle, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Tries to mark a connector for removal on the next commit.

pub fn set_connectors( &self, connectors: &[Handle], ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Tries to replace the current connector set with the newly provided one on the next commit.

Fails if one new connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn current_mode(&self) -> Mode

Returns the currently active Mode of the underlying crtc

pub fn pending_mode(&self) -> Mode

Returns the currently pending Mode to be used after the next commit.

pub fn use_mode( &mut self, mode: Mode, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Tries to set a new Mode to be used after the next commit.

Fails if the mode is not compatible with the underlying crtc or any of the pending connectors.

pub fn vrr_supported( &self, conn: Handle, ) -> Result<VrrSupport, FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Returns if Variable Refresh Rate is advertised as supported by the given connector.

See DrmSurface::vrr_supported for more details.

pub fn vrr_enabled(&self) -> bool

Returns if Variable Refresh Rate is currently enabled for frames composed by this DrmCompositor.

pub fn use_vrr( &mut self, vrr: bool, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Tries to set variable refresh rate (VRR) for the next frame.

Doing so might cause the next frame to trigger a modeset. Check DrmCompositor::vrr_supported, which indicates if VRR can be used without a modeset on the attached connectors.

pub fn set_debug_flags(&mut self, flags: DebugFlags)

Set the DebugFlags to use

Note: This will reset the primary plane swapchain if the flags differ from the current flags

pub fn debug_flags(&self) -> DebugFlags

Returns the current enabled DebugFlags

pub fn surface(&self) -> &DrmSurface

Returns a reference to the underlying drm surface

pub fn format(&self) -> DrmFourcc

Get the format of the underlying swapchain

pub fn modifiers(&self) -> &[DrmModifier]

Get the allowed modifiers of the underlying swapchain

pub fn set_format( &mut self, allocator: A, code: DrmFourcc, modifiers: impl IntoIterator<Item = DrmModifier>, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Reset the underlying swapchain and assign a new color format.

pub fn set_output_mode_source(&mut self, output_mode_source: OutputModeSource)

Change the output mode source.

pub fn clear(&mut self) -> Result<(), DrmError>

Clear the surface, setting DPMS state to off, disabling all planes, and clearing the pending frame.

Calling queue_frame will re-enable.