Struct smithay::backend::drm::compositor::RenderFrameResult

pub struct RenderFrameResult<'a, B: Buffer, F: AsRef<Handle>, E> {
    pub damage: Option<Vec<Rectangle<i32, Physical>>>,
    pub states: RenderElementStates,
    pub primary_element: PrimaryPlaneElement<'a, B, F, E>,
    pub overlay_elements: Vec<&'a E>,
    pub cursor_element: Option<&'a E>,
    /* private fields */
}

Result for DrmCompositor::render_frame

Note: This struct may contain a reference to the composited buffer of the primary display plane. Dropping it will remove said reference and allows the buffer to be reused.

Keeping the buffer longer may cause the following issues:

Too much damage - until the buffer is marked free it is not considered submitted by the swapchain, causing the age value of newly queried buffers to be lower than necessary, potentially resulting in more rendering than necessary. To avoid this make sure the buffer is dropped before starting the next render.

Exhaustion of swapchain images - Continuing rendering while holding on to too many buffers may cause the swapchain to run out of images, returning errors on rendering until buffers are freed again. The exact amount of images in a swapchain is an implementation detail, but should generally be expect to be large enough to hold onto at least one RenderFrameResult.

Fields
damage: Option<Vec<Rectangle<i32, Physical>>>

Damage of this frame

states: RenderElementStates

The render element states of this frame

primary_element: PrimaryPlaneElement<'a, B, F, E>

Element for the primary plane

overlay_elements: Vec<&'a E>

Overlay elements in front to back order

cursor_element: Option<&'a E>

Optional cursor plane element

If set always above all other elements

Implementations

impl<'a, B, F, E> RenderFrameResult<'a, B, F, E>where B: Buffer, F: AsRef<Handle>,

pub fn damage_from_age( &self, damage_tracker: &mut OutputDamageTracker, age: usize, filter: impl IntoIterator<Item = Id> ) -> Result<(Option<Vec<Rectangle<i32, Physical>>>, RenderElementStates), OutputNoMode>where E: Element,

Get the damage of this frame for the specified dtr and age

impl<'a, B, F, E> RenderFrameResult<'a, B, F, E>where B: Buffer + AsDmabuf, <B as AsDmabuf>::Error: Debug, F: AsRef<Handle>,

pub fn blit_frame_result<R>( &self, size: impl Into<Size<i32, Physical>>, transform: Transform, scale: impl Into<Scale<f64>>, renderer: &mut R, damage: impl IntoIterator<Item = Rectangle<i32, Physical>>, filter: impl IntoIterator<Item = Id> ) -> Result<(), BlitFrameResultError<<R as Renderer>::Error, <B as AsDmabuf>::Error>>where R: Renderer + Blit<Dmabuf>, <R as Renderer>::TextureId: 'static, E: Element + RenderElement<R>,

Blit the frame result into a currently bound buffer