Struct smithay::backend::drm::eglstream::EglStreamDevice

pub struct EglStreamDevice<D: RawDevice + ControlDevice + 'static> { /* fields omitted */ }

Representation of an open egl stream device to create egl rendering surfaces

Methods

impl<D: RawDevice + ControlDevice + 'static> EglStreamDevice<D>

pub fn new<L>(
raw: D,
logger: L
) -> Result<Self, Error<<<D as Device>::Surface as Surface>::Error>> where
L: Into<Option<Logger>>,

Try to create a new [EglStreamDevice] from an open device.

Returns an error if the underlying device would not support the required EGL extensions.