Enum smithay::backend::drm::error::ErrorKind

pub enum ErrorKind {
    Msg(String),
    EGL(ErrorKind),
    DrmMasterFailed,
    DrmDev(String),
    UnableToGetDeviceId,
    GbmInitFailed,
    FailedToSwap,
    DeviceInactive,
    ModeNotSuitable(Mode),
    CrtcAlreadyInUse(Handle),
    NoSuitableEncoder(Info, Handle),
    // some variants omitted
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

EGL(ErrorKind)

EGL error

DrmMasterFailed

Unable to acquire drm master

DrmDev(String)

The DrmDevice encountered an access error

UnableToGetDeviceId

Unable to determine device id of drm device

GbmInitFailed

Creation of gbm resource failed

FailedToSwap

Swapping front buffers failed

DeviceInactive

Device is currently paused

ModeNotSuitable(Mode)

Mode is not compatible with all given connectors

CrtcAlreadyInUse(Handle)

The given crtc is already in use by another backend

NoSuitableEncoder(Info, Handle)

No encoder was found for a given connector on the set crtc

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.