smithay::backend::drm::gbm
Function framebuffer_from_wayland_buffer

pub fn framebuffer_from_wayland_buffer<A: AsFd + 'static>(
    drm: &DrmDeviceFd,
    gbm: &Device<A>,
    buffer: &WlBuffer,
    use_opaque: bool,
) -> Result<Option<GbmFramebuffer>, Error>

Available on crate features backend_drm and backend_gbm and wayland_frontend only.

Attach a framebuffer for a WlBuffer

This tries to import the buffer to gbm and attach a framebuffer::Handle for the imported [BufferObject].

Returns Ok(None) for unknown buffer types and buffer types that do not support attaching a framebuffer (e.g. shm-buffers)