Struct smithay::backend::drm::gbm::GbmSurface

pub struct GbmSurface<D: RawDevice + 'static>(_);

Gbm surface for rendering

Methods

impl<D: RawDevice + 'static> GbmSurface<D>

pub unsafe fn page_flip(&self) -> Result<(), Self::Error>

Flips the underlying buffers.

The surface will report being already flipped until the matching event was processed either by calling [Device::process_events] manually after the flip (bad idea performance-wise) or by binding the device to an event-loop by using device_bind.

Note: This might trigger a full modeset on the underlying device, potentially causing some flickering. In that case this operation is blocking until the crtc is in the desired state.

Safety

When used in conjunction with an EGL context, this must be called exactly once after page-flipping the associated context.

pub fn recreate(&self) -> Result<(), Self::Error>

Recreate underlying gbm resources.

This recreates the gbm surfaces resources, which might be needed after e.g. calling Surface::use_mode. You may check if your [GbmSurface] needs recreation through needs_recreation.

pub fn needs_recreation(&self) -> bool

Check if underlying gbm resources need to be recreated.