Module smithay::backend::drm::legacy

RawDevice and RawSurface implementations using the legacy mode-setting infrastructure.

Legacy mode-setting refers to the now outdated, but still supported direct manager api of the linux kernel. Adaptations of this api can be found in BSD kernels.

The newer and objectively better api is known as atomic-modesetting (or nuclear-page-flip), however this api is not supported by every driver, so this is provided for backwards compatibility. Currenly there are no features in smithay, that are exclusive to the atomic api.

Usually this implementation will be wrapped into a GbmDevice. Take a look at anvils source code for an example of this.

For an example how to use this standalone, take a look at the raw_legacy_drm example.

For detailed overview of these abstractions take a look at the module documentation of backend::drm.

Modules

session
Support to register an open LegacyDrmDevice to an open Session.

Structs

LegacyDrmDevice
Open raw drm device utilizing legacy mode-setting

LegacyDrmSurface
Open raw crtc utilizing legacy mode-setting