Struct smithay::backend::drm::legacy::LegacyDrmDevice

pub struct LegacyDrmDevice<A: AsRawFd + 'static> { /* fields omitted */ }

Open raw drm device utilizing legacy mode-setting

Methods

impl<A: AsRawFd + 'static> LegacyDrmDevice<A>

pub fn new<L>(dev: A, logger: L) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new LegacyDrmDevice from an open drm node

Returns an error if the file is no valid drm node or context creation was not successful.