smithay::backend::drm::output

Struct DrmOutput

pub struct DrmOutput<A, F, U, G>
where
    A: Allocator,
    F: ExportFramebuffer<<A as Allocator>::Buffer>,
    <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Framebuffer: Debug + Send + Sync + 'static,
    G: AsFd + 'static,
{ /* private fields */ }

Available on crate features wayland_frontend and backend_gbm and backend_drm only.

A handle to an underlying DrmCompositor handled by an DrmOutputManager.

Implementations

impl<A, F, U, G> DrmOutput<A, F, U, G>
where A: Allocator + Clone + Debug, <A as Allocator>::Buffer: AsDmabuf, <A as Allocator>::Error: Send + Sync + 'static, <<A as Allocator>::Buffer as AsDmabuf>::Error: Send + Sync + 'static, F: ExportFramebuffer<<A as Allocator>::Buffer> + Clone, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Framebuffer: Debug + Send + Sync + 'static, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error: Send + Sync + 'static, G: AsFd + Clone + 'static, U: 'static,

pub fn set_debug_flags(&self, flags: DebugFlags)

Set the DebugFlags to use

Note: This will reset the primary plane swapchain if the flags differ from the current flags

pub fn reset_buffers(&self)

Reset the underlying buffers

pub fn frame_submitted( &self, ) -> Result<Option<U>, FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Marks the current frame as submitted.

Note: Needs to be called, after the vblank event of the matching DrmDevice was received after calling DrmOutput::queue_frame on this surface. Otherwise the underlying swapchain will run out of buffers eventually.

pub fn format(&self) -> DrmFourcc

Get the format of the underlying swapchain

pub fn render_frame<'a, R, E>( &mut self, renderer: &mut R, elements: &'a [E], clear_color: impl Into<Color32F>, frame_mode: FrameFlags, ) -> Result<RenderFrameResult<'a, A::Buffer, F::Framebuffer, E>, RenderFrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error, <R as RendererSuper>::Error>>
where E: RenderElement<R>, R: Renderer + Bind<Dmabuf>, R::TextureId: Texture + 'static, R::Error: Send + Sync + 'static,

Render the next frame

elements for this frame in front-to-back order

frame_mode specifies techniques allowed to realize the frame

pub fn queue_frame( &mut self, user_data: U, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Queues the current frame for scan-out.

If render_frame has not been called prior to this function or returned no damage this function will return FrameError::EmptyFrame. Instead of calling queue_frame it is the callers responsibility to re-schedule the frame. A simple strategy for frame re-scheduling is to queue a one-shot timer that will trigger after approximately one retrace duration.

Note: It is your responsibility to synchronize rendering if the RenderFrameResult returned by the previous render_frame call returns true on RenderFrameResult::needs_sync.

Note: This function needs to be followed up with DrmOutput::frame_submitted when a vblank event is received, that denotes successful scan-out of the frame. Otherwise the underlying swapchain will eventually run out of buffers.

user_data can be used to attach some data to a specific buffer and later retrieved with DrmCompositor::frame_submitted

pub fn commit_frame( &mut self, ) -> Result<(), FrameError<<A as Allocator>::Error, <<A as Allocator>::Buffer as AsDmabuf>::Error, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Error>>

Commits the current frame for scan-out.

If render_frame has not been called prior to this function or returned no damage this function will return FrameError::EmptyFrame. Instead of calling commit_frame it is the callers responsibility to re-schedule the frame. A simple strategy for frame re-scheduling is to queue a one-shot timer that will trigger after approximately one retrace duration.

Note: It is your responsibility to synchronize rendering if the RenderFrameResult returned by the previous render_frame call returns true on RenderFrameResult::needs_sync.

Note: This function should not be followed up with DrmOutput::frame_submitted and will not generate a vblank event on the underlying device.

pub fn use_mode<R, E>( &mut self, mode: Mode, renderer: &mut R, render_elements: &DrmOutputRenderElements<R, E>, ) -> DrmOutputManagerResult<(), A, F, R>
where E: RenderElement<R>, R: Renderer + Bind<Dmabuf>, R::TextureId: Texture + 'static, R::Error: Send + Sync + 'static,

Tries to apply a new Mode for this DrmOutput.

Fails if the mode is not compatible with the underlying crtc or any of the pending connectors.

This might cause commits on other surfaces to meet the bandwidth requirements of the new mode by temporarily disabling additional planes and forcing composition.

impl<A, F, U, G> DrmOutput<A, F, U, G>
where A: Allocator, F: ExportFramebuffer<<A as Allocator>::Buffer>, <F as ExportFramebuffer<<A as Allocator>::Buffer>>::Framebuffer: Debug + Send + Sync + 'static, G: AsFd + 'static,

pub fn crtc(&self) -> Handle

Returns the underlying crtc of this surface

pub fn with_compositor<T, R>(&self, f: T) -> R
where T: FnOnce(&mut DrmCompositor<A, F, U, G>) -> R,

Provides exclusive access to the underlying DrmCompositor