Struct smithay::backend::drm::DrmBackend

pub struct DrmBackend<A: Device + 'static> { /* fields omitted */ }

Backend based on a DrmDevice and a given crtc

Methods

impl<A: Device + 'static> DrmBackend<A>

pub fn add_connector(&mut self, connector: Handle) -> Result<()>

Add a connector to backend

Errors

Errors if the new connector does not support the currently set Mode

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn used_connectors(&self) -> &[Handle]

Returns the currently set connectors

pub fn remove_connector(&mut self, connector: Handle)

Removes a currently set connector

pub fn current_mode(&self) -> Mode

Gets the currently used mode

pub fn use_mode(&mut self, mode: Mode) -> Result<()>

Changes the currently set mode

Errors

This will fail if not all set connectors support the new Mode. Several internal resources will need to be recreated to fit the new Mode. Other errors might occur.

pub fn crtc(&self) -> Handle

Returns the crtc id used by this backend