Struct smithay::backend::drm::DrmDevice

pub struct DrmDevice<A: ControlDevice + 'static> { /* fields omitted */ }

Representation of an open drm device node to create rendering backends

Methods

impl<A: ControlDevice + 'static> DrmDevice<A>

pub fn new<L>(dev: A, logger: L) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new DrmDevice from an open drm node

Returns an error if the file is no valid drm node or context creation was not successful.

pub fn new_with_gl_attr<L>(
dev: A,
attributes: GlAttributes,
logger: L
) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new DrmDevice from an open drm node and given GlAttributes

Returns an error if the file is no valid drm node or context creation was not successful.

pub fn create_backend<I>(
&mut self,
crtc: Handle,
mode: Mode,
connectors: I
) -> Result<DrmBackend<A>> where
I: Into<Vec<Handle>>,

Create a new backend on a given crtc with a given Mode for a given amount of connectors (mirroring).

Errors if initialization fails or the mode is not available on all given connectors.

pub fn device_id(&self) -> u64

Returns an internal device id, that is unique per boot per system