smithay::backend::drm

Struct DrmDeviceFd

pub struct DrmDeviceFd(/* private fields */);

Available on crate feature backend_drm only.

Ref-counted file descriptor of an open drm device

Implementations

impl DrmDeviceFd

pub fn new(fd: DeviceFd) -> DrmDeviceFd

Create a new DrmDeviceFd.

This function will try to acquire the master lock for the underlying drm device and release the lock on drop again. For that reason you should never create multiple DrmDeviceFd out of the same DeviceFd, but instead clone the DrmDeviceFd.

Failing to do so might fail to acquire set lock and release it early, which can cause some drm ioctls to fail later.

pub fn device_fd(&self) -> DeviceFd

Returns the underlying DeviceFd

pub fn dev_id(&self) -> Result<dev_t>

Returns the dev_t of the underlying device

pub fn downgrade(&self) -> WeakDrmDeviceFd

Returns a weak reference to the underlying device