Struct smithay::backend::drm::DrmRenderSurface

pub struct DrmRenderSurface<D: AsRawFd + 'static, A: Allocator<B>, R: Bind<Dmabuf>, B: Buffer> { /* fields omitted */ }

Simplified by limited abstraction to link single DrmSurfaces to renderers.

Use-case

In some scenarios it might be enough to use of a drm-surface as the one and only target of a single renderer. In these cases DrmRenderSurface provides a way to quickly get up and running without manually handling and binding buffers.

Implementations

impl<D, A, B, R, E1, E2, E3> DrmRenderSurface<D, A, R, B> where
D: AsRawFd + 'static,
A: Allocator<B, Error = E1>,
B: Buffer + AsDmabuf<Error = E2>,
R: Bind<Dmabuf> + Renderer<Error = E3>,
E1: Error + 'static,
E2: Error + 'static,
E3: Error + 'static,

pub fn new<L: Into<Option<Logger>>>(
drm: DrmSurface<D>,
allocator: A,
renderer: R,
log: L
) -> Result<DrmRenderSurface<D, A, R, B>, Error<E1, E2, E3>>

Create a new DrmRendererSurface from a given compatible combination of a surface, an allocator and a renderer.

To sucessfully call this function, you need to have a renderer, which can render into a Dmabuf, and an allocator, which can create a buffer type, which can be converted into a Dmabuf.

The function will futhermore check for compatibility by enumerating supported pixel formats and choosing an appropriate one.

pub fn renderer(&mut self) -> &mut R

Access the underlying renderer

pub fn render<F, S>(&mut self, rendering: F) -> Result<S, Error<E1, E2, E3>> where
F: FnOnce(&mut R, &mut <R as Renderer>::Frame) -> S,

Shortcut to Renderer::render with the pending mode as dimensions and this surface set a the rendering target.

pub fn frame_submitted(&mut self) -> Result<(), Error<E1, E2, E3>>

Marks the current frame as submitted.

Needs to be called, after the vblank event of the matching DrmDevice was received after calling DrmRenderSurface::render on this surface. Otherwise the rendering will run out of buffers eventually.

pub fn crtc(&self) -> Handle

Returns the underlying crtc of this surface

pub fn plane(&self) -> Handle

Returns the underlying plane of this surface

pub fn current_connectors(&self) -> impl IntoIterator<Item = Handle>

Currently used connectors of this Surface

pub fn pending_connectors(&self) -> impl IntoIterator<Item = Handle>

Returns the pending connectors used for the next frame in render

pub fn add_connector(&self, connector: Handle) -> Result<(), Error<E1, E2, E3>>

Tries to add a new connector to be used after the next commit.

Warning: You need to make sure, that the connector is not used with another surface or was properly removed via remove_connector + commit before adding it to another surface. Behavior if failing to do so is undefined, but might result in rendering errors or the connector getting removed from the other surface without updating it’s internal state.

Fails if the connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn remove_connector(
&self,
connector: Handle
) -> Result<(), Error<E1, E2, E3>>

Tries to mark a connector for removal on the next commit.

pub fn set_connectors(
&self,
connectors: &[Handle]
) -> Result<(), Error<E1, E2, E3>>

Tries to replace the current connector set with the newly provided one on the next commit.

Fails if one new connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn current_mode(&self) -> Mode

Returns the currently active Mode of the underlying crtc

pub fn pending_mode(&self) -> Mode

Returns the currently pending Mode to be used after the next commit.

pub fn use_mode(&self, mode: Mode) -> Result<(), Error<E1, E2, E3>>

Tries to set a new Mode to be used after the next commit.

Fails if the mode is not compatible with the underlying crtc or any of the pending connectors.