Trait smithay::backend::drm::Device

pub trait Device: AsRawFd + DevPath {
    type Surface: Surface;
    fn device_id(&self) -> dev_t;

    fn set_handler(
        &mut self, 
        handler: impl DeviceHandler<Device = Self> + 'static
    );

    fn clear_handler(&mut self);

    fn create_surface(
        &mut self, 
        crtc: Handle
    ) -> Result<Self::Surface, <Self::Surface as Surface>::Error>;

    fn process_events(&mut self);

    fn resource_handles(
        &self
    ) -> Result<ResourceHandles, <Self::Surface as Surface>::Error>;

    fn get_connector_info(&self, conn: Handle) -> Result<Info, DrmError>;

    fn get_crtc_info(&self, crtc: Handle) -> Result<Info, DrmError>;

    fn get_encoder_info(&self, enc: Handle) -> Result<Info, DrmError>;

    fn get_framebuffer_info(&self, fb: Handle) -> Result<Info, DrmError>;

    fn get_plane_info(&self, plane: Handle) -> Result<Info, DrmError>;
}

An open drm device

Associated Types

type Surface: Surface

Associated [Surface] of this [Device] type

Loading content...
Required methods

fn device_id(&self) -> dev_t

Returns the id of this device node.

fn set_handler(&mut self, handler: impl DeviceHandler<Device = Self> + 'static)

Assigns a [DeviceHandler] called during event processing.

See [device_bind] and [DeviceHandler]

fn clear_handler(&mut self)

Clear a set DeviceHandler, if any

fn create_surface(
&mut self,
crtc: Handle
) -> Result<Self::Surface, <Self::Surface as Surface>::Error>

Creates a new rendering surface.

Initialization of surfaces happens through the types provided by drm-rs.

crtcs represent scanout engines of the device pointer to one framebuffer. Their responsibility is to read the data of the framebuffer and export it into an "Encoder". The number of crtc's represent the number of independant output devices the hardware may handle.

fn process_events(&mut self)

Processes any open events of the underlying file descriptor.

You should not call this function manually, but rather use [device_bind] to register the device to an EventLoop to synchronize your rendering to the vblank events of the open crtc's

fn resource_handles(
&self
) -> Result<ResourceHandles, <Self::Surface as Surface>::Error>

Attempts to acquire a copy of the [Device]'s ResourceHandle

fn get_connector_info(&self, conn: Handle) -> Result<Info, DrmError>

Retrieve the information for a connector

fn get_crtc_info(&self, crtc: Handle) -> Result<Info, DrmError>

Retrieve the information for a crtc

fn get_encoder_info(&self, enc: Handle) -> Result<Info, DrmError>

Retrieve the information for an encoder

fn get_framebuffer_info(&self, fb: Handle) -> Result<Info, DrmError>

Retrieve the information for a framebuffer

fn get_plane_info(&self, plane: Handle) -> Result<Info, DrmError>

Retrieve the information for a plane

Loading content...