Trait smithay::backend::drm::DrmHandler

pub trait DrmHandler<A: ControlDevice + 'static> {
    fn ready(
        &mut self, 
        device: &mut DrmDevice<A>, 
        crtc: Handle, 
        frame: u32, 
        duration: Duration
    );

    fn error(&mut self, device: &mut DrmDevice<A>, error: DrmError);
}

Handler for drm node events

See module-level documentation for its use

Required Methods

fn ready(
&mut self,
device: &mut DrmDevice<A>,
crtc: Handle,
frame: u32,
duration: Duration
)

The DrmBackend of crtc has finished swapping buffers and new frame can now (and should be immediately) be rendered.

fn error(&mut self, device: &mut DrmDevice<A>, error: DrmError)

The DrmDevice has thrown an error.

The related backends are most likely not usable anymore and the whole stack has to be recreated..