Trait smithay::backend::drm::RawSurface

pub trait RawSurface: Surface + ControlDevice + BasicDevice {
    fn commit_pending(&self) -> bool;

    fn commit(&self, framebuffer: Handle) -> Result<(), Self::Error>;

    fn page_flip(&self, framebuffer: Handle) -> Result<(), SwapBuffersError>;
}

An open bare crtc without any rendering abstractions

Required methods

fn commit_pending(&self) -> bool

Returns true whenever any state changes are pending to be commited

The following functions may trigger a pending commit:

add_connector

remove_connector

use_mode

fn commit(&self, framebuffer: Handle) -> Result<(), Self::Error>

Commit the pending state rendering a given framebuffer.

Note: This will trigger a full modeset on the underlying device, potentially causing some flickering. Check before performing this operation if a commit really is necessary using commit_pending.

This operation is blocking until the crtc is in the desired state.

fn page_flip(&self, framebuffer: Handle) -> Result<(), SwapBuffersError>

Page-flip the underlying crtc to a new given [framebuffer].

This will not cause the crtc to modeset.

This operation is not blocking and will produce a vblank event once swapping is done. Make sure to set a DeviceHandler and register the belonging Device before to receive the event in time.

Loading content...