smithay::backend
Module egl

Available on crate feature backend_egl only.

Common traits and types for egl rendering

This module has multiple responsibilities related to functionality provided by libEGL: Initializing EGL objects to:

initialize usage of EGL based WlBuffers via wl_drm.

initialize OpenGL contexts from.

Import/Export external resources to/from OpenGL

To use this module, you first need to create an EGLDisplay through a supported EGL platform as indicated by an implementation of the native::EGLNativeDisplay trait.

You may bind the EGLDisplay, that shall be used by clients for rendering (so pick one initialized by a fast platform) to the wayland_server::Display of your compositor. Note only one backend may be bound to any Display at any time.

You may then use the resulting display::EGLBufferReader to receive EGLBuffer of an EGL-based WlBuffer for rendering. Renderers implementing the ImportEgl-trait can manage the buffer reader for you.

Note: The support for binding the EGLDisplay for use by clients requires the use_system_lib cargo feature on Smithay.

To create OpenGL contexts you may create EGLContexts from the display and if the context is initialized with a config it may also be used to initialize an EGLSurface, which can be bound to some renderers.

Alternatively you may import dmabufs using the display, which result in an EGLImage, which can be rendered into by OpenGL. This is preferable to using surfaces as the dmabuf can be passed around freely making resource-management and more complex use-cases like Multi-GPU rendering easier to manage. Renderers based on EGL may support doing this for you by allowing you to Bind a dmabuf directly.

Re-exports

pub use self::context::EGLContext;

pub use self::display::EGLDisplay;

pub use self::surface::EGLSurface;

Modules

context

EGL context related structs

display

Type safe native types for safe egl initialisation

fence

EGL fence related structs

ffi

native

Type safe native types for safe context/surface creation

surface

EGL surface related structs

Structs

EGLBufferwayland_frontend and use_system_lib

Images of the EGL-based WlBuffer.

EGLDevice

safe EGLDevice wrapper

EglExtensionNotSupportedError

Error that can happen on optional EGL features

MakeCurrentError

Error that can happen when making a context (and surface) current on the active thread.

Enums

BufferAccessErrorwayland_frontend

Error that can occur when accessing an EGL buffer

EGLError

Raw EGL error

Error

EGL errors

Format

Texture format types

SwapBuffersError

Error that can happen when swapping buffers.

Functions

get_proc_address⚠

Returns the address of an OpenGL function.

wrap_egl_call

Wraps a raw egl call and returns error codes from eglGetError(), only if the result of the call is different from the err value.

wrap_egl_call_bool

Wraps a raw egl call and returns error codes from eglGetError(), only if the EGLBoolean returned is EGL_FALSE.

wrap_egl_call_ptr

Wraps a raw egl call and returns error codes from eglGetError(), only if the pointer returned is null.