smithay::backend::egl::context

Struct EGLContext

pub struct EGLContext { /* private fields */ }

Available on crate feature backend_egl only.

EGL context for rendering

Implementations

impl EGLContext

pub unsafe fn from_raw( display: *const c_void, config_id: *const c_void, context: *const c_void, ) -> Result<EGLContext, Error>

Creates a new EGLContext from raw handles.

Safety

The context must be created from the system default EGL library (dlopen("libEGL.so"))

The display, config, and context must be valid for the lifetime of the returned context.

pub fn new(display: &EGLDisplay) -> Result<EGLContext, Error>

Creates a new configless EGLContext from a given EGLDisplay

pub fn new_with_priority( display: &EGLDisplay, priority: ContextPriority, ) -> Result<EGLContext, Error>

Creates a new configless EGLContext with the specified priority from a given EGLDisplay

Note: The priority is a hint that might be ignored by the underlying platform. It also requires EGL_IMG_context_priority to be available, otherwise the priority will be ignored.

pub fn new_with_config( display: &EGLDisplay, attributes: GlAttributes, reqs: PixelFormatRequirements, ) -> Result<EGLContext, Error>

Create a new EGLContext from a given EGLDisplay and configuration requirements

pub fn new_with_config_and_priority( display: &EGLDisplay, attributes: GlAttributes, reqs: PixelFormatRequirements, priority: ContextPriority, ) -> Result<EGLContext, Error>

Create a new EGLContext from a given EGLDisplay, configuration requirements and priority

Note: The priority is a hint that might be ignored by the underlying platform. It also requires EGL_IMG_context_priority to be available, otherwise the priority will be ignored.

pub fn new_shared( display: &EGLDisplay, share: &EGLContext, ) -> Result<EGLContext, Error>

Create a new configless EGLContext from a given EGLDisplay sharing resources with another context

pub fn new_shared_with_priority( display: &EGLDisplay, share: &EGLContext, priority: ContextPriority, ) -> Result<EGLContext, Error>

Create a new configless EGLContext with the specified priority from a given EGLDisplay sharing resources with another context

Note: The priority is a hint that might be ignored by the underlying platform. It also requires EGL_IMG_context_priority to be available, otherwise the priority will be ignored.

pub fn new_shared_with_config( display: &EGLDisplay, share: &EGLContext, attributes: GlAttributes, reqs: PixelFormatRequirements, ) -> Result<EGLContext, Error>

Create a new EGLContext from a given EGLDisplay and configuration requirements sharing resources with another context

pub fn new_shared_with_config_and_priority( display: &EGLDisplay, share: &EGLContext, attributes: GlAttributes, reqs: PixelFormatRequirements, priority: ContextPriority, ) -> Result<EGLContext, Error>

Create a new EGLContext with the specified priority from a given EGLDisplay and configuration requirements sharing resources with another context

Note: The priority is a hint that might be ignored by the underlying platform. It also requires EGL_IMG_context_priority to be available, otherwise the priority will be ignored.

pub unsafe fn make_current(&self) -> Result<(), MakeCurrentError>

Makes the OpenGL context the current context in the current thread with no surface bound.

Safety

This function is marked unsafe, because the context cannot be made current on another thread without being unbound again (see EGLContext::unbind).

pub unsafe fn make_current_with_surface( &self, surface: &EGLSurface, ) -> Result<(), MakeCurrentError>

Makes the OpenGL context the current context in the current thread with a surface to read/draw to.

Safety

This function is marked unsafe, because the context cannot be made current on another thread without being unbound again (see EGLContext::unbind).

pub unsafe fn make_current_with_draw_and_read_surface( &self, draw_surface: &EGLSurface, read_surface: &EGLSurface, ) -> Result<(), MakeCurrentError>

Makes the OpenGL context the current context in the current thread with surfaces to read/draw to.

Safety

This function is marked unsafe, because the context cannot be made current on another thread without being unbound again (see EGLContext::unbind).

pub fn is_current(&self) -> bool

Returns true if the OpenGL context is the current one in the thread.

pub fn is_shared(&self) -> bool

Returns true if the OpenGL context is (possibly) shared with another.

Externally managed contexts created with EGLContext::from_raw are always considered shared by this function.

pub fn config_id(&self) -> EGLConfig

Returns the egl config for this context

pub fn pixel_format(&self) -> Option<PixelFormat>

Returns the pixel format of the main framebuffer of the context.

pub fn unbind(&self) -> Result<(), MakeCurrentError>

Unbinds this context from the current thread, if set.

This does nothing if this context is not the current context.

pub fn display(&self) -> &EGLDisplay

Returns the display which created this context.

pub fn dmabuf_render_formats(&self) -> &FormatSet

Returns a list of formats for dmabufs that can be rendered to.

pub fn dmabuf_texture_formats(&self) -> &FormatSet

Returns a list of formats for dmabufs that can be used as textures.

pub fn user_data(&self) -> &UserDataMap

Retrieve user_data associated with this context

Note: UserData is shared between shared context, if constructed with new_shared or new_shared_with_config.

pub fn get_context_handle(&self) -> EGLContext

Get a raw handle to the underlying context.

The pointer will become invalid, when this struct is destroyed.