Struct smithay::backend::egl::display::EGLDisplay

pub struct EGLDisplay { /* private fields */ }

EGLDisplay represents an initialised EGL environment

Implementations

impl EGLDisplay

pub unsafe fn new<N, L>(native: &N, logger: L) -> Result<EGLDisplay, Error>where
N: EGLNativeDisplay + 'static,
L: Into<Option<Logger>>,

Create a new EGLDisplay from a given EGLNativeDisplay

Safety

native: The native display handle has to be kept alive for the lifetime of the created EGLDisplay

pub unsafe fn from_raw<L>(
display: *const c_void,
config_id: *const c_void,
logger: L
) -> Result<EGLDisplay, Error>where
L: Into<Option<Logger>>,

Create a new EGLDisplay from an already initialized EGLDisplay and EGLConfig handle

Safety

The display must be created from the system default EGL library (dlopen("libEGL.so"))

The display and config must be valid for the lifetime of the returned display and any handles created by this display (using EGLDisplay::get_display_handle)

pub fn choose_config(
&self,
attributes: GlAttributes,
reqs: PixelFormatRequirements
) -> Result<(PixelFormat, EGLConfig), Error>

Finds a compatible EGLConfig for a given set of requirements

pub fn get_display_handle(&self) -> Arc<EGLDisplayHandle>

Get a handle to the underlying raw EGLDisplay handle

pub fn get_egl_version(&self) -> (i32, i32)

Returns the runtime egl version of this display

pub fn extensions(&self) -> &[String]

Returns the supported extensions of this display

pub fn dmabuf_render_formats(&self) -> &HashSet<DrmFormat>

Returns a list of formats for dmabufs that can be rendered to.

pub fn dmabuf_texture_formats(&self) -> &HashSet<DrmFormat>

Returns a list of formats for dmabufs that can be used as textures.

pub fn supports_damage(&self) -> bool

Returns when the display supports extensions required for smithays damage tracking helpers.

pub fn create_dmabuf_from_image(
&self,
image: EGLImage,
size: Size<i32, BufferCoords>,
y_inverted: bool
) -> Result<Dmabuf, Error>

Exports an EGLImage as a Dmabuf

pub fn create_image_from_dmabuf(
&self,
dmabuf: &Dmabuf
) -> Result<EGLImage, Error>

Imports a Dmabuf as an EGLImage

pub fn bind_wl_display(
&self,
display: &DisplayHandle
) -> Result<EGLBufferReader, Error>

Binds this EGL display to the given Wayland display.

This will allow clients to utilize EGL to create hardware-accelerated surfaces. The server will need to be able to handle EGL-[WlBuffer]s.

Errors

This might return EglExtensionNotSupported if binding is not supported by the EGL implementation.

This might return OtherEGLDisplayAlreadyBound if called for the same [Display] multiple times, as only one egl display may be bound at any given time.