Enum smithay::backend::egl::EGLError

pub enum EGLError {

    NotInitialized,
    BadAccess,
    BadAlloc,
    BadAttribute,
    BadContext,
    BadConfig,
    BadCurrentSurface,
    BadDevice,
    BadDisplay,
    BadSurface,
    BadMatch,
    BadParameter,
    BadNativePixmap,
    BadNativeWindow,
    ResourceBusy,
    ContextLost,
    Unknown(u32),

}

Raw EGL error

Variants

NotInitialized

EGL is not initialized, or could not be initialized, for the specified EGL display connection.

BadAccess

EGL cannot access a requested resource (for example a context is bound in another thread).

BadAlloc

EGL failed to allocate resources for the requested operation.

BadAttribute

An unrecognized attribute or attribute value was passed in the attribute list.

BadContext

An EGLContext argument does not name a valid EGL rendering context.

BadConfig

An EGLConfig argument does not name a valid EGL frame buffer configuration.

BadCurrentSurface

The current surface of the calling thread is a window, pixel buffer or pixmap that is no longer valid.

BadDevice

An EGLDevice argument is not valid for this display.

BadDisplay

An EGLDisplay argument does not name a valid EGL display connection.

BadSurface

An EGLSurface argument does not name a valid surface (window, pixel buffer or pixmap) configured for GL rendering.

BadMatch

Arguments are inconsistent (for example, a valid context requires buffers not supplied by a valid surface).

BadParameter

One or more argument values are invalid.

BadNativePixmap

A NativePixmapType argument does not refer to a valid native pixmap.

BadNativeWindow

A NativeWindowType argument does not refer to a valid native window.

ResourceBusy

The EGL operation failed due to temporary unavailability of a requested resource, but the arguments were otherwise valid, and a subsequent attempt may succeed.

ContextLost

A power management event has occurred. The application must destroy all contexts and reinitialise OpenGL ES state and objects to continue rendering.

Unknown(u32)

An unknown error