Module smithay::backend::egl::native

Type safe native types for safe context/surface creation

Structs

XlibWindow
Typed Xlib window for the X11 backend

Enums

Wayland
Wayland backend type

X11
X11 backend type

Traits

Backend
Trait for typed backend variants (X11/Wayland/GBM)

NativeDisplay
Trait for types returning Surfaces which can be used to initialize EGLSurfaces

NativeSurface
Trait for types returning valid surface pointers for initializing egl