smithay::backend::egl::native

Struct EGLPlatform

pub struct EGLPlatform<'a> {
    pub required_extensions: &'static [&'static str],
    pub platform_name: &'static str,
    pub platform: EGLenum,
    pub native_display: *mut c_void,
    pub attrib_list: Vec<EGLint>,
    /* private fields */
}

Available on crate feature backend_egl only.

Type, Raw handle and attributes used to call eglGetPlatformDisplayEXT

Fields
required_extensions: &'static [&'static str]

Required extensions to use this platform

platform_name: &'static str

Human readable name of the platform

platform: EGLenum

Platform type used to call eglGetPlatformDisplayEXT

native_display: *mut c_void

Raw native display handle used to call eglGetPlatformDisplayEXT

attrib_list: Vec<EGLint>

Attributes used to call eglGetPlatformDisplayEXT

Implementations

impl EGLPlatform<'_>

pub fn new( platform: EGLenum, platform_name: &'static str, native_display: *mut c_void, attrib_list: Vec<EGLint>, required_extensions: &'static [&'static str], ) -> Self

Create a EGLPlatform<'a> for the provided platform.

Arguments

platform - The platform defined in ffi::egl::

platform_name - A human readable representation of the platform

native_display - The native display raw pointer which can be cast to *mut c_void

attrib_list - A list of ffi::EGLint like defined in the EGL Extension

required_extensions - The names of the required EGL Extensions for this platform

Examples

use smithay::backend::egl::{ffi, native::EGLPlatform};

EGLPlatform::new(ffi::egl::PLATFORM_GBM_KHR, "PLATFORM_GBM_KHR", native_display as *mut _, vec![ffi::egl::NONE as ffi::EGLint], &["EGL_KHR_platform_gbm"]);