Trait smithay::backend::egl::native::Backend

pub trait Backend {
    type Surface: NativeSurface;
    unsafe fn get_display<F: Fn(&str) -> bool>(
        display: NativeDisplayType, 
        has_dp_extension: F, 
        log: Logger
    ) -> Result<EGLDisplay, EGLError>;
}

Trait for typed backend variants (X11/Wayland/GBM)

Associated Types

type Surface: NativeSurface

Surface type created by this backend

Loading content...
Required methods

unsafe fn get_display<F: Fn(&str) -> bool>(
display: NativeDisplayType,
has_dp_extension: F,
log: Logger
) -> Result<EGLDisplay, EGLError>

Return an EGLDisplay based on this backend

Safety

The returned EGLDisplay needs to be a valid pointer for EGL, but there is no way to test that.

Loading content...