Trait smithay::backend::egl::native::NativeDisplay

pub unsafe trait NativeDisplay<B: Backend> {
    type Arguments;
    type Error: Error + Send + 'static;
    fn is_backend(&self) -> bool;

    fn ptr(&self) -> Result<NativeDisplayType, Error>;

    fn create_surface(
        &mut self, 
        args: Self::Arguments
    ) -> Result<B::Surface, Self::Error>;
}

Trait for types returning Surfaces which can be used to initialize EGLSurfaces

Unsafety

The returned NativeDisplayType must be valid for EGL and there is no way to test that.

Associated Types

type Arguments

Arguments used to surface creation.

type Error: Error + Send + 'static

Error type thrown by the surface creation in case of failure.

Loading content...
Required methods

fn is_backend(&self) -> bool

Because one type might implement multiple [Backend]s this function must be called to check if the expected [Backend] is used at runtime.

fn ptr(&self) -> Result<NativeDisplayType, Error>

Return a raw pointer EGL will accept for context creation.

fn create_surface(
&mut self,
args: Self::Arguments
) -> Result<B::Surface, Self::Error>

Create a surface

Loading content...