Trait smithay::backend::egl::native::NativeSurface

pub unsafe trait NativeSurface {
    fn ptr(&self) -> NativeWindowType;

    fn needs_recreation(&self) -> bool { ... }

    fn recreate(&self) -> bool { ... }

    fn swap_buffers(&self) -> Result<(), SwapBuffersError> { ... }
}

Trait for types returning valid surface pointers for initializing egl

Unsafety

The returned NativeWindowType must be valid for EGL and there is no way to test that.

Required methods

fn ptr(&self) -> NativeWindowType

Return a raw pointer egl will accept for surface creation.

Loading content...
Provided methods

fn needs_recreation(&self) -> bool

Will be called to check if any internal resources will need to be recreated. Old resources must be used until recreate was called.

Only needs to be recreated, if this shall sometimes return true. The default implementation always returns false.

fn recreate(&self) -> bool

Instructs the surface to recreate internal resources

Must only be implemented if needs_recreation can return true. Returns true on success. If this call was successful ptr() should return something different.

fn swap_buffers(&self) -> Result<(), SwapBuffersError>

Adds additional semantics when calling EGLSurface::swap_buffers

Only implement if required by the backend.

Loading content...