Trait smithay::backend::egl::EGLGraphicsBackend

pub trait EGLGraphicsBackend {
    fn bind_wl_display(
        &self, 
        display: &Display
    ) -> Result<EGLBufferReader, Error>;
}

Trait any backend type may implement that allows binding a Display to create an EGLBufferReader for EGL-based [WlBuffer]s.

Required methods

fn bind_wl_display(&self, display: &Display) -> Result<EGLBufferReader, Error>

Binds this EGL context to the given Wayland display.

This will allow clients to utilize EGL to create hardware-accelerated surfaces. The server will need to be able to handle EGL-[WlBuffer]s.

Errors

This might return EglExtensionNotSupported if binding is not supported by the EGL implementation.

This might return OtherEGLDisplayAlreadyBound if called for the same [Display] multiple times, as only one context may be bound at any given time.

Loading content...