Module smithay::backend::graphics

Common traits for various ways to renderer on a given graphics backend.

Note: Not every API may be supported by every backend

Modules

gl
OpenGL rendering types

glium
Glium compatibility module

Structs

PixelFormat
Describes the pixel format of a framebuffer

Enums

SwapBuffersError
Error that can happen when swapping buffers.

Traits

CursorBackend
Functions to render cursors on graphics backend independently from it's rendering techique.