Struct smithay::backend::graphics::egl::context::EGLContext

pub struct EGLContext<B: Backend, N: NativeDisplay<B>> { /* fields omitted */ }

EGL context for rendering

Methods

impl<B: Backend, N: NativeDisplay<B>> EGLContext<B, N>

pub fn new<L>(
native: N,
attributes: GlAttributes,
reqs: PixelFormatRequirements,
logger: L
) -> Result<EGLContext<B, N>> where
L: Into<Option<Logger>>,

Create a new EGLContext from a given NativeDisplay

pub fn create_surface(
&self,
args: N::Arguments
) -> Result<EGLSurface<B::Surface>>

Creates a surface for rendering

pub unsafe fn get_proc_address(&self, symbol: &str) -> *const c_void

Returns the address of an OpenGL function.

Supposes that the context has been made current before this function is called.

pub fn is_current(&self) -> bool

Returns true if the OpenGL context is the current one in the thread.

pub fn get_pixel_format(&self) -> PixelFormat

Returns the pixel format of the main framebuffer of the context.