Enum smithay::backend::graphics::egl::ErrorKind

pub enum ErrorKind {
    Msg(String),
    OpenGlVersionNotSupported((u8, u8)),
    OpenGlesNotSupported,
    NoAvailablePixelFormat,
    NonMatchingSurfaceType(NativeType, NativeType),
    NotSupported,
    LoadingEGLFailed,
    DisplayNotSupported,
    InitFailed,
    ConfigFailed,
    CreationFailed,
    SurfaceCreationFailed,
    Unknown(u32),
    // some variants omitted
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

OpenGlVersionNotSupported((u8, u8))

The requested OpenGL version is not supported

OpenGlesNotSupported

The EGL implementation does not support creating OpenGL ES contexts

NoAvailablePixelFormat

No available pixel format matched the criteria

NonMatchingSurfaceType(NativeType, NativeType)

Surface type does not match the context type

NotSupported

Context creation is not supported on the current window system

LoadingEGLFailed

Loading libEGL failed

DisplayNotSupported

EGL was unable to optain a valid EGL Display

InitFailed

eglInitialize returned an error

ConfigFailed

Failed to configure the EGL context

CreationFailed

Context creation failed as one or more requirements could not be met. Try removing some gl attributes or pixel format requirements

SurfaceCreationFailed

eglCreateWindowSurface failed

Unknown(u32)

The reason of failure could not be determined

Methods

impl ErrorKind

fn description(&self) -> &str

A string describing the error kind.