Enum smithay::backend::graphics::egl::SwapBuffersError

pub enum SwapBuffersError {
    ContextLost,
    AlreadySwapped,
    Unknown(u32),
}

Error that can happen when swapping buffers.

Variants
ContextLost

The OpenGL context has been lost and needs to be recreated.

All the objects associated to it (textures, buffers, programs, etc.) need to be recreated from scratch.

Operations will have no effect. Functions that read textures, buffers, etc. from OpenGL will return uninitialized data instead.

A context loss usually happens on mobile devices when the user puts the application on sleep and wakes it up later. However any OpenGL implementation can theoretically lose the context at any time.

AlreadySwapped

The buffers have already been swapped.

This error can be returned when swap_buffers has been called multiple times without any modification in between.

Unknown(u32)

Unknown GL error