Enum smithay::backend::graphics::egl::error::ErrorKind

pub enum ErrorKind {
    Msg(String),
    OpenGlVersionNotSupported((u8, u8)),
    OpenGlesNotSupported,
    NoAvailablePixelFormat,
    NonMatchingBackend(&'static str),
    DisplayNotSupported,
    InitFailed,
    ConfigFailed,
    CreationFailed,
    SurfaceCreationFailed,
    EglExtensionNotSupported(&'static [&'static str]),
    OtherEGLDisplayAlreadyBound,
    NoEGLDisplayBound,
    PlaneIndexOutOfBounds,
    EGLImageCreationFailed,
    Unknown(u32),
    // some variants omitted
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

OpenGlVersionNotSupported((u8, u8))

The requested OpenGL version is not supported

OpenGlesNotSupported

The EGL implementation does not support creating OpenGL ES contexts

NoAvailablePixelFormat

No available pixel format matched the criteria

NonMatchingBackend(&'static str)

Backend does not match the context type

DisplayNotSupported

EGL was unable to optain a valid EGL Display

InitFailed

eglInitialize returned an error

ConfigFailed

Failed to configure the EGL context

CreationFailed

Context creation failed as one or more requirements could not be met. Try removing some gl attributes or pixel format requirements

SurfaceCreationFailed

eglCreateWindowSurface failed

EglExtensionNotSupported(&'static [&'static str])

The required EGL extension is not supported by the underlying EGL implementation

OtherEGLDisplayAlreadyBound

Only one EGLDisplay may be bound to a given WlDisplay at any time

NoEGLDisplayBound

No EGLDisplay is currently bound to this WlDisplay

PlaneIndexOutOfBounds

Index of plane is out of bounds for EGLImages

EGLImageCreationFailed

Failed to create EGLImages from the buffer

Unknown(u32)

The reason of failure could not be determined

Methods

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.