Trait smithay::backend::graphics::egl::native::Backend

pub trait Backend {
    type Surface: NativeSurface;
    unsafe fn get_display<F: Fn(&str) -> bool>(
        display: NativeDisplayType, 
        has_dp_extension: F, 
        log: Logger
    ) -> EGLDisplay;
}

Trait for typed backend variants (X11/Wayland/GBM)

Associated Types

type Surface: NativeSurface

Surface type created by this backend

Required Methods

unsafe fn get_display<F: Fn(&str) -> bool>(
display: NativeDisplayType,
has_dp_extension: F,
log: Logger
) -> EGLDisplay

Return an EGLDisplay based on this backend

Unsafety

The returned EGLDisplay needs to be a valid ptr for egl, but there is no way to test that.