Trait smithay::backend::graphics::egl::native::NativeDisplay

pub unsafe trait NativeDisplay<B: Backend> {
    type Arguments;
    type Error: Error + Send + 'static;
    fn is_backend(&self) -> bool;

    fn ptr(&self) -> Result<NativeDisplayType>;

    fn create_surface(
        &self, 
        args: Self::Arguments
    ) -> Result<B::Surface, Self::Error>;
}

Trait for types returning Surfaces which can be used to initialize EGLSurfaces

Unsafety

The returned NativeDisplayType must be valid for egl and there is no way to test that.

Associated Types

type Arguments

Arguments used to surface creation.

type Error: Error + Send + 'static

Error type thrown by the surface creation in case of failure.

Required Methods

fn is_backend(&self) -> bool

Because one typ might implement multiple Backend this function must be called to check if the expected Backend is used at runtime.

fn ptr(&self) -> Result<NativeDisplayType>

Return a raw pointer egl will accept for context creation.

fn create_surface(
&self,
args: Self::Arguments
) -> Result<B::Surface, Self::Error>

Create a surface