Struct smithay::backend::graphics::egl::EGLContext

pub struct EGLContext<'a, T: NativeSurface> { /* fields omitted */ }

EGL context for rendering

Methods

impl<'a> EGLContext<'a, ()>

fn new_from_winit<L>(
window: &'a WinitWindow,
attributes: GlAttributes,
reqs: PixelFormatRequirements,
logger: L
) -> Result<EGLContext<'a, WinitWindow>> where
L: Into<Option<Logger>>,

Create a new context from a given winit-Window

fn new_from_gbm<L, U: 'static>(
gbm: &'a GbmDevice<'a>,
attributes: GlAttributes,
reqs: PixelFormatRequirements,
logger: L
) -> Result<EGLContext<'a, GbmSurface<'a, U>>> where
L: Into<Option<Logger>>,

Create a new context from a given gbm::Device

impl<'a, T: NativeSurface> EGLContext<'a, T>

fn create_surface<'b>(&'a self, native: &'b T) -> Result<EGLSurface<'a, 'b, T>>

Creates a surface bound to the given egl context for rendering

unsafe fn get_proc_address(&self, symbol: &str) -> *const c_void

Returns the address of an OpenGL function.

Supposes that the context has been made current before this function is called.

fn is_current(&self) -> bool

Returns true if the OpenGL context is the current one in the thread.

fn get_pixel_format(&self) -> PixelFormat

Returns the pixel format of the main framebuffer of the context.