Struct smithay::backend::graphics::egl::surface::EGLSurface

pub struct EGLSurface<N: NativeSurface> { /* fields omitted */ }

EGL surface of a given egl context for rendering

Methods

impl<N: NativeSurface> EGLSurface<N>

pub fn swap_buffers(&self) -> Result<(), SwapBuffersError>

Swaps buffers at the end of a frame.

pub unsafe fn make_current(&self) -> Result<(), SwapBuffersError>

Makes the OpenGL context the current context in the current thread.

Unsafety

This function is marked unsafe, because the context cannot be made current on multiple threads.

pub fn is_current(&self) -> bool

Returns true if the OpenGL surface is the current one in the thread.