Trait smithay::backend::graphics::egl::NativeSurface

pub unsafe trait NativeSurface {
    type Keep: 'static;
    fn surface(
        &self, 
        backend: NativeType
    ) -> Result<(NativeSurfacePtr, Self::Keep)>;
}

Trait for supported types returning valid surface pointers for initializing egl

Safety

The returned NativeSurfacePtr must be valid for egl and there is no way to test that.

Associated Types

type Keep: 'static

Type to keep the surface valid, if needed

Required Methods

fn surface(&self, backend: NativeType) -> Result<(NativeSurfacePtr, Self::Keep)>

Return a surface for the given type if possible