Module smithay::backend::graphics::egl::wayland

Wayland specific EGL functionality - EGL based WlBuffers.

The types of this module can be used to initialize hardware acceleration rendering based on EGL for clients as it may enabled usage of EGLImage based WlBuffers.

To use it bind any backend implementing the EGLWaylandExtensions trait, that shall do the rendering (so pick a fast one), to the wayland_server::Display of your compositor. Note only one backend may be bould to any Display at any time.

You may then use the resulting EGLDisplay to recieve EGLImages of an egl-based WlBuffer for rendering.

Structs

EGLDisplay
Type to recieve EGLImages for egl-based WlBuffers.

EGLImages
Images of the egl-based WlBuffer.

Enums

BufferAccessError
Error that can occur when accessing an EGL buffer

Format
Texture format types

TextureCreationError
Error that might happen when binding an EGLImage to a gl texture

Traits

EGLWaylandExtensions
Trait any backend type may implement that allows binding a wayland_server::Display to create an EGLDisplay for egl-based WlBuffers.