Trait smithay::backend::graphics::egl::wayland::EGLWaylandExtensions

pub trait EGLWaylandExtensions {
    fn bind_wl_display(&self, display: &Display) -> Result<EGLDisplay>;
}

Trait any backend type may implement that allows binding a wayland_server::Display to create an EGLDisplay for egl-based WlBuffers.

Required Methods

fn bind_wl_display(&self, display: &Display) -> Result<EGLDisplay>

Binds this EGL context to the given Wayland display.

This will allow clients to utilize EGL to create hardware-accelerated surfaces. The server will need to be able to handle egl-wl_buffers. See the wayland::drm module.

Errors

This might return WlExtensionNotSupported if binding is not supported by the EGL implementation.

This might return OtherEGLDisplayAlreadyBound if called for the same Display multiple times, as only one context may be bound at any given time.