Trait smithay::backend::graphics::gl::GLGraphicsBackend

pub trait GLGraphicsBackend {
    fn swap_buffers(&self) -> Result<(), SwapBuffersError>;

    fn get_proc_address(&self, symbol: &str) -> *const c_void;

    fn get_framebuffer_dimensions(&self) -> (u32, u32);

    fn is_current(&self) -> bool;

    unsafe fn make_current(&self) -> Result<(), SwapBuffersError>;

    fn get_pixel_format(&self) -> PixelFormat;
}

Trait that describes objects that have an OpenGL context and can be used to render upon

Required methods

fn swap_buffers(&self) -> Result<(), SwapBuffersError>

Swaps buffers at the end of a frame.

fn get_proc_address(&self, symbol: &str) -> *const c_void

Returns the address of an OpenGL function.

fn get_framebuffer_dimensions(&self) -> (u32, u32)

Returns the dimensions of the window, or screen, etc in points.

These are the actual pixels of the underlying graphics backend. For nested compositors you will need to handle the scaling of the root compositor yourself, if you want to.

fn is_current(&self) -> bool

Returns true if the OpenGL context is the current one in the thread.

unsafe fn make_current(&self) -> Result<(), SwapBuffersError>

Makes the OpenGL context the current context in the current thread.

Safety

The context cannot be made current on multiple threads.

fn get_pixel_format(&self) -> PixelFormat

Returns the pixel format of the main framebuffer of the context.

Loading content...