Struct smithay::backend::graphics::glium::GliumGraphicsBackend

pub struct GliumGraphicsBackend<T: EGLGraphicsBackend> { /* fields omitted */ }

Wrapper to expose glium compatibility

Methods

impl<T: EGLGraphicsBackend + 'static> GliumGraphicsBackend<T>

pub fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.

pub fn borrow(&self) -> Ref<T>

Borrow the underlying backend.

This follows the same semantics as std::cell:RefCell. Multiple read-only borrows are possible. Borrowing the backend while there is a mutable reference will panic.

pub fn borrow_mut(&self) -> RefMut<T>

Borrow the underlying backend mutably.

This follows the same semantics as std::cell:RefCell. Holding any other borrow while trying to borrow the backend mutably will panic. Note that glium will borrow the backend (not mutably) during rendering.