smithay::backend::input

Enum InputEvent

pub enum InputEvent<B: InputBackend> {

    DeviceAdded {
        device: B::Device,
    },
    DeviceRemoved {
        device: B::Device,
    },
    Keyboard {
        event: B::KeyboardKeyEvent,
    },
    PointerMotion {
        event: B::PointerMotionEvent,
    },
    PointerMotionAbsolute {
        event: B::PointerMotionAbsoluteEvent,
    },
    PointerButton {
        event: B::PointerButtonEvent,
    },
    PointerAxis {
        event: B::PointerAxisEvent,
    },
    GestureSwipeBegin {
        event: B::GestureSwipeBeginEvent,
    },
    GestureSwipeUpdate {
        event: B::GestureSwipeUpdateEvent,
    },
    GestureSwipeEnd {
        event: B::GestureSwipeEndEvent,
    },
    GesturePinchBegin {
        event: B::GesturePinchBeginEvent,
    },
    GesturePinchUpdate {
        event: B::GesturePinchUpdateEvent,
    },
    GesturePinchEnd {
        event: B::GesturePinchEndEvent,
    },
    GestureHoldBegin {
        event: B::GestureHoldBeginEvent,
    },
    GestureHoldEnd {
        event: B::GestureHoldEndEvent,
    },
    TouchDown {
        event: B::TouchDownEvent,
    },
    TouchMotion {
        event: B::TouchMotionEvent,
    },
    TouchUp {
        event: B::TouchUpEvent,
    },
    TouchCancel {
        event: B::TouchCancelEvent,
    },
    TouchFrame {
        event: B::TouchFrameEvent,
    },
    TabletToolAxis {
        event: B::TabletToolAxisEvent,
    },
    TabletToolProximity {
        event: B::TabletToolProximityEvent,
    },
    TabletToolTip {
        event: B::TabletToolTipEvent,
    },
    TabletToolButton {
        event: B::TabletToolButtonEvent,
    },
    SwitchToggle {
        event: B::SwitchToggleEvent,
    },
    Special(B::SpecialEvent),

}

Different events that can be generated by an input backend

Variants

DeviceAdded

An input device was connected

Fields

device: B::Device

The added device

DeviceRemoved

An input device was disconnected

Fields

device: B::Device

The removed device

Keyboard

A keyboard event occurred

Fields

event: B::KeyboardKeyEvent

The keyboard event

PointerMotion

A relative pointer motion occurred

Fields

event: B::PointerMotionEvent

The pointer motion event

PointerMotionAbsolute

An absolute pointer motion occurs

Fields

event: B::PointerMotionAbsoluteEvent

The absolute pointer motion event

PointerButton

A pointer button was pressed or released

Fields

event: B::PointerButtonEvent

The pointer button event

PointerAxis

A pointer action occurred while scrolling on an axis

Fields

event: B::PointerAxisEvent

The pointer axis event

GestureSwipeBegin

A pointer swipe gesture began

Fields

event: B::GestureSwipeBeginEvent

The gesture event

GestureSwipeUpdate

A pointer swipe gesture updated

Fields

event: B::GestureSwipeUpdateEvent

The gesture event

GestureSwipeEnd

A pointer swipe gesture ended

Fields

event: B::GestureSwipeEndEvent

The gesture event

GesturePinchBegin

A pointer pinch gesture began

Fields

event: B::GesturePinchBeginEvent

The gesture event

GesturePinchUpdate

A pointer pinch gesture updated

Fields

event: B::GesturePinchUpdateEvent

The gesture event

GesturePinchEnd

A pointer pinch gesture ended

Fields

event: B::GesturePinchEndEvent

The gesture event

GestureHoldBegin

A pointer hold gesture began

Fields

event: B::GestureHoldBeginEvent

The gesture event

GestureHoldEnd

A pointer hold gesture ended

Fields

event: B::GestureHoldEndEvent

The gesture event

TouchDown

A new touchpoint appeared

Fields

event: B::TouchDownEvent

The touch down event

TouchMotion

A touchpoint moved

Fields

event: B::TouchMotionEvent

The touch motion event

TouchUp

A touchpoint was removed

Fields

event: B::TouchUpEvent

The touch up event

TouchCancel

A touch sequence was cancelled

Fields

event: B::TouchCancelEvent

The touch cancel event

TouchFrame

A touch frame was emitted

A set of two events received on the same seat between two frames should be interpreted as an atomic event.

Fields

event: B::TouchFrameEvent

The touch frame event

TabletToolAxis

A tablet tool axis was emitted

Fields

event: B::TabletToolAxisEvent

The tablet tool axis event

TabletToolProximity

A tablet tool proximity was emitted

Fields

event: B::TabletToolProximityEvent

The tablet tool proximity event

TabletToolTip

A tablet tool tip event was emitted

Fields

event: B::TabletToolTipEvent

The tablet tool axis event

TabletToolButton

A tablet tool button was pressed or released

Fields

event: B::TabletToolButtonEvent

The pointer button event

SwitchToggle

A switch was toggled

Fields

event: B::SwitchToggleEvent

The switch toggle event

Special(B::SpecialEvent)

Special event specific of this backend