Enum smithay::backend::input::TouchEvent

pub enum TouchEvent {
    Down {
        slot: Option<TouchSlot>,
        x: f64,
        y: f64,
    },
    Motion {
        slot: Option<TouchSlot>,
        x: f64,
        y: f64,
    },
    Up {
        slot: Option<TouchSlot>,
    },
    Cancel {
        slot: Option<TouchSlot>,
    },
    Frame,
}

Touch event

Variants
Down

The start of an event at a given position (x, y).

If the device has multi-touch capabilities a slot is given.

Fields of Down

slot: Option<TouchSlot>

TouchSlot, if the device has multi-touch capabilities

x: f64

Absolute x-coordinate of the touch position.

y: f64

Absolute y-coordinate of the touch position.

Motion

Movement of a touch on the device surface to a given position (x, y).

If the device has multi-touch capabilities a slot is given.

Fields of Motion

slot: Option<TouchSlot>

TouchSlot, if the device has multi-touch capabilities

x: f64

Absolute x-coordinate of the final touch position after the motion.

y: f64

Absolute y-coordinate of the final touch position after the motion.

Up

Stop of an event chain.

If the device has multi-touch capabilities a slot is given.

Fields of Up

slot: Option<TouchSlot>

TouchSlot, if the device has multi-touch capabilities

Cancel

Cancel of an event chain. All previous events in the chain should be ignored.

If the device has multi-touch capabilities a slot is given.

Fields of Cancel

slot: Option<TouchSlot>

TouchSlot, if the device has multi-touch capabilities

Frame

Signals the end of a set of touchpoints at one device sample time.