Struct smithay::backend::input::Seat

pub struct Seat { /* fields omitted */ }

A seat describes a group of input devices and at least one graphics device belonging together.

By default only one seat exists for most systems and smithay backends however multiseat configurations are possible and should be treated as separated users, all with their own focus, input and cursor available.

Seats referring to the same internal id will always be equal and result in the same hash, but capabilities of cloned and copied Seats will not be updated by smithay. Always referr to the Seat given by a callback for up-to-date information. You may use this to calculate the differences since the last callback.

Methods

impl Seat

pub fn capabilities(&self) -> &SeatCapabilities

Get the currently capabilities of this Seat

pub fn name(&self) -> &str

Get the name of this Seat