smithay::backend::input
Trait Device

pub trait Device:
    PartialEq
    + Eq
    + Hash {
    // Required methods
    fn id(&self) -> String;
    fn name(&self) -> String;
    fn has_capability(&self, capability: DeviceCapability) -> bool;
    fn usb_id(&self) -> Option<(u32, u32)>;
    fn syspath(&self) -> Option<PathBuf>;
}

Trait for generic functions every input device does provide

Required Methods

fn id(&self) -> String

Unique id of a single device at a point in time.

Note: This means ids may be re-used by the backend for later devices.

fn name(&self) -> String

Human-readable name of the device

fn has_capability(&self, capability: DeviceCapability) -> bool

Test if this device has a specific capability

fn usb_id(&self) -> Option<(u32, u32)>

Returns device USB (product,vendor) id

fn syspath(&self) -> Option<PathBuf>

Returns the syspath of the device.

The path is an absolute path and includes the sys mount point.

Object Safety

This trait is not object safe.