Trait smithay::backend::input::Event

pub trait Event<B: InputBackend> {
    fn time(&self) -> u32;
    fn device(&self) -> B::Device;
}

Trait for generic functions every input event does provide

Required Methods

fn time(&self) -> u32

Returns an upward counting variable useful for event ordering.

Makes no guarantees about actual time passed between events.

fn device(&self) -> B::Device

Returns the device, that generated this event