smithay::backend::input

Trait GesturePinchUpdateEvent

pub trait GesturePinchUpdateEvent<B: InputBackend>: Event<B> {
    // Required methods
    fn delta_x(&self) -> f64;
    fn delta_y(&self) -> f64;
    fn scale(&self) -> f64;
    fn rotation(&self) -> f64;

    // Provided method
    fn delta(&self) -> Point<f64, Logical> { ... }
}

Trait for gesture pinch update event.

Required Methods

fn delta_x(&self) -> f64

Delta of center on the x axis from last begin/update.

fn delta_y(&self) -> f64

Delta of center on the y axis from last begin/update.

fn scale(&self) -> f64

Absolute scale compared to begin event.

fn rotation(&self) -> f64

Relative angle in degrees from last begin/update.

Provided Methods

fn delta(&self) -> Point<f64, Logical>

Delta between the last and new gesture center.