Trait smithay::backend::input::PointerMotionEvent

pub trait PointerMotionEvent<B: InputBackend>: Event<B> {
    // Required methods
    fn delta_x(&self) -> f64;
    fn delta_y(&self) -> f64;
    fn delta_x_unaccel(&self) -> f64;
    fn delta_y_unaccel(&self) -> f64;

    // Provided methods
    fn delta(&self) -> Point<f64, Logical> { ... }
    fn delta_unaccel(&self) -> Point<f64, Logical> { ... }
}

Trait for pointer events generated by relative device movement.

Required Methods

fn delta_x(&self) -> f64

Delta on the x axis between the last and new pointer device position interpreted as pixel movement

fn delta_y(&self) -> f64

Delta on the y axis between the last and new pointer device position interpreted as pixel movement

fn delta_x_unaccel(&self) -> f64

Unaccelerated delta on the x axis between the last and new pointer device position

fn delta_y_unaccel(&self) -> f64

Unaccelerated delta on the y axis between the last and new pointer device position

Provided Methods

fn delta(&self) -> Point<f64, Logical>

Delta between the last and new pointer device position interpreted as pixel movement

fn delta_unaccel(&self) -> Point<f64, Logical>

Unaccelerated delta between the last and new pointer device position