smithay::backend::input

Trait TabletToolProximityEvent

pub trait TabletToolProximityEvent<B: InputBackend>: TabletToolEvent<B> + Event<B> {
    // Required method
    fn state(&self) -> ProximityState;
}

Trait for tablet tool proximity events.

Required Methods

fn state(&self) -> ProximityState

Returns the new proximity state of a tool from a proximity event.

Used to check whether or not a tool came in or out of proximity during an TabletToolProximityEvent.

See Handling of proximity events for recommendations on proximity handling.