Trait smithay::backend::input::TabletToolTipEvent

pub trait TabletToolTipEvent<B: InputBackend>: TabletToolEvent<B> + Event<B> {
    fn tip_state(&self) -> TabletToolTipState;
}

Signals that a tool has come in contact with the surface of a device with the DeviceCapability::TabletTool capability.

On devices without distance proximity detection, the TabletToolTipEvent is sent immediately after TabletToolProximityEvent for the tip down event, and immediately before for the tip up event.

The decision when a tip touches the surface is device-dependent and may be derived from pressure data or other means. If the tip state is changed by axes changing state, the TabletToolTipEvent includes the changed axes and no additional axis event is sent for this state change. In other words, a caller must look at both TabletToolAxisEvent and TabletToolTipEvent events to know the current state of the axes.

If a button state change occurs at the same time as a tip state change, the order of events is device-dependent.

Required methods

fn tip_state(&self) -> TabletToolTipState

Returns the new tip state of a tool from a tip event.

Used to check whether or not a tool came in contact with the tablet surface or left contact with the tablet surface during an TabletToolTipEvent.