Trait smithay::backend::input::TouchDownEvent

pub trait TouchDownEvent: Event {
    fn slot(&self) -> Option<TouchSlot>;

    fn x(&self) -> f64;

    fn y(&self) -> f64;

    fn x_transformed(&self, width: u32) -> u32;

    fn y_transformed(&self, height: u32) -> u32;

    fn position(&self) -> (f64, f64) { ... }

    fn position_transformed(&self, coordinate_space: (u32, u32)) -> (u32, u32) { ... }
}

Trait for touch events starting at a given position.

Required methods

fn slot(&self) -> Option<TouchSlot>

TouchSlot, if the device has multi-touch capabilities

fn x(&self) -> f64

Touch event's x-coordinate in the device's native coordinate space

The actual format is defined by the implementation.

fn y(&self) -> f64

Touch event's x-coordinate in the device's native coordinate space

The actual format is defined by the implementation.

fn x_transformed(&self, width: u32) -> u32

Touch event's x position converted to the targets coordinate space's width. E.g. the focused output's width.

fn y_transformed(&self, height: u32) -> u32

Touch event's y position converted to the targets coordinate space's width. E.g. the focused output's width.

Loading content...
Provided methods

fn position(&self) -> (f64, f64)

Touch position in the device's native coordinate space

The actual format is defined by the implementation.

fn position_transformed(&self, coordinate_space: (u32, u32)) -> (u32, u32)

Touch position converted into the target coordinate space. E.g. the focused output's resolution.

Loading content...