smithay::backend

Module libinput

Available on crate feature backend_libinput only.

Implementation of input backend trait for types provided by libinput

Structs

LibinputInputBackend

Libinput based InputBackend.

LibinputSessionInterfacebackend_session

Wrapper for types implementing the Session trait to provide a libinput::LibinputInterface implementation.

Enums

PointerScrollAxis

Generic pointer scroll event from libinput