Struct smithay::backend::libinput::LibinputInputBackend

pub struct LibinputInputBackend { /* private fields */ }

Libinput based InputBackend.

Tracks input of all devices given manually or via a udev seat to a provided libinput context.

Implementations

impl LibinputInputBackend

pub fn new<L>(context: Libinput, logger: L) -> Selfwhere
L: Into<Option<Logger>>,

Initialize a new LibinputInputBackend from a given already initialized libinput context.