Module smithay::backend::renderer::damage

Helper for effective damage tracked rendering

Why use this implementation

The DamageTrackedRenderer in combination with the RenderElement trait can help you to reduce resource consumption by tracking what elements have been damaged and only redraw the damaged parts on an output.

It does so by keeping track of the last used CommitCounter for all provided RenderElements and queries the element for new damage on each call to render_output.

You can initialize it with a static output by using DamageTrackedRenderer::new or allow it to track a specific Output with DamageTrackedRenderer::from_output.

See the renderer::element module for more information about how to use RenderElement.

How to use it

use smithay::{
    backend::renderer::{
        damage::DamageTrackedRenderer,
        element::memory::{MemoryRenderBuffer, MemoryRenderBufferRenderElement}
    },
    utils::{Point, Transform},
};
use std::time::{Duration, Instant};

const WIDTH: i32 = 10;
const HEIGHT: i32 = 10;

// Initialize a new damage tracked renderer
let mut damage_tracked_renderer = DamageTrackedRenderer::new((800, 600), 1.0, Transform::Normal);

// Initialize a buffer to render
let mut memory_buffer = MemoryRenderBuffer::new((WIDTH, HEIGHT), 1, Transform::Normal, None);

let mut last_update = Instant::now();

loop {
    let now = Instant::now();
    if now.duration_since(last_update) >= Duration::from_secs(3) {
        let mut render_context = memory_buffer.render();

        render_context.draw(|_buffer| {
            // Update the changed parts of the buffer

            // Return the updated parts
            Result::<_, ()>::Ok(vec![Rectangle::from_loc_and_size(Point::default(), (WIDTH, HEIGHT))])
        });

        last_update = now;
    }

    // Create a render element from the buffer
    let location = Point::from((100.0, 100.0));
    let render_element =
        MemoryRenderBufferRenderElement::from_buffer(location, &memory_buffer, None, None, None);

    // Render the output
    damage_tracked_renderer
        .render_output(
            &mut renderer,
            buffer_age,
            &[render_element],
            [0.8, 0.8, 0.9, 1.0],
            log.clone(),
        )
        .expect("failed to render the output");
}

Structs

DamageTrackedRenderer

Damage tracked renderer for a single output

OutputNoMode

Output has no active mode

Enums

DamageTrackedRendererError

Errors thrown by DamageTrackedRenderer::render_output

DamageTrackedRendererMode

Mode for the DamageTrackedRenderer