Struct smithay::backend::renderer::damage::DamageTrackedRenderer

pub struct DamageTrackedRenderer { /* private fields */ }

Damage tracked renderer for a single output

Implementations

impl DamageTrackedRenderer

pub fn new(
size: impl Into<Size<i32, Physical>>,
scale: impl Into<Scale<f64>>,
transform: Transform
) -> Self

Initialize a static DamageTrackedRenderer

pub fn from_output(output: &Output) -> Self

Initialize a new DamageTrackedRenderer from an Output

The renderer will keep track of changes to the Output and handle size and scaling changes automatically on the next call to render_output

pub fn mode(&self) -> &DamageTrackedRendererMode

Get the DamageTrackedRendererMode of the DamageTrackedRenderer

pub fn render_output<E, R>(
&mut self,
renderer: &mut R,
age: usize,
elements: &[E],
clear_color: [f32; 4],
log: impl Into<Option<Logger>>
) -> Result<(Option<Vec<Rectangle<i32, Physical>>>, RenderElementStates), DamageTrackedRendererError<R>>where
E: RenderElement<R>,
R: Renderer,
<R as Renderer>::TextureId: Texture,

Render this output

pub fn damage_output<E>(
&mut self,
age: usize,
elements: &[E],
log: impl Into<Option<Logger>>
) -> Result<(Option<Vec<Rectangle<i32, Physical>>>, RenderElementStates), OutputNoMode>where
E: Element,

Damage this output and return the damage without actually rendering the difference