Function smithay::backend::renderer::element::default_primary_scanout_output_compare

pub fn default_primary_scanout_output_compare<'a>(
    current_output: &'a Output,
    current_state: &RenderElementState,
    next_output: &'a Output,
    next_state: &RenderElementState,
) -> &'a Output

Default function for primary scan-out selection

This will prefer the next output when the visible portion of the element on screen is at least twice the size of the current visible portion. Otherwise it will prefer the output with the higher refresh rate.