Module smithay::backend::renderer::element::memory

Element to render from RGBA memory

Why use this implementation

The MemoryRenderBuffer allows for easy integration of software-rendered elements in the smithay rendering pipeline. As software-rendered elements eventually have to upload to the GPU for rendering damage tracking is a crucial part. The MemoryRenderBuffer allows for efficient damage tracking by providing a RenderContext which accumulates the software-rendering damage. It automatically uploads the damaged parts to a Renderer::TextureId during rendering.

Why not to use this implementation

As described earlier the MemoryRenderBuffer is targeted at software rendering, if you have some static content you may want to take a look at TextureBuffer or TextureRenderBuffer for hardware accelerated rendering.

How to use it

The MemoryRenderBuffer represents a buffer of your data and holds the damage as reported by the RenderContext. To render the buffer you have to create a MemoryRenderBufferRenderElement in your render loop as shown in the example.

use std::time::{Duration, Instant};

use smithay::{
    backend::{
        allocator::Fourcc,
        renderer::{
            damage::OutputDamageTracker,
            element::{
                Kind,
                memory::{MemoryRenderBuffer, MemoryRenderBufferRenderElement},
            },
        },
    },
    utils::{Point, Rectangle, Size, Transform},
};

const WIDTH: i32 = 10;
const HEIGHT: i32 = 10;

// Initialize a empty render buffer
let mut buffer = MemoryRenderBuffer::new(Fourcc::Argb8888, (WIDTH, HEIGHT), 1, Transform::Normal, None);

// Create a rendering context
let mut render_context = buffer.render();

// Draw to the buffer
render_context.draw(|buffer| {
    buffer.chunks_exact_mut(4).for_each(|chunk| {
        chunk.copy_from_slice(&[255, 231, 199, 255]);
    });

    // Return the whole buffer as damage
    Result::<_, ()>::Ok(vec![Rectangle::from_loc_and_size(Point::default(), (WIDTH, HEIGHT))])
});

// Optionally update the opaque regions
render_context.update_opaque_regions(Some(vec![Rectangle::from_loc_and_size(
    Point::default(),
    Size::from((WIDTH, HEIGHT)),
)]));

// We explicitly drop the context here to make the borrow checker happy
std::mem::drop(render_context);

// Initialize a static damage tracker
let mut damage_tracker = OutputDamageTracker::new((800, 600), 1.0, Transform::Normal);

let mut last_update = Instant::now();

loop {
    let now = Instant::now();
    if now.duration_since(last_update) >= Duration::from_secs(3) {
        let mut render_context = buffer.render();

        render_context.draw(|_buffer| {
            // Update the changed parts of the buffer

            // Return the updated parts
            Result::<_, ()>::Ok(vec![Rectangle::from_loc_and_size(Point::default(), (WIDTH, HEIGHT))])
        });

        last_update = now;
    }

    // Create a render element from the buffer
    let location = Point::from((100.0, 100.0));
    let render_element = MemoryRenderBufferRenderElement::from_buffer(&mut renderer, location, &buffer, None, None, None, Kind::Unspecified)
        .expect("Failed to upload from memory to gpu");

    // Render the element(s)
    damage_tracker
        .render_output(&mut renderer, 0, &[&render_element], [0.8, 0.8, 0.9, 1.0])
        .expect("failed to render output");
}

Structs

MemoryBuffer

A buffer storing pixel data backed by system memory

MemoryRenderBuffer

A memory backed render buffer

MemoryRenderBufferRenderElement

A render element for MemoryRenderBuffer

RenderContext

A render context for MemoryRenderBuffer