smithay::backend::renderer::element

Module texture

Element to render a texture

Why use this implementation

This module provides the following elements

Static texture

TextureBuffer represents a static texture which is expected to not change. It is possible to either use a pre-existing texture or to create the texture from RGBA memory. The TextureBuffer can be used in the smithay pipeline by using TextureRenderElement::from_texture_buffer.

Hardware accelerated rendering

TextureRenderBuffer provides a solution for hardware accelerated rending with proper damage tracking. It provides a RenderContext for rendering to the stored texture. The collected damage from RenderContext::draw is used to report the damaged regions in the TextureRenderElement. For rendering a TextureRenderBuffer you can create a render element with TextureRenderElement::from_texture_render_buffer.

Why not to use this implementation

For software rendering you may take a look at MemoryRenderBuffer which provides proper damage tracking and is agnostic over the Renderer

How to use it

Both, TextureBuffer and TextureRenderBuffer represent a buffer of your texture data. To render the buffer you have to create a TextureRenderElement in your render loop as shown in the examples.

To integrate custom texture based buffers the TextureRenderElement provides two functions to create the render element. from_static_texture can be used to create a render element from a static texture without damage tracking. from_texture_with_damage can be used to create a render element from an existing texture with custom damage tracking. In both cases you have to make sure to provide a stable Id or otherwise damage tracking will not work.

TextureBuffer

use smithay::{
    backend::{
        allocator::Fourcc,
        renderer::{
            damage::OutputDamageTracker,
            element::{
                Kind,
                texture::{TextureBuffer, TextureRenderElement},
            },
        },
    },
    utils::{Point, Transform},
};

const WIDTH: i32 = 10;
const HEIGHT: i32 = 10;

let memory = vec![0; (WIDTH * 4 * HEIGHT) as usize];

// Create the texture buffer from a chunk of memory
let texture_buffer = TextureBuffer::from_memory(
    &mut renderer,
    &memory,
    Fourcc::Argb8888,
    (WIDTH, HEIGHT),
    false,
    1,
    Transform::Normal,
    None,
)
.expect("failed to import mem");

let mut damage_tracker = OutputDamageTracker::new((800, 600), 1.0, Transform::Normal);

loop {
    // Create a render element from the buffer
    let location = Point::from((100.0, 100.0));
    let render_element =
        TextureRenderElement::from_texture_buffer(location, &texture_buffer, None, None, None, Kind::Unspecified);

    // Render the element(s)
    damage_tracker
        .render_output(&mut renderer, &mut framebuffer, 0, &[&render_element], [0.8, 0.8, 0.9, 1.0])
        .expect("failed to render output");
}

TextureRenderBuffer

use std::time::{Duration, Instant};

use smithay::{
    backend::{
        allocator::Fourcc,
        renderer::{
            damage::OutputDamageTracker,
            element::{
                Kind,
                texture::{TextureRenderBuffer, TextureRenderElement},
            },
        },
    },
    utils::{Point, Rectangle, Size, Transform},
};

const WIDTH: i32 = 10;
const HEIGHT: i32 = 10;

let memory = vec![0; (WIDTH * 4 * HEIGHT) as usize];

// Create the texture buffer from a chunk of memory
let mut texture_render_buffer = TextureRenderBuffer::from_memory(
    &mut renderer,
    &memory,
    Fourcc::Argb8888,
    (WIDTH, HEIGHT),
    false,
    1,
    Transform::Normal,
    None,
)
.expect("failed to import mem");

// Create the rendering context
let mut render_context = texture_render_buffer.render();

// Draw to the texture
render_context.draw(|_texture| {
    // Your draw code here...

    // Return the damage areas
    Result::<_, ()>::Ok(vec![Rectangle::from_size(
        Size::from((WIDTH, HEIGHT)),
    )])
});

// Optionally update the opaque regions
render_context.update_opaque_regions(Some(vec![Rectangle::from_size(
    Size::from((WIDTH, HEIGHT)),
)]));

// We explicitly drop the context here to make the borrow checker happy
std::mem::drop(render_context);

let mut damage_tracker = OutputDamageTracker::new((800, 600), 1.0, Transform::Normal);

let mut last_update = Instant::now();

loop {
    let now = Instant::now();
    if now.duration_since(last_update) >= Duration::from_secs(3) {
        let mut render_context = texture_render_buffer.render();

        render_context.draw(|_buffer| {
            // Update the changed parts of the buffer

            // Return the updated parts
            Result::<_, ()>::Ok(vec![Rectangle::from_size((WIDTH, HEIGHT).into())])
        });

        last_update = now;
    }

    // Create a render element from the buffer
    let location = Point::from((100.0, 100.0));
    let render_element = TextureRenderElement::from_texture_render_buffer(
        location,
        &texture_render_buffer,
        None,
        None,
        None,
        Kind::Unspecified,
    );

    // Render the element(s)
    damage_tracker
        .render_output(&mut renderer, &mut framebuffer, 0, &[&render_element], [0.8, 0.8, 0.9, 1.0])
        .expect("failed to render output");
}

Structs

RenderContext

A render context for TextureRenderBuffer

TextureBuffer

A single texture buffer

TextureRenderBuffer

A texture backed render buffer

TextureRenderElement

A render element for a TextureRenderBuffer