smithay::backend::renderer::element::utils

Function constrain_render_elements

pub fn constrain_render_elements<E>(
    elements: impl IntoIterator<Item = E>,
    origin: impl Into<Point<i32, Physical>>,
    constrain: Rectangle<i32, Physical>,
    reference: Rectangle<i32, Physical>,
    behavior: ConstrainScaleBehavior,
    align: ConstrainAlign,
    scale: impl Into<Scale<f64>>,
) -> impl Iterator<Item = CropRenderElement<RelocateRenderElement<RescaleRenderElement<E>>>>
where
    E: Element,

Constrain render elements on a specific location with a specific size

origin - Defines the origin for re-scaling

constrain - Defines the rectangle on screen the elements should be constrain within

reference - Defines the reference that should be used for constraining the elements, this is most commonly the bounding box or geometry of the elements

behavior - Defines the behavior for scaling the elements reference in the constrain

align - Defines how the scaled elements should be aligned within the constrain

scale - The scale that was used to create the original elements