Function smithay::backend::renderer::buffer_has_alpha

pub fn buffer_has_alpha(buffer: &WlBuffer) -> Option<bool>

Available on crate feature wayland_frontend only.

Returns if the buffer has an alpha channel

Returns None if the type is not known to smithay or otherwise not supported (e.g. not initialized using one of smithays crate::wayland-handlers).

Note: This is on a best-effort, but will never return false for a buffer with a format that supports alpha.