Struct smithay::backend::renderer::gles::element::PixelShaderElement

pub struct PixelShaderElement { /* private fields */ }

Render element for drawing with a gles2 pixel shader

Implementations

impl PixelShaderElement

pub fn new( shader: GlesPixelProgram, area: Rectangle<i32, Logical>, opaque_regions: Option<Vec<Rectangle<i32, Logical>>>, alpha: f32, additional_uniforms: Vec<Uniform<'_>> ) -> Self

Create a new PixelShaderElement from a GlesPixelProgram, which can be constructed using GlesRenderer::compile_custom_pixel_shader

pub fn resize( &mut self, area: Rectangle<i32, Logical>, opaque_regions: Option<Vec<Rectangle<i32, Logical>>> )

Resize the canvas area

pub fn update_uniforms(&mut self, additional_uniforms: Vec<Uniform<'_>>)

Update the additional uniforms (see GlesRenderer::compile_custom_pixel_shader and [GlesRenderer::render_pixel_shader_to]).

This replaces the stored uniforms, you have to update all of them, partial updates are not possible.