smithay::backend::renderer::gles

Enum UniformValue

pub enum UniformValue {

    _1f(f32),
    _2f(f32, f32),
    _3f(f32, f32, f32),
    _4f(f32, f32, f32, f32),
    _1i(i32),
    _2i(i32, i32),
    _3i(i32, i32, i32),
    _4i(i32, i32, i32, i32),
    _1ui(u32),
    _2ui(u32, u32),
    _3ui(u32, u32, u32),
    _4ui(u32, u32, u32, u32),
    Matrix2x2 {
        matrices: Vec<[f32; 4]>,
        transpose: bool,
    },
    Matrix2x3 {
        matrices: Vec<[f32; 6]>,
        transpose: bool,
    },
    Matrix2x4 {
        matrices: Vec<[f32; 8]>,
        transpose: bool,
    },
    Matrix3x2 {
        matrices: Vec<[f32; 6]>,
        transpose: bool,
    },
    Matrix3x3 {
        matrices: Vec<[f32; 9]>,
        transpose: bool,
    },
    Matrix3x4 {
        matrices: Vec<[f32; 12]>,
        transpose: bool,
    },
    Matrix4x2 {
        matrices: Vec<[f32; 8]>,
        transpose: bool,
    },
    Matrix4x3 {
        matrices: Vec<[f32; 12]>,
        transpose: bool,
    },
    Matrix4x4 {
        matrices: Vec<[f32; 16]>,
        transpose: bool,
    },

}

Available on crate feature renderer_gl only.

Value of a uniform variable

Variants

_1f(f32)

A single float

_2f(f32, f32)

Two floats

_3f(f32, f32, f32)

Three floats

_4f(f32, f32, f32, f32)

Four floats

_1i(i32)

A single signed integer

_2i(i32, i32)

Two signed integers

_3i(i32, i32, i32)

Three signed integers

_4i(i32, i32, i32, i32)

Four signed integers

_1ui(u32)

A single unsigned integer

_2ui(u32, u32)

Two unsigned integers

_3ui(u32, u32, u32)

Three signed integers

_4ui(u32, u32, u32, u32)

Four signed integers

Matrix2x2

2x2 matrices

Fields

matrices: Vec<[f32; 4]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix2x3

2x3 matrices

Fields

matrices: Vec<[f32; 6]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix2x4

2x4 matrices

Fields

matrices: Vec<[f32; 8]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix3x2

3x2 matrices

Fields

matrices: Vec<[f32; 6]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix3x3

3x3 matrices

Fields

matrices: Vec<[f32; 9]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix3x4

3x4 matrices

Fields

matrices: Vec<[f32; 12]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix4x2

4x2 matrices

Fields

matrices: Vec<[f32; 8]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix4x3

4x3 matrices

Fields

matrices: Vec<[f32; 12]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Matrix4x4

4x4 matrices

Fields

matrices: Vec<[f32; 16]>

Matrices supplied as f32-arrays

transpose: bool

If transpose is false, each matrix is assumed to be supplied in column major order. If transpose is true, each matrix is assumed to be supplied in row major order.

Implementations

impl UniformValue

pub fn matches(&self, type_: &UniformType) -> bool

Checks if the contexts of this value match the provided uniform type

pub fn type_(&self) -> UniformType

Returns the type of this uniform value

pub unsafe fn set( &self, gl: &Gles2, desc: &UniformDesc, ) -> Result<(), GlesError>

Sets the desc uniform to this value.

Safety

You have to make sure to pass a valid UniformDesc, and to only call this function when it is otherwise safe to call gl.Uniform() series of methods.