smithay::backend::renderer::gles

Struct GlesFrame

pub struct GlesFrame<'frame, 'buffer> { /* private fields */ }

Available on crate feature renderer_gl only.

Handle to the currently rendered frame during GlesRenderer::render.

Leaking this frame will cause a variety of problems:

It might prevent the frame from synchronizing the rendered framebuffer causing glitches.

Depending on the bound target this can deadlock, if the same target is used later in any way.

Additionally parts of the GL state might not be reset correctly, causing unexpected results for later render commands.

The internal GL context and framebuffer will remain valid, no re-creation will be necessary.

Implementations

impl GlesFrame<'_, '_>

pub fn with_context<F, R>(&mut self, func: F) -> Result<R, GlesError>
where F: FnOnce(&Gles2) -> R,

Run custom code in the GL context owned by this renderer.

The OpenGL state of the renderer is considered an implementation detail and no guarantee is made about what can or cannot be changed, as such you should reset everything you change back to its previous value or check the source code of the version of Smithay you are using to ensure your changes don’t interfere with the renderer’s behavior. Doing otherwise can lead to rendering errors while using other functions of this renderer.

pub fn with_profiled_context<F, R>( &mut self, location: SpanLocation, func: F, ) -> Result<R, GlesError>
where F: FnOnce(&Gles2) -> R,

Run custom code in the GL context with GPU profiling.

Calls with_context() inside with_gpu_span().

pub fn with_gpu_span<F, R>(&mut self, location: SpanLocation, func: F) -> R
where F: FnOnce(&mut Self) -> R,

Run a function in a GPU profiling span.

impl GlesFrame<'_, '_>

pub fn override_default_tex_program( &mut self, program: GlesTexProgram, additional_uniforms: Vec<Uniform<'static>>, )

Overrides the default texture shader used, if none is specified.

This affects calls to Frame::render_texture_at or Frame::render_texture_from_to as well as calls to GlesFrame::render_texture_from_to or GlesFrame::render_texture, if the passed in program is None.

Override is active only for the lifetime of this GlesFrame and can be reset via GlesFrame::clear_tex_program_override.

pub fn clear_tex_program_override(&mut self)

Resets a texture shader override previously set by GlesFrame::override_default_tex_program.

pub fn draw_solid( &mut self, dest: Rectangle<i32, Physical>, damage: &[Rectangle<i32, Physical>], color: Color32F, ) -> Result<(), GlesError>

Draw a solid color to the current target at the specified destination with the specified color.

pub fn render_texture_from_to( &mut self, texture: &GlesTexture, src: Rectangle<f64, BufferCoord>, dest: Rectangle<i32, Physical>, damage: &[Rectangle<i32, Physical>], opaque_regions: &[Rectangle<i32, Physical>], transform: Transform, alpha: f32, program: Option<&GlesTexProgram>, additional_uniforms: &[Uniform<'_>], ) -> Result<(), GlesError>

Render part of a texture as given by src to the current target into the rectangle described by dst as a flat 2d-plane after applying the inverse of the given transformation. (Meaning src_transform should match the orientation of surface being rendered).

Optionally allows a custom texture program and matching additional uniforms to be passed in.

pub fn render_texture( &mut self, tex: &GlesTexture, tex_matrix: Matrix3<f32>, matrix: Matrix3<f32>, instances: Option<impl IntoIterator<Item = GLfloat>>, alpha: f32, program: Option<&GlesTexProgram>, additional_uniforms: &[Uniform<'_>], ) -> Result<(), GlesError>

Render a texture to the current target using given projection matrix and alpha.

The instances are used to define the regions which should get drawn. Each instance has to define 4 GLfloat which define the relative offset and scale for the vertex position and range from 0.0 to 1.0. The first 2 GLfloat define the relative x and y offset. The remaining 2 GLfloat define the x and y scale. This can be used to only update parts of the texture on screen.

The given texture matrix is used to transform the instances into texture coordinates. In case the texture is rotated, flipped or y-inverted the matrix has to be set up accordingly. Additionally the matrix can be used to crop the texture.

Optionally allows a custom texture program and matching additional uniforms to be passed in.

pub fn render_pixel_shader_to( &mut self, pixel_shader: &GlesPixelProgram, src: Rectangle<f64, BufferCoord>, dest: Rectangle<i32, Physical>, size: Size<i32, BufferCoord>, damage: Option<&[Rectangle<i32, Physical>]>, alpha: f32, additional_uniforms: &[Uniform<'_>], ) -> Result<(), GlesError>

Render a pixel shader into the current target at a given dest-region.

pub fn projection(&self) -> &[f32; 9]

Projection matrix for this frame

pub fn egl_context(&self) -> &EGLContext

Get access to the underlying EGLContext.

Note: Modifying the context state, might result in rendering issues. The context state is considered an implementation detail and no guarantee is made about what can or cannot be changed. To make sure a certain modification does not interfere with the renderer’s behaviour, check the source.

pub fn capabilities(&self) -> &[Capability]

Returns the supported Capabilities of the underlying renderer.

pub fn debug_flags(&self) -> DebugFlags

Returns the current enabled DebugFlags of the underlying renderer.