Struct smithay::backend::renderer::gles2::Gles2Frame

pub struct Gles2Frame<'frame> { /* private fields */ }

Handle to the currently rendered frame during Gles2Renderer::render.

Leaking this frame will prevent it from synchronizing the rendered framebuffer, which might cause glitches. Additionally parts of the GL state might not be reset correctly, causing unexpected results for later render commands. The internal GL context and framebuffer will remain valid, no re-creation will be necessary.

Implementations

impl<'frame> Gles2Frame<'frame>

pub fn with_context<F, R>(&mut self, func: F) -> Result<R, Gles2Error>where F: FnOnce(&Gles2) -> R,

Run custom code in the GL context owned by this renderer.

The OpenGL state of the renderer is considered an implementation detail and no guarantee is made about what can or cannot be changed, as such you should reset everything you change back to its previous value or check the source code of the version of Smithay you are using to ensure your changes don’t interfere with the renderer’s behavior. Doing otherwise can lead to rendering errors while using other functions of this renderer.

impl<'frame> Gles2Frame<'frame>

pub fn render_texture( &mut self, tex: &Gles2Texture, tex_matrix: Matrix3<f32>, matrix: Matrix3<f32>, instances: Option<&[GLfloat]>, alpha: f32 ) -> Result<(), Gles2Error>

Render a texture to the current target using given projection matrix and alpha.

The instances are used to define the regions which should get drawn. Each instance has to define 4 GLfloat which define the relative offset and scale for the vertex position and range from 0.0 to 1.0. The first 2 GLfloat define the relative x and y offset. The remaining 2 GLfloat define the x and y scale. This can be used to only update parts of the texture on screen.

The given texture matrix is used to transform the instances into texture coordinates. In case the texture is rotated, flipped or y-inverted the matrix has to be set up accordingly. Additionally the matrix can be used to crop the texture.

pub fn projection(&self) -> &[f32; 9]

Projection matrix for this frame