Struct smithay::backend::renderer::glow::GlowFrame

pub struct GlowFrame<'a> { /* private fields */ }

Frame implementation of a GlowRenderer.

Leaking the frame will cause the same problems as leaking a Gles2Frame.

Implementations

impl<'frame> GlowFrame<'frame>

pub fn with_context<F, R>(&mut self, func: F) -> Result<R, Gles2Error>where
F: FnOnce(&Arc<Context>) -> R,

Run custom code in the GL context owned by this renderer.

The OpenGL state of the renderer is considered an implementation detail and no guarantee is made about what can or cannot be changed, as such you should reset everything you change back to its previous value or check the source code of the version of Smithay you are using to ensure your changes don’t interfere with the renderer’s behavior. Doing otherwise can lead to rendering errors while using other functions of this renderer.