smithay::backend::renderer::glow

Struct GlowRenderer

pub struct GlowRenderer { /* private fields */ }

Available on crate feature renderer_glow only.

A renderer utilizing OpenGL ES 2 and [glow] on top for easier custom rendering.

Implementations

impl GlowRenderer

pub unsafe fn supported_capabilities( context: &EGLContext, ) -> Result<Vec<Capability>, GlesError>

Get the supported Capabilities of the renderer

Safety

This operation will cause undefined behavior if the given EGLContext is active in another thread.

pub unsafe fn new(context: EGLContext) -> Result<GlowRenderer, GlesError>

Creates a new OpenGL ES 2 + Glow renderer from a given EGLContext with all supported capabilities.

Safety

This operation will cause undefined behavior if the given EGLContext is active in another thread.

See: with_capabilities for more information

pub unsafe fn with_capabilities( context: EGLContext, capabilities: impl IntoIterator<Item = Capability>, ) -> Result<GlowRenderer, GlesError>

Creates a new OpenGL ES 2 + Glow renderer from a given EGLContext with the specified Capabilities. If a requested Capability is not supported an error will be returned.

Safety

This operation will cause undefined behavior if the given EGLContext is active in another thread.

Implementation details

Texture handles created by the resulting renderer are valid for every rendered created with an EGLContext shared with the given one (see EGLContext::new_shared) and can be used on any of these renderers.

This renderer has no default framebuffer, use Bind::bind before rendering.

Binding a new target, while another one is already bound, will replace the current target.

Shm buffers can be released after a successful import, without the texture handle becoming invalid.

Texture filtering starts with Linear-downscaling and Linear-upscaling

pub fn egl_context(&self) -> &EGLContext

Get access to the underlying EGLContext.

Note: Modifying the context state, might result in rendering issues. The context state is considered an implementation detail and no guarantee is made about what can or cannot be changed. To make sure a certain modification does not interfere with the renderer’s behaviour, check the source.

pub fn with_context<F, R>(&mut self, func: F) -> Result<R, GlesError>
where F: FnOnce(&Arc<Context>) -> R,

Run custom code in the GL context owned by this renderer.

The OpenGL state of the renderer is considered an implementation detail and no guarantee is made about what can or cannot be changed, as such you should reset everything you change back to its previous value or check the source code of the version of Smithay you are using to ensure your changes don’t interfere with the renderer’s behavior. Doing otherwise can lead to rendering errors while using other functions of this renderer.