Struct smithay::backend::renderer::multigpu::MultiFrame

pub struct MultiFrame<'render, 'target, 'frame, R, T: GraphicsApi>
where
    R: 'static + GraphicsApi + 'frame,
    R::Error: 'static,
    T::Error: 'static,
    <R::Device as ApiDevice>::Renderer: ExportMem + ImportDma + ImportMem,
    <T::Device as ApiDevice>::Renderer: ImportDma + ImportMem,
    <<R::Device as ApiDevice>::Renderer as Renderer>::Error: 'static,
    <<T::Device as ApiDevice>::Renderer as Renderer>::Error: 'static,
{ /* private fields */ }

Available on crate feature renderer_multi only.

Frame implementation of a MultiRenderer.

Leaking the frame will potentially keep it from doing necessary copies of the internal framebuffer for some multi-gpu configurations. The result would be no updated framebuffer contents. Additionally all problems related to the Frame-implementation of the underlying GraphicsApi will be present.