smithay::backend::renderer::multigpu

Trait ApiDevice

pub trait ApiDevice: Debug {
    type Renderer: Renderer;

    // Required methods
    fn renderer(&self) -> &Self::Renderer;
    fn renderer_mut(&mut self) -> &mut Self::Renderer;
    fn allocator(
        &mut self,
    ) -> &mut dyn Allocator<Buffer = Dmabuf, Error = AnyError>;
    fn node(&self) -> &DrmNode;
    fn can_do_cross_device_imports(&self) -> bool;
}

Available on crate feature renderer_multi only.

A device produced by a GraphicsApi.

Required Associated Types

type Renderer: Renderer

The Renderer this devices contains

Required Methods

fn renderer(&self) -> &Self::Renderer

Returns a reference to the underlying renderer

fn renderer_mut(&mut self) -> &mut Self::Renderer

Returns a mutable reference to the underlying renderer

fn allocator(&mut self) -> &mut dyn Allocator<Buffer = Dmabuf, Error = AnyError>

Returns an allocator for buffers renderers of this device can sample from and render to

fn node(&self) -> &DrmNode

Returns a DrmNode representing the graphics device

fn can_do_cross_device_imports(&self) -> bool

Returns whether the underlying renderer can in principle do cross-device imports. (With no guarantee on being able to import a specific buffer.)