smithay::backend::renderer

Trait BlitFrame

pub trait BlitFrame<Framebuffer>
where
    Self: Frame,
{
    // Required methods
    fn blit_to(
        &mut self,
        to: &mut Framebuffer,
        src: Rectangle<i32, Physical>,
        dst: Rectangle<i32, Physical>,
        filter: TextureFilter,
    ) -> Result<(), Self::Error>;
    fn blit_from(
        &mut self,
        from: &Framebuffer,
        src: Rectangle<i32, Physical>,
        dst: Rectangle<i32, Physical>,
        filter: TextureFilter,
    ) -> Result<(), Self::Error>;
}

Trait for frames supporting blitting contents from/to the current framebuffer to/from another.

Required Methods

fn blit_to( &mut self, to: &mut Framebuffer, src: Rectangle<i32, Physical>, dst: Rectangle<i32, Physical>, filter: TextureFilter, ) -> Result<(), Self::Error>

Copies the contents of the bound framebuffer to dst in the provided framebuffer, applying filter if necessary.

This operation is non destructive, the contents of the current framebuffer are kept intact as is any region not in dst for the target framebuffer.

This function may fail, if (but not limited to):

The bound framebuffer is not readable

The destination framebuffer is not writable

src is out of bounds for the bound framebuffer

dst is out of bounds for the destination framebuffer

src and dst sizes are different and interpolation is not supported by this renderer.

bound and target framebuffer are the same

fn blit_from( &mut self, from: &Framebuffer, src: Rectangle<i32, Physical>, dst: Rectangle<i32, Physical>, filter: TextureFilter, ) -> Result<(), Self::Error>

Copies the contents of the provided framebuffer to dst in the bound framebuffer, applying filter if necessary.

This operation is non destructive, the contents of the source framebuffer are kept intact as is any region not in dst for the bound framebuffer.

This function may fail, if (but not limited to):

The source framebuffer is not readable

The bound framebuffer is not writable

src is out of bounds for the source framebuffer

dst is out of bounds for the bound framebuffer

src and dst sizes are different and interpolation is not supported by this renderer.

source and bound framebuffer are the same

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.