Trait smithay::backend::renderer::ExportDma

pub trait ExportDma: Renderer {
    fn export_framebuffer(
        &mut self,
        size: Size<i32, BufferCoord>
    ) -> Result<Dmabuf, <Self as Renderer>::Error>;
    fn export_texture(
        &mut self,
        texture: &<Self as Renderer>::TextureId
    ) -> Result<Dmabuf, <Self as Renderer>::Error>;
}

Trait for renderers supporting exporting contents of framebuffers or textures as dmabufs.

Required Methods

fn export_framebuffer(
&mut self,
size: Size<i32, BufferCoord>
) -> Result<Dmabuf, <Self as Renderer>::Error>

Exports the currently bound framebuffer as a dmabuf.

This operation is not destructive, the contents of the framebuffer keep being valid.

This function may fail, if (but not limited to):

The framebuffer is not readable

The size is larger than the framebuffer

There is not enough space to create a copy

fn export_texture(
&mut self,
texture: &<Self as Renderer>::TextureId
) -> Result<Dmabuf, <Self as Renderer>::Error>

Exports the given texture as a dmabuf.

This operation is not destructive, the contents of the texture keep being valid.

This function may fail, if (but not limited to):

The texture cannot be exported