smithay::backend::renderer

Trait FrameContext

pub trait FrameContext<'a, 'frame, 'buffer, R: Renderer>: Frame {
    type Guard: Debug + AsRef<R> + AsMut<R> + 'a;

    // Required method
    fn renderer(&'a mut self) -> Self::Guard;
}

Helper trait for Frames, that allow referencing the underlying renderer to create additional frames for offscreen targets.

Required Associated Types

type Guard: Debug + AsRef<R> + AsMut<R> + 'a

Type returned by FrameContext::renderer which derefs into the underlying Renderer.

Required Methods

fn renderer(&'a mut self) -> Self::Guard

Receive the underlying Renderer

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.