Trait smithay::backend::renderer::ImportDma

pub trait ImportDma: Renderer {
    // Required method
    fn import_dmabuf(
        &mut self,
        dmabuf: &Dmabuf,
        damage: Option<&[Rectangle<i32, BufferCoord>]>
    ) -> Result<<Self as Renderer>::TextureId, <Self as Renderer>::Error>;

    // Provided methods
    fn dmabuf_formats(&self) -> Box<dyn Iterator<Item = Format>> { ... }
    fn has_dmabuf_format(&self, format: Format) -> bool { ... }
}

Trait for Renderers supporting importing dmabufs.

Required Methods

fn import_dmabuf( &mut self, dmabuf: &Dmabuf, damage: Option<&[Rectangle<i32, BufferCoord>]> ) -> Result<<Self as Renderer>::TextureId, <Self as Renderer>::Error>

Import a given raw dmabuf into the renderer.

Returns a texture_id, which can be used with Frame::render_texture_from_to (or Frame::render_texture_at) or implementation-specific functions.

If not otherwise defined by the implementation, this texture id is only valid for the renderer, that created it.

This operation needs no bound or default rendering target.

The implementation defines, if the id keeps being valid, if the buffer is released, to avoid relying on implementation details, keep the buffer alive, until you destroyed this texture again.

Provided Methods

fn dmabuf_formats(&self) -> Box<dyn Iterator<Item = Format>>

Returns supported formats for dmabufs.

fn has_dmabuf_format(&self, format: Format) -> bool

Test if a specific dmabuf Format is supported

Object Safety

This trait is not object safe.