Trait smithay::backend::renderer::ImportDmaWl

pub trait ImportDmaWl: ImportDma {
    // Provided method
    fn import_dma_buffer(
        &mut self,
        buffer: &WlBuffer,
        _surface: Option<&SurfaceData>,
        damage: &[Rectangle<i32, BufferCoord>]
    ) -> Result<<Self as Renderer>::TextureId, <Self as Renderer>::Error> { ... }
}

Available on crate feature wayland_frontend only.

Trait for Renderers supporting importing dmabuf-based wl_buffers

Provided Methods

fn import_dma_buffer( &mut self, buffer: &WlBuffer, _surface: Option<&SurfaceData>, damage: &[Rectangle<i32, BufferCoord>] ) -> Result<<Self as Renderer>::TextureId, <Self as Renderer>::Error>

Import a given dmabuf-based buffer into the renderer (see buffer_type).

Returns a texture_id, which can be used with Frame::render_texture_from_to (or Frame::render_texture_at) or implementation-specific functions.

If not otherwise defined by the implementation, this texture id is only valid for the renderer, that created it.

This operation needs no bound or default rendering target.

The implementation defines, if the id keeps being valid, if the buffer is released, to avoid relying on implementation details, keep the buffer alive, until you destroyed this texture again.

Object Safety

This trait is not object safe.