smithay::backend::renderer

Trait ImportMem

pub trait ImportMem: Renderer {
    // Required methods
    fn import_memory(
        &mut self,
        data: &[u8],
        format: Fourcc,
        size: Size<i32, BufferCoord>,
        flipped: bool,
    ) -> Result<Self::TextureId, Self::Error>;
    fn update_memory(
        &mut self,
        texture: &Self::TextureId,
        data: &[u8],
        region: Rectangle<i32, BufferCoord>,
    ) -> Result<(), Self::Error>;
    fn mem_formats(&self) -> Box<dyn Iterator<Item = Fourcc>>;
}

Trait for Renderers supporting importing bitmaps from memory.

Required Methods

fn import_memory( &mut self, data: &[u8], format: Fourcc, size: Size<i32, BufferCoord>, flipped: bool, ) -> Result<Self::TextureId, Self::Error>

Import a given chunk of memory into the renderer.

Returns a texture_id, which can be used with Frame::render_texture_from_to (or Frame::render_texture_at) or implementation-specific functions.

If not otherwise defined by the implementation, this texture id is only valid for the renderer, that created it. This operation needs no bound or default rendering target.

Settings flipped to true will cause the buffer to be interpreted like the y-axis is flipped (opposed to the lower left begin (0, 0)). This is a texture specific property, so future uploads to the same texture via ImportMem::update_memory will also be interpreted as flipped.

The provided data slice needs to be in a format supported as indicated by ImportMem::mem_formats. Its length should thus be size.w * size.h * bits_per_pixel. Anything beyond will be truncated, if the buffer is too small an error will be returned.

fn update_memory( &mut self, texture: &Self::TextureId, data: &[u8], region: Rectangle<i32, BufferCoord>, ) -> Result<(), Self::Error>

Update a portion of a given chunk of memory into an existing texture.

This operation needs no bound or default rendering target.

The provided data slice needs to be in the same format used to create the texture and the same size of the texture. Its length should this be texture.size().w * texture.size().h * bits_per_pixel. Anything beyond will be ignored, if the buffer is too small an error will be returned.

This function may error, if (but not limited to):

The texture was not created using either ImportMemWl::import_shm_buffer or ImportMem::import_memory. External textures imported by other means (e.g. via ImportDma) may not be writable. This property is defined by the implementation.

The region is out of bounds of the initial size the texture was created with. Implementations are not required to support resizing the original texture.

fn mem_formats(&self) -> Box<dyn Iterator<Item = Fourcc>>

Returns supported formats for memory imports.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.