Trait smithay::backend::renderer::ImportMemWl

pub trait ImportMemWl: ImportMem {
    fn import_shm_buffer(
        &mut self,
        buffer: &WlBuffer,
        surface: Option<&SurfaceData>,
        damage: &[Rectangle<i32, BufferCoord>]
    ) -> Result<<Self as Renderer>::TextureId, <Self as Renderer>::Error>;

    fn shm_formats(&self) -> &[Format] { ... }
}

Trait for Renderers supporting importing wl_buffers using shared memory.

Required Methods

fn import_shm_buffer(
&mut self,
buffer: &WlBuffer,
surface: Option<&SurfaceData>,
damage: &[Rectangle<i32, BufferCoord>]
) -> Result<<Self as Renderer>::TextureId, <Self as Renderer>::Error>

Import a given shm-based buffer into the renderer (see buffer_type).

Returns a texture_id, which can be used with Frame::render_texture_from_to (or Frame::render_texture_at) or implementation-specific functions.

If not otherwise defined by the implementation, this texture id is only valid for the renderer, that created it. This operation needs no bound or default rendering target.

The implementation defines, if the id keeps being valid, if the buffer is released, to avoid relying on implementation details, keep the buffer alive, until you destroyed this texture again.

If provided the SurfaceAttributes can be used to do caching of rendering resources and is generally recommended.

The damage argument provides a list of rectangle locating parts of the buffer that need to be updated. When provided with an empty list &[], the renderer is allowed to not update the texture at all.

Provided Methods

fn shm_formats(&self) -> &[Format]

Returns supported formats for shared memory buffers.

Will always contain At least Argb8888 and Xrgb8888.