smithay::backend::renderer

Trait Offscreen

pub trait Offscreen<Target>: Renderer + Bind<Target> {
    // Required method
    fn create_buffer(
        &mut self,
        format: Fourcc,
        size: Size<i32, BufferCoord>,
    ) -> Result<Target, Self::Error>;
}

Trait for renderers that support creating offscreen framebuffers to render into.

Usually also implement ExportMem to receive the framebuffers contents.

Required Methods

fn create_buffer( &mut self, format: Fourcc, size: Size<i32, BufferCoord>, ) -> Result<Target, Self::Error>

Create a new instance of a framebuffer.

This call may fail, if (but not limited to):

The maximum amount of framebuffers for this renderer would be exceeded

The format is not supported to be rendered into

The size is too large for a framebuffer

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.