smithay::backend::renderer
Trait Renderer

pub trait Renderer: Debug {
    type Error: Error;
    type TextureId: Texture;
    type Frame<'frame>: Frame<Error = Self::Error, TextureId = Self::TextureId> + 'frame
       where Self: 'frame;

    // Required methods
    fn id(&self) -> usize;
    fn downscale_filter(
        &mut self,
        filter: TextureFilter,
    ) -> Result<(), Self::Error>;
    fn upscale_filter(
        &mut self,
        filter: TextureFilter,
    ) -> Result<(), Self::Error>;
    fn set_debug_flags(&mut self, flags: DebugFlags);
    fn debug_flags(&self) -> DebugFlags;
    fn render(
        &mut self,
        output_size: Size<i32, Physical>,
        dst_transform: Transform,
    ) -> Result<Self::Frame<'_>, Self::Error>;
    fn wait(&mut self, sync: &SyncPoint) -> Result<(), Self::Error>;

    // Provided method
    fn cleanup_texture_cache(&mut self) -> Result<(), Self::Error> { ... }
}

Abstraction of commonly used rendering operations for compositors.

Required Associated Types

type Error: Error

Error type returned by the rendering operations of this renderer.

type TextureId: Texture

Texture Handle type used by this renderer.

type Frame<'frame>: Frame<Error = Self::Error, TextureId = Self::TextureId> + 'frame where Self: 'frame

Type representing a currently in-progress frame during the Renderer::render-call

Required Methods

fn id(&self) -> usize

Returns an id, that is unique to all renderers, that can use TextureIds originating from any of these renderers.

fn downscale_filter(&mut self, filter: TextureFilter) -> Result<(), Self::Error>

Set the filter method to be used when rendering a texture into a smaller area than its size

fn upscale_filter(&mut self, filter: TextureFilter) -> Result<(), Self::Error>

Set the filter method to be used when rendering a texture into a larger area than its size

fn set_debug_flags(&mut self, flags: DebugFlags)

Set the enabled DebugFlags

fn debug_flags(&self) -> DebugFlags

Returns the current enabled DebugFlags

fn render( &mut self, output_size: Size<i32, Physical>, dst_transform: Transform, ) -> Result<Self::Frame<'_>, Self::Error>

Initialize a rendering context on the current rendering target with given dimensions and transformation.

The output_size specifies the dimensions of the display before the dst_transform is applied.

This function may error, if:

The given dimensions are unsupported (too large) for this renderer

The given Transformation is not supported by the renderer (Transform::Normal is always supported).

This renderer implements Bind, no target was bound and has no default target.

(Renderers not implementing Bind always have a default target.)

fn wait(&mut self, sync: &SyncPoint) -> Result<(), Self::Error>

Wait for a SyncPoint to be signaled

Provided Methods

fn cleanup_texture_cache(&mut self) -> Result<(), Self::Error>

Forcibly clean up the renderer internal texture cache

Note: Resources used by the renderer will be implicitly cleaned-up after finishing a Frame by either dropping the Frame or explicitly calling Frame::finish. This call can be used to clean-up resources in cases where either no Frame is used at all to prevent resource pile-up or in case of only infrequent access to lower system resource usage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.