Module smithay::backend::renderer::utils

Utility module for helpers around drawing WlSurfaces and RenderElements with Renderers.

Structs

Bufferwayland_frontend

A wayland buffer

CommitCounter

A simple wrapper for counting commits

DamageBag

A tracker for holding damage

DamageSet

A set of damage returned from DamageBag::damage_since of DamageSnapshot::damage_since

DamageSetIter

Iterator for DamageSet::into_iter

DamageSnapshot

A snapshot of the current state of a DamageBag

OpaqueRegions

Wrapper for a set of opaque regions

OpaqueRegionsIter

Iterator for OpaqueRegions::into_iter

RendererSurfaceStatewayland_frontend

Surface state for rendering related data

SurfaceView

Defines a view into the surface

Functions

draw_render_elementswayland_frontend

Draws the render elements using a given Renderer and Frame

import_surfacewayland_frontend

Imports buffers of a surface using a given Renderer

import_surface_treewayland_frontend

Imports buffers of a surface and its subsurfaces using a given Renderer.

on_commit_buffer_handlerwayland_frontend

Handler to let smithay take over buffer management.

with_renderer_surface_statewayland_frontend

Access the buffer related states associated to this surface

Type Aliases

RendererSurfaceStateUserDatawayland_frontend

Type stored in WlSurface states data_map