Function smithay::backend::renderer::utils::draw_render_elements

pub fn draw_render_elements<'a, R, S, E>(
    frame: &mut <R as Renderer>::Frame<'a>,
    scale: S,
    elements: &[E],
    damage: &[Rectangle<i32, Physical>]
) -> Result<Option<Vec<Rectangle<i32, Physical>>>, <R as Renderer>::Error>
where
    R: Renderer,
    <R as Renderer>::TextureId: 'static,
    S: Into<Scale<f64>>,
    E: RenderElement<R>,

Available on crate feature wayland_frontend only.

Draws the render elements using a given Renderer and Frame

scale needs to be equivalent to the fractional scale the rendered result should have.

location is the position the surface should be drawn at.

damage is the set of regions that should be drawn relative to the same origin as the location.

Note: This element will render nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.