Function smithay::backend::renderer::utils::draw_surface_tree

pub fn draw_surface_tree<R, E, F, T>(
    renderer: &mut R, 
    frame: &mut F, 
    surface: &WlSurface, 
    scale: f64, 
    location: Point<i32, Logical>, 
    damage: &[Rectangle<i32, Logical>], 
    log: &Logger
) -> Result<(), R::Error> where
    R: Renderer<Error = E, TextureId = T, Frame = F> + ImportAll,
    F: Frame<Error = E, TextureId = T>,
    E: Error,
    T: Texture + 'static, 

Draws a surface and its subsurfaces using a given Renderer and Frame.

scale needs to be equivalent to the fractional scale the rendered result should have.

location is the position the surface should be drawn at.

damage is the set of regions of the surface that should be drawn.

Note: This element will render nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.