Function smithay::backend::renderer::utils::import_surface

pub fn import_surface<R>(
    renderer: &mut R,
    states: &SurfaceData,
) -> Result<(), <R as Renderer>::Error>
where
    R: Renderer + ImportAll,
    <R as Renderer>::TextureId: 'static,

Available on crate feature wayland_frontend only.

Imports buffers of a surface using a given Renderer

This (or import_surface_tree) need to be called beforedraw_render_elements, if used later.

Note: This will do nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.