smithay::backend::renderer::utils

Function on_commit_buffer_handler

pub fn on_commit_buffer_handler<D: 'static>(surface: &WlSurface)

Available on crate feature wayland_frontend only.

Handler to let smithay take over buffer management.

Needs to be called first on the commit-callback of crate::wayland::compositor::CompositorHandler::commit.

Consumes the buffer of SurfaceAttributes, the buffer will not be accessible anymore, but draw_render_elements and other draw_* helpers of the desktop module will become usable for surfaces handled this way.