smithay::backend
Module session

Available on crate feature backend_session only.

Abstraction of different session APIs.

Sessions provide a way for multiple graphical systems to run in parallel by providing mechanisms to switch between and handle device access and permissions for every running instance. They are crucial to allow unprivileged processes to use graphical or input devices.

The functions of the provided interfaces lies in two main components: the ability to open privileged devices, and be notified about when the session is paused and resumed (for example when the user switches to an other TTY, or is the computer goes to sleep).

General use

Session handling in Smithay is done through a pair of types that each session provider implements.

The first is a handle implementing the Session trait, which allows you to request the opening of devices, a VT change, or information about the session state.

The second is a notifier which informs you when the session is enabled or disabled by the system. This notifier takes the form of a calloop event source to deliver pause and activation events.

Available providers

This module provides just one session implementation, through libseat, gated by the backend_session_libseat cargo feature.

Other implementations can be provided out-of-tree.

Modules

libseatbackend_session_libseat

Implementation of the Session trait through the libseat.

Enums

Event

Events that can be generated by a session

Traits

AsErrno

Allows errors to be described by an error number

Session

General session interface.