Module smithay::backend::session::auto

Implementation of the Session trait through various implementations automatically choosing the best available interface.

How to use it

Initialization

To initialize a session just call AutoSession::new. A new session will be opened, if the any available interface is successful and will be closed once the AutoSessionNotifier is dropped.

Usage of the session

The session may be used to open devices manually through the Session interface or be passed to other objects that need it to open devices themselves. The AutoSession is clonable and may be passed to multiple devices easily.

Examples for those are e.g. the LibinputInputBackend (its context might be initialized through a Session via the LibinputSessionInterface) or the UdevBackend.

Usage of the session notifier

The notifier might be used to pause device access, when the session gets paused (e.g. by switching the tty via AutoSession::change_vt) and to automatically enable it again, when the session becomes active again.

It is crutial to avoid errors during that state. Examples for object that might be registered for notifications are the Libinput context, the UdevBackend or a DrmDevice (handled automatically by the UdevBackend, if not done manually).

Structs

AutoId
Id's used by the AutoSessionNotifier internally.

Error
The Error type.

Enums

AutoSession
Session using the best available inteface

AutoSessionNotifier
SessionNotifier using the best available inteface

BoundAutoSession
Bound session that is driven by the wayland_server::EventLoop.

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

auto_session_bind
Bind an AutoSessionNotifier to an EventLoop.

Type Definitions

Result
Convenient wrapper around std::Result.