Function smithay::backend::session::auto::auto_session_bind

pub fn auto_session_bind<Data: 'static>(
    notifier: AutoSessionNotifier, 
    handle: &LoopHandle<Data>
) -> Result<BoundAutoSession, (IoError, AutoSessionNotifier)>

Bind an AutoSessionNotifier to an EventLoop.

Allows the AutoSessionNotifier to listen for incoming signals signalling the session state. If you don't use this function AutoSessionNotifier will not correctly tell you the session state and call its SessionObservers.