Trait smithay::backend::session::AsErrno

pub trait AsErrno: Debug {
    fn as_errno(&self) -> Option<i32>;
}

Allows errors to be described by an error number

Required Methods

fn as_errno(&self) -> Option<i32>

Returns the error number representing this error if any