Trait smithay::backend::session::SessionNotifier

pub trait SessionNotifier {
    type Id: PartialEq + Eq;
    fn register<S: SessionObserver + 'static, A: AsSessionObserver<S>>(
        &mut self, 
        signal: &mut A
    ) -> Self::Id;

    fn unregister(&mut self, signal: Self::Id);

    fn is_active(&self) -> bool;

    fn seat(&self) -> &str;
}

Interface for registering for notifications for a given session.

Part of the session api which allows to get notified, when the given session gets paused or becomes active again. Any object implementing the SessionObserver trait may be registered.

Associated Types

type Id: PartialEq + Eq

Id type of registered observers

Required Methods

fn register<S: SessionObserver + 'static, A: AsSessionObserver<S>>(
&mut self,
signal: &mut A
) -> Self::Id

Registers a given SessionObserver.

Returns an id of the inserted observer, can be used to remove it again.

fn unregister(&mut self, signal: Self::Id)

Removes an observer by its given id from SessionNotifier::register.

fn is_active(&self) -> bool

Check if this session is currently active

fn seat(&self) -> &str

Which seat this session is on