Trait smithay::backend::session::SessionObserver

pub trait SessionObserver {
    fn pause(&mut self, device: Option<(u32, u32)>);

    fn activate(&mut self, device: Option<(u32, u32, Option<RawFd>)>);
}

Trait describing the ability to be notified when the session pauses or becomes active again.

It might be impossible to interact with devices while the session is disabled. This interface provides callbacks for when that happens.

Required methods

fn pause(&mut self, device: Option<(u32, u32)>)

Session/Device is about to be paused.

If only a specific device shall be closed a device number in the form of (major, minor) is provided. All observers not using the specified device should ignore the signal in that case.

fn activate(&mut self, device: Option<(u32, u32, Option<RawFd>)>)

Session/Device got active again

If only a specific device shall be activated again a device number in the form of (major, major, Option<RawFd>) is provided. Optionally the session may decide to replace the currently open file descriptor of the device with a new one. In that case the old one should not be used anymore and be closed. All observers not using the specified device should ignore the signal in that case.

Loading content...