Struct smithay::backend::udev::UdevBackend

pub struct UdevBackend<H: DrmHandler<SessionFdDrmDevice> + 'static, S: Session + 'static, T: UdevHandler<H> + 'static> { /* fields omitted */ }

Graphical backend that monitors available drm devices.

Provides a way to automatically initialize a DrmDevice for available gpus and notifies the given handler of any changes. Can be used to provide hot-plug functionality for gpus and attached monitors.

Methods

impl<H: DrmHandler<SessionFdDrmDevice> + 'static, S: Session + 'static, T: UdevHandler<H> + 'static> UdevBackend<H, S, T>

pub fn new<'a, L>(
token: LoopToken,
context: &Context,
session: S,
handler: T,
logger: L
) -> Result<UdevBackend<H, S, T>> where
L: Into<Option<Logger>>,

Creates a new UdevBackend and adds it to the given EventLoop's state.

Arguments

evlh - An event loop to use for binding DrmDevices context - An initialized udev context session - A session used to open and close devices as they become available handler - User-provided handler to respond to any detected changes logger - slog Logger to be used by the backend and its DrmDevices.

pub fn close(&mut self)

Closes the udev backend and frees all remaining open devices.