Trait smithay::backend::udev::UdevHandler

pub trait UdevHandler<H: DrmHandler<SessionFdDrmDevice> + 'static> {
    fn device_added(
        &mut self, 
        device: &mut DrmDevice<SessionFdDrmDevice>
    ) -> Option<H>;

    fn device_changed(&mut self, device: &mut DrmDevice<SessionFdDrmDevice>);

    fn device_removed(&mut self, device: &mut DrmDevice<SessionFdDrmDevice>);

    fn error(&mut self, error: IoError);
}

Handler for the UdevBackend, allows to open, close and update drm devices as they change during runtime.

Required Methods

fn device_added(
&mut self,
device: &mut DrmDevice<SessionFdDrmDevice>
) -> Option<H>

Called on initialization for every known device and when a new device is detected.

Returning a DrmHandler will initialize the device, returning None will ignore the device.

Panics

Panics if you try to borrow the token of the belonging UdevBackend using this StateProxy.

fn device_changed(&mut self, device: &mut DrmDevice<SessionFdDrmDevice>)

Called when an open device is changed.

This usually indicates that some connectors did become available or were unplugged. The handler should scan again for connected monitors and mode switch accordingly.

Panics

Panics if you try to borrow the token of the belonging UdevBackend using this StateProxy.

fn device_removed(&mut self, device: &mut DrmDevice<SessionFdDrmDevice>)

Called when a device was removed.

The device will not accept any operations anymore and its file descriptor will be closed once this function returns, any open references/tokens to this device need to be released.

Panics

Panics if you try to borrow the token of the belonging UdevBackend using this StateProxy.

fn error(&mut self, error: IoError)

Called when the udev context has encountered and error.

Panics

Panics if you try to borrow the token of the belonging UdevBackend using this StateProxy.