Struct smithay::backend::vulkan::PhysicalDevice

pub struct PhysicalDevice { /* private fields */ }

A Vulkan physical device.

A physical device refers to a Vulkan implementation. A physical device has no associated resources and may be used to create a logical device.

Implementations

impl PhysicalDevice

pub fn enumerate(
instance: &Instance
) -> VkResult<impl Iterator<Item = PhysicalDevice>>

Enumerates over all physical devices available on the system, returning an iterator of PhysicalDevice

pub fn name(&self) -> &str

Returns the name of the device.

pub fn api_version(&self) -> Version

Returns the version of Vulkan supported by this device.

Unlike the api_version property, which is the version reported by the device directly, this function returns the version the device can actually support, based on the instance’s, api_version.

The Vulkan specification provides more information about the version requirements: https://www.khronos.org/registry/vulkan/specs/1.3-extensions/html/vkspec.html#fundamentals-validusage-versions

pub fn ty(&self) -> PhysicalDeviceType

Returns the device type.

This may be used during device selection to choose a higher performance GPU.

pub fn features(&self) -> PhysicalDeviceFeatures

Returns the Vulkan 1.0 physical device features.

pub fn properties(&self) -> PhysicalDeviceProperties

Returns the physical device properties.

Some properties such as the device name can be obtained using other functions defined on PhysicalDevice.

pub fn properties_maintenance_3(&self) -> PhysicalDeviceMaintenance3Properties

Returns the device’s descriptor set properties.

This also describes the maximum memory allocation size.

pub fn id_properties(&self) -> PhysicalDeviceIDProperties

Information about universally unique identifiers (UUIDs) that identify this device.

pub fn limits(&self) -> PhysicalDeviceLimits

Returns the physical device limits.

pub fn driver(&self) -> Option<&DriverInfo>

Information about the Vulkan driver.

This may return None for a few reasons:

The Vulkan implementation is not at least 1.2

If the Vulkan implementation is not at least Vulkan 1.2, the VK_KHR_driver_properties device extension is not available.

pub fn primary_node(&self) -> Result<Option<DrmNode>, UnsupportedProperty>

Returns the major and minor numbers of the primary node which corresponds to this physical device’s DRM device.

pub fn render_node(&self) -> Result<Option<DrmNode>, UnsupportedProperty>

Returns the major and minor numbers of the render node which corresponds to this physical device’s DRM device.

Note that not every device has a render node. If there is no render node (this function returns None) then try to use the primary node.

pub unsafe fn get_properties(&self, props: &mut PhysicalDeviceProperties2)

Get physical device properties.

This function is equivalent to calling vkGetPhysicalDeviceProperties2.

Safety

All valid usage requirements for vkGetPhysicalDeviceProperties2 apply. Read the specification for more information.

pub unsafe fn get_format_properties(
&self,
format: Format,
props: &mut FormatProperties2
)

Get physical device format properties.

This function is equivalent to calling [vkGetPhysicalDeviceFormatProperties2].

Safety

All valid usage requirements for [vkGetPhysicalDeviceFormatProperties2] apply. Read the specification for more information.

pub fn get_format_modifier_properties(
&self,
format: Format
) -> Result<Vec<DrmFormatModifierPropertiesEXT>, UnsupportedProperty>

Returns properties for each supported DRM modifier for the specified format.

Returns Err if the VK_EXT_image_drm_format_modifier extension is not supported.

pub fn device_extensions(&self) -> impl Iterator<Item = &CStr>

Returns the device extensions supported by the physical device.

pub fn has_device_extension(&self, extension: &CStr) -> bool

Returns true if this device supports the specified device extension.

pub fn handle(&self) -> PhysicalDevice

Returns a handle to the underlying [vk::PhysicalDevice].

The handle refers to a specific physical device advertised by the instance. This handle is only valid for the lifetime of the instance.

pub fn instance(&self) -> &Instance

The instance which provided this physical device.