Function smithay::backend::winit::init_from_builder_with_gl_attr

pub fn init_from_builder_with_gl_attr<R>(
    builder: WindowBuilder,
    attributes: GlAttributes
) -> Result<(WinitGraphicsBackend<R>, WinitEventLoop), Error>where
    R: From<GlesRenderer> + Bind<Rc<EGLSurface>>,
    SwapBuffersError: From<<R as Renderer>::Error>,

Create a new WinitGraphicsBackend, which implements the Renderer trait, from a given WindowBuilder struct, as well as given GlAttributes for further customization of the rendering pipeline and a corresponding WinitEventLoop.