smithay::backend::winit

Struct WinitEventLoop

pub struct WinitEventLoop { /* private fields */ }

Available on crate feature backend_winit only.

Abstracted event loop of a WinitWindow.

You can register it into calloop or call dispatch_new_events periodically to receive any events.

Implementations

impl WinitEventLoop

pub fn dispatch_new_events<F>(&mut self, callback: F) -> PumpStatus
where F: FnMut(WinitEvent),

Processes new events of the underlying event loop and calls the provided callback.

You need to periodically call this function to keep the underlying event loop and WinitWindow active. Otherwise the window may not respond to user interaction.

Returns an error if the WinitWindow the window has been closed. Calling dispatch_new_events again after the WinitWindow has been closed is considered an application error and unspecified behaviour may occur.

The linked WinitGraphicsBackend will error with a lost context and should not be used anymore as well.