Struct smithay::backend::winit::WinitInputBackend

pub struct WinitInputBackend { /* fields omitted */ }

Abstracted event loop of a winit Window implementing the InputBackend trait

You need to call dispatch_new_events periodically to receive any events.

Methods

impl WinitInputBackend

pub fn set_events_handler<H: WinitEventsHandler + 'static>(
&mut self,
handler: H
)

Set the events handler

pub fn get_events_handler(&mut self) -> Option<&mut WinitEventsHandler>

Get a reference to the set events handler, if any

pub fn clear_events_handler(&mut self)

Clear out the currently set events handler