Enum smithay::compositor::TraversalAction

pub enum TraversalAction<T> {
    DoChildren(T),
    SkipChildren,
    Break,
}

Possible actions to do after handling a node diring tree traversal

Variants
DoChildren(T)

Traverse its children as well, providing them the data T

SkipChildren

Skip its children

Break

Stop traversal completely