Function smithay::compositor::compositor_init

pub fn compositor_init<U, R, ID, L>(
    evl: &mut EventLoop, 
    implem: SurfaceUserImplementation<U, R, ID>, 
    idata: ID, 
    logger: L
) -> (CompositorToken<U, R, ID>, Global<WlCompositor, SurfaceIData<U, R, ID>>, Global<WlSubcompositor, ()>) where
    L: Into<Option<Logger>>,
    U: Default + 'static,
    R: Default + RoleType + Role<SubsurfaceRole> + 'static,
    ID: 'static, 

Create new wl_compositor and wl_subcompositor globals.

The globals are directly registered into the eventloop, and this function returns a CompositorToken which you'll need access the data associated to the wl_surfaces.

It also returns the two global handles, in case you whish to remove these globals from the event loop in the future.