Struct smithay::compositor::SurfaceUserImplementation

pub struct SurfaceUserImplementation<U, R, ID> {
    pub commit: fn(_: &mut EventLoopHandle, _: &mut ID, _: &WlSurface, _: CompositorToken<U, R, ID>),
    pub frame: fn(_: &mut EventLoopHandle, _: &mut ID, _: &WlSurface, _: WlCallback, _: CompositorToken<U, R, ID>),
}

Sub-implementation for surface event handling

The global provided by Smithay cannot process these events for you, so they are forwarded directly to this implementation that you must provide at creation of the compositor global.

Fields
commit: fn(_: &mut EventLoopHandle, _: &mut ID, _: &WlSurface, _: CompositorToken<U, R, ID>)

The double-buffered state has been validated by the client

At this point, the pending state that has been accumulated in the SurfaceAttributes associated to this surface should be integrated into the current state of the surface.

See wayland_server::protocol::wl_surface::Implementation::commit for more details

frame: fn(_: &mut EventLoopHandle, _: &mut ID, _: &WlSurface, _: WlCallback, _: CompositorToken<U, R, ID>)

The client asks to be notified when would be a good time to update the contents of this surface

You must keep the provided WlCallback and trigger it at the appropriate time by calling its done() method.

See wayland_server::protocol::wl_surface::Implementation::frame for more details