Function smithay::desktop::draw_layer_surface

pub fn draw_layer_surface<R, E, F, T, P>(
    renderer: &mut R, 
    frame: &mut F, 
    layer: &LayerSurface, 
    scale: f64, 
    location: P, 
    damage: &[Rectangle<i32, Logical>], 
    log: &Logger
) -> Result<(), R::Error> where
    R: Renderer<Error = E, TextureId = T, Frame = F> + ImportAll,
    F: Frame<Error = E, TextureId = T>,
    E: Error,
    T: Texture + 'static,
    P: Into<Point<i32, Logical>>, 

Renders a given LayerSurface using a provided renderer and frame.

scale needs to be equivalent to the fractional scale the rendered result should have.

location is the position the layer surface should be drawn at.

damage is the set of regions of the layer surface that should be drawn.

Note: This function will render nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.