Function smithay::desktop::layer_map_for_output

pub fn layer_map_for_output(o: &Output) -> RefMut<'_, LayerMap>

Retrieve a LayerMap for a given Output.

If none existed before a new empty LayerMap is attached to the output and returned on subsequent calls.

Note: This function internally uses a RefCell per Output as exposed by its return type. Therefor trying to hold on to multiple references of a LayerMap of the same output using this function will result in a panic.