Struct smithay::desktop::space::SurfaceTree

pub struct SurfaceTree {
    pub surface: WlSurface,
    pub position: Point<i32, Logical>,
}

Generic helper for drawing WlSurfaces and their subsurfaces as custom elements via RenderElement.

For example useful for cursor or drag-and-drop surfaces.

Note: This element will render nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.

Fields
surface: WlSurface

Surface to be drawn

position: Point<i32, Logical>

Position to draw add