smithay::desktop

Struct LayerMap

pub struct LayerMap { /* private fields */ }

Available on crate features wayland_frontend and desktop only.

Map of LayerSurfaces on an Output

Implementations

impl LayerMap

pub fn map_layer(&mut self, layer: &LayerSurface) -> Result<(), LayerError>

Map a LayerSurface to this LayerMap.

pub fn unmap_layer(&mut self, layer: &LayerSurface)

Remove a LayerSurface from this LayerMap.

pub fn non_exclusive_zone(&self) -> Rectangle<i32, Logical>

Return the area of this output, that is not exclusive to any LayerSurfaces.

pub fn layer_geometry( &self, layer: &LayerSurface, ) -> Option<Rectangle<i32, Logical>>

Returns the geometry of a given mapped LayerSurface.

If the surface was not previously mapped onto this layer map, this function return None.

pub fn layer_under<P: Into<Point<f64, Logical>>>( &self, layer: WlrLayer, point: P, ) -> Option<&LayerSurface>

Returns a LayerSurface under a given point and on a given layer, if any.

pub fn layers(&self) -> impl DoubleEndedIterator<Item = &LayerSurface>

Iterator over all LayerSurfaces currently mapped.

pub fn layers_on( &self, layer: WlrLayer, ) -> impl DoubleEndedIterator<Item = &LayerSurface>

Iterator over all LayerSurfaces currently mapped on a given layer.

pub fn layer_for_surface( &self, surface: &WlSurface, surface_type: WindowSurfaceType, ) -> Option<&LayerSurface>

Returns the LayerSurface matching a given WlSurface, if any.

surface_type can be used to limit the types of surfaces queried for equality.

pub fn arrange(&mut self) -> bool

Force re-arranging the layer surfaces, e.g. when the output size changes.

Note: Mapping or unmapping a layer surface will automatically cause a re-arrangement.

Return whenever any position or size changes of existing surfaces were necessary.

pub fn cleanup(&mut self)

Cleanup some internally used resources.

This function needs to be called periodically (though not necessarily frequently) to be able cleanup internally used resources.

pub fn len(&self) -> usize

Returns layers count