Struct smithay::desktop::LayerSurface

pub struct LayerSurface(_);

A LayerSurface represents a single layer surface as given by the wlr-layer-shell protocol.

Implementations

impl LayerSurface

pub fn new(surface: WlrLayerSurface, namespace: String) -> LayerSurface

Create a new LayerSurface from a given WlrLayerSurface and its namespace.

pub fn layer_surface(&self) -> &WlrLayerSurface

Returns the underlying WlrLayerSurface

pub fn wl_surface(&self) -> &WlSurface

Returns the underlying [WlSurface]

pub fn cached_state(&self) -> LayerSurfaceCachedState

Returns the cached protocol state

pub fn can_receive_keyboard_focus(&self) -> bool

Returns true, if the surface has indicated, that it is able to process keyboard events.

pub fn layer(&self) -> WlrLayer

Returns the layer this surface resides on, if any yet.

pub fn namespace(&self) -> &str

Returns the namespace of this surface

pub fn bbox(&self) -> Rectangle<i32, Logical>

Returns the bounding box over this layer surface and its subsurfaces.

pub fn bbox_with_popups(&self) -> Rectangle<i32, Logical>

Returns the bounding box over this layer surface, it subsurfaces as well as any popups.

Note: You need to use a PopupManager to track popups, otherwise the bounding box will not include the popups.

pub fn surface_under<P: Into<Point<f64, Logical>>>( &self, point: P, surface_type: WindowSurfaceType ) -> Option<(WlSurface, Point<i32, Logical>)>

Finds the topmost surface under this point if any and returns it together with the location of this surface.

point needs to be relative to (0,0) of the layer surface.

pub fn send_frame<T, F>( &self, output: &Output, time: T, throttle: Option<Duration>, primary_scan_out_output: F )where T: Into<Duration>, F: FnMut(&WlSurface, &SurfaceData) -> Option<Output> + Copy,

Sends the frame callback to all the subsurfaces in this layer that requested it

See send_frames_surface_tree for more information

pub fn send_dmabuf_feedback<'a, P, F>( &self, output: &Output, primary_scan_out_output: P, select_dmabuf_feedback: F )where P: FnMut(&WlSurface, &SurfaceData) -> Option<Output> + Copy, F: Fn(&WlSurface, &SurfaceData) -> &'a DmabufFeedback + Copy,

Sends the dmabuf feedback to all the subsurfaces in this window that requested it

See send_dmabuf_feedback_surface_tree for more information

pub fn take_presentation_feedback<F1, F2>( &self, output_feedback: &mut OutputPresentationFeedback, primary_scan_out_output: F1, presentation_feedback_flags: F2 )where F1: FnMut(&WlSurface, &SurfaceData) -> Option<Output> + Copy, F2: FnMut(&WlSurface, &SurfaceData) -> Kind + Copy,

Takes the [PresentationFeedbackCallback]s from all subsurfaces in this layer

see take_presentation_feedback_surface_tree for more information

pub fn with_surfaces<F>(&self, processor: F)where F: FnMut(&WlSurface, &SurfaceData) + Copy,

Run a closure on all surfaces in this layer (including it’s popups, if PopupManager is used)

pub fn user_data(&self) -> &UserDataMap

Returns a UserDataMap to allow associating arbitrary data with this surface.