smithay::desktop

Struct PopupGrab

pub struct PopupGrab<D>
where
    D: SeatHandler + 'static,
    <D as SeatHandler>::KeyboardFocus: WaylandFocus,
    <D as SeatHandler>::PointerFocus: From<<D as SeatHandler>::KeyboardFocus> + WaylandFocus,
{ /* private fields */ }

Available on crate features wayland_frontend and desktop only.

Represents the explicit grab a client requested for a popup

An explicit grab can be used by a client to redirect all keyboard input to a single popup. The focus of the keyboard will stay on the popup for as long as the grab is valid, that is as long as the compositor did not call ungrab or the client did not destroy the popup. A grab can be nested by requesting a grab on a popup whoโs parent is the currently grabbed popup. The grab will be returned to the parent after the popup has been dismissed.

This module also provides default implementations for KeyboardGrab and PointerGrab that implement the behavior described in the xdg-shell specification. See PopupKeyboardGrab and PopupPointerGrab for more information on the default implementations.

In case the implemented behavior is not suited for your use-case the grab can be either decorated or a custom KeyboardGrab/PointerGrab can use the methods on the PopupGrab to implement a custom behavior.

One example would be to use a timer to automatically dismiss the popup after some timeout.

The grab is obtained by calling PopupManager::grab_popup.

Implementations

impl<D> PopupGrab<D>
where D: SeatHandler + 'static, <D as SeatHandler>::KeyboardFocus: WaylandFocus + From<PopupKind>, <D as SeatHandler>::PointerFocus: From<<D as SeatHandler>::KeyboardFocus> + WaylandFocus,

pub fn serial(&self) -> Serial

Returns the serial that was used to grab the popup

pub fn previous_serial(&self) -> Option<Serial>

Returns the previous serial that was used to grab the parent popup in case of nested grabs

pub fn has_ended(&self) -> bool

Check if this grab has ended

A grab has ended if either all popups associated with the grab have been dismissed by the server with PopupGrab::ungrab or by the client by destroying the popup.

This will also return false if the root of the grab has been destroyed.

pub fn current_grab(&self) -> Option<<D as SeatHandler>::KeyboardFocus>

Returns the current grabbed WlSurface.

If the grab has ended this will return the root surface so that the client expected focus can be restored

pub fn ungrab(&mut self, strategy: PopupUngrabStrategy) -> Option<WlSurface>

Ungrab and dismiss a popup

This will dismiss either the topmost or all popups according to the specified PopupUngrabStrategy

Returns the new topmost popup in case of nested popups or if the grab has ended the root surface

pub fn keyboard_grab_start_data(&self) -> &KeyboardGrabStartData<D>

Convenience method for getting a KeyboardGrabStartData for this grab.

The focus of the KeyboardGrabStartData will always be the root of the popup grab, e.g. the surface of the toplevel, to make sure the grab is not automatically unset.

pub fn pointer_grab_start_data(&self) -> &PointerGrabStartData<D>

Convenience method for getting a PointerGrabStartData for this grab.

The focus of the PointerGrabStartData will always be the root of the popup grab, e.g. the surface of the toplevel, to make sure the grab is not automatically unset.