Struct smithay::desktop::PopupPointerGrab

pub struct PopupPointerGrab<D>
where
    D: SeatHandler + 'static,
    <D as SeatHandler>::KeyboardFocus: WaylandFocus,
    <D as SeatHandler>::PointerFocus: From<<D as SeatHandler>::KeyboardFocus> + WaylandFocus,
{ /* private fields */ }

Available on crate features wayland_frontend and desktop only.

Default implementation of a PointerGrab for PopupGrab

The PopupPointerGrab will make sure that the pointer focus stays on the same client as the grabbed popup (similar to an “owner-events” grab in X11 parlance). If an input event happens outside of the grabbed WlSurface the popup will be dismissed and the grab ends. In case of a nested grab all parent grabs will also be dismissed.

If the grab has ended the pointer focus is restored and the PointerGrab is unset. Additional it will unset an active KeyboardGrab that matches the Serial of this grab and restore the keyboard focus like described in PopupKeyboardGrab

Implementations

impl<D> PopupPointerGrab<D>
where D: SeatHandler + 'static, <D as SeatHandler>::KeyboardFocus: WaylandFocus, <D as SeatHandler>::PointerFocus: From<<D as SeatHandler>::KeyboardFocus> + WaylandFocus,

pub fn new(popup_grab: &PopupGrab<D>) -> Self

Create a PopupPointerGrab for the provided PopupGrab