smithay::desktop::utils

Function bbox_from_surface_tree

pub fn bbox_from_surface_tree<P>(
    surface: &WlSurface,
    location: P,
) -> Rectangle<i32, Logical>
where
    P: Into<Point<i32, Logical>>,

Available on crate features wayland_frontend and desktop only.

Returns the bounding box of a given surface and all its subsurfaces.

location can be set to offset the returned bounding box.