Function smithay::desktop::utils::send_dmabuf_feedback_surface_tree

pub fn send_dmabuf_feedback_surface_tree<'a, P, F>(
    surface: &WlSurface,
    output: &Output,
    primary_scan_out_output: P,
    select_dmabuf_feedback: F
)
where
    P: FnMut(&WlSurface, &SurfaceData) -> Option<Output>,
    F: Fn(&WlSurface, &SurfaceData) -> &'a DmabufFeedback,

Available on crate features wayland_frontend and desktop only.

Sends dmabuf feedback for a surface and its subsurfaces with the given select function.

The dmabuf feedback for a WlSurface will only be sent if the primary scan-out output equals the provided output and the surface has requested dmabuf feedback.