smithay::desktop::utils

Function send_frames_surface_tree

pub fn send_frames_surface_tree<T, F>(
    surface: &WlSurface,
    output: &Output,
    time: T,
    throttle: Option<Duration>,
    primary_scan_out_output: F,
)
where
    T: Into<Duration>,
    F: FnMut(&WlSurface, &SurfaceData) -> Option<Output>,

Available on crate features wayland_frontend and desktop only.

Sends frame callbacks for a surface and its subsurfaces with the given time.

The frame callbacks for a WlSurface will only be sent if the primary scan-out output equals the provided output or if the surface has no primary scan-out output and the frame callback is overdue. A frame callback is considered overdue if the last time a frame callback has been sent is greater than the provided throttle threshold. If the threshold is None this will never send frame callbacks for a surface that is not visible. Specifying Duration::ZERO as the throttle threshold will always send frame callbacks for non visible surfaces.