smithay::desktop::utils

Function under_from_surface_tree

pub fn under_from_surface_tree<P>(
    surface: &WlSurface,
    point: Point<f64, Logical>,
    location: P,
    surface_type: WindowSurfaceType,
) -> Option<(WlSurface, Point<i32, Logical>)>
where
    P: Into<Point<i32, Logical>>,

Available on crate features wayland_frontend and desktop only.

Returns the topmost (sub-)surface under a given position matching the input regions of the surface.

In case no surface input region matches the point None is returned.

point has to be the position to query, relative to (0, 0) of the given surface + location.

location can be used to offset the returned point.