smithay::input::dnd

Struct DnDGrab

pub struct DnDGrab<D: SeatHandler, S: Source, F: DndFocus<D> + 'static> { /* private fields */ }

Grab during a client-initiated DnD operation.

Implementations

impl<D: SeatHandler, S: Source> DnDGrab<D, S, D::PointerFocus>
where D::PointerFocus: DndFocus<D>,

pub fn new_pointer( dh: &DisplayHandle, start_data: PointerGrabStartData<D>, source: S, seat: Seat<D>, ) -> Self

Create a new DnDGrab from an implicit pointer grab for a given source

impl<D: SeatHandler, S: Source> DnDGrab<D, S, D::TouchFocus>
where D::TouchFocus: DndFocus<D>,

pub fn new_touch( dh: &DisplayHandle, start_data: TouchGrabStartData<D>, source: S, seat: Seat<D>, ) -> Self

Create a new DnDGrab from an implicit touch grab for a given source