smithay::input::dnd

Trait DndFocus

pub trait DndFocus<D: SeatHandler>:
    WaylandFocus
    + IsAlive
    + PartialEq {
    type OfferData<S>: OfferData
       where S: Source;

    // Required methods
    fn enter<S: Source>(
        &self,
        data: &mut D,
        dh: &DisplayHandle,
        source: Arc<S>,
        seat: &Seat<D>,
        location: Point<f64, Logical>,
        serial: &Serial,
    ) -> Option<Self::OfferData<S>>;
    fn motion<S: Source>(
        &self,
        data: &mut D,
        offer: Option<&mut Self::OfferData<S>>,
        seat: &Seat<D>,
        location: Point<f64, Logical>,
        time: u32,
    );
    fn leave<S: Source>(
        &self,
        data: &mut D,
        offer: Option<&mut Self::OfferData<S>>,
        seat: &Seat<D>,
    );
    fn drop<S: Source>(
        &self,
        data: &mut D,
        offer: Option<&mut Self::OfferData<S>>,
        seat: &Seat<D>,
    );
}

Available on crate feature xwayland only.

A potential Drag’n’Drop target

Required Associated Types

type OfferData<S>: OfferData where S: Source

OfferData implementation returned by this target

Required Methods

fn enter<S: Source>( &self, data: &mut D, dh: &DisplayHandle, source: Arc<S>, seat: &Seat<D>, location: Point<f64, Logical>, serial: &Serial, ) -> Option<Self::OfferData<S>>

An active Drag’n’Drop operation has entered the client

fn motion<S: Source>( &self, data: &mut D, offer: Option<&mut Self::OfferData<S>>, seat: &Seat<D>, location: Point<f64, Logical>, time: u32, )

An active Drag’n’Drop operation, which has previously entered the client, has been moved

fn leave<S: Source>( &self, data: &mut D, offer: Option<&mut Self::OfferData<S>>, seat: &Seat<D>, )

An active Drag’n’Drop operation, which has previously entered the client, left again.

fn drop<S: Source>( &self, data: &mut D, offer: Option<&mut Self::OfferData<S>>, seat: &Seat<D>, )

An active Drag’n’Drop operation, which has previously entered the client, has been dropped.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.