smithay::input::dnd

Trait DndGrabHandler

pub trait DndGrabHandler: SeatHandler + Sized {
    // Provided method
    fn dropped(
        &mut self,
        target: Option<DndTarget<'_, Self>>,
        validated: bool,
        seat: Seat<Self>,
        location: Point<f64, Logical>,
    ) { ... }
}

Events that are generated during drag’n’drop

Provided Methods

fn dropped( &mut self, target: Option<DndTarget<'_, Self>>, validated: bool, seat: Seat<Self>, location: Point<f64, Logical>, )

The drag’n’drop action was finished by the user releasing the pointer button / touch inputs.

At this point, any icon should be removed.

target - The target that the contents were dropped on.

validated - Whether the drop offer was negotiated and accepted. If false, the drop was cancelled or otherwise not successful.

seat - The seat on which the DnD action was finished.

location - The location the drop was finished at

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.