smithay::input::dnd

Trait Source

pub trait Source:
    IsAlive
    + Send
    + Sync
    + 'static {
    // Required methods
    fn metadata(&self) -> Option<SourceMetadata>;
    fn choose_action(&self, action: DndAction);
    fn send(&self, mime_type: &str, fd: OwnedFd);
    fn drop_performed(&self);
    fn cancel(&self);
    fn finished(&self);

    // Provided method
    fn is_client_local(&self, target: &dyn Any) -> bool { ... }
}

A Drag’n’Drop data source

Required Methods

fn metadata(&self) -> Option<SourceMetadata>

Access the metadata associated with this source.

If this returns None the source is not managed by smithay (e.g. client_local)

fn choose_action(&self, action: DndAction)

An action was selected by the target

fn send(&self, mime_type: &str, fd: OwnedFd)

The target requests data to be transferred to the given file descriptor for the given mime-type

fn drop_performed(&self)

A drop was performed

fn cancel(&self)

The source is cancelled

fn finished(&self)

The source is done

Provided Methods

fn is_client_local(&self, target: &dyn Any) -> bool

Method specifically for implementing drag’n’drop operations, which are only visible to a particular client with data being transferred out-of-band.