Struct smithay::input::keyboard::KeyboardHandle

pub struct KeyboardHandle<D: SeatHandler> { /* private fields */ }

An handle to a keyboard handler

It can be cloned and all clones manipulate the same internal state.

This handle gives you 2 main ways to interact with the keyboard handling:

set the current focus for this keyboard: designing the surface that will receive the key inputs using the KeyboardHandle::set_focus method.

process key inputs from the input backend, allowing them to be caught at the compositor-level or forwarded to the client. See the documentation of the KeyboardHandle::input method for details.

Implementations

impl<D: SeatHandler + 'static> KeyboardHandle<D>

pub fn set_grab<G: KeyboardGrab<D> + 'static>(&self, grab: G, serial: Serial)

Change the current grab on this keyboard to the provided grab

Overwrites any current grab.

pub fn unset_grab(&self)

Remove any current grab on this keyboard, resetting it to the default behavior

pub fn has_grab(&self, serial: Serial) -> bool

Check if this keyboard is currently grabbed with this serial

pub fn is_grabbed(&self) -> bool

Check if this keyboard is currently being grabbed

pub fn grab_start_data(&self) -> Option<GrabStartData<D>>

Returns the start data for the grab, if any.

pub fn input<T, F>(
&self,
data: &mut D,
keycode: u32,
state: KeyState,
serial: Serial,
time: u32,
filter: F
) -> Option<T>where
F: FnOnce(&mut D, &ModifiersState, KeysymHandle<'_>) -> FilterResult<T>,

Handle a keystroke

All keystrokes from the input backend should be fed in order to this method of the keyboard handler. It will internally track the state of the keymap.

The filter argument is expected to be a closure which will peek at the generated input as interpreted by the keymap before it is forwarded to the focused client. If this closure returns FilterResult::Forward, the input will not be sent to the client. If it returns FilterResult::Intercept a value can be passed to be returned by the whole function. This mechanism can be used to implement compositor-level key bindings for example.

The module [crate::wayland::seat::keysyms] exposes definitions of all possible keysyms to be compared against. This includes non-character keysyms, such as XF86 special keys.

pub fn set_focus(
&self,
data: &mut D,
focus: Option<<D as SeatHandler>::KeyboardFocus>,
serial: Serial
)

Set the current focus of this keyboard

If the new focus is different from the previous one, any previous focus will be sent a wl_keyboard::Event::Leave event, and if the new focus is not None, a wl_keyboard::Event::Enter event will be sent.

pub fn is_focused(&self) -> bool

Check if keyboard has focus

pub fn change_repeat_info(&self, rate: i32, delay: i32)

Change the repeat info configured for this keyboard

impl<D> KeyboardHandle<D>where
D: SeatHandler,
<D as SeatHandler>::KeyboardFocus: Clone,

pub fn current_focus(&self) -> Option<<D as SeatHandler>::KeyboardFocus>

Retrieve the current keyboard focus

impl<D: SeatHandler + 'static> KeyboardHandle<D>where
D: SeatHandler + 'static,
<D as SeatHandler>::KeyboardFocus: WaylandFocus,

pub fn client_of_object_has_focus(&self, id: &ObjectId) -> bool

Check if client of given resource currently has keyboard focus