Struct smithay::input::keyboard::KeysymHandle

pub struct KeysymHandle<'a> { /* private fields */ }

Handle to the underlying keycode to allow for different conversions

Implementations

impl<'a> KeysymHandle<'a>

pub fn modified_sym(&self) -> Keysym

Returns the sym for the underlying keycode with all modifications by the current keymap state applied.

This function is similar to KeysymHandle::modified_syms, but is intended for cases where the user does not want to or cannot handle multiple keysyms.

If the key does not have exactly one keysym, returns keysyms::KEY_NoSymbol.

pub fn modified_syms(&self) -> &[Keysym]

Returns the syms for the underlying keycode with all modifications by the current keymap state applied.

pub fn raw_syms(&self) -> &[Keysym]

Returns the syms for the underlying keycode without any modifications by the current keymap state applied.

pub fn raw_latin_sym_or_raw_current_sym(&self) -> Option<Keysym>

Get the raw latin keysym or fallback to current raw keysym.

This method is handy to implement layout agnostic bindings. Keep in mind that it could be not-ideal to use just this function, since some layouts utilize non-standard shift levels and you should look into Self::modified_sym first.

The None is returned when the underlying keycode doesn’t produce a valid keysym.

pub fn raw_code(&'a self) -> Keycode

Returns the raw code in X keycode system (shifted by 8)