Trait smithay::input::keyboard::KeyboardTarget

pub trait KeyboardTarget<D>: IsAlive + PartialEq + Clone + Debug + Send
where
    D: SeatHandler,
{
    // Required methods
    fn enter(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        keys: Vec<KeysymHandle<'_>>,
        serial: Serial
    );
    fn leave(&self, seat: &Seat<D>, data: &mut D, serial: Serial);
    fn key(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        key: KeysymHandle<'_>,
        state: KeyState,
        serial: Serial,
        time: u32
    );
    fn modifiers(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        modifiers: ModifiersState,
        serial: Serial
    );
}

Trait representing object that can receive keyboard interactions

Required Methods

fn enter( &self, seat: &Seat<D>, data: &mut D, keys: Vec<KeysymHandle<'_>>, serial: Serial )

Keyboard focus of a given seat was assigned to this handler

fn leave(&self, seat: &Seat<D>, data: &mut D, serial: Serial)

The keyboard focus of a given seat left this handler

fn key( &self, seat: &Seat<D>, data: &mut D, key: KeysymHandle<'_>, state: KeyState, serial: Serial, time: u32 )

A key was pressed on a keyboard from a given seat

fn modifiers( &self, seat: &Seat<D>, data: &mut D, modifiers: ModifiersState, serial: Serial )

Hold modifiers were changed on a keyboard from a given seat

Object Safety

This trait is not object safe.