Function smithay::input::keyboard::xkb::utf32_to_keysym

pub fn utf32_to_keysym(ucs: u32) -> Keysym

Get the keysym corresponding to a Unicode/UTF-32 codepoint.

Returns the keysym corresponding to the specified Unicode codepoint, or KEY_NoSymbol if there is none.

This function is the inverse of keysym_to_utf32. In cases where a single codepoint corresponds to multiple keysyms, returns the keysym with the lowest value.

Unicode codepoints which do not have a special (legacy) keysym encoding use a direct encoding scheme. These keysyms don’t usually have an associated keysym constant (XKB_KEY_*).

For noncharacter Unicode codepoints and codepoints outside of the defined Unicode planes this function returns KEY_NoSymbol.