smithay::input::pointer

Struct AxisFrame

pub struct AxisFrame {
    pub source: Option<AxisSource>,
    pub relative_direction: (AxisRelativeDirection, AxisRelativeDirection),
    pub time: u32,
    pub axis: (f64, f64),
    pub v120: Option<(i32, i32)>,
    pub stop: (bool, bool),
}

A frame of pointer axis events. Frames of axis events should be considers as one logical action.

Can be used with the builder pattern, e.g.:

AxisFrame::new()
    .source(AxisSource::Wheel)
    .discrete(Axis::Vertical, 6)
    .value(Axis::Vertical, 30, time)
    .stop(Axis::Vertical);

Fields
source: Option<AxisSource>

Source of the axis event, if known

relative_direction: (AxisRelativeDirection, AxisRelativeDirection)

Direction of the physical motion that caused axis event

time: u32

Time of the axis event

axis: (f64, f64)

Raw scroll value per axis of the event

v120: Option<(i32, i32)>

Discrete representation of scroll value per axis, if available

stop: (bool, bool)

If the axis is considered having stopped movement

Only useful in conjunction of AxisSource::Finger events

Implementations

impl AxisFrame

pub fn new(time: u32) -> Self

Create a new frame of axis events

pub fn source(self, source: AxisSource) -> Self

Specify the source of the axis events

This event is optional, if no source is known, you can ignore this call. Only one source event is allowed per frame.

Using the AxisSource::Finger requires a stop event to be send, when the user lifts off the finger (not necessarily in the same frame).

pub fn relative_direction( self, axis: Axis, relative_direction: AxisRelativeDirection, ) -> Self

Specify the direction of the physical motion, relative to axis direction

pub fn v120(self, axis: Axis, steps: i32) -> Self

Specify discrete scrolling steps additionally to the computed value.

This event is optional and gives the client additional information about the nature of the axis event. E.g. a scroll wheel might issue separate steps, while a touchpad may never issue this event as it has no steps.

pub fn value(self, axis: Axis, value: f64) -> Self

The actual scroll value. This event is the only required one, but can also be send multiple times. The values off one frame will be accumulated by the client.

pub fn stop(self, axis: Axis) -> Self

Notification of stop of scrolling on an axis.

This event is required for sources of the AxisSource::Finger type and otherwise optional.