smithay::input::pointer

Struct PointerHandle

pub struct PointerHandle<D: SeatHandler> { /* private fields */ }

An handle to a pointer handler

It can be cloned and all clones manipulate the same internal state.

This handle gives you access to an interface to send pointer events to your clients.

When sending events using this handle, they will be intercepted by a pointer grab if any is active. See the PointerGrab trait for details.

Implementations

impl<D: SeatHandler + 'static> PointerHandle<D>

pub fn set_grab<G: PointerGrab<D> + 'static>( &self, data: &mut D, grab: G, serial: Serial, focus: Focus, )

Change the current grab on this pointer to the provided grab

If focus is set to Focus::Clear any currently focused surface will be unfocused.

Overwrites any current grab.

pub fn unset_grab(&self, data: &mut D, serial: Serial, time: u32)

Remove any current grab on this pointer, resetting it to the default behavior

pub fn has_grab(&self, serial: Serial) -> bool

Check if this pointer is currently grabbed with this serial

pub fn is_grabbed(&self) -> bool

Check if this pointer is currently being grabbed

pub fn grab_start_data(&self) -> Option<GrabStartData<D>>

Returns the start data for the grab, if any.

pub fn with_grab<T>( &self, f: impl FnOnce(Serial, &dyn PointerGrab<D>) -> T, ) -> Option<T>

Calls f with the active grab, if any.

pub fn motion( &self, data: &mut D, focus: Option<(<D as SeatHandler>::PointerFocus, Point<f64, Logical>)>, event: &MotionEvent, )

Notify that the pointer moved

You provide the new location of the pointer, in the form of:

The coordinates of the pointer in the global compositor space

The surface on top of which the cursor is, and the coordinates of its origin in the global compositor space (or None of the pointer is not on top of a client surface).

This will internally take care of notifying the appropriate client objects of enter/motion/leave events.

pub fn relative_motion( &self, data: &mut D, focus: Option<(<D as SeatHandler>::PointerFocus, Point<f64, Logical>)>, event: &RelativeMotionEvent, )

Notify about relative pointer motion

This will internally send the appropriate button event to the client objects matching with the currently focused surface, if the client uses the relative pointer protocol.

pub fn button(&self, data: &mut D, event: &ButtonEvent)

Notify that a button was pressed

This will internally send the appropriate button event to the client objects matching with the currently focused surface.

pub fn axis(&self, data: &mut D, details: AxisFrame)

Start an axis frame

A single frame will group multiple scroll events as if they happened in the same instance.

pub fn frame(&self, data: &mut D)

End of a pointer frame

A frame groups associated events. This terminates the frame.

pub fn gesture_swipe_begin(&self, data: &mut D, event: &GestureSwipeBeginEvent)

Notify about swipe gesture begin

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_swipe_update( &self, data: &mut D, event: &GestureSwipeUpdateEvent, )

Notify about swipe gesture update

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_swipe_end(&self, data: &mut D, event: &GestureSwipeEndEvent)

Notify about swipe gesture end

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_pinch_begin(&self, data: &mut D, event: &GesturePinchBeginEvent)

Notify about pinch gesture begin

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_pinch_update( &self, data: &mut D, event: &GesturePinchUpdateEvent, )

Notify about pinch gesture update

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_pinch_end(&self, data: &mut D, event: &GesturePinchEndEvent)

Notify about pinch gesture end

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_hold_begin(&self, data: &mut D, event: &GestureHoldBeginEvent)

Notify about hold gesture begin

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn gesture_hold_end(&self, data: &mut D, event: &GestureHoldEndEvent)

Notify about hold gesture end

This will internally send the appropriate event to the client objects matching with the currently focused surface, if the client uses the pointer gestures protocol.

pub fn current_location(&self) -> Point<f64, Logical>

Access the current location of this pointer in the global space

pub fn set_location(&self, location: Point<f64, Logical>)

Update the current location of this pointer in the global space, without sending any event and without updating the focus.

If you want to update the location, and update the focus, and send events, use Self::motion instead of this.

This is useful when the pointer is only moved by relative events, such as when a pointer lock is held by the focused surface. The client can give us a cursor position hint, which corresponds to the actual location the client may be rendering a pointer at. This position hint should be used as the initial location when the pointer lock is deactivated.

The next time Self::motion is called, the focus will be updated accordingly as if this function was never called. Clients will never be notified of a location hint.

pub fn last_enter(&self) -> Option<Serial>

Available on crate feature wayland_frontend only.

Access the Serial of the last pointer_enter event, if that focus is still active.

In other words this will return None again, once a pointer_leave event occurred.

impl<D> PointerHandle<D>
where D: SeatHandler, <D as SeatHandler>::PointerFocus: Clone,

pub fn current_focus(&self) -> Option<<D as SeatHandler>::PointerFocus>

Retrieve the current pointer focus

impl<D: SeatHandler + 'static> PointerHandle<D>

pub fn from_resource(seat: &WlPointer) -> Option<Self>

Available on crate feature wayland_frontend only.

Attempt to retrieve a PointerHandle from an existing resource

May return None for a valid WlPointer that was created without the pointer capability.

pub fn client_pointers<'a>( &'a self, client: &Client, ) -> impl Iterator<Item = WlPointer> + 'a

Available on crate feature wayland_frontend only.

Return all raw WlPointer instances for a particular Client