Trait smithay::input::pointer::PointerGrab

pub trait PointerGrab<D: SeatHandler>: Send {
    fn motion(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        focus: Option<(<D as SeatHandler>::PointerFocus, Point<i32, Logical>)>,
        event: &MotionEvent
    );
    fn button(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &ButtonEvent
    );
    fn axis(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        details: AxisFrame
    );
    fn start_data(&self) -> &GrabStartData<D>;
}

A trait to implement a pointer grab

In some context, it is necessary to temporarily change the behavior of the pointer. This is typically known as a pointer grab. A typical example would be, during a drag’n’drop operation, the underlying surfaces will no longer receive classic pointer event, but rather special events.

This trait is the interface to intercept regular pointer events and change them as needed, its interface mimics the PointerHandle interface.

Any interactions with PointerHandle should be done using PointerInnerHandle, as handle is borrowed/locked before grab methods are called, so calling methods on PointerHandle would result in a deadlock.

If your logic decides that the grab should end, both PointerInnerHandle and PointerHandle have a method to change it.

When your grab ends (either as you requested it or if it was forcefully cancelled by the server), the struct implementing this trait will be dropped. As such you should put clean-up logic in the destructor, rather than trying to guess when the grab will end.

Required Methods

fn motion(
&mut self,
data: &mut D,
handle: &mut PointerInnerHandle<'_, D>,
focus: Option<(<D as SeatHandler>::PointerFocus, Point<i32, Logical>)>,
event: &MotionEvent
)

A motion was reported

This method allows you attach additional behavior to a motion event, possibly altering it. You generally will want to invoke PointerInnerHandle::motion() as part of your processing. If you don’t, the rest of the compositor will behave as if the motion event never occurred.

Some grabs (such as drag’n’drop, shell resize and motion) unset the focus while they are active, this is achieved by just setting the focus to None when invoking PointerInnerHandle::motion().

fn button(
&mut self,
data: &mut D,
handle: &mut PointerInnerHandle<'_, D>,
event: &ButtonEvent
)

A button press was reported

This method allows you attach additional behavior to a button event, possibly altering it. You generally will want to invoke PointerInnerHandle::button() as part of your processing. If you don’t, the rest of the compositor will behave as if the button event never occurred.

fn axis(
&mut self,
data: &mut D,
handle: &mut PointerInnerHandle<'_, D>,
details: AxisFrame
)

An axis scroll was reported

This method allows you attach additional behavior to an axis event, possibly altering it. You generally will want to invoke PointerInnerHandle::axis() as part of your processing. If you don’t, the rest of the compositor will behave as if the axis event never occurred.

fn start_data(&self) -> &GrabStartData<D>

The data about the event that started the grab.