Struct smithay::input::Seat

pub struct Seat<D: SeatHandler> { /* private fields */ }

A Seat handle

This struct gives you access to the control of the capabilities of the associated seat.

This is an handle to the inner logic, it can be cloned.

See module-level documentation for details of use.

Implementations

impl<D: SeatHandler + 'static> Seat<D>

pub fn user_data(&self) -> &UserDataMap

Access the UserDataMap associated with this Seat

pub fn add_pointer(&mut self) -> PointerHandle<D>

Adds the pointer capability to this seat

You are provided a PointerHandle, which allows you to send input events to this pointer. This handle can be cloned.

Calling this method on a seat that already has a pointer capability will overwrite it, and will be seen by the clients as if the mouse was unplugged and a new one was plugged.

Examples

let pointer_handle = seat.add_pointer();

pub fn get_pointer(&self) -> Option<PointerHandle<D>>

Access the pointer of this seat if any

pub fn remove_pointer(&mut self)

Remove the pointer capability from this seat

Clients will be appropriately notified.

pub fn add_keyboard(
&mut self,
xkb_config: XkbConfig<'_>,
repeat_delay: i32,
repeat_rate: i32
) -> Result<KeyboardHandle<D>, KeyboardError>

Adds the keyboard capability to this seat

You are provided a KeyboardHandle, which allows you to send input events to this keyboard. This handle can be cloned.

You also provide a Model/Layout/Variant/Options specification of the keymap to be used for this keyboard, as well as any repeat-info that will be forwarded to the clients.

Calling this method on a seat that already has a keyboard capability will overwrite it, and will be seen by the clients as if the keyboard was unplugged and a new one was plugged.

Examples

let keyboard = seat
    .add_keyboard(
        XkbConfig {
            layout: "de",
            variant: "nodeadkeys",
            ..XkbConfig::default()
        },
        200,
        25,
    )
    .expect("Failed to initialize the keyboard");

pub fn get_keyboard(&self) -> Option<KeyboardHandle<D>>

Access the keyboard of this seat if any

pub fn remove_keyboard(&mut self)

Remove the keyboard capability from this seat

Clients will be appropriately notified.

impl<D: SeatHandler + 'static> Seat<D>

pub fn owns(&self, seat: &WlSeat) -> bool

Checks whether a given WlSeat is associated with this Seat

pub fn from_resource(seat: &WlSeat) -> Option<Self>

Attempt to retrieve a Seat from an existing resource

pub fn global(&self) -> Option<GlobalId>

Get the id of WlSeat global

pub fn add_touch(&mut self) -> TouchHandle

Adds the touch capability to this seat

You are provided a TouchHandle, which allows you to send input events to this pointer. This handle can be cloned.

Calling this method on a seat that already has a touch capability will overwrite it, and will be seen by the clients as if the touchscreen was unplugged and a new one was plugged in.

Examples

let touch_handle = seat.add_touch();

pub fn get_touch(&self) -> Option<TouchHandle>

Access the touch device of this seat, if any.

pub fn remove_touch(&mut self)

Remove the touch capability from this seat

Clients will be appropriately notified.