Module smithay::keyboard

Utilities for keyboard handling

This module provides utilities for keyboardand keymap handling: keymap interpretation and forwarding keystrokes to clients using xkbcommon.

You can first create a KbdHandle using the create_keyboard_handler function in this module. The handle you obtained can be cloned to access this keyboard state from different places. It is expected that such a context is created for each keyboard the compositor has access to.

This handle gives you 3 main ways to interact with the keymap handling:

send the keymap information to a client using the KbdHandle::send_keymap method.

set the current focus for this keyboard: designing the client that will receive the key inputs using the KbdHandle::set_focus method.

process key inputs from the input backend, allowing them to be catched at the compositor-level or forwarded to the client. See the documentation of the KbdHandle::input method for details.

Modules

keysyms

Structs

KbdHandle
An handle to a keyboard handler

ModifiersState
Represents the current state of the keyboard modifiers

Enums

Error
Errors that can be encountered when creating a keyboard handler

Functions

create_keyboard_handler
Create a keyboard handler from a set of RMLVO rules

Type Definitions

Keysym
A number used to represent the symbols generated from a key on a keyboard.