Struct smithay::keyboard::KbdHandle

pub struct KbdHandle { /* fields omitted */ }

An handle to a keyboard handler

It can be cloned and all clones manipulate the same internal state. Clones can also be sent across threads.

See module-level documentation for details of its use.

Methods

impl KbdHandle

fn input<F>(&self, keycode: u32, state: KeyState, serial: u32, filter: F) where
F: FnOnce(&ModifiersState, Keysym) -> bool,

Handle a keystroke

All keystrokes from the input backend should be fed in order to this method of the keyboard handler. It will internally track the state of the keymap.

The filter argument is expected to be a closure which will peek at the generated input as interpreted by the keymap befor it is forwarded to the focused client. If this closure returns false, the input will not be sent to the client. This mechanism can be used to implement compositor-level key bindings for example.

The module smithay::keyboard::keysyms exposes definitions of all possible keysyms to be compared against. This includes non-characted keysyms, such as XF86 special keys.

fn set_focus(&self, focus: Option<(WlSurface, WlKeyboard)>, serial: u32)

Set the current focus of this keyboard

Any previous focus will be sent a wl_keyboard::leave event, and if the new focus is not None, a wl_keyboard::enter event will be sent.

fn send_keymap(&self, kbd: &WlKeyboard)

Send the keymap to this keyboard

This should be done first, before anything else is done with this keyboard.