Macro smithay::custom_elements

macro_rules! custom_elements {
    ($vis:vis $name:ident<=$renderer:ty>; $($tail:tt)*) => { ... };
    ($vis:vis $name:ident<$renderer:ident>; $($tail:tt)*) => { ... };
    ($vis:vis $name:ident; $($tail:tt)*) => { ... };
}

Macro to collate multiple [smithay::desktop::RenderElement]-implementations into one type to be used with Space::render_output.

Example

use smithay::{
    backend::renderer::{Texture, Renderer, ImportAll},
    desktop::space::{SurfaceTree, Space, SpaceOutputTuple, RenderElement},
    utils::{Point, Size, Rectangle, Transform, Logical},
};
use slog::Logger;

smithay::custom_elements! {
    CustomElem; // name of the new type
    SurfaceTree=SurfaceTree, // <variant name> = <type to collate>
};

smithay::custom_elements! {
    CustomElemGeneric<R>; // You can make it generic over renderers!
    SurfaceTree=SurfaceTree,
    PointerElement=PointerElement<<R as Renderer>::TextureId>, // and then use R in your types
};

smithay::custom_elements! {
    CustomElemExplicit<=DummyRenderer>; // You can make it only usable with one renderer
    // that is particulary useful, if your renderer has lifetimes (e.g. MultiRenderer)
    SurfaceTree=SurfaceTree,
    PointerElement=PointerElement<DummyTexture>, // and then you use a concrete and matching texture type
};
// in case your renderer wraps another renderer and implements `std::convert::AsMut`
// you can also use `RenderElement`-implementations requiring the wrapped renderer
// by writing something like: `EguiFrame=EguiFrame as <Gles2Renderer>` where `as <renderer>`
// denotes the type of the wrapped renderer.

pub struct PointerElement<T: Texture> {
   texture: T,
   position: Point<i32, Logical>,
   size: Size<i32, Logical>,
}

impl<T: Texture> PointerElement<T> {
   pub fn new(texture: T, pointer_pos: Point<i32, Logical>) -> PointerElement<T> {
       let size = texture.size().to_logical(1, Transform::Normal);
       PointerElement {
           texture,
           position: pointer_pos,
           size,
       }
   }
}

let elements = [CustomElem::from(surface_tree)];
space.render_output(&mut renderer, &output, age, [0.0, 0.0, 0.0, 1.0], &elements);