Module smithay::seat

Seat global utilities

This module provides you with utilities for handling the seat globals and the associated input wayland objects.

How to use it

Initialization

use smithay::seat::Seat;

// insert the seat:
let (seat_state_token, seat_global) = Seat::new(
    &mut event_loop,
    "seat-0".into(), // the name of the seat, will be advertize to clients
    None /* insert a logger here*/
);

Run usage

Once the seat is initialized, you can add capabilities to it.

Currently, only pointer and keyboard capabilities are supported by smithay.

You can add these capabilities via methods of the Seat struct that was inserted in the event loop, that you can retreive via its token:

let pointer_handle = event_loop.state().get_mut(&seat_state_token).add_pointer();

These handles can be cloned and sent accross thread, so you can keep one around in your event-handling code to forward inputs to your clients.

Structs

KeyboardHandle
An handle to a keyboard handler

PointerHandle
An handle to a keyboard handler

Seat
Internal data of a seat global

Enums

KeyboardError
Errors that can be encountered when creating a keyboard handler