Struct smithay::seat::PointerHandle

pub struct PointerHandle { /* fields omitted */ }

An handle to a keyboard handler

It can be cloned and all clones manipulate the same internal state. Clones can also be sent across threads.

This handle gives you access to an interface to send pointer events to your clients.

Methods

impl PointerHandle

fn motion(
&self,
location: Option<(&WlSurface, f64, f64)>,
serial: u32,
time: u32
)

Notify that the pointer moved

You provide the new location of the pointer, in the form of:

None if the pointer is not on top of a client surface

Some(surface, x, y) if the pointer is focusing surface surface, at location (x, y) relative to this surface

This will internally take care of notifying the appropriate client objects of enter/motion/leave events.

fn button(&self, button: u32, state: ButtonState, serial: u32, time: u32)

Notify that a button was pressed

This will internally send the appropriate button event to the client objects matching with the currently focused surface.