Module smithay::shell

Utilities for handling shell surfaces, toplevel and popups

This module provides automatic handling of shell surfaces objects, by being registered as a global handler for wl_shell and xdg_shell.

Why use this implementation

This implementation can track for you the various shell surfaces defined by the clients by handling the xdg_shell protocol. It also includes a compatibility layer for the deprecated wl_shell global.

It allows you to easily access a list of all shell surfaces defined by your clients access their associated metadata and underlying wl_surfaces.

This handler only handles the protocol exchanges with the client to present you the information in a coherent and relatively easy to use maneer. All the actual drawing and positioning logic of windows is out of its scope.

How to use it

Initialization

To initialize this handler, simple use the shell_init function provided in this module. You will need to provide it the CompositorToken you retrieved from an instanciation of the CompositorHandler provided by smithay.

use smithay::compositor::roles::*;
use smithay::compositor::CompositorToken;
use smithay::shell::{shell_init, ShellSurfaceRole, ShellSurfaceUserImplementation};
use wayland_server::protocol::wl_shell::WlShell;
use wayland_protocols::unstable::xdg_shell::server::zxdg_shell_v6::ZxdgShellV6;
use wayland_server::{EventLoop, EventLoopHandle};

// define the roles type. You need to integrate the ShellSurface role:
define_roles!(MyRoles =>
    [ShellSurface, ShellSurfaceRole]
);

// define the metadata you want associated with the shell clients
#[derive(Default)]
struct MyShellData {
    /* ... */
}

// define your implementation for shell
let my_shell_implementation = ShellSurfaceUserImplementation {
    new_client: |evlh, idata, client| { unimplemented!() },
    client_pong: |evlh, idata, client| { unimplemented!() },
    new_toplevel: |evlh, idata, toplevel| { unimplemented!() },
    new_popup: |evlh, idata, popup| { unimplemented!() },
    move_: |evlh, idata, toplevel, seat, serial| { unimplemented!() },
    resize: |evlh, idata, toplevel, seat, serial, edges| { unimplemented!() },
    grab: |evlh, idata, popup, seat, serial| { unimplemented!() },
    change_display_state: |evlh, idata, toplevel, maximized, minimized, fullscreen, output| {
        unimplemented!()
    },
    show_window_menu: |evlh, idata, toplevel, seat, serial, x, y| { unimplemented!() },
};

// define your implementation data
let my_shell_implementation_data = ();

let (shell_state_token, _, _) = shell_init::<_, _, _, _, MyShellData, _>(
    &mut event_loop,
    compositor_token,             // token from the compositor implementation
    my_shell_implementation,      // instance of shell::ShellSurfaceUserImplementation
    my_shell_implementation_data, // whatever data you need here
    None                          // put a logger if you want
);

// You're now ready to go!

Access to shell surface and clients data

There are mainly 3 kind of objects that you'll manipulate from this implementation:

ShellClient: This is a handle representing an isntanciation of a shell global you can associate client-wise metadata to it (this is the MyShellData type in the example above).

ToplevelSurface: This is a handle representing a toplevel surface, you can retrive a list of all currently alive toplevel surface from the ShellState.

PopupSurface: This is a handle representing a popup/tooltip surface. Similarly, you can get a list of all currently alive popup surface from the ShellState.

You'll obtain these objects though two means: either via the callback methods of the subhandler you provided, or via methods on the ShellState that you can access from the state() of the event loop and the token returned by the init function.

Structs

PopupConfigure
A configure message for popup surface

PopupState
The pending state of a popup surface

PopupSurface
A handle to a popup surface

PositionerState
The state of a positioner, as set by the client

ShellClient
A shell client

ShellState
Shell global state

ShellSurfaceIData
Internal implementation data of shell surfaces

ShellSurfaceRole
Metadata associated with the shell_surface role

ShellSurfaceUserImplementation
A sub-implementation for the shell

ToplevelConfigure
A configure message for toplevel surfaces

ToplevelState
State of a regular toplevel surface

ToplevelSurface
A handle to a toplevel surface

Enums

ShellSurfacePendingState
Contents of the pending state of a shell surface, depending on its role

Functions

shell_init
Create new xdg_shell and wl_shell globals.