Function smithay::shell::shell_init

pub fn shell_init<U, R, CID, SID, SD, L>(
    evl: &mut EventLoop, 
    token: CompositorToken<U, R, CID>, 
    implementation: ShellSurfaceUserImplementation<U, R, CID, SID, SD>, 
    idata: SID, 
    logger: L
) -> (StateToken<ShellState<U, R, CID, SD>>, Global<WlShell, ShellSurfaceIData<U, R, CID, SID, SD>>, Global<ZxdgShellV6, ShellSurfaceIData<U, R, CID, SID, SD>>) where
    U: 'static,
    R: Role<ShellSurfaceRole> + 'static,
    CID: 'static,
    SID: 'static,
    SD: Default + 'static,
    L: Into<Option<Logger>>, 

Create new xdg_shell and wl_shell globals.

The globals are directly registered into the eventloop, and this function returns a StateToken<_> which you'll need access the list of shell surfaces created by your clients.

It also returns the two global handles, in case you whish to remove these globals from the event loop in the future.