Struct smithay::shell::PopupState

pub struct PopupState {
    pub parent: WlSurface,
    pub positioner: PositionerState,
}

The pending state of a popup surface

Fields
parent: WlSurface

Parent of this popup surface

positioner: PositionerState

The positioner specifying how this tooltip should be placed relative to its parent.

Methods

impl PopupState

fn clone(&self) -> Option<PopupState>

Clone this PopupState

If the parent surface refers to a surface that no longer exists, this will return None, as the popup can no longer be meaningfully displayed.