Struct smithay::shell::PopupSurface

pub struct PopupSurface<U, R, CID, SD> { /* fields omitted */ }

A handle to a popup surface

This is an unified abstraction over the popup surfaces of both wl_shell and xdg_shell.

Methods

impl<U, R, CID, SD> PopupSurface<U, R, CID, SD> where
U: 'static,
R: Role<ShellSurfaceRole> + 'static,
CID: 'static,
SD: 'static,

fn alive(&self) -> bool

Is the popup surface refered by this handle still alive?

fn equals(&self, other: &Self) -> bool

Do this handle and the other one actually refer to the same popup surface?

fn client(&self) -> Option<ShellClient<SD>>

Retrieve the shell client owning this popup surface

Returns None if the surface does actually no longer exist.

fn send_configure(&self, cfg: PopupConfigure) -> EventResult<()>

Send a configure event to this toplevel surface to suggest it a new configuration

The serial of this configure will be tracked waiting for the client to ACK it.

fn ensure_configured(&self) -> bool

Make sure this surface was configured

Returns true if it was, if not, returns false and raise a protocol error to the associated client. Also returns false if the surface is already destroyed.

xdg_shell mandates that a client acks a configure before commiting anything.

fn send_popup_done(&self) -> EventResult<()>

Send a 'popup_done' event to the popup surface

It means that the use has dismissed the popup surface, or that the pointer has left the area of popup grab if there was a grab.

fn get_surface(&self) -> Option<&WlSurface>

Access the underlying wl_surface of this toplevel surface

Returns None if the toplevel surface actually no longer exists.

fn get_pending_state(&self) -> Option<PopupState>

Retrieve a copy of the pending state of this popup surface

Returns None of the popup surface actually no longer exists.