Struct smithay::shell::ToplevelSurface

pub struct ToplevelSurface<U, R, CID, SD> { /* fields omitted */ }

A handle to a toplevel surface

This is an unified abstraction over the toplevel surfaces of both wl_shell and xdg_shell.

Methods

impl<U, R, CID, SD> ToplevelSurface<U, R, CID, SD> where
U: 'static,
R: Role<ShellSurfaceRole> + 'static,
CID: 'static,
SD: 'static,

fn alive(&self) -> bool

Is the toplevel surface refered by this handle still alive?

fn equals(&self, other: &Self) -> bool

Do this handle and the other one actually refer to the same toplevel surface?

fn client(&self) -> Option<ShellClient<SD>>

Retrieve the shell client owning this toplevel surface

Returns None if the surface does actually no longer exist.

fn send_configure(&self, cfg: ToplevelConfigure) -> EventResult<()>

Send a configure event to this toplevel surface to suggest it a new configuration

The serial of this configure will be tracked waiting for the client to ACK it.

fn ensure_configured(&self) -> bool

Make sure this surface was configured

Returns true if it was, if not, returns false and raise a protocol error to the associated client. Also returns false if the surface is already destroyed.

xdg_shell mandates that a client acks a configure before commiting anything.

fn get_surface(&self) -> Option<&WlSurface>

Access the underlying wl_surface of this toplevel surface

Returns None if the toplevel surface actually no longer exists.

fn get_pending_state(&self) -> Option<ToplevelState>

Retrieve a copy of the pending state of this toplevel surface

Returns None of the toplevel surface actually no longer exists.