Function smithay::shm::with_buffer_contents

pub fn with_buffer_contents<F>(
    buffer: &WlBuffer, 
    f: F
) -> Result<(), BufferAccessError> where
    F: FnOnce(&[u8], BufferData), 

Call given closure with the contents of the given buffer

If the buffer is managed by the provided ShmGlobal, its contents are extracted and the closure is extracted with them:

The first argument is a data slice of the contents of the pool

The second argument is the specification of this buffer is this pool

If the buffer is not managed by the provided ShmGlobal, the closure is not called and this method will return Err(()) (this will be the case for an EGL buffer for example).