Struct smithay::shm::ShmGlobal

pub struct ShmGlobal { /* fields omitted */ }

A global for handling SHM pool and buffers

You must register it to an event loop using register_with_init, or it will quickly panic.

Methods

impl ShmGlobal

fn new<L>(formats: Vec<Format>, logger: L) -> ShmGlobal where L: Into<Option<Logger>>

Create a new SHM global advertizing given supported formats.

This global will always advertize ARGB8888 and XRGB8888 format as they are required by the protocol. Formats given as argument as additionnaly advertized.

fn get_token(&self) -> ShmGlobalToken

Retreive a token from the SHM global.

This can only be called once the ShmGlobal has been added to and event loop and has been initialized. If it is not the case, this method will panic.

This is needed to retrieve the contents of the shm pools and buffers.