Trait smithay::signaling::Linkable

pub trait Linkable<S> {
    fn link(&mut self, signaler: Signaler<S>);
}

Trait representing the capability of an object to listen for some signals

It is provided so that the signaling system can play nicely into generic constructs.

Required methods

fn link(&mut self, signaler: Signaler<S>)

Make this object listen for signals from given signaler

Loading content...