Module smithay::utils

Various utilities functions and types

Modules

signaling

A general purpose signaling mechanism

user_data

Various utilities used for user data implementations

x11rbx11rb_event_source

Helper utilities for using x11rb as an event source in calloop.

Structs

Buffer

Type-level marker for the buffer coordinate space

Clock

Defines a clock with a specific kind

DeadResource

This resource has been destroyed and can no longer be used.

DeviceFd

Ref-counted file descriptor of an open device node

Logical

Type-level marker for the logical coordinate space

Monotonic

Monotonic clock

Physical

Type-level marker for the physical coordinate space

Point

A point as defined by its x and y coordinates

Raw

Type-level marker for raw coordinate space, provided by input devices

Realtime

Realtime clock

Rectangle

A rectangle defined by its top-left corner and dimensions

Scale

A two-dimensional scale that can be used to scale Points, Sizes and Rectangles

Serial

A serial type, whose comparison takes into account the wrapping-around behavior of the underlying counter.

SerialCounter

A counter for generating serials, for use in the client protocol

Size

A size as defined by its width and height

Time

A point in time for a clock with a specific kind

UnmanagedResource

This resource is not managed by Smithay

Enums

Transform

Possible transformations to two-dimensional planes

Statics

SERIAL_COUNTER

A global SerialCounter for use in your compositor.

Traits

ClockSource

Id for a clock according to unix clockid_t

Coordinate

Trait for types serving as a coordinate for other geometry utils

DevPath

Trait representing open devices that may return a Path

IsAlive

Trait that is implemented on wayland objects tracked by Smithay

NonNegativeClockSource

Marker for clock source that never returns a negative Time