smithay::utils::signaling

Trait Linkable

pub trait Linkable<S> {
    // Required method
    fn link(&mut self, signaler: Signaler<S>);
}

Trait representing the capability of an object to listen for some signals

It is provided so that the signaling system can play nicely into generic constructs.

Required Methods

fn link(&mut self, signaler: Signaler<S>)

Make this object listen for signals from given signaler