Struct smithay::utils::Point

#[repr(C)]
pub struct Point<N, Kind> {
    pub x: N,
    pub y: N,
    /* private fields */
}

A point as defined by its x and y coordinates

Operations on points are saturating.

Fields
x: N

horizontal coordinate

y: N

vertical coordinate

Implementations

impl<N: Coordinate, Kind> Point<N, Kind>

pub fn to_size(self) -> Size<N, Kind>

Convert this Point to a Size with the same coordinates

Checks that the coordinates are positive with a debug_assert!().

pub fn to_size_abs(self) -> Size<N, Kind>

Convert this Point to a Size with the same coordinates

Ensures that the coordinates are positive by taking their absolute value

pub fn upscale(self, scale: impl Into<Scale<N>>) -> Point<N, Kind>

Upscale this Point by a specified Scale

pub fn downscale(self, scale: impl Into<Scale<N>>) -> Point<N, Kind>

Downscale this Point by a specified Scale

impl<N: Coordinate, Kind> Point<N, Kind>

pub fn constrain(self, rect: impl Into<Rectangle<N, Kind>>) -> Point<N, Kind>

Constrain this Point within a Rectangle with the same coordinates

The Point returned is guaranteed to be not smaller than the Rectangle location and not greater than the Rectangle location plus size.

impl<N: Coordinate, Kind> Point<N, Kind>

pub fn to_f64(self) -> Point<f64, Kind>

Convert the underlying numerical type to f64 for floating point manipulations

impl<Kind> Point<f64, Kind>

pub fn to_i32_round<N: Coordinate>(self) -> Point<N, Kind>

Convert to i32 for integer-space manipulations by rounding float values

pub fn to_i32_floor<N: Coordinate>(self) -> Point<N, Kind>

Convert to i32 for integer-space manipulations by flooring float values

pub fn to_i32_ceil<N: Coordinate>(self) -> Point<N, Kind>

Convert to i32 for integer-space manipulations by ceiling float values

impl<N: Coordinate> Point<N, Logical>

pub fn to_physical(self, scale: impl Into<Scale<N>>) -> Point<N, Physical>

Convert this logical point to physical coordinate space according to given scale factor

pub fn to_physical_precise_round<S: Coordinate, R: Coordinate>( self, scale: impl Into<Scale<S>>, ) -> Point<R, Physical>

Convert this logical point to physical coordinate space according to given scale factor and round the result

pub fn to_physical_precise_ceil<S: Coordinate, R: Coordinate>( &self, scale: impl Into<Scale<S>>, ) -> Point<R, Physical>

Convert this logical point to physical coordinate space according to given scale factor and ceil the result

pub fn to_physical_precise_floor<S: Coordinate, R: Coordinate>( &self, scale: impl Into<Scale<S>>, ) -> Point<R, Physical>

Convert this logical point to physical coordinate space according to given scale factor and floor the result

pub fn to_buffer( self, scale: impl Into<Scale<N>>, transformation: Transform, area: &Size<N, Logical>, ) -> Point<N, Buffer>

Convert this logical point to buffer coordinate space according to given scale factor

impl<N: Coordinate> Point<N, Physical>

pub fn to_logical(self, scale: impl Into<Scale<N>>) -> Point<N, Logical>

Convert this physical point to logical coordinate space according to given scale factor

impl<N: Coordinate> Point<N, Buffer>

pub fn to_logical( self, scale: impl Into<Scale<N>>, transform: Transform, area: &Size<N, Buffer>, ) -> Point<N, Logical>

Convert this physical point to logical coordinate space according to given scale factor