smithay::utils

Struct SealedFile

pub struct SealedFile { /* private fields */ }

A file whose fd cannot be written by other processes

This mechanism is useful for giving clients access to large amounts of information such as keymaps without them being able to write to the handle.

On Linux, Android, and FreeBSD, this uses a sealed memfd. On other platforms it creates a POSIX shared memory object with shm_open, opens a read-only copy, and unlinks it.

Implementations

impl SealedFile

pub fn with_content(name: &CStr, contents: &CStr) -> Result<Self, Error>

Create a [SealedFile] with the given nul-terminated C string.

pub fn with_data(name: &CStr, data: &[u8]) -> Result<Self, Error>

Available on Linux or FreeBSD or Android only.

Create a [SealedFile] with the given binary data.

pub fn size(&self) -> usize

Size of the data contained in the sealed file.