smithay::utils

Struct Size

pub struct Size<N, Kind> {
    pub w: N,
    pub h: N,
    /* private fields */
}

A size as defined by its width and height

Constructors of this type ensure that the values are always positive via debug_assert!(), however manually changing the values of the fields can break this invariant.

Operations on sizes are saturating.

Fields
w: N

horizontal coordinate

h: N

vertical coordinate

Implementations

impl<N: Coordinate, Kind> Size<N, Kind>

pub fn new(w: N, h: N) -> Size<N, Kind>

Create a new Size

impl<N: Coordinate, Kind> Size<N, Kind>

pub fn to_point(self) -> Point<N, Kind>

Convert this Size to a Point with the same coordinates

impl<N: Coordinate, Kind> Size<N, Kind>

pub fn clamp( self, min: impl Into<Size<N, Kind>>, max: impl Into<Size<N, Kind>>, ) -> Size<N, Kind>

Restrict this Size to min and max Size with the same coordinates

impl<N: Coordinate, Kind> Size<N, Kind>

pub fn to_f64(self) -> Size<f64, Kind>

Convert the underlying numerical type to f64 for floating point manipulations

impl<N: Coordinate, Kind> Size<N, Kind>

pub fn upscale(self, scale: impl Into<Scale<N>>) -> Size<N, Kind>

Upscale this Size by a specified Scale

pub fn downscale(self, scale: impl Into<Scale<N>>) -> Size<N, Kind>

Downscale this Size by a specified Scale

pub fn is_empty(&self) -> bool

Check if this Size is empty

Returns true if either the width or the height is zero

impl<Kind> Size<f64, Kind>

pub fn to_i32_round<N: Coordinate>(self) -> Size<N, Kind>

Convert to i32 for integer-space manipulations by rounding float values

pub fn to_i32_floor<N: Coordinate>(self) -> Size<N, Kind>

Convert to i32 for integer-space manipulations by flooring float values

pub fn to_i32_ceil<N: Coordinate>(self) -> Size<N, Kind>

Convert to i32 for integer-space manipulations by ceiling float values

impl<N: Coordinate> Size<N, Logical>

pub fn to_physical(self, scale: impl Into<Scale<N>>) -> Size<N, Physical>

Convert this logical size to physical coordinate space according to given scale factor

pub fn to_physical_precise_round<S: Coordinate, R: Coordinate>( self, scale: impl Into<Scale<S>>, ) -> Size<R, Physical>

Convert this logical size to physical coordinate space according to given scale factor and round the result

pub fn to_physical_precise_ceil<S: Coordinate, R: Coordinate>( &self, scale: impl Into<Scale<S>>, ) -> Size<R, Physical>

Convert this logical size to physical coordinate space according to given scale factor and ceil the result

pub fn to_physical_precise_floor<S: Coordinate, R: Coordinate>( &self, scale: impl Into<Scale<S>>, ) -> Size<R, Physical>

Convert this logical size to physical coordinate space according to given scale factor and floor the result

pub fn to_buffer( self, scale: impl Into<Scale<N>>, transformation: Transform, ) -> Size<N, Buffer>

Convert this logical size to buffer coordinate space according to given scale factor

impl<N: Coordinate> Size<N, Physical>

pub fn to_logical(self, scale: impl Into<Scale<N>>) -> Size<N, Logical>

Convert this physical point to logical coordinate space according to given scale factor

impl<N: Coordinate> Size<N, Buffer>

pub fn to_logical( self, scale: impl Into<Scale<N>>, transformation: Transform, ) -> Size<N, Logical>

Convert this physical point to logical coordinate space according to given scale factor