smithay::utils

Trait DevPath

pub trait DevPath {
    // Required method
    fn dev_path(&self) -> Option<PathBuf>;
}

Trait representing open devices that may return a Path

Required Methods

fn dev_path(&self) -> Option<PathBuf>

Returns the path of the open device if possible