smithay::wayland::background_effect

Trait ExtBackgroundEffectHandler

pub trait ExtBackgroundEffectHandler:
    GlobalDispatch<ExtBackgroundEffectManagerV1, ()>
    + Dispatch<ExtBackgroundEffectManagerV1, ()>
    + Dispatch<ExtBackgroundEffectSurfaceV1, BackgroundEffectSurfaceUserData>
    + 'static {
    // Provided methods
    fn capabilities(&self) -> Capability { ... }
    fn set_blur_region(
        &mut self,
        wl_surface: WlSurface,
        region: RegionAttributes,
    ) { ... }
    fn unset_blur_region(&mut self, wl_surface: WlSurface) { ... }
}

Available on crate feature wayland_frontend only.

Handler trait for ext background effect events.

Provided Methods

fn capabilities(&self) -> Capability

Getter for background-effect capabilities of the compositor

fn set_blur_region(&mut self, wl_surface: WlSurface, region: RegionAttributes)

Called when blur becomes pending, and awaits surface commit. Blur region is stored in wl_surface BackgroundEffectSurfaceCachedState

fn unset_blur_region(&mut self, wl_surface: WlSurface)

Called when blur unset becomes pending, and awaits surface commit.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.