smithay::wayland

Module commit_timing

Available on crate feature wayland_frontend only.

Utilities for handling the wp-commit-timing protocol

How to use it

Initialization

To initialize this implementation create the CommitTimingManagerState and store it inside your State struct.

use smithay::delegate_commit_timing;
use smithay::wayland::compositor;
use smithay::wayland::commit_timing::CommitTimingManagerState;

// Create the compositor state
let commit_timing_manager_state = CommitTimingManagerState::new::<State>(
    &display.handle(),
);

// insert the CommitTimingManagerState into your state
// ..

delegate_commit_timing!(State);

// You're now ready to go!

Use the commit timer state

Whenever the client commits a surface content update containing a commit timer timestamp set through wp_commit_timer_v1::Request::SetTimestamp the implementation will place a Blocker on the surface and register it in the CommitTimerBarrierState. It is your responsibility to query for pending commit timers and signal them to allow client to make forward progress.

You can query the pending commit timers and signal them as shown in the following example:

use smithay::wayland::compositor;
use smithay::wayland::commit_timing::CommitTimerBarrierStateUserData;
compositor::with_states(&surface, |states| {
    if let Some(mut commit_timer_state) = states
        .data_map
        .get::<CommitTimerBarrierStateUserData>()
        .map(|commit_timer| commit_timer.lock().unwrap())
        {
            if commit_timer_state.signal_until(frame_target) {
                // ..signal blocker cleared
            }
        }
});

Unmanaged mode

If for some reason the integrated solution for commit timers does not suit your needs you can create an unmanaged version with CommitTimingManagerState::unmanaged. In this case it is your responsibility to listen for surface commits and place blockers accordingly to the spec.

The commit timer timestamp can be retrieved like shown in the following example:

use smithay::wayland::compositor;
use smithay::wayland::commit_timing::CommitTimerStateUserData;

let timestamp = compositor::with_states(&surface, |states| {
    states
        .data_map
        .get::<CommitTimerStateUserData>()
        .and_then(|state| state.borrow_mut().timestamp.take())
});

Structs

CommitTimerBarrierState

Per surface barrier state using the managed mode

CommitTimerState

Per surface state for WpCommitTimerV1

CommitTimingManagerState

State for the WpCommitTimingManagerV1 global

Timestamp

Timestamp set through wp_commit_timer_v1::Request::SetTimestamp for a surface

Type Aliases

CommitTimerBarrierStateUserData

Per surface barrier state stored in the surface user data

CommitTimerStateUserData

Per surface WpCommitTimerV1 state stored in the surface user data