Function smithay::wayland::compositor::compositor_init

pub fn compositor_init<U, R, Impl, L>(
    display: &mut Display, 
    token: LoopToken, 
    implem: Impl, 
    logger: L
) -> (CompositorToken<U, R>, Global<WlCompositor>, Global<WlSubcompositor>) where
    L: Into<Option<Logger>>,
    U: Default + 'static,
    R: Default + RoleType + Role<SubsurfaceRole> + 'static,
    Impl: Implementation<(Resource<WlSurface>, CompositorToken<U, R>), SurfaceEvent> + 'static, 

Create new wl_compositor and wl_subcompositor globals.

The globals are directly registered into the eventloop, and this function returns a CompositorToken which you'll need access the data associated to the wl_surfaces.

It also returns the two global handles, in case you whish to remove these globals from the event loop in the future.