Module smithay::wayland::compositor::roles

Tools for handling surface roles

In the Wayland protocol, surfaces can have several different roles, which define how they are to be used. The core protocol defines 3 of these roles:

shell_surface: This surface is to be considered as what is most often called a "window".

pointer_surface: This surface represent the contents of a pointer icon and replaces the default pointer.

subsurface: This surface is part of a subsurface tree, and as such has a parent surface.

A surface can have only one role at any given time. To change he role of a surface, the client must first remove the previous role before assigning the new one. A surface without a role is not displayed at all.

This module provides tools to manage roles of a surface in a composable way allowing all handlers of smithay to manage surface roles while being aware of the possible role conflicts.

General mechanism

First, all roles need to have an unique type, holding its metadata and identifying it to the type-system. Even if your role does not hold any metadata, you still need its unique type, using a unit-like struct rather than ().

You then need a type for managing the roles of a surface. This type holds information about what is the current role of a surface, and what is the metadata associated with it.

For convenience, you can use the define_roles! macro provided by Smithay to define this type. You can call it like this:

// Metadata for a first role
#[derive(Default)]
pub struct MyRoleMetadata {
}

// Metadata for a second role
#[derive(Default)]
pub struct MyRoleMetadata2 {
}

define_roles!(Roles =>
    // You can put several roles like this
    // first identifier is the name of the variant for this
    // role in the generated enum, second is the token type
    // for this role
    [MyRoleName, MyRoleMetadata]
    [MyRoleName2, MyRoleMetadata2]
    /* ... */
);

And this will expand to an enum like this:

pub enum Roles {
    NoRole,
    // The subsurface role is always inserted, as it is required
    // by the CompositorHandler
    Subsurface(::smithay::compositor::SubsurfaceAttributes),
    // all your other roles come here
    MyRoleName(MyRoleMetadata),
    MyRoleName2(MyRoleMetadata2),
    /* ... */
}

as well as implement a few trait for it, allowing it to be used by all smithay handlers:

The trait RoleType, which defines it as a type handling roles

For each of your roles, the trait Role<Token> (where Token is your token type), marking its ability to handle this given role.

All handlers that handle a specific role will require you to provide them with a CompositorToken<U, R, H> where R: Role<TheToken>.

See the documentation of these traits for their specific definition and capabilities.

Structs

WrongRole
An error type signifying that the surface does not have expected role

Traits

Role
A trait representing the capability of a [RoleType] to handle a given role

RoleType
A trait representing a type that can manage surface roles