smithay::wayland::compositor

Struct CompositorState

pub struct CompositorState { /* private fields */ }

Available on crate feature wayland_frontend only.

State of a compositor

Implementations

impl CompositorState

pub fn new<D>(display: &DisplayHandle) -> Self
where D: GlobalDispatch<WlCompositor, ()> + GlobalDispatch<WlSubcompositor, ()> + 'static,

Create new wl_compositor version 5 and wl_subcompositor globals.

It returns the two global handles, in case you wish to remove these globals from the event loop in the future.

pub fn new_v6<D>(display: &DisplayHandle) -> Self
where D: GlobalDispatch<WlCompositor, ()> + GlobalDispatch<WlSubcompositor, ()> + 'static,

The same as new, but binds at least version 6 of wl_compositor.

This means that for clients to scale and apply transformation with non-default values send_surface_state must be used.

pub fn compositor_global(&self) -> GlobalId

Get id of compositor global

pub fn subcompositor_global(&self) -> GlobalId

Get id of subcompositor global