Struct smithay::wayland::compositor::CompositorToken

pub struct CompositorToken<U, R> { /* fields omitted */ }

A Compositor global token

This token can be cloned at will, and is the entry-point to access data associated with the wl_surface and wl_region managed by the CompositorGlobal that provided it.

Methods

impl<U: 'static, R: 'static> CompositorToken<U, R>

pub fn with_surface_data<F, T>(&self, surface: &Resource<WlSurface>, f: F) -> T where
F: FnOnce(&mut SurfaceAttributes<U>) -> T,

Access the data of a surface

The closure will be called with the contents of the data associated with this surface.

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

impl<U, R> CompositorToken<U, R> where
U: 'static,
R: RoleType + Role<SubsurfaceRole> + 'static,

pub fn with_surface_tree_upward<F, T>(
&self,
surface: &Resource<WlSurface>,
initial: T,
f: F
) -> Result<(), ()> where
F: FnMut(&Resource<WlSurface>, &mut SurfaceAttributes<U>, &mut R, &T) -> TraversalAction<T>,

Access the data of a surface tree from bottom to top

The provided closure is called successively on the surface and all its child subsurfaces, in a depth-first order. This matches the order in which the surfaces are supposed to be drawn: top-most last.

The arguments provided to the closure are, in this order:

The surface object itself

a mutable reference to its surface attribute data

a mutable reference to its role data,

a custom value that is passed in a fold-like maneer, but only from the output of a parent to its children. See TraversalAction for details.

If the surface not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn with_surface_tree_downward<F, T>(
&self,
surface: &Resource<WlSurface>,
initial: T,
f: F
) -> Result<(), ()> where
F: FnMut(&Resource<WlSurface>, &mut SurfaceAttributes<U>, &mut R, &T) -> TraversalAction<T>,

Access the data of a surface tree from top to bottom

The provided closure is called successively on the surface and all its child subsurfaces, in a depth-first order. This matches the reverse of the order in which the surfaces are supposed to be drawn: top-most first.

Behavior is the same as with_surface_tree_upward.

pub fn get_parent(
&self,
surface: &Resource<WlSurface>
) -> Option<Resource<WlSurface>>

Retrieve the parent of this surface

Returns None is this surface is a root surface

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_children(
&self,
surface: &Resource<WlSurface>
) -> Vec<Resource<WlSurface>>

Retrieve the children of this surface

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

impl<U: 'static, R: RoleType + 'static> CompositorToken<U, R>

pub fn has_a_role(&self, surface: &Resource<WlSurface>) -> bool

Check wether this surface as a role or not

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn has_role<RoleData>(&self, surface: &Resource<WlSurface>) -> bool where
R: Role<RoleData>,

Check wether this surface as a specific role

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn give_role<RoleData>(
&self,
surface: &Resource<WlSurface>
) -> Result<(), ()> where
R: Role<RoleData>,
RoleData: Default,

Register that this surface has given role with default data

Fails if the surface already has a role.

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn give_role_with<RoleData>(
&self,
surface: &Resource<WlSurface>,
data: RoleData
) -> Result<(), RoleData> where
R: Role<RoleData>,

Register that this surface has given role with given data

Fails if the surface already has a role and returns the data.

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn with_role_data<RoleData, F, T>(
&self,
surface: &Resource<WlSurface>,
f: F
) -> Result<T, WrongRole> where
R: Role<RoleData>,
F: FnOnce(&mut RoleData) -> T,

Access the role data of a surface

Fails and don't call the closure if the surface doesn't have this role

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn remove_role<RoleData>(
&self,
surface: &Resource<WlSurface>
) -> Result<RoleData, WrongRole> where
R: Role<RoleData>,

Register that this surface does not have a role any longer and retrieve the data

Fails if the surface didn't already have this role.

If the surface is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).

pub fn get_region_attributes(
&self,
region: &Resource<WlRegion>
) -> RegionAttributes

Retrieve the metadata associated with a wl_region

If the region is not managed by the CompositorGlobal that provided this token, this will panic (having more than one compositor is not supported).