smithay::wayland

Module cursor_shape

Available on crate feature wayland_frontend only.

Utilities for compositor side cursor theming support

This protocol allows clients to request compositor to draw a cursor for them, thus resulting in more consistent look and feel of the cursor across the applications.

Initialization

extern crate smithay;
extern crate wayland_server;

use smithay::wayland::cursor_shape::CursorShapeManagerState;
use smithay::delegate_cursor_shape;

let state = CursorShapeManagerState::new::<State>(&display.handle());

delegate_cursor_shape!(State);

Structs

CursorShapeManagerState

State of the cursor shape manager.