Module smithay::wayland::data_device

Utilities for manipulating the data devices

The data device is wayland’s abstraction to represent both selection (copy/paste) and drag’n’drop actions. This module provides logic to handle this part of the protocol. Selection and drag’n’drop are per-seat notions.

This module provides the freestanding set_data_device_focus function: This function sets the data device focus for a given seat; you’d typically call it whenever the keyboard focus changes, to follow it (for example in the focus hook of your keyboards).

Using these two functions is enough for your clients to be able to interact with each other using the data devices.

The module also provides additional mechanisms allowing your compositor to see and interact with the contents of the data device:

the freestanding function set_data_device_selection allows you to set the contents of the selection for your clients

the freestanding function start_dnd allows you to initiate a drag’n’drop event from the compositor itself and receive interactions of clients with it via an other dedicated callback.

The module defines the role "dnd_icon" that is assigned to surfaces used as drag’n’drop icons.

Initialization

To initialize this implementation, create the DataDeviceState, store it inside your State struct and implement the DataDeviceHandler, as shown in this example:

use smithay::delegate_data_device;
use smithay::wayland::data_device::{ClientDndGrabHandler, DataDeviceState, DataDeviceHandler, ServerDndGrabHandler};

// Create the data_device state
let data_device_state = DataDeviceState::new::<State, _>(
    &display.handle(),
    None // We don't add a logger in this example
);

// insert the DataDeviceState into your state
// ..

// implement the necessary traits
impl ClientDndGrabHandler for State {}
impl ServerDndGrabHandler for State {}
impl DataDeviceHandler for State {
    fn data_device_state(&self) -> &DataDeviceState { &self.data_device_state }
    // ... override default implementations here to customize handling ...
}
delegate_data_device!(State);

// You're now ready to go!

Structs

DataDeviceState

State of data device

SourceMetadata

The metadata describing a data source

Constants

DND_ICON_ROLE

WlSurface role of drag and drop icon

Traits

ClientDndGrabHandler

Events that are generated during client initiated drag’n’drop

DataDeviceHandler

Events that are generated by interactions of the clients with the data device

ServerDndGrabHandler

Event generated by the interactions of clients with a server initiated drag’n’drop

Functions

default_action_chooser

A simple action chooser for DnD negociation

set_data_device_focus

Set the data device focus to a certain client for a given seat

set_data_device_selection

Set a compositor-provided selection for this seat

start_dnd

Start a drag’n’drop from a resource controlled by the compositor

with_source_metadata

Access the metadata of a data source