Enum smithay::wayland::data_device::DataDeviceEvent

pub enum DataDeviceEvent {
    NewSelection(Option<WlDataSource>),
    DnDStarted {
        source: Option<WlDataSource>,
        icon: Option<WlSurface>,
    },
    DnDDropped,
    SendSelection {
        mime_type: String,
        fd: RawFd,
    },
}

Events that are generated by interactions of the clients with the data device

Variants

NewSelection(Option<WlDataSource>)

A client has set the selection

DnDStarted

A client started a drag'n'drop as response to a user pointer action

Fields of DnDStarted

source: Option<WlDataSource>

The data source provided by the client

If it is None, this means the DnD is restricted to surfaces of the same client and the client will manage data transfert by itself.

icon: Option<WlSurface>

The icon the client requested to be used to be associated with the cursor icon during the drag'n'drop.

DnDDropped

The drag'n'drop action was finished by the user releasing the buttons

At this point, any pointer icon should be removed.

Note that this event will only be genrated for client-initiated drag'n'drop session.

SendSelection

A client requested to read the server-set selection

Fields of SendSelection

mime_type: String

the requested mime type

fd: RawFd

the fd to write into