Function smithay::wayland::data_device::start_dnd

pub fn start_dnd<D, C>(
    dh: &DisplayHandle,
    seat: &Seat<D>,
    data: &mut D,
    serial: Serial,
    start_data: PointerGrabStartData<D>,
    metadata: SourceMetadata
) where
    D: SeatHandler + DataDeviceHandler + 'static,
    <D as SeatHandler>::PointerFocus: WaylandFocus, 

Start a drag’n’drop from a resource controlled by the compositor

You’ll receive events generated by the interaction of clients with your drag’n’drop in the provided callback. See ServerDndGrabHandler for details about which events can be generated and what response is expected from you to them.