Trait smithay::wayland::data_device::ClientDndGrabHandler

pub trait ClientDndGrabHandler: SeatHandler + Sized {
    fn started(
        &mut self,
        source: Option<WlDataSource>,
        icon: Option<WlSurface>,
        seat: Seat<Self>
    ) { ... }
    fn dropped(&mut self, seat: Seat<Self>) { ... }
}

Events that are generated during client initiated drag’n’drop

Provided Methods

fn started(
&mut self,
source: Option<WlDataSource>,
icon: Option<WlSurface>,
seat: Seat<Self>
)

A client started a drag’n’drop as response to a user pointer action

source - The data source provided by the client. If it is None, this means the DnD is restricted to surfaces of the same client and the client will manage data transfer by itself.

icon - The icon the client requested to be used to be associated with the cursor icon during the drag’n’drop.

seat - The seat on which the DnD operation was started

fn dropped(&mut self, seat: Seat<Self>)

The drag’n’drop action was finished by the user releasing the buttons

At this point, any pointer icon should be removed.

Note that this event will only be generated for client-initiated drag’n’drop session.

seat - The seat on which the DnD action was finished.