smithay::wayland::dmabuf

Function get_dmabuf

pub fn get_dmabuf(buffer: &WlBuffer) -> Result<&Dmabuf, UnmanagedResource>

Available on crate feature wayland_frontend only.

Gets the contents of a Dmabuf backed WlBuffer.

If the buffer is managed by the dmabuf handler, the Dmabuf is returned.

If the buffer is not managed by the dmabuf handler (whether the buffer is a different kind of buffer, such as an shm buffer or is not managed by smithay), this function will return an UnmanagedResource error.