Function smithay::wayland::dmabuf::init_dmabuf_global

pub fn init_dmabuf_global<F, L>(
    display: &mut Display, 
    formats: Vec<Format>, 
    handler: F, 
    logger: L
) -> Global<ZwpLinuxDmabufV1> where
    L: Into<Option<Logger>>,
    F: for<'a> FnMut(&Dmabuf, DispatchData<'a>) -> bool + 'static, 

Initialize a dmabuf global.

You need to provide a vector of the supported formats, as well as a closure, that will validate the parameters provided by the client and tests the import as a dmabuf.