Struct smithay::wayland::dmabuf::DmabufState

pub struct DmabufState { /* private fields */ }

Delegate type for all dmabuf globals.

Dmabuf globals are created using this type and events will be forwarded to an instance of the dmabuf global.

Implementations

impl DmabufState

pub fn new() -> DmabufState

Creates a new DmabufState delegate type.

pub fn create_global<D>( &mut self, display: &DisplayHandle, formats: Vec<Format> ) -> DmabufGlobalwhere D: GlobalDispatch<ZwpLinuxDmabufV1, DmabufGlobalData> + BufferHandler + DmabufHandler + 'static,

Creates a dmabuf global with the specified supported formats.

Note: This function will create a version 3 dmabuf global and thus not call DmabufHandler::new_surface_feedback, if you want to create a version 4 global you need to call DmabufState::create_global_with_default_feedback.

pub fn create_global_with_filter<D, F>( &mut self, display: &DisplayHandle, formats: Vec<Format>, filter: F ) -> DmabufGlobalwhere D: GlobalDispatch<ZwpLinuxDmabufV1, DmabufGlobalData> + BufferHandler + DmabufHandler + 'static, F: for<'c> Fn(&'c Client) -> bool + Send + Sync + 'static,

Creates a dmabuf global with the specified supported formats.

This function unlike DmabufState::create_global also allows you to specify a filter function to determine which clients may see this global. This functionality may be used on multi-gpu systems in order to make a client choose the correct gpu.

Note: This function will create a version 3 dmabuf global and thus not call DmabufHandler::new_surface_feedback, if you want to create a version 4 global you need to call DmabufState::create_global_with_filter_and_default_feedback

pub fn create_global_with_default_feedback<D>( &mut self, display: &DisplayHandle, default_feedback: &DmabufFeedback ) -> DmabufGlobalwhere D: GlobalDispatch<ZwpLinuxDmabufV1, DmabufGlobalData> + BufferHandler + DmabufHandler + 'static,

Creates a dmabuf global with the specified default feedback.

Clients binding to version 3 or lower will receive the formats from the main tranche.

pub fn create_global_with_filter_and_default_feedback<D, F>( &mut self, display: &DisplayHandle, default_feedback: &DmabufFeedback, filter: F ) -> DmabufGlobalwhere D: GlobalDispatch<ZwpLinuxDmabufV1, DmabufGlobalData> + BufferHandler + DmabufHandler + 'static, F: for<'c> Fn(&'c Client) -> bool + Send + Sync + 'static,

Creates a dmabuf global with the specified supported formats and default feedback

This function unlike DmabufState::create_global_with_default_feedback also allows you to specify a filter function to determine which clients may see this global. This functionality may be used on multi-gpu systems in order to make a client choose the correct gpu.

Clients binding to version 3 or lower will receive the formats from the main tranche.

pub fn set_default_feedback( &self, global: &DmabufGlobal, default_feedback: &DmabufFeedback )

Set the default DmabufFeedback for the specified global and send it to all known default feedbacks.

Note: This will do nothing if the global is not found, the global has been initialized without feedback or the feedback equals the current default feedback.

pub fn disable_global<D: 'static>( &mut self, display: &DisplayHandle, global: &DmabufGlobal )

Disables a dmabuf global.

This operation is permanent and there is no way to re-enable a global.

pub fn destroy_global<D: 'static>( &mut self, display: &DisplayHandle, global: DmabufGlobal )

Destroys a dmabuf global.

It is highly recommended you disable the global before destroying it and ensure all child objects have been destroyed.