smithay::wayland::dmabuf

Trait DmabufHandler

pub trait DmabufHandler: BufferHandler {
    // Required methods
    fn dmabuf_state(&mut self) -> &mut DmabufState;
    fn dmabuf_imported(
        &mut self,
        global: &DmabufGlobal,
        dmabuf: Dmabuf,
        notifier: ImportNotifier,
    );

    // Provided method
    fn new_surface_feedback(
        &mut self,
        _surface: &WlSurface,
        _global: &DmabufGlobal,
    ) -> Option<DmabufFeedback> { ... }
}

Available on crate feature wayland_frontend only.

Handler trait for Dmabuf import from the compositor.

Required Methods

fn dmabuf_state(&mut self) -> &mut DmabufState

Returns a mutable reference to the DmabufState delegate type.

fn dmabuf_imported( &mut self, global: &DmabufGlobal, dmabuf: Dmabuf, notifier: ImportNotifier, )

This function is called when a client has imported a Dmabuf.

The global indicates which DmabufGlobal the buffer was imported to. You should import the dmabuf into your renderer to ensure the dmabuf may be used later when rendering.

Whether dmabuf import succeeded is notified through the ImportNotifier object provided in this function.

Provided Methods

fn new_surface_feedback( &mut self, _surface: &WlSurface, _global: &DmabufGlobal, ) -> Option<DmabufFeedback>

This function allows to override the default DmabufFeedback for a surface

Note: This will only be called if this will be the first time a surface requests feedback.

Returning None will use the default DmabufFeedback from the global