smithay::wayland::drm_lease

Struct DrmLeaseState

pub struct DrmLeaseState { /* private fields */ }

Available on crate features wayland_frontend and backend_drm only.

Delegate type for a drm_lease global

Implementations

impl DrmLeaseState

pub fn new<D>( display: &DisplayHandle, drm_node: &DrmNode, ) -> Result<DrmLeaseState, Error>
where D: DrmLeaseHandler + GlobalDispatch<WpDrmLeaseDeviceV1, DrmLeaseDeviceGlobalData> + Dispatch<WpDrmLeaseConnectorV1, DrmNode, D> + Dispatch<WpDrmLeaseDeviceV1, DrmNode, D> + Dispatch<WpDrmLeaseRequestV1, DrmLeaseRequestData, D> + Dispatch<WpDrmLeaseV1, DrmLeaseData, D> + 'static,

Create a new DRM lease global for a given DrmNode.

pub fn new_with_filter<D, F>( display: &DisplayHandle, drm_node: &DrmNode, filter: F, ) -> Result<DrmLeaseState, Error>
where D: DrmLeaseHandler + GlobalDispatch<WpDrmLeaseDeviceV1, DrmLeaseDeviceGlobalData> + Dispatch<WpDrmLeaseConnectorV1, DrmNode, D> + Dispatch<WpDrmLeaseDeviceV1, DrmNode, D> + Dispatch<WpDrmLeaseRequestV1, DrmLeaseRequestData, D> + Dispatch<WpDrmLeaseV1, DrmLeaseData, D> + 'static, F: for<'c> Fn(&'c Client) -> bool + Send + Sync + 'static,

Create a new DRM lease global for a given DrmNode and a filter.

Filters can be used to limit visibility of a global to certain clients.

pub fn add_connector<D>( &mut self, connector: Handle, name: String, description: String, )
where D: DrmLeaseHandler + GlobalDispatch<WpDrmLeaseDeviceV1, DrmLeaseDeviceGlobalData> + Dispatch<WpDrmLeaseConnectorV1, DrmNode, D> + Dispatch<WpDrmLeaseDeviceV1, DrmNode, D> + Dispatch<WpDrmLeaseRequestV1, DrmLeaseRequestData, D> + Dispatch<WpDrmLeaseV1, DrmLeaseData, D> + 'static,

Add a connector, that is free to be leased to clients.

pub fn withdraw_connector(&mut self, connector: Handle)

Withdraw a connector from the set of connectors available for leasing

pub fn suspend(&mut self)

Suspend all connectors temporarily (e.g. upon losing DRM master as the session becomes inactive)

pub fn resume<D>(&mut self)
where D: DrmLeaseHandler + GlobalDispatch<WpDrmLeaseDeviceV1, DrmLeaseDeviceGlobalData> + Dispatch<WpDrmLeaseConnectorV1, DrmNode, D> + Dispatch<WpDrmLeaseDeviceV1, DrmNode, D> + Dispatch<WpDrmLeaseRequestV1, DrmLeaseRequestData, D> + Dispatch<WpDrmLeaseV1, DrmLeaseData, D> + 'static,

Resume all connectors temporarily (e.g. upon gaining DRM master as the session becomes active)

pub fn node(&self) -> DrmNode

DrmNode belonging to this DRM lease global

pub fn disable_global<D>(&mut self)
where D: DrmLeaseHandler + GlobalDispatch<WpDrmLeaseDeviceV1, DrmLeaseDeviceGlobalData> + Dispatch<WpDrmLeaseConnectorV1, DrmNode, D> + Dispatch<WpDrmLeaseDeviceV1, DrmNode, D> + Dispatch<WpDrmLeaseRequestV1, DrmLeaseRequestData, D> + Dispatch<WpDrmLeaseV1, DrmLeaseData, D> + 'static,

Disable the global, it will no longer be advertised to new clients