smithay::wayland

Module drm_syncobj

Available on crate features wayland_frontend and backend_drm only.

DRM syncobj protocol

This module implement the linux-drm-syncobj-v1 protocol, used to support explicit sync.

Currently, the implementation here assumes acquire fences are already signalled when the surface transaction is ready. Use DrmSyncPointBlocker.

The server should only expose the protocol if supports_syncobj_eventfd returns true. Or it won’t be possible to create the blocker. This is similar to other implementations.

The release fence is signalled when all references to a Buffer are dropped.

pub struct State {
    syncobj_state: Option<DrmSyncobjState>,
}

impl DrmSyncobjHandler for State {
    fn drm_syncobj_state(&mut self) -> Option<&mut DrmSyncobjState> {
        self.syncobj_state.as_mut()
    }
}

let syncobj_state = if supports_syncobj_eventfd(&import_device) {
    Some(DrmSyncobjState::new::<State>(&display_handle, import_device))
} else {
    None
};

delegate_drm_syncobj!(State);

Structs

DrmSyncPoint

Point on a DRM timeline syncobj

DrmSyncPointBlocker

Blocker implementation for an accompaning DrmSyncPointSource

DrmSyncPointSource

Event source generating an event when a DrmSyncPoint is signalled..

DrmSyncobjCachedState

Pending DRM syncobj sync point state

DrmSyncobjGlobalData

Data associated with a drm syncobj global

DrmSyncobjState

Delegate type for a wp_linux_drm_syncobj_manager_v1 global

DrmSyncobjSurfaceData

Data attached to wp_linux_drm_syncobj_surface_v1 objects

DrmSyncobjTimelineData

Data attached to wp_linux_drm_syncobj_timeline_v1 objects

DrmTimeline

DRM timeline syncobj

Traits

DrmSyncobjHandler

Handler trait for DRM syncobj protocol.

Functions

supports_syncobj_eventfd

Test if DRM device supports syncobj_eventfd.