Function smithay::wayland::explicit_synchronization::get_explicit_synchronization_state

pub fn get_explicit_synchronization_state(
    attrs: &mut SurfaceAttributes
) -> Result<ExplicitSyncState, ()>

Retrieve the explicit synchronization state commited by the client

This state can contain an acquire fence and a release object, for synchronization (see module-level docs).

This function will clear the pending state, preparing the surface for the next commit, as a result you should always call it on surface commit to avoid getting out-of-sync with the client.

This function returns an error if the client has not setup explicit synchronization for this surface.