Function smithay::wayland::explicit_synchronization::send_explicit_synchronization_error

pub fn send_explicit_synchronization_error(
    attrs: &SurfaceAttributes, 
    error: ExplicitSyncError
)

Send a synchronization error to a client

See the enum definition for possible errors. These errors are protocol errors, meaning that the client associated with this SurfaceAttributes will be killed as a result of calling this function.