smithay::wayland

Module fifo

Available on crate feature wayland_frontend only.

Utilities for handling the wp-fifo protocol

How to use it

Initialization

To initialize this implementation create the FifoManagerState and store it inside your State struct.

use smithay::delegate_fifo;
use smithay::wayland::compositor;
use smithay::wayland::fifo::FifoManagerState;

// Create the compositor state
let fifo_manager_state = FifoManagerState::new::<State>(
    &display.handle(),
);

// insert the FifoManagerState into your state
// ..

delegate_fifo!(State);

// You're now ready to go!

Use the fifo state

Whenever the client commits a surface content update requesting to wait on a previously set barrier through wp_fifo_v1::Request::WaitBarrier the implementation will place a Blocker on the surface. A surface content update requesting to set a barrier through wp_fifo_v1::Request::SetBarrier will update the barrier a later commit can use to wait on.

It is your responsibility to query for a set barrier on a surface and signal it to allow clients to make forward progress.

You can query the current barrier and signal it as shown in the following example:

use smithay::wayland::compositor;
use smithay::wayland::fifo::FifoBarrierCachedState;
compositor::with_states(&surface, |states| {
    let fifo_barrier = states
        .cached_state
        .get::<FifoBarrierCachedState>()
        .current()
        .barrier
        .take();
     
    if let Some(fifo_barrier) = fifo_barrier {
        fifo_barrier.signal();
        // ..signal blocker cleared
    }
});

Unmanaged mode

If for some reason the integrated solution for fifo does not suit your needs you can create an unmanaged version with FifoManagerState::unmanaged. In this case it is your responsibility to listen for surface commits and place blockers accordingly to the spec.

The double buffered fifo state can be retrieved like shown in the following example:

use smithay::wayland::compositor;
use smithay::wayland::fifo::FifoCachedState;

let fifo_state = compositor::with_states(&surface, |states| {
    *states.cached_state.get::<FifoCachedState>().pending()
});

Structs

FifoBarrierCachedState

Fifo barrier per surface state

FifoCachedState

State for the WpFifoV1 object

FifoManagerState

State for the WpFifoManagerV1 global