smithay::wayland
Module foreign_toplevel_list

Available on crate feature wayland_frontend only.

Foreign toplevel list

The purpose of this protocol is to provide protocol object handles for toplevels.

use smithay::wayland::foreign_toplevel_list::{ForeignToplevelListState, ForeignToplevelListHandler};

pub struct State {
    foreign_toplevel_list: ForeignToplevelListState,
}

smithay::delegate_foreign_toplevel_list!(State);

impl ForeignToplevelListHandler for State {
    fn foreign_toplevel_list_state(&mut self) -> &mut ForeignToplevelListState {
        &mut self.foreign_toplevel_list
    }
}

let mut state = State {
    foreign_toplevel_list: ForeignToplevelListState::new::<State>(&display_handle),
};

let handle = state.foreign_toplevel_list.new_toplevel::<State>("Window Title", "com.example");

// Handle can be used to update title and app_id
handle.send_title("Window title has changed");
handle.send_done();

// Handle can also be used to close the window, after this call the handle will become inert,
// and the handle will no longer be announced to clients
handle.send_closed();

Structs

ForeignToplevelHandle

Handle of a toplevel, used to update title or app_id, after initial handle creation

ForeignToplevelListGlobalData

Glabal data of ExtForeignToplevelListV1

ForeignToplevelListState

State of the ExtForeignToplevelListV1 global

ForeignToplevelWeakHandle

Weak version of ForeignToplevelHandle

Traits

ForeignToplevelListHandler

Handler for foreign toplevel list protocol