smithay::wayland::idle_inhibit

Trait IdleInhibitHandler

pub trait IdleInhibitHandler {
    // Required methods
    fn inhibit(&mut self, surface: WlSurface);
    fn uninhibit(&mut self, surface: WlSurface);
}

Available on crate feature wayland_frontend only.

Handler trait for idle-inhibit.

Required Methods

fn inhibit(&mut self, surface: WlSurface)

Enable idle inhibition for the output of the provided surface.

fn uninhibit(&mut self, surface: WlSurface)

Stop inhibition for the provided surface.

This function is only called when a client explicitly removes the session inhibition. It is up to the compositor to ignore inhibiting surfaces which are invisible or dead.