smithay::wayland::image_capture_source

Trait ImageCaptureSourceHandler

pub trait ImageCaptureSourceHandler: Dispatch<ExtImageCaptureSourceV1, ImageCaptureSourceData> + 'static {
    // Provided method
    fn source_destroyed(&mut self, source: ImageCaptureSource) { ... }
}

Available on crate feature wayland_frontend only.

Core handler for image capture sources.

This is required by both OutputCaptureSourceHandler and ToplevelCaptureSourceHandler.

Provided Methods

fn source_destroyed(&mut self, source: ImageCaptureSource)

Called when a capture source is destroyed.

Use this to clean up any compositor-side state associated with the source.

Note: Destruction might happen explicitly by the client, or implicitly when the client quits. In case of implicit destruction the order the callbacks are called in is undefined.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.