Struct smithay::wayland::output::Output

pub struct Output { /* fields omitted */ }

An output as seen by the clients

This handle is stored in the events loop, and allows you to notify clients about any change in the properties of this output.

Methods

impl Output

pub fn new<L>(
evlh: &mut EventLoopHandle,
name: String,
physical: PhysicalProperties,
logger: L
) -> (StateToken<Output>, Global<WlOutput, StateToken<Output>>) where
L: Into<Option<Logger>>,

Create a new output global with given name and physical properties

The global is directly registered into the eventloop, and this function returns the state token allowing you to access it, as well as the global handle, in case you whish to remove this global in the future.

pub fn set_preferred(&mut self, mode: Mode)

Sets the preferred mode of this output

If the provided mode was not previously known to this output, it is added to its internal list.

pub fn add_mode(&mut self, mode: Mode)

Adds a mode to the list of known modes to this output

pub fn delete_mode(&mut self, mode: Mode)

Removes a mode from the list of known modes

It will not de-advertize it from existing clients (the protocol does not allow it), but it won't be advertized to now clients from now on.

pub fn change_current_state(
&mut self,
new_mode: Option<Mode>,
new_transform: Option<Transform>,
new_scale: Option<i32>
)

Change the current state of this output

You can changed the current mode, transform status or scale of this output. Providing None to any of these field means that the value does not change.

If the provided mode was not previously known to this output, it is added to its internal list.

By default, transform status is Normal, and scale is 1.

pub fn owns(&self, output: &WlOutput) -> bool

Chech is given wl_output instance is managed by this Output.

pub fn cleanup(&mut self)

Cleanup internal wl_output instances list

Clients do not necessarily notify the server on the destruction of their wl_output instances. This can lead to accumulation of stale values in the internal instances list. This methods delete them.

It can be good to call this regularly (but not necessarily very often).