smithay::wayland

Module pointer_constraints

Available on crate feature wayland_frontend only.

Protocol for confining the pointer.

This provides a way for the client to request that the pointer is confined to a region or locked in place.

Structs

ConfinedPointer

Constraint confining pointer to a region of the surface

LockedPointer

Constraint locking pointer in place

PointerConstraintRef

A reference to a pointer constraint that can be activated or deactivated.

PointerConstraintsState

Pointer constraints state.

Enums

PointerConstraint

A constraint imposed on the pointer instance

Traits

PointerConstraintsHandler

Handler for pointer constraints

Functions

with_pointer_constraint

Get constraint for surface and pointer, if any