smithay::wayland::pointer_warp

Trait PointerWarpHandler

pub trait PointerWarpHandler:
    SeatHandler
    + GlobalDispatch<WpPointerWarpV1, ()>
    + Dispatch<WpPointerWarpV1, ()>
    + 'static {
    // Provided method
    fn warp_pointer(
        &mut self,
        surface: WlSurface,
        pointer: WlPointer,
        pos: Point<f64, Logical>,
        serial: Serial,
    ) { ... }
}

Available on crate feature wayland_frontend only.

Handler trait for pointer warp events.

Provided Methods

fn warp_pointer( &mut self, surface: WlSurface, pointer: WlPointer, pos: Point<f64, Logical>, serial: Serial, )

Request the compositor to move the pointer to a surface-local position. Whether or not the compositor honors the request is implementation defined, but it should

honor it if the surface has pointer focus, including when it has an implicit pointer grab

reject it if the enter serial is incorrect

reject it if the requested position is outside of the surface

Note that the enter serial is valid for any surface of the client, and does not have to be from the surface the pointer is warped to.

serial - serial number of the surface enter event

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.