Struct smithay::wayland::seat::PointerAxisHandle

pub struct PointerAxisHandle<'a> { /* fields omitted */ }

A frame of pointer axis events.

Can be used with the builder pattern, e.g.:

pointer.axis()
    .source(AxisSource::Wheel)
    .discrete(Axis::Vertical, 6)
    .value(Axis::Vertical, 30, time)
    .stop(Axis::Vertical);

Methods

impl<'a> PointerAxisHandle<'a>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn source(&mut self, source: AxisSource) -> &mut Self

Specify the source of the axis events

This event is optional, if no source is known, you can ignore this call. Only one source event is allowed per frame.

Using the AxisSource::Finger requires a stop event to be send, when the user lifts off the finger (not necessarily in the same frame).

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn discrete(&mut self, axis: Axis, steps: i32) -> &mut Self

Specify discrete scrolling steps additionally to the computed value.

This event is optional and gives the client additional information about the nature of the axis event. E.g. a scroll wheel might issue separate steps, while a touchpad may never issue this event as it has no steps.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn value(&mut self, axis: Axis, value: f64, time: u32) -> &mut Self

The actual scroll value. This event is the only required one, but can also be send multiple times. The values off one frame will be accumulated by the client.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn stop(&mut self, axis: Axis, time: u32) -> &mut Self

Notification of stop of scrolling on an axis.

This event is required for sources of the AxisSource::Finger type and otherwise optional.

pub fn done(&mut self)

Finish this event

This will group all axis calls together. Note: They are already submitted to the client, obmitting this call just leaves room for additional events.