Struct smithay::wayland::seat::PointerInnerHandle

pub struct PointerInnerHandle<'a> { /* fields omitted */ }

This inner handle is accessed from inside a pointer grab logic, and directly sends event to the client

Methods

impl<'a> PointerInnerHandle<'a>

pub fn set_grab<G: PointerGrab + 'static>(&mut self, serial: u32, grab: G)

Change the current grab on this pointer to the provided grab

Overwrites any current grab.

pub fn unset_grab(&mut self)

Remove any current grab on this pointer, reseting it to the default behavior

pub fn current_focus(&self) -> Option<&(Resource<WlSurface>, (f64, f64))>

Access the current focus of this pointer

pub fn current_location(&self) -> (f64, f64)

Access the current location of this pointer in the global space

pub fn motion(
&mut self,
(x, y): (f64, f64),
focus: Option<(Resource<WlSurface>, (f64, f64))>,
serial: u32,
time: u32
)

Notify that the pointer moved

You provide the new location of the pointer, in the form of:

The coordinates of the pointer in the global compositor space

The surface on top of which the cursor is, and the coordinates of its origin in the global compositor space (or None of the pointer is not on top of a client surface).

This will internally take care of notifying the appropriate client objects of enter/motion/leave events.

pub fn button(&self, button: u32, state: ButtonState, serial: u32, time: u32)

Notify that a button was pressed

This will internally send the appropriate button event to the client objects matching with the currently focused surface.

pub fn axis(&mut self, details: AxisFrame)

Notify that an axis was scrolled

This will internally send the appropriate axis events to the client objects matching with the currently focused surface.