Struct smithay::wayland::seat::Seat

pub struct Seat { /* fields omitted */ }

Internal data of a seat global

This struct gives you access to the control of the capabilities of the associated seat.

It is directly inserted in the event loop by its new method.

See module-level documentation for details of use.

Methods

impl Seat

pub fn new<L>(
evlh: &mut EventLoopHandle,
name: String,
logger: L
) -> (StateToken<Seat>, Global<WlSeat, StateToken<Seat>>) where
L: Into<Option<Logger>>,

Create a new seat global

A new seat global is created with given name and inserted into this event loop.

You are provided with the state token to retrieve it (allowing you to add or remove capabilities from it), and the global handle, in case you want to remove it.

pub fn add_pointer(&mut self) -> PointerHandle

Adds the pointer capability to this seat

You are provided a PointerHandle, which allows you to send input events to this keyboard. This handle can be cloned and sent accross threads.

Calling this method on a seat that already has a pointer capability will overwrite it, and will be seen by the clients as if the mouse was unplugged and a new one was plugged.

pub fn remove_pointer(&mut self)

Remove the pointer capability from this seat

Clients will be appropriately notified.

pub fn add_keyboard(
&mut self,
model: &str,
layout: &str,
variant: &str,
options: Option<String>,
repeat_delay: i32,
repeat_rate: i32
) -> Result<KeyboardHandle, KeyboardError>

Adds the keyboard capability to this seat

You are provided a KbdHandle, which allows you to send input events to this keyboard. This handle can be cloned and sent accross threads.

You also provide a Model/Layout/Variant/Options specification of the keymap to be used for this keyboard, as well as any repeat-info that will be forwarded to the clients.

Calling this method on a seat that already has a keyboard capability will overwrite it, and will be seen by the clients as if the keyboard was unplugged and a new one was plugged.

pub fn remove_keyboard(&mut self)

Remove the keyboard capability from this seat

Clients will be appropriately notified.

pub fn owns(&self, seat: &WlSeat) -> bool

Checks wether a given WlSeat is associated with this Seat

pub fn cleanup(&mut self)

Cleanup internal states from old resources

Deletes all remnnant of ressources from clients that are now disconnected.

It can be wise to run this from time to time.