Trait smithay::wayland::seat::KeyboardGrab

pub trait KeyboardGrab {
    fn input(
        &mut self, 
        handle: &mut KeyboardInnerHandle<'_>, 
        keycode: u32, 
        key_state: WlKeyState, 
        modifiers: Option<(u32, u32, u32, u32)>, 
        serial: Serial, 
        time: u32
    );

    fn set_focus(
        &mut self, 
        handle: &mut KeyboardInnerHandle<'_>, 
        focus: Option<&WlSurface>, 
        serial: Serial
    );

    fn start_data(&self) -> &GrabStartData;
}

A trait to implement a keyboard grab

In some context, it is necessary to temporarily change the behavior of the keyboard. This is typically known as a keyboard grab. A example would be, during a popup grab the keyboard focus will not be changed and stay on the grabbed popup.

This trait is the interface to intercept regular keyboard events and change them as needed, its interface mimics the KeyboardHandle interface.

If your logic decides that the grab should end, both KeyboardInnerHandle and KeyboardHandle have a method to change it.

When your grab ends (either as you requested it or if it was forcefully cancelled by the server), the struct implementing this trait will be dropped. As such you should put clean-up logic in the destructor, rather than trying to guess when the grab will end.

Required methods

fn input(
&mut self,
handle: &mut KeyboardInnerHandle<'_>,
keycode: u32,
key_state: WlKeyState,
modifiers: Option<(u32, u32, u32, u32)>,
serial: Serial,
time: u32
)

An input was reported

fn set_focus(
&mut self,
handle: &mut KeyboardInnerHandle<'_>,
focus: Option<&WlSurface>,
serial: Serial
)

A focus change was requested

fn start_data(&self) -> &GrabStartData

The data about the event that started the grab.