smithay::wayland

Module selection

Available on crate feature wayland_frontend only.

Handler utilities for the various selection protocols.

Wayland originally only had a singe way to deal with the user clipboard with the help of wl_data_device protocol. However the demand for primary selection as well as writing clipboard managers made it into developing protocol extension for the said cases.

The following selection providers are supported:

The data_device module to work with the traditional clipboard selection.

The primary_selection module to work with the primary selection.

The wlr_data_control module to hook data control into clipboard and primary selection

Modules

data_device

Utilities for manipulating the data devices

ext_data_control

Automatic handling of the ext_data_control protocol

primary_selection

Utilities for manipulating the primary selection

wlr_data_control

Automatic handling of the wlr_data_control protocol

Structs

SelectionSource

The source of the selection data.

Enums

SelectionTarget

The target for the selection request.

Traits

SelectionHandler

Events that are generated by interactions of the clients with the data device.