smithay::wayland::selection

Module ext_data_control

Available on crate feature wayland_frontend only.

Automatic handling of the ext_data_control protocol

Initialization

To initialize this implementation, create DataControlState, store it in your State struct, and implement the required trait, as shown in the example:

use smithay::wayland::selection::SelectionHandler;
use smithay::wayland::selection::ext_data_control::{DataControlState, DataControlHandler};

// Create the data_control state
let data_control_state = DataControlState::new::<State, _>(
    &display.handle(), None, |_| true
);

// insert the DataControlState into your state
// ..

// implement the necessary traits
impl SelectionHandler for State {
    type SelectionUserData = ();
}
impl DataControlHandler for State {
    fn data_control_state(&mut self) -> &mut DataControlState { &mut self.data_control_state }
    // ... override default implementations here to customize handling ...
}
delegate_ext_data_control!(State);

// You're now ready to go!

Be aware that data control clients rely on other selection providers to be implemneted, like wl_data_device or zwp_primary_selection.

Structs

DataControlState

State of the data control.

Traits

DataControlHandler

Access the data control state.