smithay::wayland::selection::primary_selection

Function request_primary_client_selection

pub fn request_primary_client_selection<D>(
    seat: &Seat<D>,
    mime_type: String,
    fd: OwnedFd,
) -> Result<(), SelectionRequestError>
where
    D: SeatHandler + PrimarySelectionHandler + 'static,

Available on crate feature wayland_frontend only.

Request the current primary selection of the given seat to be written to the provided file descriptor with the given mime type.