smithay::wayland
Module shell

Available on crate feature wayland_frontend only.

Handler utilities for the various shell protocols

Wayland, via its different protocol extensions, supports different kind of shells. Here a shell represent the logic associated to displaying windows and arranging them on the screen.

The shell protocols thus define what kind of interactions a client can have with the compositor to properly display its contents on the screen.

Smithay currently provides three of them:

The xdg module provides handlers for the xdg_shell protocol, which is the current standard for desktop apps

The wlr_layer module provides handlers for the wlr_layer_shell protocol, which is for windows rendering above/below normal XDG windows

The kde module provides handlers for KDE-specific protocols

Modules

kde

Handler utilities for KDE shell protocols.

wlr_layer

Utilities for handling shell surfaces with the wlr_layer_shell protocol

xdg

Utilities for handling shell surfaces with the xdg_shell protocol

Enums

PingError

Represents the possible errors returned from a surface ping

Functions

is_toplevel_equivalent

Returns true if the surface is toplevel equivalent.