Module smithay::wayland::shell::kde::decoration

KDE Window decoration manager

This interface allows a compositor to announce support for KDE’s legacy server-side decorations.

A client can use this protocol to request being decorated by a supporting compositor.

extern crate wayland_server;
extern crate smithay;

use smithay::delegate_kde_decoration;
use smithay::wayland::shell::kde::decoration::{KdeDecorationHandler, KdeDecorationState};
use wayland_protocols_misc::server_decoration::server::org_kde_kwin_server_decoration_manager::Mode;

// Create the new KdeDecorationState.
let state = KdeDecorationState::new::<State, _>(&display.handle(), Mode::Server, None);

// Insert KdeDecorationState into your compositor state.
// …

// Implement KDE server decoration handlers.
impl KdeDecorationHandler for State {
    fn kde_decoration_state(&self) -> &KdeDecorationState {
        &self.kde_decoration_state
    }
}

delegate_kde_decoration!(State);

Structs

KdeDecorationState

KDE server decoration state.

Traits

KdeDecorationHandler

KDE server decoration handler.