smithay::wayland::shell::kde::decoration

Struct KdeDecorationState

pub struct KdeDecorationState { /* private fields */ }

Available on crate feature wayland_frontend only.

KDE server decoration state.

Implementations

impl KdeDecorationState

pub fn new<D>(display: &DisplayHandle, default_mode: DefaultMode) -> Self
where D: GlobalDispatch<OrgKdeKwinServerDecorationManager, KdeDecorationManagerGlobalData> + Dispatch<OrgKdeKwinServerDecorationManager, ()> + Dispatch<OrgKdeKwinServerDecoration, WlSurface> + KdeDecorationHandler + 'static,

Create a new KDE server decoration global.

pub fn new_with_filter<D, F>( display: &DisplayHandle, default_mode: DefaultMode, filter: F, ) -> Self
where D: GlobalDispatch<OrgKdeKwinServerDecorationManager, KdeDecorationManagerGlobalData> + Dispatch<OrgKdeKwinServerDecorationManager, ()> + Dispatch<OrgKdeKwinServerDecoration, WlSurface> + KdeDecorationHandler + 'static, F: for<'c> Fn(&'c Client) -> bool + Send + Sync + 'static,

Create a new KDE server decoration global with a filter.

Filters can be used to limit visibility of a global to certain clients.

pub fn global(&self) -> GlobalId

Returns the id of the [OrgKdeKwinServerDecorationManager] global.

pub fn set_default_mode(&mut self, default_mode: DefaultMode)

Updates the default decoration mode.

The new default mode will only be sent to new binds to the manager object; it is up to you to update any existing decoration instances with the new mode, if desired.