Module smithay::wayland::shell::legacy

Utilities for handling shell surfaces with the wl_shell protocol

This module provides automatic handling of shell surfaces objects, by being registered as a global handler for wl_shell. This protocol is deprecated in favor of xdg_shell, thus this module is provided as a compatibility layer with older clients. As a consequence, you can as a compositor-writer decide to only support its functionality in a best-effort maneer: as this global is part of the core protocol, you are still required to provide some support for it.

Why use this implementation

This implementation can track for you the various shell surfaces defined by the clients by handling the wl_shell protocol.

It allows you to easily access a list of all shell surfaces defined by your clients access their associated metadata and underlying wl_surfaces.

This handler only handles the protocol exchanges with the client to present you the information in a coherent and relatively easy to use maneer. All the actual drawing and positioning logic of windows is out of its scope.

How to use it

Initialization

To initialize this handler, simple use the wl_shell_init function provided in this module. You will need to provide it the CompositorToken you retrieved from an instanciation of the compositor handler provided by smithay.

use smithay::wayland::compositor::roles::*;
use smithay::wayland::compositor::CompositorToken;
use smithay::wayland::shell::legacy::{wl_shell_init, ShellSurfaceRole, ShellRequest};
use wayland_server::{EventLoop, LoopToken};

// define the metadata you want associated with the shell surfaces
#[derive(Default)]
pub struct MyShellSurfaceData {
    /* ... */
}

// define the roles type. You need to integrate the XdgSurface role:
define_roles!(MyRoles =>
    [ShellSurface, ShellSurfaceRole<MyShellSurfaceData>]
);

let (shell_state, _) = wl_shell_init(
    &mut display,
    event_loop.token(),
    // token from the compositor implementation
    compositor_token,
    // your implementation, can also be a strucy implementing the
    // appropriate Implementation<(), ShellRequest<_, _, _>> trait
    |event: ShellRequest<_, _, MyShellSurfaceData>, ()| { /* ... */ },
    None  // put a logger if you want
);

// You're now ready to go!

Structs

ShellState
Shell global state

ShellSurface
A handle to a shell surface

ShellSurfaceRole
Metadata associated with the xdg_surface role

Enums

ShellRequest
A request triggered by a wl_shell_surface

ShellSurfaceKind
Possible kinds of shell surface of the wl_shell protocol

Functions

wl_shell_init
Create a new wl_shell global