Enum smithay::wayland::shell::legacy::ShellRequest

pub enum ShellRequest<U, R, D> {
    NewShellSurface {
        surface: ShellSurface<U, R, D>,
    },
    Pong {
        surface: ShellSurface<U, R, D>,
    },
    Move {
        surface: ShellSurface<U, R, D>,
        serial: u32,
        seat: Resource<WlSeat>,
    },
    Resize {
        surface: ShellSurface<U, R, D>,
        serial: u32,
        seat: Resource<WlSeat>,
        edges: Resize,
    },
    SetKind {
        surface: ShellSurface<U, R, D>,
        kind: ShellSurfaceKind,
    },
}

A request triggered by a wl_shell_surface

Variants
NewShellSurface

A new shell surface was created

by default it has no kind and this should not be displayed

Fields of NewShellSurface

surface: ShellSurface<U, R, D>

The created surface

Pong

A pong event

The surface responded to its pending ping. If you receive this event, smithay has already checked that the responded serial was valid.

Fields of Pong

surface: ShellSurface<U, R, D>

The surface that sent the pong

Move

Start of an interactive move

The surface requests that an interactive move is started on it

Fields of Move

surface: ShellSurface<U, R, D>

The surface requesting the move

serial: u32

Serial of the implicit grab that initiated the move

seat: Resource<WlSeat>

Seat associated with the move

Resize

Start of an interactive resize

The surface requests that an interactive resize is started on it

Fields of Resize

surface: ShellSurface<U, R, D>

The surface requesting the resize

serial: u32

Serial of the implicit grab that initiated the resize

seat: Resource<WlSeat>

Seat associated with the resize

edges: Resize

Direction of the resize

SetKind

The surface changed its kind

Fields of SetKind

surface: ShellSurface<U, R, D>

The surface

kind: ShellSurfaceKind

Its new kind