Enum smithay::wayland::shell::legacy::ShellSurfaceKind

pub enum ShellSurfaceKind {
    Toplevel,
    Transient {
        parent: Resource<WlSurface>,
        location: (i32, i32),
        inactive: bool,
    },
    Fullscreen {
        method: FullscreenMethod,
        framerate: u32,
        output: Option<Resource<WlOutput>>,
    },
    Popup {
        parent: Resource<WlSurface>,
        serial: u32,
        inactive: bool,
        location: (i32, i32),
        seat: Resource<WlSeat>,
    },
    Maximized {
        output: Option<Resource<WlOutput>>,
    },
}

Possible kinds of shell surface of the wl_shell protocol

Variants
Toplevel

Toplevel, a regular window displayed somewhere in the compositor space

Transient

Transient, this surface has a parent surface

These are sub-windows of an application (for example a configuration window), and as such should only be visible in their parent window is, and on top of it.

Fields of Transient

parent: Resource<WlSurface>

The surface considered as parent

location: (i32, i32)

Location relative to the parent

inactive: bool

Wether this window should be marked as inactive

Fullscreen

Fullscreen surface, covering an entire output

Fields of Fullscreen

method: FullscreenMethod

Method used for fullscreen

framerate: u32

Framerate (relevant only for driver fullscreen)

output: Option<Resource<WlOutput>>

Requested output if any

Popup

A popup surface

Short-lived surface, typically refrered as "tooltips" in many contexts.

Fields of Popup

parent: Resource<WlSurface>

The parent surface of this popup

serial: u32

The serial of the input event triggering the creation of this popup

inactive: bool

Wether this popup should be marked as inactive

location: (i32, i32)

Location of the popup relative to its parent

seat: Resource<WlSeat>

Seat associated this the input that triggered the creation of the popup. Used to define when the "popup done" event is sent.

Maximized

A maximized surface

Like a toplevel surface, but as big as possible on a single output while keeping any relevant desktop-environment interface visible.

Fields of Maximized

output: Option<Resource<WlOutput>>

Requested output for maximization