Struct smithay::wayland::shell::ShellSurfaceUserImplementation

pub struct ShellSurfaceUserImplementation<U, R, CID, SID, SD> {
    pub new_client: fn(evlh: &mut EventLoopHandle, idata: &mut SID, client: ShellClient<SD>),
    pub client_pong: fn(evlh: &mut EventLoopHandle, idata: &mut SID, client: ShellClient<SD>),
    pub new_toplevel: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>) -> ToplevelConfigure,
    pub new_popup: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: PopupSurface<U, R, CID, SD>) -> PopupConfigure,
    pub move_: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32),
    pub resize: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32, edges: ResizeEdge),
    pub grab: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: PopupSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32),
    pub change_display_state: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, maximized: Option<bool>, minimized: Option<bool>, fullscreen: Option<bool>, output: Option<&WlOutput>) -> ToplevelConfigure,
    pub show_window_menu: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32, x: i32, y: i32),
}

A sub-implementation for the shell

You need to provide this to handle events that the provided implementation cannot process for you directly.

Depending on what you want to do, you might implement some of these functions as doing nothing.

Fields
new_client: fn(evlh: &mut EventLoopHandle, idata: &mut SID, client: ShellClient<SD>)

A new shell client was instanciated

client_pong: fn(evlh: &mut EventLoopHandle, idata: &mut SID, client: ShellClient<SD>)

The pong for a pending ping of this shell client was received

The ShellHandler already checked for you that the serial matches the one from the pending ping.

new_toplevel: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>) -> ToplevelConfigure

A new toplevel surface was created

You need to return a ToplevelConfigure from this function, which will be sent to the client to configure this surface

new_popup: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: PopupSurface<U, R, CID, SD>) -> PopupConfigure

A new popup surface was created

You need to return a PopupConfigure from this function, which will be sent to the client to configure this surface

move_: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32)

The client requested the start of an interactive move for this surface

resize: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32, edges: ResizeEdge)

The client requested the start of an interactive resize for this surface

The edges argument specifies which part of the window's border is being dragged.

grab: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: PopupSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32)

This popup requests a grab of the pointer

This means it requests to be sent a popup_done event when the pointer leaves the grab area.

change_display_state: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, maximized: Option<bool>, minimized: Option<bool>, fullscreen: Option<bool>, output: Option<&WlOutput>) -> ToplevelConfigure

A toplevel surface requested its display state to be changed

Each field represents the request of the client for a specific property:

None: no request is made to change this property

Some(true): this property should be enabled

Some(false): this property should be disabled

For fullscreen/maximization, the client can also optionnaly request a specific output.

You are to answer with a ToplevelConfigure that will be sent to the client in response.

show_window_menu: fn(evlh: &mut EventLoopHandle, idata: &mut SID, surface: ToplevelSurface<U, R, CID, SD>, seat: &WlSeat, serial: u32, x: i32, y: i32)

The client requests the window menu to be displayed on this surface at this location

This menu belongs to the compositor. It is typically expected to contain options for control of the window (maximize/minimize/close/move/etc...).