Enum smithay::wayland::shell::xdg::XdgRequest

pub enum XdgRequest<U, R, SD> {
    NewClient {
        client: ShellClient<SD>,
    },
    ClientPong {
        client: ShellClient<SD>,
    },
    NewToplevel {
        surface: ToplevelSurface<U, R, SD>,
    },
    NewPopup {
        surface: PopupSurface<U, R, SD>,
    },
    Move {
        surface: ToplevelSurface<U, R, SD>,
        seat: Resource<WlSeat>,
        serial: u32,
    },
    Resize {
        surface: ToplevelSurface<U, R, SD>,
        seat: Resource<WlSeat>,
        serial: u32,
        edges: ResizeEdge,
    },
    Grab {
        surface: PopupSurface<U, R, SD>,
        seat: Resource<WlSeat>,
        serial: u32,
    },
    Maximize {
        surface: ToplevelSurface<U, R, SD>,
    },
    UnMaximize {
        surface: ToplevelSurface<U, R, SD>,
    },
    Fullscreen {
        surface: ToplevelSurface<U, R, SD>,
        output: Option<Resource<WlOutput>>,
    },
    UnFullscreen {
        surface: ToplevelSurface<U, R, SD>,
    },
    Minimize {
        surface: ToplevelSurface<U, R, SD>,
    },
    ShowWindowMenu {
        surface: ToplevelSurface<U, R, SD>,
        seat: Resource<WlSeat>,
        serial: u32,
        location: (i32, i32),
    },
}

Events generated by xdg shell surfaces

These are events that the provided implementation cannot process for you directly.

Depending on what you want to do, you might ignore some of them

Variants
NewClient

A new shell client was instanciated

Fields of NewClient

client: ShellClient<SD>

the client

ClientPong

The pong for a pending ping of this shell client was received

The ShellHandler already checked for you that the serial matches the one from the pending ping.

Fields of ClientPong

client: ShellClient<SD>

the client

NewToplevel

A new toplevel surface was created

You likely need to send a ToplevelConfigure to the surface, to hint the client as to how its window should be

Fields of NewToplevel

surface: ToplevelSurface<U, R, SD>

the surface

NewPopup

A new popup surface was created

You likely need to send a PopupConfigure to the surface, to hint the client as to how its popup should be

Fields of NewPopup

surface: PopupSurface<U, R, SD>

the surface

Move

The client requested the start of an interactive move for this surface

Fields of Move

surface: ToplevelSurface<U, R, SD>

the surface

seat: Resource<WlSeat>

the seat associated to this move

serial: u32

the grab serial

Resize

The client requested the start of an interactive resize for this surface

Fields of Resize

surface: ToplevelSurface<U, R, SD>

The surface

seat: Resource<WlSeat>

The seat associated with this resize

serial: u32

The grab serial

edges: ResizeEdge

Specification of which part of the window's border is being dragged

Grab

This popup requests a grab of the pointer

This means it requests to be sent a popup_done event when the pointer leaves the grab area.

Fields of Grab

surface: PopupSurface<U, R, SD>

The surface

seat: Resource<WlSeat>

The seat to grab

serial: u32

The grab serial

Maximize

A toplevel surface requested to be maximized

Fields of Maximize

surface: ToplevelSurface<U, R, SD>

The surface

UnMaximize

A toplevel surface requested to stop being maximized

Fields of UnMaximize

surface: ToplevelSurface<U, R, SD>

The surface

Fullscreen

A toplevel surface requested to be set fullscreen

Fields of Fullscreen

surface: ToplevelSurface<U, R, SD>

The surface

output: Option<Resource<WlOutput>>

The output (if any) on which the fullscreen is requested

UnFullscreen

A toplevel surface request to stop being fullscreen

Fields of UnFullscreen

surface: ToplevelSurface<U, R, SD>

The surface

Minimize

A toplevel surface requested to be minimized

Fields of Minimize

surface: ToplevelSurface<U, R, SD>

The surface

ShowWindowMenu

The client requests the window menu to be displayed on this surface at this location

This menu belongs to the compositor. It is typically expected to contain options for control of the window (maximize/minimize/close/move/etc...).

Fields of ShowWindowMenu

surface: ToplevelSurface<U, R, SD>

The surface

seat: Resource<WlSeat>

The seat associated with this input grab

serial: u32

the grab serial

location: (i32, i32)

location of the menu request