smithay::wayland::shell::xdg

Struct PopupSurface

pub struct PopupSurface { /* private fields */ }

Available on crate feature wayland_frontend only.

A handle to a popup surface

This is an unified abstraction over the popup surfaces of both wl_shell and xdg_shell.

Implementations

impl PopupSurface

pub fn alive(&self) -> bool

Is the toplevel surface referred by this handle still alive?

pub fn get_parent_surface(&self) -> Option<WlSurface>

Gets a reference of the parent WlSurface of this popup.

pub fn client(&self) -> ShellClient

Retrieve the shell client owning this popup surface

pub fn send_pending_configure( &self, ) -> Result<Option<Serial>, PopupConfigureError>

Send a pending configure event to this popup surface to suggest it a new configuration

If changes have occurred a configure event will be send to the clients and the serial will be returned (for tracking the configure in XdgShellHandler::ack_configure if desired). If no changes occurred no event will be send and Ok(None) will be returned.

See send_configure and has_pending_changes for more information.

pub fn send_configure(&self) -> Result<Serial, PopupConfigureError>

Send a configure event to this popup surface to suggest it a new configuration

The serial of this configure will be tracked waiting for the client to ACK it.

You can manipulate the state that will be sent to the client with the with_pending_state method.

Returns Err(PopupConfigureError) if the initial configure has already been sent and the client protocol version disallows a re-configure or the current PositionerState is not reactive.

Note: This will always send a configure event, if you intend to only send a configure event on changes take a look at send_pending_configure

pub fn is_initial_configure_sent(&self) -> bool

Did the surface sent the initial configure event to the client.

Calls compositor::with_states internally.

pub fn send_repositioned(&self, token: u32) -> Serial

Send a configure event, including the repositioned event to the client in response to a reposition request.

For further information see send_configure

pub fn send_popup_done(&self)

Send a popup_done event to the popup surface

It means that the use has dismissed the popup surface, or that the pointer has left the area of popup grab if there was a grab.

pub fn wl_surface(&self) -> &WlSurface

Access the underlying wl_surface of this popup surface

pub fn xdg_popup(&self) -> &XdgPopup

Access the underlying xdg_popup of this popup surface

pub fn with_pending_state<F, T>(&self, f: F) -> T
where F: FnOnce(&mut PopupState) -> T,

Allows the pending state of this popup to be manipulated.

This should be used to inform the client about size and position changes, for example after a move of the parent toplevel.

The state will be sent to the client when calling send_configure.

pub fn has_pending_changes(&self) -> bool

Tests this PopupSurface for pending changes

Returns true if with_pending_state was used to manipulate the state and resulted in a different state or if the initial configure is still pending.

pub fn with_cached_state<F, T>(&self, f: F) -> T
where F: FnOnce(&PopupCachedState) -> T,

Provides access to the current committed cached state.

pub fn with_committed_state<F, T>(&self, f: F) -> T
where F: FnOnce(Option<&PopupState>) -> T,

Provides access to the current committed state.

This is the state that the client last acked before making the current commit.