Struct smithay::wayland::shell::xdg::PositionerState

pub struct PositionerState {
    pub rect_size: Size<i32, Logical>,
    pub anchor_rect: Rectangle<i32, Logical>,
    pub anchor_edges: Anchor,
    pub gravity: Gravity,
    pub constraint_adjustment: ConstraintAdjustment,
    pub offset: Point<i32, Logical>,
    pub reactive: bool,
    pub parent_size: Option<Size<i32, Logical>>,
    pub parent_configure: Option<Serial>,
}

The state of a positioner, as set by the client

Fields
rect_size: Size<i32, Logical>

Size of the rectangle that needs to be positioned

anchor_rect: Rectangle<i32, Logical>

Anchor rectangle in the parent surface coordinates relative to which the surface must be positioned

anchor_edges: Anchor

Edges defining the anchor point

gravity: Gravity

Gravity direction for positioning the child surface relative to its anchor point

constraint_adjustment: ConstraintAdjustment

Adjustments to do if previous criteria constrain the surface

offset: Point<i32, Logical>

Offset placement relative to the anchor point

reactive: bool

When set reactive, the surface is reconstrained if the conditions used for constraining changed, e.g. the parent window moved.

If the conditions changed and the popup was reconstrained, an xdg_popup.configure event is sent with updated geometry, followed by an xdg_surface.configure event.

parent_size: Option<Size<i32, Logical>>

The parent window geometry the compositor should use when positioning the popup. The compositor may use this information to determine the future state the popup should be constrained using. If this doesn’t match the dimension of the parent the popup is eventually positioned against, the behavior is undefined.

The arguments are given in the surface-local coordinate space.

parent_configure: Option<Serial>

The serial of an xdg_surface.configure event this positioner will be used in response to. The compositor may use this information together with set_parent_size to determine what future state the popup should be constrained using.

Implementations

impl PositionerState

pub fn get_geometry(&self) -> Rectangle<i32, Logical>

Get the geometry for a popup as defined by this positioner.

Rectangle::width and Rectangle::height corresponds to the size set by xdg_positioner.set_size.

Rectangle::x and Rectangle::y define the position of the popup relative to it’s parent surface window_geometry. The position is calculated according to the rules defined in the xdg_shell protocol. The constraint_adjustment will not be considered by this implementation and the position and size should be re-calculated in the compositor if the compositor implements constraint_adjustment