Struct smithay::wayland::shell::xdg::XdgShellState

pub struct XdgShellState { /* private fields */ }

Shell global state

This state allows you to retrieve a list of surfaces currently known to the shell global.

Implementations

impl XdgShellState

pub fn new<D>(display: &DisplayHandle) -> XdgShellStatewhere D: GlobalDispatch<XdgWmBase, ()> + 'static,

Create a new xdg_shell global with all WmCapabilities

pub fn new_with_capabilities<D>( display: &DisplayHandle, capabilities: impl Into<WmCapabilitieSet> ) -> XdgShellStatewhere D: GlobalDispatch<XdgWmBase, ()> + 'static,

Create a new xdg_shell global with a specific set of WmCapabilities

pub fn replace_capabilities( &mut self, capabilities: impl Into<WmCapabilitieSet> )

Replace the capabilities of this global

Note: This does not update the capabilities on existing toplevels, only new toplevels are affected. To update existing toplevels iterate over them, update their capabilities and send a configure.

pub fn toplevel_surfaces(&self) -> &[ToplevelSurface]

Access all the shell surfaces known by this handler

pub fn get_toplevel(&self, toplevel: &XdgToplevel) -> Option<ToplevelSurface>

Returns a ToplevelSurface from an underlying toplevel surface.

pub fn popup_surfaces(&self) -> &[PopupSurface]

Access all the popup surfaces known by this handler

pub fn get_popup(&self, popup: &XdgPopup) -> Option<PopupSurface>

Returns a PopupSurface from an underlying popup surface.

pub fn global(&self) -> GlobalId

Returns the xdg shell global.