Struct smithay::wayland::shell::xdg::XdgSurfaceRole

pub struct XdgSurfaceRole {
    pub pending_state: XdgSurfacePendingState,
    pub window_geometry: Option<Rectangle>,
    pub pending_configures: Vec<u32>,
    pub configured: bool,
}

Metadata associated with the xdg_surface role

Fields
pending_state: XdgSurfacePendingState

Pending state as requested by the client

The data in this field are double-buffered, you should apply them on a surface commit.

window_geometry: Option<Rectangle>

Geometry of the surface

Defines, in surface relative coordinates, what should be considered as "the surface itself", regarding focus, window alignment, etc...

By default, you should consider the full contents of the buffers of this surface and its subsurfaces.

pending_configures: Vec<u32>

List of non-acked configures pending

Whenever a configure is acked by the client, all configure older than it are discarded as well. As such, this vec contains the serials of all the configure send to this surface that are newer than the last ack received.

configured: bool

Has this surface acked at least one configure?

xdg_shell defines it as illegal to commit on a surface that has not yet acked a configure.