smithay::wayland::shell::xdg

Struct XdgToplevelSurfaceRoleAttributes

pub struct XdgToplevelSurfaceRoleAttributes {
    pub initial_configure_sent: bool,
    pub server_pending: Option<ToplevelState>,
    pub last_acked: Option<ToplevelConfigure>,
    pub parent: Option<WlSurface>,
    pub dialog_hint: ToplevelDialogHint,
    pub title: Option<String>,
    pub app_id: Option<String>,
    pub initial_decoration_configure_sent: bool,
    /* private fields */
}

Available on crate feature wayland_frontend only.

Role specific attributes for xdg_toplevel

This interface defines an xdg_surface role which allows a surface to, among other things, set window-like properties such as maximize, fullscreen, and minimize, set application-specific metadata like title and id, and well as trigger user interactive operations such as interactive resize and move.

Unmapping an xdg_toplevel means that the surface cannot be shown by the compositor until it is explicitly mapped again. All active operations (e.g., move, resize) are cancelled and all attributes (e.g. title, state, stacking, …) are discarded for an xdg_toplevel surface when it is unmapped. The xdg_toplevel returns to the state it had right after xdg_surface.get_toplevel. The client can re-map the toplevel by performing a commit without any buffer attached, waiting for a configure event and handling it as usual (see xdg_surface description).

Attaching a null buffer to a toplevel unmaps the surface.

Fields
initial_configure_sent: bool

Holds the state if the surface has sent the initial configure event to the client. It is expected that during the first commit a initial configure event is sent to the client

server_pending: Option<ToplevelState>

Holds the pending state as set by the server.

last_acked: Option<ToplevelConfigure>

Holds the last configure that has been acknowledged by the client. This state should be cloned to the current during a commit. Note that this state can be newer than the last acked state at the time of the last commit.

parent: Option<WlSurface>

The parent field of a toplevel should be used by the compositor to determine which toplevel should be brought to front. If the parent is focused all of it’s child should be brought to front.

dialog_hint: ToplevelDialogHint

Hint related to the toplevel’s “dialog” status. See ToplevelDialogHint for more information.

This value has no effect on toplevels that are not attached to a parent toplevel.

title: Option<String>

Holds the optional title the client has set for this toplevel. For example a web-browser will most likely set this to include the current uri.

This string may be used to identify the surface in a task bar, window list, or other user interface elements provided by the compositor.

app_id: Option<String>

Holds the optional app ID the client has set for this toplevel.

The app ID identifies the general class of applications to which the surface belongs. The compositor can use this to group multiple surfaces together, or to determine how to launch a new application.

For D-Bus activatable applications, the app ID is used as the D-Bus service name.

initial_decoration_configure_sent: bool

An zxdg_toplevel_decoration_v1::configure event has been sent to the client.

Implementations

impl XdgToplevelSurfaceRoleAttributes

pub fn current_server_state(&self) -> ToplevelState

Gets the latest state that has been configured on the server and sent to the client.

The state includes all changes that have been made on the server, including all not yet acked or committed changes, but excludes the current server_pending state.

This can be used for example to check if the server_pending state is different from the last configured.

pub fn has_pending_changes(&self) -> bool

Check if the state has pending changes that have not been sent to the client.

This differs from just checking if the server_pending state is Some in that it also checks if a current pending state is different from the current_server_state.

pub fn pending_configures(&self) -> &[ToplevelConfigure]

Returns a list of configures sent to, but not yet acknowledged by the client.

The list is ordered by age, so the last configure in the list is the last one sent to the client.