smithay::wayland::shell::xdg

Trait XdgShellHandler

pub trait XdgShellHandler {

    // Required methods
    fn xdg_shell_state(&mut self) -> &mut XdgShellState;
    fn new_toplevel(&mut self, surface: ToplevelSurface);
    fn new_popup(&mut self, surface: PopupSurface, positioner: PositionerState);
    fn grab(&mut self, surface: PopupSurface, seat: WlSeat, serial: Serial);
    fn reposition_request(
        &mut self,
        surface: PopupSurface,
        positioner: PositionerState,
        token: u32,
    );

    // Provided methods
    fn new_client(&mut self, client: ShellClient) { ... }
    fn client_pong(&mut self, client: ShellClient) { ... }
    fn move_request(
        &mut self,
        surface: ToplevelSurface,
        seat: WlSeat,
        serial: Serial,
    ) { ... }
    fn resize_request(
        &mut self,
        surface: ToplevelSurface,
        seat: WlSeat,
        serial: Serial,
        edges: ResizeEdge,
    ) { ... }
    fn maximize_request(&mut self, surface: ToplevelSurface) { ... }
    fn unmaximize_request(&mut self, surface: ToplevelSurface) { ... }
    fn fullscreen_request(
        &mut self,
        surface: ToplevelSurface,
        output: Option<WlOutput>,
    ) { ... }
    fn unfullscreen_request(&mut self, surface: ToplevelSurface) { ... }
    fn minimize_request(&mut self, surface: ToplevelSurface) { ... }
    fn show_window_menu(
        &mut self,
        surface: ToplevelSurface,
        seat: WlSeat,
        serial: Serial,
        location: Point<i32, Logical>,
    ) { ... }
    fn ack_configure(&mut self, surface: WlSurface, configure: Configure) { ... }
    fn client_destroyed(&mut self, client: ShellClient) { ... }
    fn toplevel_destroyed(&mut self, surface: ToplevelSurface) { ... }
    fn popup_destroyed(&mut self, surface: PopupSurface) { ... }
    fn app_id_changed(&mut self, surface: ToplevelSurface) { ... }
    fn title_changed(&mut self, surface: ToplevelSurface) { ... }
    fn parent_changed(&mut self, surface: ToplevelSurface) { ... }

}

Available on crate feature wayland_frontend only.

Xdg Shell handler type

Required Methods

fn xdg_shell_state(&mut self) -> &mut XdgShellState

XdgShellState getter

fn new_toplevel(&mut self, surface: ToplevelSurface)

A new toplevel surface was created

You likely need to send a ToplevelConfigure to the surface, to hint the client as to how its window should be sized.

fn new_popup(&mut self, surface: PopupSurface, positioner: PositionerState)

A new popup surface was created

You likely need to send a PopupConfigure to the surface, to hint the client as to how its popup should be sized.

Arguments

positioner - The state of the positioner at the timethe popup was requested. The positioner state can be used by the compositor to calculate the best placement for the popup. For example the compositor should prevent that a popup is placed outside the visible rectangle of a output.

fn grab(&mut self, surface: PopupSurface, seat: WlSeat, serial: Serial)

This popup requests a grab of the pointer

This means it requests to be sent a popup_done event when the pointer leaves the grab area.

fn reposition_request( &mut self, surface: PopupSurface, positioner: PositionerState, token: u32, )

A client requested a reposition, providing a new positioner for a popup.

To confirm the new popup position, PopupSurface::send_repositioned must be called on the provided surface with the token.

Arguments

positioner - The state of the positioner at the time the reposition request was made. The positioner state can be used by the compositor to calculate the best placement for the popup. For example the compositor should prevent that a popup is placed outside the visible rectangle of a output.

token - The passed token will be sent in the corresponding xdg_popup.repositioned event. The new popup position will not take effect until the corresponding configure event is acknowledged by the client. See xdg_popup.repositioned for details. The token itself is opaque, and has no other special meaning.

Provided Methods

fn new_client(&mut self, client: ShellClient)

A new shell client was instantiated

fn client_pong(&mut self, client: ShellClient)

The pong for a pending ping of this shell client was received

The ShellHandler already checked for you that the serial matches the one from the pending ping.

fn move_request( &mut self, surface: ToplevelSurface, seat: WlSeat, serial: Serial, )

The client requested the start of an interactive move for this surface

fn resize_request( &mut self, surface: ToplevelSurface, seat: WlSeat, serial: Serial, edges: ResizeEdge, )

The client requested the start of an interactive resize for this surface

fn maximize_request(&mut self, surface: ToplevelSurface)

A toplevel surface requested to be maximized

fn unmaximize_request(&mut self, surface: ToplevelSurface)

A toplevel surface requested to stop being maximized

fn fullscreen_request( &mut self, surface: ToplevelSurface, output: Option<WlOutput>, )

A toplevel surface requested to be set fullscreen

fn unfullscreen_request(&mut self, surface: ToplevelSurface)

A toplevel surface request to stop being fullscreen

fn minimize_request(&mut self, surface: ToplevelSurface)

A toplevel surface requested to be minimized

fn show_window_menu( &mut self, surface: ToplevelSurface, seat: WlSeat, serial: Serial, location: Point<i32, Logical>, )

The client requests the window menu to be displayed on this surface at this location

This menu belongs to the compositor. It is typically expected to contain options for control of the window (maximize/minimize/close/move/etc…).

fn ack_configure(&mut self, surface: WlSurface, configure: Configure)

A surface has acknowledged a configure serial.

fn client_destroyed(&mut self, client: ShellClient)

A shell client was destroyed.

Note: Destruction might happen explicitly by the client, or implicitly when the client quits. In case of implicit destruction the order the callbacks are called in is undefined.

fn toplevel_destroyed(&mut self, surface: ToplevelSurface)

A toplevel surface was destroyed.

Note: Destruction might happen explicitly by the client, or implicitly when the client quits. In case of implicit destruction the order the callbacks are called in is undefined.

fn popup_destroyed(&mut self, surface: PopupSurface)

A popup surface was destroyed.

Note: Destruction might happen explicitly by the client, or implicitly when the client quits. In case of implicit destruction the order the callbacks are called in is undefined.

fn app_id_changed(&mut self, surface: ToplevelSurface)

The toplevel surface set a different app id.

fn title_changed(&mut self, surface: ToplevelSurface)

The toplevel surface set a different title.

fn parent_changed(&mut self, surface: ToplevelSurface)

The parent of a toplevel surface has changed.