Function smithay::wayland::shm::init_shm_global

pub fn init_shm_global<L>(
    display: &mut Display, 
    formats: Vec<Format>, 
    logger: L
) -> Global<WlShm> where
    L: Into<Option<Logger>>, 

Create a new SHM global advertizing given supported formats.

This global will always advertize ARGB8888 and XRGB8888 format as they are required by the protocol. Formats given as argument as additionally advertized.

The global is directly created on the provided Display, and this function returns the global handle, in case you wish to remove this global in the future.