Function smithay::wayland::shm::with_buffer_contents

pub fn with_buffer_contents<F, T>(
    buffer: &WlBuffer,
    f: F
) -> Result<T, BufferAccessError>
where
    F: FnOnce(*const u8, usize, BufferData) -> T,

Available on crate feature wayland_frontend only.

Call given closure with the contents of the given buffer

If the buffer is managed by the provided ShmGlobal, its contents are extracted and the closure is extracted with them:

The first argument is a pointer to the contents of the pool

The second argument is the length of the contents of the pool

The third argument is the specification of this buffer is this pool

The pool cannot be provided as a slice since it is shared memory that could be mutated by the client.

If the buffer is not managed by the provided ShmGlobal, the closure is not called and this method will return Err(BufferAccessError::NotManaged) (this will be the case for an EGL buffer for example).

Safety

The pointer passed to the callback is only valid while the callback is running. The shared memory may be mutated by the client at any time. Creating a reference or slice into this memory is undefined behavior if it is mutated by the client while the reference exists.