smithay::wayland

Trait Dispatch2

pub trait Dispatch2<I: Resource, State> {
    // Required method
    fn request(
        &self,
        state: &mut State,
        client: &Client,
        resource: &I,
        request: I::Request,
        dhandle: &DisplayHandle,
        data_init: &mut DataInit<'_, State>,
    );

    // Provided method
    fn destroyed(&self, _state: &mut State, _client: ClientId, _resource: &I) { ... }
}

Available on crate feature wayland_frontend only.

A simplified version of wayland_server::Dispatch

A future version of wayland-server will replace Dispatch with this.

Required Methods

fn request( &self, state: &mut State, client: &Client, resource: &I, request: I::Request, dhandle: &DisplayHandle, data_init: &mut DataInit<'_, State>, )

Called when a request from a client is processed.

Provided Methods

fn destroyed(&self, _state: &mut State, _client: ClientId, _resource: &I)

Called when the object this user data is associated with has been destroyed.