Module smithay::wayland::viewporter

Available on crate feature wayland_frontend only.

Utilities for handling the wp_viewporter protocol

How to use it

Initialization

To initialize this implementation, create ViewporterState, store it in your State struct and implement the required traits, as shown in this example:

use smithay::wayland::viewporter::ViewporterState;
use smithay::delegate_viewporter;

// Create the viewporter state:
let viewporter_state = ViewporterState::new::<State>(
    &display.handle(), // the display
);

// implement Dispatch for the Viewporter types
delegate_viewporter!(State);

// You're now ready to go!

Use the viewport state

The viewport state is double-buffered and can be accessed by using the with_states function

let viewport = with_states(surface, |states| {
    states.cached_state.current::<ViewportCachedState>();
});

Before accessing the state you should call ensure_viewport_valid to ensure the viewport is valid.

Note: If you already hand over buffer management to smithay by using on_commit_buffer_handler the implementation will already call ensure_viewport_valid for you.

Structs

ViewportCachedState

Represents the double-buffered viewport state of a WlSurface

ViewportState

State of a single viewport attached to a surface

ViewporterState

State of the wp_viewporter Global

Functions

ensure_viewport_valid

Ensures that the viewport, if any, is valid accordingly to the protocol specification.