Trait smithay::wayland::xdg_activation::XdgActivationHandler

pub trait XdgActivationHandler {
    // Required methods
    fn activation_state(&mut self) -> &mut XdgActivationState;
    fn request_activation(
        &mut self,
        token: XdgActivationToken,
        token_data: XdgActivationTokenData,
        surface: WlSurface
    );

    // Provided method
    fn token_created(
        &mut self,
        token: XdgActivationToken,
        data: XdgActivationTokenData
    ) -> bool { ... }
}

Available on crate feature wayland_frontend only.

A trait implemented to be notified of activation requests using the xdg activation protocol.

Required Methods

fn activation_state(&mut self) -> &mut XdgActivationState

Returns the activation state.

fn request_activation( &mut self, token: XdgActivationToken, token_data: XdgActivationTokenData, surface: WlSurface )

A client has requested surface activation.

The compositor may know which client requested this by checking the token data and may decide whether or not to follow through with the activation if it’s considered unwanted.

The token remains in the pool and might be used to issue other requests until the compositor decides to remove it using XdgActivationState::remove_token or XdgActivationState::retain_tokens.

Provided Methods

fn token_created( &mut self, token: XdgActivationToken, data: XdgActivationTokenData ) -> bool

A client has created a new token.

If the token isn’t considered valid, it can be immediately untracked by returning false. The default implementation considers every token valid and will always return true.

This method may also be used to attach user_data to the token.