smithay::wayland

Module xdg_system_bell

Available on crate feature wayland_frontend only.

XDG System Bell

This protocol enables clients to ring the system bell.

In order to advertise system bell global call XdgSystemBellState::new and delegate events to it with delegate_xdg_system_bell.

use smithay::wayland::xdg_system_bell::{XdgSystemBellState, XdgSystemBellHandler};
use wayland_server::protocol::wl_surface::WlSurface;
use smithay::delegate_xdg_system_bell;

XdgSystemBellState::new::<State>(
    &display.handle(),
);

// provide the necessary trait implementations
impl XdgSystemBellHandler for State {
    fn ring(&mut self, surface: Option<WlSurface>) {
        println!("Ring got called");
    }
}

delegate_xdg_system_bell!(State);

Structs

XdgSystemBellState

State of the xdg system bell

Traits

XdgSystemBellHandler

Handler for xdg ring request