smithay::wayland

Module xdg_toplevel_tag

Available on crate feature wayland_frontend only.

XDG Toplevel Tag

This protocol allows clients to set a tag and description of a toplevel.

In order to advertise toplevel tag global call XdgToplevelTagManager::new and delegate events to it with delegate_xdg_toplevel_tag. Currently attached tag is available either via XdgToplevelTagHandler or in XdgToplevelTagSurfaceData

use smithay::wayland::xdg_toplevel_tag::{XdgToplevelTagManager, XdgToplevelTagHandler};
use wayland_protocols::xdg::shell::server::xdg_toplevel::XdgToplevel;
use wayland_server::protocol::wl_surface::WlSurface;
use smithay::delegate_xdg_toplevel_tag;

XdgToplevelTagManager::new::<State>(
    &display.handle(),
);

impl XdgToplevelTagHandler for State {
    fn set_tag(&mut self, toplevel: XdgToplevel, tag: String) {
        dbg!(tag);
    }

    fn set_description(&mut self, toplevel: XdgToplevel, description: String) {
        dbg!(description);
    }
}

delegate_xdg_toplevel_tag!(State);

Structs

XdgToplevelTagManager

Delegate type for handling xdg toplevel tag events.

XdgToplevelTagSurfaceData

Data associated with WlSurface Represents the client pending state

Traits

XdgToplevelTagHandler

Handler trait for xdg toplevel tag events.