smithay::xwayland
Module xwm

Available on crate feature xwayland only.

Xwayland Window Manager module

Provides an X11Wm type, which will register itself as a window manager for a previously spawned Xwayland instances, allowing backwards-compatibility by seamlessly integrating X11 windows into a wayland compositor.

To use this functionality you must first spawn an XWayland instance to attach a X11Wm to.

Associating an X11Surface with a [WlSurface]

When an X11 window is created, XWayland will associate it with a wl_surface that it creates via the wayland connection. This happens in two steps:

Via the xwayland shell protocol, a serial number is set on the surface, and it is given the “xwayland_shell” role.

Via the X11 connection, a matching WL_SURFACE_SERIAL atom is set on the X11 window, which can be queried with X11Surface::wl_surface_serial().

Note that these two steps can happen in any order.

Example

struct State { /* ... */ }
impl XwmHandler for State {
    fn xwm_state(&mut self, xwm: XwmId) -> &mut X11Wm {
        // ...
    }
    fn new_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn new_override_redirect_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn map_window_request(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn mapped_override_redirect_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn unmapped_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn destroyed_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn configure_request(&mut self, xwm: XwmId, window: X11Surface, x: Option<i32>, y: Option<i32>, w: Option<u32>, h: Option<u32>, reorder: Option<Reorder>) { /* ... */ }
    fn configure_notify(&mut self, xwm: XwmId, window: X11Surface, geometry: Rectangle<i32, Logical>, above: Option<u32>) { /* ... */ }
    fn resize_request(&mut self, xwm: XwmId, window: X11Surface, button: u32, resize_edge: ResizeEdge) { /* ... */ }
    fn move_request(&mut self, xwm: XwmId, window: X11Surface, button: u32) { /* ... */ }
    fn send_selection(&mut self, xwm: XwmId, selection: SelectionTarget, mime_type: String, fd: OwnedFd) { /* ... */ }
}

let (xwayland, client) = XWayland::spawn(
    &dh,
    None,
    std::iter::empty::<(String, String)>(),
    true,
    Stdio::null(),
    Stdio::null(),
    |_| (),
)
.expect("failed to start XWayland");
let ret = handle.insert_source(xwayland, move |event, _, data| match event {
    XWaylandEvent::Ready {
        x11_socket,
        display_number: _,
    } => {
        let wm = X11Wm::start_wm(
            handle.clone(),
            x11_socket,
            client.clone(),
        )
        .expect("Failed to attach X11 Window Manager");
         
        // store the WM somewhere
    }
    XWaylandEvent::Error => eprintln!("XWayland failed to start!"),
}); if let Err(e) = ret { tracing::error!( "Failed to insert the
XWaylandSource into the event loop: {}", e ); }

Modules

settings

Types related to XSETTINGS handling

Structs

Atoms

Atoms used by the XWM and X11Surface types

X11Surface

X11 window managed by an X11Wm

X11Wm

The runtime state of an reparenting XWayland window manager.

XwmId

Id of an X11 WM

Enums

Reorder

Window asks to be re-stacked

ResizeEdge

Edge values for resizing

SelectionError

Errors generated working with Xwm Selections

SettingsError

Errors generated updating XSETTINGS

WmWindowProperty

Window properties of X11Surfaces

WmWindowType

Window types of X11Surfaces

X11SurfaceError

Errors that can happen for operations on an X11Surface

Traits

X11Relatable

Trait for objects, that represent an x11 window in some shape or form and can be tested for equality.

XwmHandler

Handler trait for X11Wm interactions

Type Aliases

X11Window