Crate smithay

Smithay: the wayland compositor smithy

Most entry points in the modules can take an optionnal slog::Logger as argument that will be used as a drain for logging. If None is provided, they'll log to slog-stdlog.

Modules

backend
Backend (rendering/input) creation helpers

compositor
Utilities for handling surfaces, subsurfaces and regions

seat
Seat global utilities

shell
Utilities for handling shell surfaces, toplevel and popups

shm
SHM handling helpers

utils
Various utilities functions and types

Macros

define_roles