Module smithay::backend

Backend (rendering/input) creation helpers

Collection of common traits and implementation about rendering onto various targets and receiving input from various sources.

Supported graphics backends:

glutin (headless/windowed)

Supported input backends:

glutin (windowed)

Reexports

pub use glium::*;

Modules

glutin
Implementation of backend traits for types provided by glutin

graphics
Common traits for various ways to renderer on a given graphics backend.

input
Common traits for input backends to receive input from.

Macros

assert_no_gl_error
Calls the assert_no_error method on a glium::Display instance with file and line number information.

implement_buffer_content
Implements the glium::buffer::Content trait for the given type.

implement_uniform_block
Implements the glium::uniforms::UniformBlock trait for the given type.

implement_vertex
Implements the glium::vertex::Vertex trait for the given type.

program
Builds a program depending on the GLSL version supported by the backend.

uniform
Returns an implementation-defined type which implements the Uniform trait.