Module smithay::backend

Backend (rendering/input) creation helpers

Collection of common traits and implementation about rendering onto various targets and receiving input from various sources.

Supported graphics backends:

winit

drm

Supported input backends:

winit

libinput

Modules

allocator
Buffer allocation and management.

drm
This module represents abstraction on top the linux direct rendering manager api (drm).

egl
Common traits and types for egl rendering

input
Common traits for input backends to receive input from.

libinput
Implementation of input backend trait for types provided by libinput

renderer
Rendering functionality and abstractions

session
Abstraction of different session APIs.

udev
Provides udev related functionality for automated device scanning.

winit
Implementation of backend traits for types provided by winit

Enums

SwapBuffersError
Error that can happen when swapping buffers.