Module smithay::backend

Backend (rendering/input) creation helpers

Collection of common traits and implementation about rendering onto various targets and receiving input from various sources.

Supported graphics backends:

winit

drm

Supported input backends:

winit

libinput

Modules

drm
This module provides Traits reprensentating open devices and their surfaces to render contents.

egl
Common traits and types for egl rendering

graphics
Common traits for various ways to renderer on a given graphics backend.

input
Common traits for input backends to receive input from.

libinput
Implementation of input backend trait for types provided by libinput

session
Abstraction of different session APIs.

udev
Provides udev related functionality for automated device scanning.

winit
Implementation of backend traits for types provided by winit