Module smithay::backend::allocator::dmabuf

Module for dmabuf buffers.

Dmabufs act alike to smart pointers and can be freely cloned and passed around. Once the last Dmabuf reference is dropped, its file descriptor is closed and underlying resources are freed.

If you want to hold on to a potentially alive dmabuf without blocking the free up of the underlying resources, you may downgrade a Dmabuf reference to a WeakDmabuf.

This can be especially useful in resources where other parts of the stack should decide upon the lifetime of the buffer. E.g. when you are only caching associated resources for a dmabuf.

Structs

Dmabuf

Strong reference to a dmabuf handle

DmabufBuilder

Builder for Dmabufs

DmabufFlags

Possible flags for a DMA buffer

WeakDmabuf

Weak reference to a dmabuf handle

Constants

MAX_PLANES

Maximum amount of planes this implementation supports

Traits

AsDmabuf

Buffer that can be exported as Dmabufs