Module smithay::backend::allocator::dmabuf

Module for dmabuf buffers.

Dmabufs act alike to smart pointers and can be freely cloned and passed around. Once the last Dmabuf reference is dropped, its file descriptor is closed and underlying resources are freed.

If you want to hold on to a potentially alive dmabuf without blocking the free up of the underlying resources, you may downgrade a Dmabuf reference to a WeakDmabuf.

This can be especially useful in resources where other parts of the stack should decide upon the lifetime of the buffer. E.g. when you are only caching associated resources for a dmabuf.

Structs

AlreadyReady

Dmabuf is already ready for the given interest

AnyError

Type erased error

Dmabuf

Strong reference to a dmabuf handle

DmabufAllocator

Wrapper for Allocators, whos buffer types implement AsDmabuf.

DmabufBlocker

crate::wayland::compositor::Blocker implementation for an accompaning DmabufSource

DmabufBuilder

Builder for Dmabufs

DmabufFlags

Possible flags for a DMA buffer

DmabufSource

Dmabuf-based event source. Can be used to monitor implicit fences of a dmabuf.

WeakDmabuf

Weak reference to a dmabuf handle

Constants

MAX_PLANES

Maximum amount of planes this implementation supports

Traits

AsDmabuf

Buffer that can be exported as Dmabufs