Struct smithay::backend::allocator::dmabuf::Dmabuf

pub struct Dmabuf(_);

Strong reference to a dmabuf handle

Implementations

impl Dmabuf

pub fn builder_from_buffer(
src: &impl Buffer,
flags: DmabufFlags
) -> DmabufBuilder

Create a new Dmabuf by initializing with values from an existing buffer

pub fn builder(
size: impl Into<Size<i32, BufferCoords>>,
format: Fourcc,
flags: DmabufFlags
) -> DmabufBuilder

Create a new Dmabuf builder

pub fn num_planes(&self) -> usize

The amount of planes this Dmabuf has

pub fn handles(&self) -> impl Iterator<Item = RawFd> + '_

Returns raw handles of the planes of this buffer

pub fn offsets(&self) -> impl Iterator<Item = u32> + '_

Returns offsets for the planes of this buffer

pub fn strides(&self) -> impl Iterator<Item = u32> + '_

Returns strides for the planes of this buffer

pub fn has_modifier(&self) -> bool

Returns if this buffer format has any vendor-specific modifiers set or is implicit/linear

pub fn y_inverted(&self) -> bool

Returns if the buffer is stored inverted on the y-axis

pub fn weak(&self) -> WeakDmabuf

Create a weak reference to this dmabuf

impl Dmabuf

pub fn import_to<A: AsRawFd + 'static, T>(
&self,
gbm: &GbmDevice<A>,
usage: GbmBufferFlags
) -> Result<GbmBuffer<T>>

Import a Dmabuf using libgbm, creating a gbm Buffer Object to the same underlying data.