Module smithay::backend::allocator::format

Format info tables for DRM formats.

This module provides three functions, get_opaque, has_alpha and get_bpp.

get_opaque returns the opaque alternative of a DRM format with an alpha channel.

assert_eq!(Some(Fourcc::Xrgb8888), get_opaque(Fourcc::Argb8888));

has_alpha returns true if the format has an alpha channel.

assert!(has_alpha(Fourcc::Argb8888));
assert!(!has_alpha(Fourcc::Xrgb8888));

get_bpp returns the number of bits per pixel of a format.

assert_eq!(get_bpp(Fourcc::Argb8888), Some(32));

get_depth returns the number of used bits per pixel of a format (excluding padding or non-alpha “X” parts of the format).

assert_eq!(get_depth(Fourcc::Argb8888), Some(32));
assert_eq!(get_depth(Fourcc::Xrgb8888), Some(24));

Functions

get_bpp

Returns the bits per pixel of the specified format.

get_depth

Returns the depth of the specified format.

get_opaque

Returns the opaque alternative of the specified format.

get_transparent

Returns the transparent alternative of the specified format.

has_alpha

Returns true if the format has an alpha channel.