smithay::backend::allocator::gbm

Struct GbmDevice

pub struct GbmDevice<T>
where
    T: AsFd,
{ /* private fields */ }

Available on crate feature backend_gbm only.

An open GBM device

Implementations

impl<T> Device<T>
where T: AsFd,

pub fn new(fd: T) -> Result<Device<T>, Error>

Open a GBM device from a given open DRM device.

The underlying file descriptor passed in is used by the backend to communicate with platform for allocating the memory. For allocations using DRI this would be the file descriptor returned when opening a device such as /dev/dri/card0.

pub fn backend_name(&self) -> &str

Get the backend name

pub fn is_format_supported( &self, format: DrmFourcc, usage: BufferObjectFlags, ) -> bool

Test if a format is supported for a given set of usage flags

pub fn format_modifier_plane_count( &self, format: DrmFourcc, modifier: DrmModifier, ) -> Option<u32>

Get the required number of planes for a given format and modifier

Some combination (e.g. when using a Modifier::Invalid) might not have a defined/fixed number of planes. In these cases the function might return Option::None.

pub fn create_surface<U>( &self, width: u32, height: u32, format: DrmFourcc, usage: BufferObjectFlags, ) -> Result<Surface<U>, Error>
where U: 'static,

Allocate a new surface object

pub fn create_surface_with_modifiers<U>( &self, width: u32, height: u32, format: DrmFourcc, modifiers: impl Iterator<Item = DrmModifier>, ) -> Result<Surface<U>, Error>
where U: 'static,

Allocate a new surface object with explicit modifiers

pub fn create_surface_with_modifiers2<U>( &self, width: u32, height: u32, format: DrmFourcc, modifiers: impl Iterator<Item = DrmModifier>, usage: BufferObjectFlags, ) -> Result<Surface<U>, Error>
where U: 'static,

Allocate a new surface object with explicit modifiers and flags

pub fn create_buffer_object<U>( &self, width: u32, height: u32, format: DrmFourcc, usage: BufferObjectFlags, ) -> Result<BufferObject<U>, Error>
where U: 'static,

Allocate a buffer object for the given dimensions

pub fn create_buffer_object_with_modifiers<U>( &self, width: u32, height: u32, format: DrmFourcc, modifiers: impl Iterator<Item = DrmModifier>, ) -> Result<BufferObject<U>, Error>
where U: 'static,

Allocate a buffer object for the given dimensions with explicit modifiers

pub fn create_buffer_object_with_modifiers2<U>( &self, width: u32, height: u32, format: DrmFourcc, modifiers: impl Iterator<Item = DrmModifier>, usage: BufferObjectFlags, ) -> Result<BufferObject<U>, Error>
where U: 'static,

Allocate a buffer object for the given dimensions with explicit modifiers and flags

pub fn import_buffer_object_from_wayland<U>( &self, buffer: &WlBuffer, usage: BufferObjectFlags, ) -> Result<BufferObject<U>, Error>
where U: 'static,

Create a GBM buffer object from a wayland buffer

This function imports a foreign WlBuffer object and creates a new GBM buffer object for it. This enables using the foreign object with a display API such as KMS.

The GBM bo shares the underlying pixels but its life-time is independent of the foreign object.

pub fn import_buffer_object_from_dma_buf<U>( &self, buffer: BorrowedFd<'_>, width: u32, height: u32, stride: u32, format: DrmFourcc, usage: BufferObjectFlags, ) -> Result<BufferObject<U>, Error>
where U: 'static,

Create a GBM buffer object from a dma buffer

This function imports a foreign dma buffer from an open file descriptor and creates a new GBM buffer object for it. This enables using the foreign object with a display API such as KMS.

The GBM bo shares the underlying pixels but its life-time is independent of the foreign object.

pub fn import_buffer_object_from_dma_buf_with_modifiers<U>( &self, len: u32, buffers: [Option<BorrowedFd<'_>>; 4], width: u32, height: u32, format: DrmFourcc, usage: BufferObjectFlags, strides: [i32; 4], offsets: [i32; 4], modifier: DrmModifier, ) -> Result<BufferObject<U>, Error>
where U: 'static,

Create a GBM buffer object from a dma buffer with explicit modifiers

This function imports a foreign dma buffer from an open file descriptor and creates a new GBM buffer object for it. This enables using the foreign object with a display API such as KMS.

The GBM bo shares the underlying pixels but its life-time is independent of the foreign object.