Module smithay::backend::allocator::vulkan

Module for Buffers created using Vulkan.

The VulkanAllocator type implements the Allocator trait and VulkanImage implements Buffer. A VulkanImage may be exported as a dmabuf.

The Vulkan allocator supports up to Vulkan 1.3.

The Vulkan allocator requires the following device extensions (and their dependencies):

VK_EXT_image_drm_format_modifier

VK_EXT_external_memory_dmabuf

VK_KHR_external_memory_fd

Additionally the Vulkan allocator may enable the following extensions if available:

VK_EXT_4444_formats

To get the required extensions a device must support to use the Vulkan allocator, use VulkanAllocator::required_extensions.

Modules

format

Format conversions between Vulkan and DRM formats.

Structs

ImageUsageFlags

Flags to indicate the intended usage for a buffer.

VulkanAllocator

An allocator which uses Vulkan to create buffers.

VulkanImage

Vulkan image object.

Enums

Error

Error type for VulkanAllocator.

ExportError

The error type for exporting a VulkanImage as a Dmabuf.