smithay::backend::allocator::vulkan::format

Function get_vk_format

pub const fn get_vk_format(fourcc: Fourcc) -> Option<Format>

Available on crate feature backend_vulkan only.

Converts a FourCC format code to a Vulkan format code.

This will return None if the format is not known.

These format conversions will return all known FourCC and Vulkan format conversions. However a Vulkan implementation may not support some Vulkan format. One notable example of this are the formats introduced in VK_EXT_4444_formats. The corresponding FourCC codes will return the formats from VK_EXT_4444_formats, but the caller is responsible for testing that a Vulkan device supports these formats.