Module smithay::backend::drm

Drm/Kms types and backend implementations

This module provide a DrmDevice which acts as a reprensentation for any drm device and can be used to create the second provided structure a DrmBackend.

Initialization happens through the types provided by drm-rs.

Three entities are relevant for the initialization procedure.

"Crtc"s represent scanout engines of the device pointer to one framebuffer. There responsibility is to read the data of the framebuffer and export it into an "Encoder". The number of crtc's represent the number of independant output devices the hardware may handle.

An "Encoder" encodes the data of connected crtcs into a video signal for a fixed set of connectors. E.g. you might have an analog encoder based on a DAG for VGA ports, but another one for digital ones. Also not every encoder might be connected to every crtc.

The last entity the "Connector" represents a port on your computer, possibly with a connected monitor, TV, capture card, etc.

The DrmBackend created from a DrmDevice represents a crtc of the device you can render to and that feeds a given set of connectors, that can be manipulated at runtime.

From these circumstances it becomes clear, that one crtc might only send it's data to a connector, that is attached to any encoder that is attached to the crtc itself. It is the responsibility of the user to ensure that a given set of a crtc with it's connectors is valid or an error will be thrown.

For more details refer to the drm-rs documentation.

How to use it

Initialization

To initialize the DrmDevice you need either a RawFd or a File of your drm node. The File is recommended as it represents the save api.

Once you got your DrmDevice you can then use it to create DrmBackends. You will need to use the drm crate to provide the required types to create a backend.

extern crate drm;
extern crate smithay;
extern crate wayland_server;

use drm::Device as BasicDevice;
use drm::control::{Device as ControlDevice, ResourceInfo};
use drm::control::connector::{Info as ConnectorInfo, State as ConnectorState};
use drm::control::encoder::{Info as EncoderInfo};
use std::fs::{File, OpenOptions};
use std::os::unix::io::RawFd;
use std::os::unix::io::AsRawFd;
use smithay::backend::drm::{DrmDevice, DrmBackend};

#[derive(Debug)]
pub struct Card(File);

impl AsRawFd for Card {
    fn as_raw_fd(&self) -> RawFd {
        self.0.as_raw_fd()
    }
}

impl BasicDevice for Card {}
impl ControlDevice for Card {}

// Open the drm device
let mut options = OpenOptions::new();
options.read(true);
options.write(true);
let mut device = DrmDevice::new(
    Card(options.open("/dev/dri/card0").unwrap()), // try to detect it properly
    None /*put a logger here*/
).unwrap();

// Get a set of all modesetting resource handles
let res_handles = device.resource_handles().unwrap();

// Use first connected connector for this example
let connector_info = res_handles.connectors().iter()
    .map(|conn| ConnectorInfo::load_from_device(&device, *conn).unwrap())
    .find(|conn| conn.connection_state() == ConnectorState::Connected)
    .unwrap();

// Use the first encoder
let encoder_info = EncoderInfo::load_from_device(&device, connector_info.encoders()[0]).unwrap();

// use the connected crtc if any
let crtc = encoder_info.current_crtc()
    // or use the first one that is compatible with the encoder
    .unwrap_or_else(||
        *res_handles.filter_crtcs(encoder_info.possible_crtcs())
          .iter()
          .next()
          .unwrap());

// Use first mode (usually the highest resolution)
let mode = connector_info.modes()[0];

// Create the backend
let backend = device.create_backend(
    crtc,
    mode,
    vec![connector_info.handle()]
).unwrap();

Page Flips / Tear-free video

Calling the usual EglGraphicsBackend::swap_buffers function on a DrmBackend works the same to finish the rendering, but will return SwapBuffersError::AlreadySwapped for any new calls until the page flip of the crtc has happened.

You can monitor the page flips by registering the DrmDevice as and FdEventSourceHandler and setting a DrmHandler on it. You will be notified whenever a page flip has happend, so you can render the next frame immediately and get a tear-free reprensentation on the display.

You need to render at least once to successfully trigger the first event.

use drm::control::crtc::{Handle as CrtcHandle};
use drm::result::Error as DrmError;
use smithay::backend::drm::{DrmDevice, DrmBackend, DrmHandler, drm_device_bind};
use smithay::backend::graphics::egl::EGLGraphicsBackend;

struct MyDrmHandler(DrmBackend<Card>);

impl DrmHandler<Card> for MyDrmHandler {
    fn ready(
        &mut self,
        _device: &mut DrmDevice<Card>,
        _crtc: CrtcHandle,
        _frame: u32,
        _duration: Duration)
    {
        // render surfaces and swap again
        self.0.swap_buffers().unwrap();
    }
    fn error(
        &mut self,
        device: &mut DrmDevice<Card>,
        error: DrmError)
    {
        panic!("DrmDevice errored: {}", error);
    }
}

// render something (like clear_color)
backend.swap_buffers().unwrap();

let (_source, _device_rc) = drm_device_bind(
    &event_loop.token(),
    device,
    MyDrmHandler(backend)
).map_err(|(err, _)| err).unwrap();

event_loop.run().unwrap();

Modules

error
Errors thrown by the DrmDevice and DrmBackend

Structs

DrmBackend
Backend based on a DrmDevice and a given crtc

DrmDevice
Representation of an open drm device node to create rendering backends

DrmDeviceObserver
SessionObserver linked to the DrmDevice it was created from.

Traits

DevPath
Trait for types representing open devices

DrmHandler
Handler for drm node events

Functions

drm_device_bind
Bind a DrmDevice to an EventLoop,