Module smithay::backend::drm

This module provides Traits reprensentating open devices and their surfaces to render contents.

Initialization of devices happens through an open file descriptor of a drm device.

Initialization of surfaces happens through the types provided by drm-rs.

Four entities are relevant for the initialization procedure.

crtcs represent scanout engines of the device pointer to one framebuffer. Their responsibility is to read the data of the framebuffer and export it into an "Encoder". The number of crtc's represent the number of independant output devices the hardware may handle.

An encoder encodes the data of connected crtcs into a video signal for a fixed set of connectors. E.g. you might have an analog encoder based on a DAG for VGA ports, but another one for digital ones. Also not every encoder might be connected to every crtc.

A connector represents a port on your computer, possibly with a connected monitor, TV, capture card, etc.

On surface creation a matching encoder for your encoder-connector is automatically selected, if it exists, which means you still need to check your configuration.

At last a Mode needs to be selected, supported by the crtc in question.

Modules

egl
Device and Surface implementations using egl contexts and surfaces for efficient rendering.

gbm
Device and Surface implementations using gbm buffers for efficient rendering.

legacy
RawDevice and RawSurface implementations using the legacy mode-setting infrastructure.

Traits

DevPath
Trait representing open devices that may return a Path

Device
An open drm device

DeviceHandler
Trait to receive events of a bound [Device]

RawDevice
Marker trait for [Device]s able to provide [RawSurface]s

RawSurface
An open bare crtc without any rendering abstractions

Surface
An open crtc that can be used for rendering

Functions

device_bind
Bind a Device to an EventLoop,