smithay::backend::drm

Module compositor

Available on crate features wayland_frontend and backend_gbm and backend_drm only.

Composition for Elements using drm planes

When possible composition can be (partially) offloaded to the display driver by assigning elements to drm planes. This is especially important for latency intensive fullscreen clients like video renderers or games.

The DrmCompositor does so by walking the stack of provided Elements from front to back while trying to assign each element to a drm overlay plane. Each item that fails the plane test will be rendered on the primary plane using the provided Renderer. Additionally it will try to assign the top most element that fit’s into the cursor size (as specified by the DrmDevice) on the cursor plane. If the element can not be directly scanned out, pixman will be used to render the element.

Note: While the DrmCompositor also works on legacy drm the use of overlay and cursor planes is disabled in that case. Direct scan-out will only work with an atomic DrmSurface.

What makes a Element eligible for direct scan-out

General

First the element has to provide a UnderlyingStorage which can be exported as a drm framebuffer. Currently this is limited to wayland buffers, but may be extended in the future. This module provides a default exporter based on gbm which should fit most use-cases.

If a certain combination of elements works can only be determined by asking the driver by submitting a atomic commit test. If that test fails the element is scheduled to be rendered on the primary plane.

Overlay planes

The element can only be directly scanned out if it’s geometry does not overlap with an already assigned element on a plane higher in the stack.

Underlay planes

An underlay plane is only used if it does not overlap with an already assigned plane lower in the stack and the element is fully opaque.

Primary plane

For an element to be considered to be directly scanned out on the primary plane it has to be the last remaining visible element on the output and no other element has been assigned to the primary plane. If there are multiple element assigned to the primary plane the renderer will be used to composite the primary plane into a allocator provided buffer. Additionally the element has to be either fully opaque or the clear color has to match the CRTC background color and no overlap with an underlay is found.

How to use it

use smithay::{
    backend::drm::{
        compositor::{DrmCompositor, FrameFlags},
        exporter::gbm::GbmFramebufferExporter,
        DrmSurface,
    },
    output::{Output, PhysicalProperties, Subpixel},
    utils::Size,
};

// ...initialize the output, drm device, drm surface and allocator
let output = Output::new(
    "e-DP".into(),
    PhysicalProperties {
        size: Size::from((800, 600)),
        make: "N/A".into(),
        model: "N/A".into(),
        subpixel: Subpixel::Unknown,
        serial_number: "N/A".into(),
    },
);

let mut compositor: DrmCompositor<_, _, (), _> = DrmCompositor::new(
    &output,
    surface,
    None,
    allocator,
    exporter,
    color_formats,
    renderer_formats,
    device.cursor_size(),
    Some(gbm),
)
.expect("failed to initialize drm compositor");

let render_frame_result = compositor
    .render_frame::<_, _>(&mut renderer, &elements, CLEAR_COLOR, FrameFlags::DEFAULT)
    .expect("failed to render frame");

if !render_frame_result.is_empty {
    compositor.queue_frame(()).expect("failed to queue frame");

    // ...wait for VBlank event

    compositor
        .frame_submitted()
        .expect("failed to mark frame as submitted");
} else {
    // ...re-schedule frame
}

Structs

DrmCompositor

Composite an output using a combination of planes and rendering

FrameFlags

Possible flags for a DMA buffer

PrimarySwapchainElement

Defines the element for the primary plane in cases where a composited buffer was used.

RenderFrameResult

Result for DrmCompositor::render_frame

Enums

BlitFrameResultError

Error for RenderFrameResult::blit_frame_result

FrameError

Errors thrown by a DrmCompositor

PrimaryPlaneElement

Defines the element for the primary plane

RenderFrameError

Error returned from DrmCompositor::render_frame