Struct smithay::backend::drm::DrmBackend

pub struct DrmBackend { /* fields omitted */ }

Backend based on a DrmDevice and a given crtc

Methods

impl DrmBackend

fn add_connector(&mut self, connector: Handle) -> Result<()>

Add a connector to backend

Errors

Errors if the new connector does not support the currently set Mode

fn used_connectors(&self) -> &[Handle]

Returns the currently set connectors

fn remove_connector(&mut self, connector: Handle)

Removes a currently set connector

fn use_mode(&mut self, mode: Mode) -> Result<()>

Changes the currently set mode

Errors

This will fail if not all set connectors support the new Mode. Several internal resources will need to be recreated to fit the new Mode. Other errors might occur.