Struct smithay::backend::drm::DrmDevice

pub struct DrmDevice { /* private fields */ }

An open drm device

Implementations

impl DrmDevice

pub fn new(
fd: DrmDeviceFd,
disable_connectors: bool,
logger: impl Into<Option<Logger>>
) -> Result<Self, Error>

Create a new DrmDevice from an open drm node

Arguments

fd - Open drm node

disable_connectors - Setting this to true will initialize all connectors
as disabled on device creation. smithay enables connectors, when attached
to a surface, and disables them, when detached. Setting this to false
requires usage of drm-rs to disable unused connectors to prevent them
showing garbage, but will also prevent flickering of already turned on
connectors (assuming you wonโt change the resolution).

logger - Optional [slog::Logger] to be used by this device.

Return

Returns an error if the file is no valid drm node or the device is not accessible.

pub fn is_atomic(&self) -> bool

Returns if the underlying implementation uses atomic-modesetting or not.

pub fn crtcs(&self) -> &[Handle]

Returns a list of crtcs for this device

pub fn planes(&self, crtc: &Handle) -> Result<Planes, Error>

Returns a set of available planes for a given crtc

pub fn cursor_size(&self) -> Size<u32, Physical>

Returns the size of the hardware cursor

Note: In case of universal planes this is the maximum size of a buffer that can be used on the cursor plane.

pub fn create_surface(
&self,
crtc: Handle,
mode: Mode,
connectors: &[Handle]
) -> Result<DrmSurface, Error>

Creates a new rendering surface.

Arguments

Initialization of surfaces happens through the types provided by drm-rs.

crtcs represent scanout engines of the device pointing to one framebuffer.
Their responsibility is to read the data of the framebuffer and export it into an โEncoderโ.
The number of crtcโs represent the number of independent output devices the hardware may handle.

mode describes the resolution and rate of images produced by the crtc and
has to be compatible with the provided connectors.

connectors - List of connectors driven by the crtc. At least one(!) connector needs to be
attached to a crtc in smithay.

pub fn device_id(&self) -> dev_t

Returns the device_id of the underlying drm node

pub fn device_fd(&self) -> DrmDeviceFd

Returns the underlying file descriptor

pub fn pause(&self)

Pauses the device.

This will cause the DrmDevice to avoid making calls to the file descriptor e.g. on drop. Note that calls directly utilizing the underlying file descriptor, like the traits of the drm-rs crate, will ignore this state. Use DrmDevice::is_active to guard these calls.

pub fn activate(&self)

Actives a previously paused device.

pub fn is_active(&self) -> bool

Returns if the device is currently paused or not.