Struct smithay::backend::drm::DrmDevice

pub struct DrmDevice<B: Borrow<DrmBackend> + 'static> { /* fields omitted */ }

Representation of an open drm device node to create rendering backends

Methods

impl<B: From<DrmBackend> + Borrow<DrmBackend> + 'static> DrmDevice<B>

unsafe fn new_from_fd<L>(fd: RawFd, logger: L) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new DrmDevice from a raw file descriptor

Returns an error of opening the device failed or context creation was not successful.

Safety

The file descriptor might not be valid and needs to be owned by smithay, make sure not to share it. Otherwise undefined behavior might occur.

unsafe fn new_from_fd_with_gl_attr<L>(
fd: RawFd,
attributes: GlAttributes,
logger: L
) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new DrmDevice from a raw file descriptor and given GlAttributes

Returns an error of opening the device failed or context creation was not successful.

Safety

The file descriptor might not be valid and needs to be owned by smithay, make sure not to share it. Otherwise undefined behavior might occur.

fn new_from_file<L>(file: File, logger: L) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new DrmDevice from a File of an open drm node

Returns an error if the file is no valid drm node or context creation was not successful.

fn new_from_file_with_gl_attr<L>(
file: File,
attributes: GlAttributes,
logger: L
) -> Result<Self> where
L: Into<Option<Logger>>,

Create a new DrmDevice from a File of an open drm node and given GlAttributes

Returns an error if the file is no valid drm node or context creation was not successful.

fn create_backend<I>(
&mut self,
evlh: &mut EventLoopHandle,
crtc: Handle,
mode: Mode,
connectors: I
) -> Result<StateToken<B>> where
I: Into<Vec<Handle>>,

Create a new backend on a given crtc with a given Mode for a given amount of connectors (mirroring).

Errors if initialization fails or the mode is not available on all given connectors.