Struct smithay::backend::drm::DrmDevice

pub struct DrmDevice<A: AsRawFd + 'static> { /* fields omitted */ }

An open drm device

Implementations

impl<A: AsRawFd + 'static> DrmDevice<A>

pub fn new<L>(fd: A, disable_connectors: bool, logger: L) -> Result<Self, Error> where
A: AsRawFd + 'static,
L: Into<Option<Logger>>,

Create a new DrmDevice from an open drm node

Arguments

fd - Open drm node

disable_connectors - Setting this to true will initialize all connectors
as disabled on device creation. smithay enables connectors, when attached
to a surface, and disables them, when detached. Setting this to false
requires usage of drm-rs to disable unused connectors to prevent them
showing garbage, but will also prevent flickering of already turned on
connectors (assuming you won’t change the resolution).

logger - Optional slog::Logger to be used by this device.

Return

Returns an error if the file is no valid drm node or the device is not accessible.

pub fn is_atomic(&self) -> bool

Returns if the underlying implementation uses atomic-modesetting or not.

pub fn crtcs(&self) -> &[Handle]

Returns a list of crtcs for this device

pub fn planes(&self, crtc: &Handle) -> Result<Planes, Error>

Returns a set of available planes for a given crtc

pub fn create_surface(
&self,
crtc: Handle,
mode: Mode,
connectors: &[Handle]
) -> Result<DrmSurface<A>, Error>

Creates a new rendering surface.

Arguments

Initialization of surfaces happens through the types provided by drm-rs.

crtcs represent scanout engines of the device pointing to one framebuffer.
Their responsibility is to read the data of the framebuffer and export it into an “Encoder”.
The number of crtc’s represent the number of independent output devices the hardware may handle.

mode describes the resolution and rate of images produced by the crtc and
has to be compatible with the provided connectors.

connectors - List of connectors driven by the crtc. At least one(!) connector needs to be
attached to a crtc in smithay.

pub fn device_id(&self) -> dev_t

Returns the device_id of the underlying drm node