Struct smithay::backend::drm::DrmSurface

pub struct DrmSurface<A: AsRawFd + 'static> { /* fields omitted */ }

An open crtc + plane combination that can be used for scan-out

Implementations

impl<A: AsRawFd + 'static> DrmSurface<A>

pub fn crtc(&self) -> Handle

Returns the underlying crtc of this surface

pub fn plane(&self) -> Handle

Returns the underlying primary plane of this surface

pub fn current_connectors(&self) -> impl IntoIterator<Item = Handle>

Currently used connectors of this surface

pub fn pending_connectors(&self) -> impl IntoIterator<Item = Handle>

Returns the pending connectors used after the next commit of this surface

pub fn add_connector(&self, connector: Handle) -> Result<(), Error>

Tries to add a new connector to be used after the next commit.

Warning: You need to make sure, that the connector is not used with another surface or was properly removed via remove_connector + commit before adding it to another surface. Behavior if failing to do so is undefined, but might result in rendering errors or the connector getting removed from the other surface without updating it’s internal state.

Fails if the connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn remove_connector(&self, connector: Handle) -> Result<(), Error>

Tries to mark a connector for removal on the next commit.

pub fn set_connectors(&self, connectors: &[Handle]) -> Result<(), Error>

Tries to replace the current connector set with the newly provided one on the next commit.

Fails if one new connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn current_mode(&self) -> Mode

Returns the currently active Mode of the underlying crtc

pub fn pending_mode(&self) -> Mode

Returns the currently pending Mode to be used after the next commit.

pub fn use_mode(&self, mode: Mode) -> Result<(), Error>

Tries to set a new Mode to be used after the next commit.

Fails if the mode is not compatible with the underlying crtc or any of the pending connectors.

pub fn use_plane(
&self,
plane: Handle,
position: (i32, i32),
size: (u32, u32)
) -> Result<(), Error>

Tries to setup a cursor or overlay Plane to be set at the next commit/page_flip with the given position and size.

Planes can have arbitrary hardware constraints, that cannot be expressed in the api, like supporting only positions at even or odd values, allowing only certain sizes or disallowing overlapping planes. Using planes should therefor be done in a best-efford manner. Failures on page_flip or commit should be expected and alternative code paths without the usage of planes prepared.

Fails if tests for the given plane fail, if the underlying implementation does not support the use of planes or if the plane is not supported by this crtc.

pub fn clear_plane(&self, plane: Handle) -> Result<(), Error>

Disables the given plane.

Errors if the plane is not supported by this crtc or if the underlying implementation does not support the use of planes.

pub fn commit_pending(&self) -> bool

Returns true whenever any state changes are pending to be commited

The following functions may trigger a pending commit:

add_connector

remove_connector

use_mode

pub fn commit<'a>(
&self,
framebuffers: impl Iterator<Item = &'a (Handle, Handle)>,
event: bool
) -> Result<(), Error>

Commit the pending state rendering a given framebuffer.

Note: This will trigger a full modeset on the underlying device, potentially causing some flickering. Check before performing this operation if a commit really is necessary using commit_pending.

This operation is not necessarily blocking until the crtc is in the desired state, but will trigger a vblank event once done. Make sure to have the device registered in your event loop prior to invoking this, to not miss any generated event.

pub fn page_flip<'a>(
&self,
framebuffers: impl Iterator<Item = &'a (Handle, Handle)>,
event: bool
) -> Result<(), Error>

Page-flip the underlying crtc to a new given framebuffer.

This will not cause the crtc to modeset.

This operation is not blocking and will produce a vblank event once swapping is done. Make sure to have the device registered in your event loop to not miss the event.

pub fn supported_formats(&self, plane: Handle) -> Result<HashSet<Format>, Error>

Returns a set of supported pixel formats for attached buffers

pub fn planes(&self) -> Result<Planes, Error>

Returns a set of available planes for this surface

pub fn test_buffer(
&self,
fb: Handle,
mode: &Mode,
allow_screen_change: bool
) -> Result<bool, Error>

Tests is a framebuffer can be used with this surface.

Arguments

fb - Framebuffer handle that has an attached buffer, that shall be tested

mode - The mode that should be used to display the buffer

allow_screen_change - If an actual screen change is permitted to carry out this test. If the test cannot be performed otherwise, this function returns false.

pub fn test_plane_buffer(
&self,
fb: Handle,
plane: Handle,
position: (i32, i32),
size: (u32, u32)
) -> Result<bool, Error>

Tests is a framebuffer can be used with this surface and a given plane.

Arguments

fb - Framebuffer handle that has an attached buffer, that shall be tested

plane - The plane that should be used to display the buffer (only works for cursor and overlay planes - for primary planes use test_buffer)

position - The position of the plane

size - The size of the plane

If the test cannot be performed, this function returns false. This is always the case for non-atomic surfaces.

pub fn reset_state(&self) -> Result<(), Error>

Re-evaluates the current state of the crtc.

Usually you do not need to call this, but if the state of the crtc is modified elsewhere and you need to reset the initial state of this surface, you may call this function.