Struct smithay::backend::drm::GbmBufferedSurface

pub struct GbmBufferedSurface<A: Allocator<Buffer = BufferObject<()>> + 'static, U> { /* private fields */ }

Simplified abstraction of a swapchain for gbm-buffers displayed on a DrmSurface.

Implementations

impl<A, U> GbmBufferedSurface<A, U>where A: Allocator<Buffer = BufferObject<()>>, A::Error: Error + Send + Sync,

pub fn new( drm: DrmSurface, allocator: A, color_formats: &[Fourcc], renderer_formats: HashSet<Format> ) -> Result<GbmBufferedSurface<A, U>, Error<A::Error>>

Create a new GbmBufferedSurface from a given compatible combination of a surface, an allocator and renderer formats.

The provided color_formats are tested in order until a working configuration is found.

To successfully call this function, you need to have a renderer, which can render into a Dmabuf, and a gbm allocator that can produce buffers of a supported format for rendering.

pub fn next_buffer(&mut self) -> Result<(Dmabuf, u8), Error<A::Error>>

Retrieves the next buffer to be rendered into and it’s age.

Note: This function can be called multiple times and will return the same buffer until it is queued (see GbmBufferedSurface::queue_buffer).

pub fn queue_buffer( &mut self, damage: Option<Vec<Rectangle<i32, Physical>>>, user_data: U ) -> Result<(), Error<A::Error>>

Queues the current buffer for rendering.

Returns Error::NoBuffer in case GbmBufferedSurface::next_buffer has not been called prior to this function.

Note: This function needs to be followed up with GbmBufferedSurface::frame_submitted when a vblank event is received, that denotes successful scanout of the buffer. Otherwise the underlying swapchain will eventually run out of buffers.

user_data can be used to attach some data to a specific buffer and later retrieved with GbmBufferedSurface::frame_submitted

pub fn frame_submitted(&mut self) -> Result<Option<U>, Error<A::Error>>

Marks the current frame as submitted.

Note: Needs to be called, after the vblank event of the matching DrmDevice was received after calling GbmBufferedSurface::queue_buffer on this surface. Otherwise the underlying swapchain will run out of buffers eventually.

Returns the user data that was stored with GbmBufferedSurface::queue_buffer if a buffer was pending, otherwise None is returned.

pub fn reset_buffers(&mut self)

Reset the underlying buffers

pub fn crtc(&self) -> Handle

Returns the underlying crtc of this surface

pub fn plane(&self) -> Handle

Returns the underlying plane of this surface

pub fn current_connectors(&self) -> impl IntoIterator<Item = Handle>

Currently used connectors of this Surface

pub fn pending_connectors(&self) -> impl IntoIterator<Item = Handle>

Returns the pending connectors used for the next frame queued via queue_buffer.

pub fn add_connector(&self, connector: Handle) -> Result<(), Error<A::Error>>

Tries to add a new connector to be used after the next commit.

Warning: You need to make sure, that the connector is not used with another surface or was properly removed via remove_connector + commit before adding it to another surface. Behavior if failing to do so is undefined, but might result in rendering errors or the connector getting removed from the other surface without updating it’s internal state.

Fails if the connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn remove_connector(&self, connector: Handle) -> Result<(), Error<A::Error>>

Tries to mark a connector for removal on the next commit.

pub fn set_connectors( &self, connectors: &[Handle] ) -> Result<(), Error<A::Error>>

Tries to replace the current connector set with the newly provided one on the next commit.

Fails if one new connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

pub fn current_mode(&self) -> Mode

Returns the currently active Mode of the underlying crtc

pub fn pending_mode(&self) -> Mode

Returns the currently pending Mode to be used after the next commit.

pub fn use_mode(&mut self, mode: Mode) -> Result<(), Error<A::Error>>

Tries to set a new Mode to be used after the next commit.

Fails if the mode is not compatible with the underlying crtc or any of the pending connectors.

pub fn surface(&self) -> &DrmSurface

Returns a reference to the underlying drm surface

pub fn format(&self) -> Fourcc

Get the format of the underlying swapchain