Trait smithay::backend::drm::DrmHandler

pub trait DrmHandler<B: Borrow<DrmBackend> + 'static> {
    fn ready(
        &mut self, 
        evlh: &mut EventLoopHandle, 
        device: &mut DrmDevice<B>, 
        backend: &StateToken<B>, 
        frame: u32, 
        duration: Duration
    );

    fn error(
        &mut self, 
        evlh: &mut EventLoopHandle, 
        device: &mut DrmDevice<B>, 
        error: IoError
    );
}

Handler for drm node events

See module-level documentation for its use

Required Methods

fn ready(
&mut self,
evlh: &mut EventLoopHandle,
device: &mut DrmDevice<B>,
backend: &StateToken<B>,
frame: u32,
duration: Duration
)

A DrmBackend has finished swapping buffers and new frame can now (and should be immediately) be rendered.

The id argument is the Id of the DrmBackend that finished rendering, check using DrmBackend::is.

fn error(
&mut self,
evlh: &mut EventLoopHandle,
device: &mut DrmDevice<B>,
error: IoError
)

The DrmDevice has thrown an error.

The related backends are most likely not usable anymore and the whole stack has to be recreated.