Trait smithay::backend::drm::Surface

pub trait Surface {
    type Connectors: IntoIterator<Item = Handle>;
    type Error: Error + Send + 'static;
    fn crtc(&self) -> Handle;

    fn current_connectors(&self) -> Self::Connectors;

    fn pending_connectors(&self) -> Self::Connectors;

    fn add_connector(&self, connector: Handle) -> Result<(), Self::Error>;

    fn remove_connector(&self, connector: Handle) -> Result<(), Self::Error>;

    fn set_connectors(&self, connectors: &[Handle]) -> Result<(), Self::Error>;

    fn current_mode(&self) -> Mode;

    fn pending_mode(&self) -> Mode;

    fn use_mode(&self, mode: Mode) -> Result<(), Self::Error>;
}

An open crtc that can be used for rendering

Associated Types

type Connectors: IntoIterator<Item = Handle>

Type repesenting a collection of connectors returned by current_connectors and pending_connectors

type Error: Error + Send + 'static

Error type returned by methods of this trait

Loading content...
Required methods

fn crtc(&self) -> Handle

Returns the underlying crtc of this surface

fn current_connectors(&self) -> Self::Connectors

Currently used connectors of this Surface

fn pending_connectors(&self) -> Self::Connectors

Returns the pending connectors used after the next commit of this [Surface]

Note: Only on a [RawSurface] you may directly trigger a commit. Other Surfaces provide their own methods that may trigger a commit, you will need to read their docs.

fn add_connector(&self, connector: Handle) -> Result<(), Self::Error>

Tries to add a new connector to be used after the next commit.

Warning: You need to make sure, that the connector is not used with another surface or was properly removed via remove_connector + commit before adding it to another surface. Behavior if failing to do so is undefined, but might result in rendering errors or the connector getting removed from the other surface without updating it's internal state.

Fails if the connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

fn remove_connector(&self, connector: Handle) -> Result<(), Self::Error>

Tries to mark a connector for removal on the next commit.

fn set_connectors(&self, connectors: &[Handle]) -> Result<(), Self::Error>

Tries to replace the current connector set with the newly provided one on the next commit.

Fails if one new connector is not compatible with the underlying crtc (e.g. no suitable encoder may be found) or is not compatible with the currently pending Mode.

fn current_mode(&self) -> Mode

Returns the currently active Mode of the underlying crtc

fn pending_mode(&self) -> Mode

Returns the currently pending Mode to be used after the next commit.

fn use_mode(&self, mode: Mode) -> Result<(), Self::Error>

Tries to set a new Mode to be used after the next commit.

Fails if the mode is not compatible with the underlying crtc or any of the pending connectors.

Note: Only on a [RawSurface] you may directly trigger a commit. Other [Surface]s provide their own methods that may trigger a commit, you will need to read their docs.

Loading content...