Module smithay::backend::egl

Common traits and types for egl rendering

Large parts of this module are taken from glutin src/api/egl

It therefore falls under glutin's Apache 2.0 license

Wayland specific EGL functionality - EGL based WlBuffers.

The types of this module can be used to initialize hardware acceleration rendering based on EGL for clients as it may enabled usage of EGLImage based WlBuffers.

To use it bind any backend implementing the EGLGraphicsBackend trait, that shall do the rendering (so pick a fast one), to the [wayland_server::Display] of your compositor. Note only one backend may be bound to any Display at any time.

You may then use the resulting EGLDisplay to receive EGLImages of an EGL-based WlBuffer for rendering.

Re-exports

pub use self::context::EGLContext;

pub use self::surface::EGLSurface;

Modules

context
EGL context related structs

display
Type safe native types for safe egl initialisation

ffi

native
Type safe native types for safe context/surface creation

surface
EGL surface related structs

Structs

EGLImages
Images of the EGL-based [WlBuffer].

EglExtensionNotSupportedError
Error that can happen on optional EGL features

Enums

BufferAccessError
Error that can occur when accessing an EGL buffer

Error
EGL errors

Format
Texture format types

TextureCreationError
Error that might happen when binding an EGLImage to a GL texture

Traits

EGLGraphicsBackend
Trait any backend type may implement that allows binding a Display to create an EGLBufferReader for EGL-based [WlBuffer]s.

Functions

get_proc_address
Returns the address of an OpenGL function.