Struct smithay::backend::egl::context::EGLContext

pub struct EGLContext { /* private fields */ }

EGL context for rendering

Implementations

impl EGLContext

pub unsafe fn from_raw<L>(
display: *const c_void,
config_id: *const c_void,
context: *const c_void,
log: L
) -> Result<EGLContext, Error> where
L: Into<Option<Logger>>,

Creates a new EGLContext from raw handles.

Safety

The context must be created from the system default EGL library (dlopen("libEGL.so"))

The display, config, and context must be valid for the lifetime of the returned context.

pub fn new<L>(display: &EGLDisplay, log: L) -> Result<EGLContext, Error> where
L: Into<Option<Logger>>,

Creates a new configless EGLContext from a given EGLDisplay

pub fn new_with_config<L>(
display: &EGLDisplay,
attributes: GlAttributes,
reqs: PixelFormatRequirements,
log: L
) -> Result<EGLContext, Error> where
L: Into<Option<Logger>>,

Create a new EGLContext from a given EGLDisplay and configuration requirements

pub fn new_shared<L>(
display: &EGLDisplay,
share: &EGLContext,
log: L
) -> Result<EGLContext, Error> where
L: Into<Option<Logger>>,

Create a new configless EGLContext from a given EGLDisplay sharing resources with another context

pub fn new_shared_with_config<L>(
display: &EGLDisplay,
share: &EGLContext,
attributes: GlAttributes,
reqs: PixelFormatRequirements,
log: L
) -> Result<EGLContext, Error> where
L: Into<Option<Logger>>,

Create a new EGLContext from a given EGLDisplay and configuration requirements sharing resources with another context

pub unsafe fn make_current_with_surface(
&self,
surface: &EGLSurface
) -> Result<(), MakeCurrentError>

Makes the OpenGL context the current context in the current thread with a surface to read/write to.

Safety

This function is marked unsafe, because the context cannot be made current on another thread without being unbound again (see EGLContext::unbind).

pub unsafe fn make_current(&self) -> Result<(), MakeCurrentError>

Makes the OpenGL context the current context in the current thread with no surface bound.

Safety

This function is marked unsafe, because the context cannot be made current on another thread without being unbound again (see EGLContext::unbind).

pub fn is_current(&self) -> bool

Returns true if the OpenGL context is the current one in the thread.

pub fn config_id(&self) -> EGLConfig

Returns the egl config for this context

pub fn pixel_format(&self) -> Option<PixelFormat>

Returns the pixel format of the main framebuffer of the context.

pub fn unbind(&self) -> Result<(), MakeCurrentError>

Unbinds this context from the current thread, if set.

This does nothing if this context is not the current context.

pub fn display(&self) -> &EGLDisplay

Returns the display which created this context.

pub fn dmabuf_render_formats(&self) -> &HashSet<DrmFormat>

Returns a list of formats for dmabufs that can be rendered to.

pub fn dmabuf_texture_formats(&self) -> &HashSet<DrmFormat>

Returns a list of formats for dmabufs that can be used as textures.

pub fn user_data(&self) -> &UserDataMap

Retrieve user_data associated with this context

Note: UserData is shared between shared context, if constructed with new_shared or new_shared_with_config.

pub fn get_context_handle(&self) -> EGLContext

Get a raw handle to the underlying context.

The pointer will become invalid, when this struct is destroyed.