smithay::backend::egl::display

Struct EGLDisplay

pub struct EGLDisplay { /* private fields */ }

Available on crate feature backend_egl only.

EGLDisplay represents an initialised EGL environment

Implementations

impl EGLDisplay

pub unsafe fn new<N>(native: N) -> Result<EGLDisplay, Error>
where N: EGLNativeDisplay + 'static,

Create a new EGLDisplay from a given EGLNativeDisplay.

Safety

smithay internally tracks EGLDisplay instances as calls with the same parameters to eglGetPlatformDisplay will return references to the same underlying display.

You donโt have to worry about this, when using just smithay, but if other code calls eglGetPlatformDisplay it is possible to close the resulting EGLDisplay via eglTerminate without invalidating smithayโs instances.

If you are using other code creating EGLDisplays, donโt use this method, but use the external code to create the display and use EGLDisplay::from_raw to make it usable in smithay. This way eglTerminate will be skipped and you can clean up the display externally.

pub unsafe fn from_raw( display: *const c_void, config_id: *const c_void, ) -> Result<EGLDisplay, Error>

Create a new EGLDisplay from an already initialized EGLDisplay and EGLConfig handle

Safety

The display must be created from the system default EGL library (dlopen("libEGL.so"))

The display and config must be valid for the lifetime of the returned display and any handles created by this display (using EGLDisplay::get_display_handle)

smithay canโt track the parameters used to create this display, which may cause two Displays to point to the same underlying instance. For this reason displays created using this method will not call eglTerminate on destruction. You will have to cleanup manually.

pub fn choose_config( &self, attributes: GlAttributes, reqs: PixelFormatRequirements, ) -> Result<(PixelFormat, EGLConfig), Error>

Finds a compatible EGLConfig for a given set of requirements

pub fn get_display_handle(&self) -> Arc<EGLDisplayHandle>

Get a handle to the underlying raw EGLDisplay handle

pub fn get_egl_version(&self) -> (i32, i32)

Returns the runtime egl version of this display

pub fn extensions(&self) -> &[String]

Returns the supported extensions of this display

pub fn dmabuf_render_formats(&self) -> &FormatSet

Returns a list of formats for dmabufs that can be rendered to.

pub fn dmabuf_texture_formats(&self) -> &FormatSet

Returns a list of formats for dmabufs that can be used as textures.

pub fn supports_damage(&self) -> bool

Returns when the display supports extensions required for smithays damage tracking helpers.

pub fn create_dmabuf_from_image( &self, image: EGLImage, size: Size<i32, BufferCoords>, y_inverted: bool, ) -> Result<Dmabuf, Error>

Exports an EGLImage as a Dmabuf

pub fn create_image_from_dmabuf( &self, dmabuf: &Dmabuf, ) -> Result<EGLImage, Error>

Imports a Dmabuf as an EGLImage

pub fn bind_wl_display( &self, display: &DisplayHandle, ) -> Result<EGLBufferReader, Error>

Available on crate features use_system_lib and wayland_frontend only.

Binds this EGL display to the given Wayland display.

This will allow clients to utilize EGL to create hardware-accelerated surfaces. The server will need to be able to handle EGL-WlBuffers.

Errors

This might return EglExtensionNotSupported if binding is not supported by the EGL implementation.

This might return OtherEGLDisplayAlreadyBound if called for the same DisplayHandle multiple times, as only one egl display may be bound at any given time.