Struct smithay::backend::egl::display::EGLDisplay

pub struct EGLDisplay<B: Backend, N: NativeDisplay<B>> { /* fields omitted */ }

[EGLDisplay] represents an initialised EGL environment

Methods

impl<B: Backend, N: NativeDisplay<B>> EGLDisplay<B, N>

pub fn new<L>(native: N, logger: L) -> Result<EGLDisplay<B, N>, Error> where
L: Into<Option<Logger>>,

Create a new [EGLDisplay] from a given NativeDisplay

pub fn choose_config(
&self,
attributes: GlAttributes,
reqs: PixelFormatRequirements
) -> Result<(PixelFormat, EGLConfig), Error>

Finds a compatible [EGLConfig] for a given set of requirements

pub fn create_context(
&self,
attributes: GlAttributes,
reqs: PixelFormatRequirements
) -> Result<EGLContext, Error>

Create a new EGLContext

pub fn create_surface(
&self,
pixel_format: PixelFormat,
double_buffer: Option<bool>,
config: EGLConfig,
args: N::Arguments
) -> Result<EGLSurface<B::Surface>, SurfaceCreationError<B::Error>>

Creates a surface for rendering

pub fn get_egl_version(&self) -> (i32, i32)

Returns the runtime egl version of this display

pub fn get_extensions(&self) -> Vec<String>

Returns the supported extensions of this display

pub fn borrow(&self) -> Ref<N>

Borrow the underlying native display.

This follows the same semantics as std::cell:RefCell. Multiple read-only borrows are possible. Borrowing the backend while there is a mutable reference will panic.

pub fn borrow_mut(&self) -> RefMut<N>

Borrow the underlying native display mutably.

This follows the same semantics as std::cell:RefCell. Holding any other borrow while trying to borrow the backend mutably will panic. Note that EGL will borrow the display mutably during surface creation.