Struct smithay::backend::egl::display::EGLDisplay

pub struct EGLDisplay { /* fields omitted */ }

EGLDisplay represents an initialised EGL environment

Implementations

impl EGLDisplay

pub fn new<N, L>(native: &N, logger: L) -> Result<EGLDisplay, Error> where
N: EGLNativeDisplay + 'static,
L: Into<Option<Logger>>,

Create a new EGLDisplay from a given EGLNativeDisplay

pub fn choose_config(
&self,
attributes: GlAttributes,
reqs: PixelFormatRequirements
) -> Result<(PixelFormat, EGLConfig), Error>

Finds a compatible EGLConfig for a given set of requirements

pub fn get_display_handle(&self) -> Arc<EGLDisplayHandle>

Get a handle to the underlying raw EGLDisplay handle

pub fn get_egl_version(&self) -> (i32, i32)

Returns the runtime egl version of this display

pub fn get_extensions(&self) -> Vec<String>

Returns the supported extensions of this display

pub fn create_image_from_dmabuf(
&self,
dmabuf: &Dmabuf
) -> Result<EGLImage, Error>

Imports a Dmabuf as an EGLImage

pub fn bind_wl_display(
&self,
display: &Display
) -> Result<EGLBufferReader, Error>

Binds this EGL display to the given Wayland display.

This will allow clients to utilize EGL to create hardware-accelerated surfaces. The server will need to be able to handle EGL-WlBuffers.

Errors

This might return EglExtensionNotSupported if binding is not supported by the EGL implementation.

This might return OtherEGLDisplayAlreadyBound if called for the same Display multiple times, as only one egl display may be bound at any given time.