Trait smithay::backend::egl::native::EGLNativeDisplay

pub trait EGLNativeDisplay: Send {
    // Required method
    fn supported_platforms(&self) -> Vec<EGLPlatform<'_>>;

    // Provided methods
    fn surface_type(&self) -> EGLint { ... }
    fn identifier(&self) -> Option<String> { ... }
}

Trait describing platform specific functionality to create a valid EGLDisplay using the EGL_EXT_platform_base extension.

Required Methods

fn supported_platforms(&self) -> Vec<EGLPlatform<'_>>

List of supported platforms that can be used to create a display using eglGetPlatformDisplayEXT

Provided Methods

fn surface_type(&self) -> EGLint

Type of surfaces created

fn identifier(&self) -> Option<String>

String identifying this native display from its counterparts of the same platform, if applicable.