Trait smithay::backend::egl::native::EGLNativeDisplay

pub trait EGLNativeDisplay: Send {
    fn supported_platforms(&self) -> Vec<EGLPlatform<'_>>;

    fn surface_type(&self) -> EGLint { ... }
}

Trait describing platform specific functionality to create a valid EGLDisplay using the EGL_EXT_platform_base extension.

Required methods

fn supported_platforms(&self) -> Vec<EGLPlatform<'_>>

List of supported platforms that can be used to create a display using eglGetPlatformDisplayEXT

Provided methods

fn surface_type(&self) -> EGLint

Type of surfaces created