Trait smithay::backend::egl::native::NativeSurface

pub unsafe trait NativeSurface {
    type Error: Error + Send + 'static;
    unsafe fn create(
        &self, 
        display: &Arc<EGLDisplayHandle>, 
        config_id: EGLConfig, 
        surface_attributes: &[c_int]
    ) -> Result<*const c_void, SurfaceCreationError<Self::Error>>;

    fn needs_recreation(&self) -> bool { ... }

    fn swap_buffers(
        &self, 
        display: &Arc<EGLDisplayHandle>, 
        surface: EGLSurface
    ) -> Result<(), SwapBuffersError<Self::Error>> { ... }
}

Trait for types returning valid surface pointers for initializing egl

Unsafety

The returned NativeWindowType must be valid for EGL and there is no way to test that.

Associated Types

type Error: Error + Send + 'static

Error type thrown by the surface creation in case of failure.

Loading content...
Required methods

unsafe fn create(
&self,
display: &Arc<EGLDisplayHandle>,
config_id: EGLConfig,
surface_attributes: &[c_int]
) -> Result<*const c_void, SurfaceCreationError<Self::Error>>

Create an EGLSurface from the internal native type.

Must be able to deal with re-creation of existing resources, if needs_recreation can return true.

Safety

This is usually an unsafe operation returning a raw pointer.

Loading content...
Provided methods

fn needs_recreation(&self) -> bool

Will be called to check if any internal resources will need to be recreated. Old resources must be used until create was called again and a new surface was optained.

Only needs to be recreated, if this may return true. The default implementation always returns false.

fn swap_buffers(
&self,
display: &Arc<EGLDisplayHandle>,
surface: EGLSurface
) -> Result<(), SwapBuffersError<Self::Error>>

Adds additional semantics when calling EGLSurface::swap_buffers

Only implement if required by the backend.

Loading content...