Trait smithay::backend::graphics::egl::native::NativeSurface

pub unsafe trait NativeSurface {
    fn ptr(&self) -> NativeWindowType;
}

Trait for types returning valid surface pointers for initializing egl

Unsafety

The returned NativeWindowType must be valid for egl and there is no way to test that.

Required Methods

fn ptr(&self) -> NativeWindowType

Return a raw pointer egl will accept for surface creation.