Struct smithay::backend::graphics::glium::GliumGraphicsBackend

pub struct GliumGraphicsBackend<T: EGLGraphicsBackend> { /* fields omitted */ }

Wrapper to expose glium compatibility

Methods

impl<T: EGLGraphicsBackend + 'static> GliumGraphicsBackend<T>

pub fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.