Module smithay::backend::input

Common traits for input backends to receive input from.

Structs

Seat
A seat describes a group of input devices and at least one graphics device belonging together.

SeatCapabilities
Describes capabilities a Seat has.

TouchSlot
Slot of a different touch event.

Enums

Axis
Axis when scrolling

AxisSource
Source of an axis when scrolling

KeyState
State of key on a keyboard. Either pressed or released

MouseButton
A particular mouse button

MouseButtonState
State of a button on a mouse. Either pressed or released

TouchEvent
Touch event

Traits

InputBackend
Trait that describes objects providing a source of input events. All input backends need to implemenent this and provide the same base gurantees about the presicion of given events.

InputHandler
Implement to receive input events from any InputBackend.