Module smithay::backend::input

Common traits for input backends to receive input from.

Structs

TabletToolCapabilities

Describes extra capabilities on a tablet.

TabletToolDescriptor

Description of physical tablet tool

TouchSlot

Slot of a different touch event.

Enums

Axis

Axis when scrolling

AxisSource

Source of an axis when scrolling

ButtonState

State of a button on a pointer device, like mouse or tablet tool. Either pressed or released

DeviceCapability

Set of input types a device may provide

InputEvent

Different events that can be generated by an input backend

KeyState

State of key on a keyboard. Either pressed or released

MouseButton

A particular mouse button

ProximityState

The state of proximity for a tool on a device.

TabletToolTipState

The tip contact state for a tool on a device.

TabletToolType

Describes the physical type of tool. The physical type of tool generally defines its base usage.

UnusedEvent

Used to mark events never emitted by an InputBackend implementation.

Traits

AbsolutePositionEvent

Input event with absolute location data.

Device

Trait for generic functions every input device does provide

Event

Trait for generic functions every input event does provide

GestureBeginEvent

Trait for gesture begin events.

GestureEndEvent

Trait for gesture end events.

GestureHoldBeginEvent

Trait for gesture hold begin event.

GestureHoldEndEvent

Trait for gesture hold end event.

GesturePinchBeginEvent

Trait for gesture pinch begin event.

GesturePinchEndEvent

Trait for gesture pinch end event.

GesturePinchUpdateEvent

Trait for gesture pinch update event.

GestureSwipeBeginEvent

Trait for gesture swipe begin event.

GestureSwipeEndEvent

Trait for gesture swipe end event.

GestureSwipeUpdateEvent

Trait for gesture swipe update event.

InputBackend

Trait that describes objects providing a source of input events. All input backends need to implement this and provide the same base guarantees about the precision of given events.

KeyboardKeyEvent

Trait for keyboard event

PointerAxisEvent

Trait for pointer events generated by scrolling on an axis.

PointerButtonEvent

Common methods pointer event generated by pressed buttons do implement

PointerMotionAbsoluteEvent

Trait for pointer events generated by absolute device positioning.

PointerMotionEvent

Trait for pointer events generated by relative device movement.

TabletToolAxisEvent

Trait for axis tablet tool events.

TabletToolButtonEvent

Signals that a tool has changed a logical button state on a device with the DeviceCapability::TabletTool capability.

TabletToolEvent

Tablet tool event

TabletToolProximityEvent

Trait for tablet tool proximity events.

TabletToolTipEvent

Signals that a tool has come in contact with the surface of a device with the DeviceCapability::TabletTool capability.

TouchCancelEvent

Trait for touch events canceling the chain

TouchDownEvent

Trait for touch events starting at a given position.

TouchEvent

Trait with functions available for all touch events.

TouchFrameEvent

Trait for touch frame events

TouchMotionEvent

Trait for touch events regarding movement on the screen

TouchUpEvent

Trait for touch events finishing.