Module smithay::backend::input

Common traits for input backends to receive input from.

Structs

Seat
A seat describes a group of input devices and at least one graphics device belonging together.

SeatCapabilities
Describes capabilities a Seat has.

TouchSlot
Slot of a different touch event.

Enums

Axis
Axis when scrolling

AxisSource
Source of an axis when scrolling

KeyState
State of key on a keyboard. Either pressed or released

MouseButton
A particular mouse button

MouseButtonState
State of a button on a mouse. Either pressed or released

UnusedEvent
Used to mark events never emitted by an InputBackend implementation.

Traits

Event
Trait for generic functions every input event does provide/

InputBackend
Trait that describes objects providing a source of input events. All input backends need to implemenent this and provide the same base gurantees about the presicion of given events.

InputHandler
Implement to receive input events from any InputBackend.

KeyboardKeyEvent
Trait for keyboard event

PointerAxisEvent
Trait for pointer events generated by scrolling on an axis.

PointerButtonEvent
Common methods pointer event generated by pressed buttons do implement

PointerMotionAbsoluteEvent
Trait for pointer events generated by absolute device positioning.

PointerMotionEvent
Trait for pointer events generated by relative device movement.

TouchCancelEvent
Trait for touch events cancelling the chain

TouchDownEvent
Trait for touch events starting at a given position.

TouchFrameEvent
Trait for touch frame events

TouchMotionEvent
Trait for touch events regarding movement on the screen

TouchUpEvent
Trait for touch events finishing.