Enum smithay::backend::input::InputEvent

pub enum InputEvent<B: InputBackend> {
    NewSeat(Seat),
    SeatChanged(Seat),
    SeatRemoved(Seat),
    Keyboard {
        seat: Seat,
        event: B::KeyboardKeyEvent,
    },
    PointerMotion {
        seat: Seat,
        event: B::PointerMotionEvent,
    },
    PointerMotionAbsolute {
        seat: Seat,
        event: B::PointerMotionAbsoluteEvent,
    },
    PointerButton {
        seat: Seat,
        event: B::PointerButtonEvent,
    },
    PointerAxis {
        seat: Seat,
        event: B::PointerAxisEvent,
    },
    TouchDown {
        seat: Seat,
        event: B::TouchDownEvent,
    },
    TouchMotion {
        seat: Seat,
        event: B::TouchMotionEvent,
    },
    TouchUp {
        seat: Seat,
        event: B::TouchUpEvent,
    },
    TouchCancel {
        seat: Seat,
        event: B::TouchCancelEvent,
    },
    TouchFrame {
        seat: Seat,
        event: B::TouchFrameEvent,
    },
    Special(B::SpecialEvent),
}

Different events that can be generated by an input backend

Variants

NewSeat(Seat)

A new seat has been created

SeatChanged(Seat)

A seat has changed

SeatRemoved(Seat)

A seat has been removed

Keyboard

A keyboard event occured

Fields of Keyboard

seat: Seat

Seat that generated the event

event: B::KeyboardKeyEvent

The keyboard event

PointerMotion

A relative pointer motion occured

Fields of PointerMotion

seat: Seat

Seat that generated the event

event: B::PointerMotionEvent

The pointer motion event

PointerMotionAbsolute

An absolute pointer motion occures

Fields of PointerMotionAbsolute

seat: Seat

Seat that generated the event

event: B::PointerMotionAbsoluteEvent

The absolute pointer motion event

PointerButton

A pointer button was pressed or released

Fields of PointerButton

seat: Seat

Seat that generated the event

event: B::PointerButtonEvent

The pointer button event

PointerAxis

A pointer axis was actionned

Fields of PointerAxis

seat: Seat

Seat that generated the event

event: B::PointerAxisEvent

The pointer axis event

TouchDown

A new touchpoint appeared

Fields of TouchDown

seat: Seat

Seat that generated the event

event: B::TouchDownEvent

The touch down event

TouchMotion

A touchpoint moved

Fields of TouchMotion

seat: Seat

Seat that generated the event

event: B::TouchMotionEvent

The touch motion event

TouchUp

A touchpoint was removed

Fields of TouchUp

seat: Seat

Seat that generated the event

event: B::TouchUpEvent

The touch up event

TouchCancel

A touch sequence was cancelled

Fields of TouchCancel

seat: Seat

Seat that generated the event

event: B::TouchCancelEvent

The touch cancel event

TouchFrame

A touch frame was emmited

A set of two events received on the same seat between two frames should be interpreted as an atomic event.

Fields of TouchFrame

seat: Seat

Seat that generated the event

event: B::TouchFrameEvent

The touch frame event

Special(B::SpecialEvent)

Special event specific of this backend