Enum smithay::backend::input::InputEvent

pub enum InputEvent<B: InputBackend> {

    DeviceAdded {
        device: B::Device,
    },
    DeviceRemoved {
        device: B::Device,
    },
    Keyboard {
        event: B::KeyboardKeyEvent,
    },
    PointerMotion {
        event: B::PointerMotionEvent,
    },
    PointerMotionAbsolute {
        event: B::PointerMotionAbsoluteEvent,
    },
    PointerButton {
        event: B::PointerButtonEvent,
    },
    PointerAxis {
        event: B::PointerAxisEvent,
    },
    TouchDown {
        event: B::TouchDownEvent,
    },
    TouchMotion {
        event: B::TouchMotionEvent,
    },
    TouchUp {
        event: B::TouchUpEvent,
    },
    TouchCancel {
        event: B::TouchCancelEvent,
    },
    TouchFrame {
        event: B::TouchFrameEvent,
    },
    TabletToolAxis {
        event: B::TabletToolAxisEvent,
    },
    TabletToolProximity {
        event: B::TabletToolProximityEvent,
    },
    TabletToolTip {
        event: B::TabletToolTipEvent,
    },
    TabletToolButton {
        event: B::TabletToolButtonEvent,
    },
    Special(B::SpecialEvent),

}

Different events that can be generated by an input backend

Variants

DeviceAdded

An input device was connected

Fields of DeviceAdded

device: B::Device

The added device

DeviceRemoved

An input device was disconnected

Fields of DeviceRemoved

device: B::Device

The removed device

Keyboard

A keyboard event occurred

Fields of Keyboard

event: B::KeyboardKeyEvent

The keyboard event

PointerMotion

A relative pointer motion occurred

Fields of PointerMotion

event: B::PointerMotionEvent

The pointer motion event

PointerMotionAbsolute

An absolute pointer motion occurs

Fields of PointerMotionAbsolute

event: B::PointerMotionAbsoluteEvent

The absolute pointer motion event

PointerButton

A pointer button was pressed or released

Fields of PointerButton

event: B::PointerButtonEvent

The pointer button event

PointerAxis

A pointer action occurred while scrolling on an axis

Fields of PointerAxis

event: B::PointerAxisEvent

The pointer axis event

TouchDown

A new touchpoint appeared

Fields of TouchDown

event: B::TouchDownEvent

The touch down event

TouchMotion

A touchpoint moved

Fields of TouchMotion

event: B::TouchMotionEvent

The touch motion event

TouchUp

A touchpoint was removed

Fields of TouchUp

event: B::TouchUpEvent

The touch up event

TouchCancel

A touch sequence was cancelled

Fields of TouchCancel

event: B::TouchCancelEvent

The touch cancel event

TouchFrame

A touch frame was emitted

A set of two events received on the same seat between two frames should be interpreted as an atomic event.

Fields of TouchFrame

event: B::TouchFrameEvent

The touch frame event

TabletToolAxis

A tablet tool axis was emitted

Fields of TabletToolAxis

event: B::TabletToolAxisEvent

The tablet tool axis event

TabletToolProximity

A tablet tool proximity was emitted

Fields of TabletToolProximity

event: B::TabletToolProximityEvent

The tablet tool proximity event

TabletToolTip

A tablet tool tip event was emitted

Fields of TabletToolTip

event: B::TabletToolTipEvent

The tablet tool axis event

TabletToolButton

A tablet tool button was pressed or released

Fields of TabletToolButton

event: B::TabletToolButtonEvent

The pointer button event

Special(B::SpecialEvent)

Special event specific of this backend

Tuple Fields of Special

0: B::SpecialEvent