Enum smithay::backend::input::UnusedEvent

pub enum UnusedEvent {}

Used to mark events never emitted by an InputBackend implementation.

Implements all event types and can be used in place for any Event type, that is not used by an InputBackend implementation. Initialization is not possible, making accidential use impossible and enabling a lot of possible compiler optimizations.