smithay::backend::input

Trait AbsolutePositionEvent

pub trait AbsolutePositionEvent<B: InputBackend>: Event<B> {
    // Required methods
    fn x(&self) -> f64;
    fn y(&self) -> f64;
    fn x_transformed(&self, width: i32) -> f64;
    fn y_transformed(&self, height: i32) -> f64;

    // Provided methods
    fn position(&self) -> Point<f64, Raw> { ... }
    fn position_transformed(
        &self,
        coordinate_space: Size<i32, Logical>,
    ) -> Point<f64, Logical> { ... }
}

Input event with absolute location data.

Required Methods

fn x(&self) -> f64

Device x position in it’s original coordinate space.

The format is defined by the backend implementation.

fn y(&self) -> f64

Device y position in it’s original coordinate space.

The format is defined by the backend implementation.

fn x_transformed(&self, width: i32) -> f64

Device x position converted to the targets coordinate space’s width. E.g. the focused output’s width.

fn y_transformed(&self, height: i32) -> f64

Device y position converted to the targets coordinate space’s height. E.g. the focused output’s height.

Provided Methods

fn position(&self) -> Point<f64, Raw>

Device position in it’s original coordinate space.

The format is defined by the backend implementation.

fn position_transformed( &self, coordinate_space: Size<i32, Logical>, ) -> Point<f64, Logical>

Device position converted to the targets coordinate space. E.g. the focused output’s resolution.