Trait smithay::backend::input::Device

pub trait Device: PartialEq + Eq + Hash {
    fn id(&self) -> String;

    fn name(&self) -> String;

    fn has_capability(&self, capability: DeviceCapability) -> bool;
}

Trait for generic functions every input device does provide

Required methods

fn id(&self) -> String

Unique id of a single device at a point in time.

Note: This means ids may be re-used by the backend for later devices.

fn name(&self) -> String

Human-readable name of the device

fn has_capability(&self, capability: DeviceCapability) -> bool

Test if this device has a specific capability