Trait smithay::backend::input::Event

pub trait Event<B: InputBackend> {
    // Required methods
    fn time(&self) -> u64;
    fn device(&self) -> B::Device;

    // Provided method
    fn time_msec(&self) -> u32 { ... }
}

Trait for generic functions every input event does provide

Required Methods

fn time(&self) -> u64

Timestamp in microseconds, with an undefined base.

Libinput does not guarantee that timestamps always increase monotonically.

fn device(&self) -> B::Device

Returns the device, that generated this event

Provided Methods

fn time_msec(&self) -> u32

Timestamp in milliseconds