Trait smithay::backend::input::Event

pub trait Event {
    fn time(&self) -> u32;
}

Trait for generic functions every input event does provide

Required methods

fn time(&self) -> u32

Returns an upward counting variable useful for event ordering.

Makes no guarantees about actual time passed between events.