Trait smithay::backend::input::InputBackend

pub trait InputBackend: Sized {
    type InputConfig;
    type EventError: Error;
    fn set_handler<H: InputHandler<Self> + 'static>(&mut self, handler: H);
    fn get_handler(&mut self) -> Option<&mut InputHandler<Self>>;
    fn clear_handler(&mut self);
    fn input_config(&mut self) -> &mut Self::InputConfig;
    fn dispatch_new_events(&mut self) -> Result<(), Self::EventError>;
    fn set_cursor_position(&mut self, x: u32, y: u32) -> Result<(), ()>;
}

Trait that describes objects providing a source of input events. All input backends need to implemenent this and provide the same base gurantees about the presicion of given events.

Associated Types

type InputConfig

Type of input device associated with the backend

type EventError: Error

Type representing errors that may be returned when processing events

Required Methods

fn set_handler<H: InputHandler<Self> + 'static>(&mut self, handler: H)

Sets a new handler for this InputBackend

fn get_handler(&mut self) -> Option<&mut InputHandler<Self>>

Get a reference to the currently set handler, if any

fn clear_handler(&mut self)

Clears the currently handler, if one is set

fn input_config(&mut self) -> &mut Self::InputConfig

Get current InputConfig

fn dispatch_new_events(&mut self) -> Result<(), Self::EventError>

Processes new events of the underlying backend and drives the InputHandler.

fn set_cursor_position(&mut self, x: u32, y: u32) -> Result<(), ()>

Sets the cursor position, useful for e.g. pointer wrapping.

Not guaranteed to be supported on every backend. The result usually depends on the capability of the backend, but may also fail for certain specific actions. See the backends documentation.