smithay::backend::input

Trait InputBackend

pub trait InputBackend: Sized {

    type Device: Device;
    type KeyboardKeyEvent: KeyboardKeyEvent<Self>;
    type PointerAxisEvent: PointerAxisEvent<Self>;
    type PointerButtonEvent: PointerButtonEvent<Self>;
    type PointerMotionEvent: PointerMotionEvent<Self>;
    type PointerMotionAbsoluteEvent: PointerMotionAbsoluteEvent<Self>;
    type GestureSwipeBeginEvent: GestureSwipeBeginEvent<Self>;
    type GestureSwipeUpdateEvent: GestureSwipeUpdateEvent<Self>;
    type GestureSwipeEndEvent: GestureSwipeEndEvent<Self>;
    type GesturePinchBeginEvent: GesturePinchBeginEvent<Self>;
    type GesturePinchUpdateEvent: GesturePinchUpdateEvent<Self>;
    type GesturePinchEndEvent: GesturePinchEndEvent<Self>;
    type GestureHoldBeginEvent: GestureHoldBeginEvent<Self>;
    type GestureHoldEndEvent: GestureHoldEndEvent<Self>;
    type TouchDownEvent: TouchDownEvent<Self>;
    type TouchUpEvent: TouchUpEvent<Self>;
    type TouchMotionEvent: TouchMotionEvent<Self>;
    type TouchCancelEvent: TouchCancelEvent<Self>;
    type TouchFrameEvent: TouchFrameEvent<Self>;
    type TabletToolAxisEvent: TabletToolAxisEvent<Self>;
    type TabletToolProximityEvent: TabletToolProximityEvent<Self>;
    type TabletToolTipEvent: TabletToolTipEvent<Self>;
    type TabletToolButtonEvent: TabletToolButtonEvent<Self>;
    type SwitchToggleEvent: SwitchToggleEvent<Self>;
    type SpecialEvent;

}

Trait that describes objects providing a source of input events. All input backends need to implement this and provide the same base guarantees about the precision of given events.

Required Associated Types

type Device: Device

Type representing input devices

type KeyboardKeyEvent: KeyboardKeyEvent<Self>

Type representing keyboard events

type PointerAxisEvent: PointerAxisEvent<Self>

Type representing axis events on pointer devices

type PointerButtonEvent: PointerButtonEvent<Self>

Type representing button events on pointer devices

type PointerMotionEvent: PointerMotionEvent<Self>

Type representing motion events of pointer devices

type PointerMotionAbsoluteEvent: PointerMotionAbsoluteEvent<Self>

Type representing motion events of pointer devices

type GestureSwipeBeginEvent: GestureSwipeBeginEvent<Self>

Type representing swipe begin events of pointer devices

type GestureSwipeUpdateEvent: GestureSwipeUpdateEvent<Self>

Type representing swipe update events of pointer devices

type GestureSwipeEndEvent: GestureSwipeEndEvent<Self>

Type representing swipe end events of pointer devices

type GesturePinchBeginEvent: GesturePinchBeginEvent<Self>

Type representing pinch begin events of pointer devices

type GesturePinchUpdateEvent: GesturePinchUpdateEvent<Self>

Type representing pinch update events of pointer devices

type GesturePinchEndEvent: GesturePinchEndEvent<Self>

Type representing pinch end events of pointer devices

type GestureHoldBeginEvent: GestureHoldBeginEvent<Self>

Type representing hold begin events of pointer devices

type GestureHoldEndEvent: GestureHoldEndEvent<Self>

Type representing hold end events of pointer devices

type TouchDownEvent: TouchDownEvent<Self>

Type representing touch events starting

type TouchUpEvent: TouchUpEvent<Self>

Type representing touch events ending

type TouchMotionEvent: TouchMotionEvent<Self>

Type representing touch events from moving

type TouchCancelEvent: TouchCancelEvent<Self>

Type representing canceling of touch events

type TouchFrameEvent: TouchFrameEvent<Self>

Type representing touch frame events

type TabletToolAxisEvent: TabletToolAxisEvent<Self>

Type representing axis events on tablet devices

type TabletToolProximityEvent: TabletToolProximityEvent<Self>

Type representing proximity events on tablet devices

type TabletToolTipEvent: TabletToolTipEvent<Self>

Type representing tip events on tablet devices

type TabletToolButtonEvent: TabletToolButtonEvent<Self>

Type representing button events on tablet tool devices

type SwitchToggleEvent: SwitchToggleEvent<Self>

Type representing switch toggle events

type SpecialEvent

Special events that are custom to this backend

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.