Trait smithay::backend::input::InputBackend

pub trait InputBackend: Sized {
    type InputConfig: ?Sized;
    type EventError: Error;
    type KeyboardKeyEvent: KeyboardKeyEvent;
    type PointerAxisEvent: PointerAxisEvent;
    type PointerButtonEvent: PointerButtonEvent;
    type PointerMotionEvent: PointerMotionEvent;
    type PointerMotionAbsoluteEvent: PointerMotionAbsoluteEvent;
    type TouchDownEvent: TouchDownEvent;
    type TouchUpEvent: TouchUpEvent;
    type TouchMotionEvent: TouchMotionEvent;
    type TouchCancelEvent: TouchCancelEvent;
    type TouchFrameEvent: TouchFrameEvent;
    fn set_handler<H: InputHandler<Self> + 'static>(&mut self, handler: H);

    fn get_handler(&mut self) -> Option<&mut InputHandler<Self>>;

    fn clear_handler(&mut self);

    fn input_config(&mut self) -> &mut Self::InputConfig;

    fn dispatch_new_events(&mut self) -> Result<(), Self::EventError>;
}

Trait that describes objects providing a source of input events. All input backends need to implemenent this and provide the same base gurantees about the presicion of given events.

Associated Types

type InputConfig: ?Sized

Type of input device associated with the backend

type EventError: Error

Type representing errors that may be returned when processing events

type KeyboardKeyEvent: KeyboardKeyEvent

Type representing keyboard events

type PointerAxisEvent: PointerAxisEvent

Type representing axis events on pointer devices

type PointerButtonEvent: PointerButtonEvent

Type representing button events on pointer devices

type PointerMotionEvent: PointerMotionEvent

Type representing motion events of pointer devices

type PointerMotionAbsoluteEvent: PointerMotionAbsoluteEvent

Type representing motion events of pointer devices

type TouchDownEvent: TouchDownEvent

Type representing touch events starting

type TouchUpEvent: TouchUpEvent

Type representing touch events ending

type TouchMotionEvent: TouchMotionEvent

Type representing touch events from moving

type TouchCancelEvent: TouchCancelEvent

Type representing cancelling of touch events

type TouchFrameEvent: TouchFrameEvent

Type representing touch frame events

Required Methods

fn set_handler<H: InputHandler<Self> + 'static>(&mut self, handler: H)

Sets a new handler for this InputBackend

fn get_handler(&mut self) -> Option<&mut InputHandler<Self>>

Get a reference to the currently set handler, if any

fn clear_handler(&mut self)

Clears the currently handler, if one is set

fn input_config(&mut self) -> &mut Self::InputConfig

Get current InputConfig

fn dispatch_new_events(&mut self) -> Result<(), Self::EventError>

Processes new events of the underlying backend and drives the InputHandler.