Trait smithay::backend::input::InputBackend

pub trait InputBackend: Sized {
    type EventError: Error;
    type KeyboardKeyEvent: KeyboardKeyEvent;
    type PointerAxisEvent: PointerAxisEvent;
    type PointerButtonEvent: PointerButtonEvent;
    type PointerMotionEvent: PointerMotionEvent;
    type PointerMotionAbsoluteEvent: PointerMotionAbsoluteEvent;
    type TouchDownEvent: TouchDownEvent;
    type TouchUpEvent: TouchUpEvent;
    type TouchMotionEvent: TouchMotionEvent;
    type TouchCancelEvent: TouchCancelEvent;
    type TouchFrameEvent: TouchFrameEvent;
    type SpecialEvent;
    type InputConfig: ?Sized;
    fn seats(&self) -> Vec<Seat>;

    fn input_config(&mut self) -> &mut Self::InputConfig;

    fn dispatch_new_events<F>(
        &mut self, 
        callback: F
    ) -> Result<(), Self::EventError>
    where
        F: FnMut(InputEvent<Self>, &mut Self::InputConfig);
}

Trait that describes objects providing a source of input events. All input backends need to implement this and provide the same base guarantees about the precision of given events.

Associated Types

type EventError: Error

Type representing errors that may be returned when processing events

type KeyboardKeyEvent: KeyboardKeyEvent

Type representing keyboard events

type PointerAxisEvent: PointerAxisEvent

Type representing axis events on pointer devices

type PointerButtonEvent: PointerButtonEvent

Type representing button events on pointer devices

type PointerMotionEvent: PointerMotionEvent

Type representing motion events of pointer devices

type PointerMotionAbsoluteEvent: PointerMotionAbsoluteEvent

Type representing motion events of pointer devices

type TouchDownEvent: TouchDownEvent

Type representing touch events starting

type TouchUpEvent: TouchUpEvent

Type representing touch events ending

type TouchMotionEvent: TouchMotionEvent

Type representing touch events from moving

type TouchCancelEvent: TouchCancelEvent

Type representing cancelling of touch events

type TouchFrameEvent: TouchFrameEvent

Type representing touch frame events

type SpecialEvent

Special events that are custom to this backend

type InputConfig: ?Sized

Backend-specific type allowing you to configure it

Loading content...
Required methods

fn seats(&self) -> Vec<Seat>

Get the list of currently known Seats

fn input_config(&mut self) -> &mut Self::InputConfig

Access the input configuration interface

fn dispatch_new_events<F>(
&mut self,
callback: F
) -> Result<(), Self::EventError> where
F: FnMut(InputEvent<Self>, &mut Self::InputConfig),

Processes new events of the underlying backend and drives the [InputHandler].

The callback can only assume its second argument to be usable if the event is InputEvent::ConfigChanged.

Loading content...