Trait smithay::backend::input::InputHandler

pub trait InputHandler<B: InputBackend> {
    fn on_seat_created(&mut self, seat: &Seat);
    fn on_seat_destroyed(&mut self, seat: &Seat);
    fn on_seat_changed(&mut self, seat: &Seat);
    fn on_keyboard_key(&mut self,
                       seat: &Seat,
                       time: u32,
                       key_code: u32,
                       state: KeyState,
                       count: u32);
    fn on_pointer_move(&mut self, seat: &Seat, time: u32, to: (u32, u32));
    fn on_pointer_button(&mut self,
                         seat: &Seat,
                         time: u32,
                         button: MouseButton,
                         state: MouseButtonState);
    fn on_pointer_scroll(&mut self,
                         seat: &Seat,
                         time: u32,
                         axis: Axis,
                         source: AxisSource,
                         amount: f64);
    fn on_touch(&mut self, seat: &Seat, time: u32, event: TouchEvent);
    fn on_input_config_changed(&mut self, config: &mut B::InputConfig);
}

Implement to receive input events from any InputBackend.

Required Methods

fn on_seat_created(&mut self, seat: &Seat)

Called when a new Seat has been created

fn on_seat_destroyed(&mut self, seat: &Seat)

Called when an existing Seat has been destroyed.

fn on_seat_changed(&mut self, seat: &Seat)

Called when a Seat's properties have changed.

fn on_keyboard_key(&mut self,
seat: &Seat,
time: u32,
key_code: u32,
state: KeyState,
count: u32)

Called when a new keyboard event was received.

Arguments

seat - The Seat the event belongs to

time - A upward counting variable useful for event ordering. Makes no gurantees about actual time passed between events.

key_code - Code of the pressed key. See linux/input-event-codes.h

state - KeyState of the event

count - Total number of keys pressed on all devices on the associated Seat

TODO:

check if events can arrive out of order.

Make stronger time guarantees

fn on_pointer_move(&mut self, seat: &Seat, time: u32, to: (u32, u32))

Called when a new pointer movement event was received.

Arguments

seat - The Seat the event belongs to

time - A upward counting variable useful for event ordering. Makes no gurantees about actual time passed between events.

to - Absolute screen coordinates of the pointer moved to.

TODO:

check if events can arrive out of order.

Make stronger time guarantees

fn on_pointer_button(&mut self,
seat: &Seat,
time: u32,
button: MouseButton,
state: MouseButtonState)

Called when a new pointer button event was received.

Arguments

seat - The Seat the event belongs to

time - A upward counting variable useful for event ordering. Makes no gurantees about actual time passed between events.

button - Which button was pressed..

state - MouseButtonState of the event

TODO:

check if events can arrive out of order.

Make stronger time guarantees

fn on_pointer_scroll(&mut self,
seat: &Seat,
time: u32,
axis: Axis,
source: AxisSource,
amount: f64)

Called when a new pointer scroll event was received.

Arguments

seat - The Seat the event belongs to

time - A upward counting variable useful for event ordering. Makes no gurantees about actual time passed between events.

axis - Axis this event was generated for.

source - Source of the scroll event. Important for interpretation of amount.

amount - Amount of scrolling on the given Axis. See source for interpretation.

TODO:

check if events can arrive out of order.

Make stronger time guarantees

fn on_touch(&mut self, seat: &Seat, time: u32, event: TouchEvent)

Called when a new touch event was received.

Arguments

seat - The Seat the event belongs to

time - A upward counting variable useful for event ordering. Makes no gurantees about actual time passed between events.

event - Touch event recieved. See TouchEvent.

TODO:

check if events can arrive out of order.

Make stronger time guarantees

fn on_input_config_changed(&mut self, config: &mut B::InputConfig)

Called when the InputConfig was changed through an external event.

What kind of events can trigger this call is completely backend dependent. E.g. an input devices was attached/detached or changed it's own configuration.