Trait smithay::backend::input::InputHandler

pub trait InputHandler<B: InputBackend> {
    fn on_seat_created(&mut self, seat: &Seat);

    fn on_seat_destroyed(&mut self, seat: &Seat);

    fn on_seat_changed(&mut self, seat: &Seat);

    fn on_keyboard_key(&mut self, seat: &Seat, event: B::KeyboardKeyEvent);

    fn on_pointer_move(&mut self, seat: &Seat, event: B::PointerMotionEvent);

    fn on_pointer_move_absolute(
        &mut self, 
        seat: &Seat, 
        event: B::PointerMotionAbsoluteEvent
    );

    fn on_pointer_button(&mut self, seat: &Seat, event: B::PointerButtonEvent);

    fn on_pointer_axis(&mut self, seat: &Seat, event: B::PointerAxisEvent);

    fn on_touch_down(&mut self, seat: &Seat, event: B::TouchDownEvent);

    fn on_touch_motion(&mut self, seat: &Seat, event: B::TouchMotionEvent);

    fn on_touch_up(&mut self, seat: &Seat, event: B::TouchUpEvent);

    fn on_touch_cancel(&mut self, seat: &Seat, event: B::TouchCancelEvent);

    fn on_touch_frame(&mut self, seat: &Seat, event: B::TouchFrameEvent);

    fn on_input_config_changed(&mut self, config: &mut B::InputConfig);
}

Implement to receive input events from any InputBackend.

Required methods

fn on_seat_created(&mut self, seat: &Seat)

Called when a new Seat has been created

fn on_seat_destroyed(&mut self, seat: &Seat)

Called when an existing Seat has been destroyed.

fn on_seat_changed(&mut self, seat: &Seat)

Called when a Seat's properties have changed.

Note:

It is not guaranteed that any change has actually happened.

fn on_keyboard_key(&mut self, seat: &Seat, event: B::KeyboardKeyEvent)

Called when a new keyboard event was received.

Arguments

seat - The Seat the event belongs to

event - The keyboard event

fn on_pointer_move(&mut self, seat: &Seat, event: B::PointerMotionEvent)

Called when a new pointer movement event was received.

Arguments

seat - The Seat the event belongs to

event - The pointer movement event

fn on_pointer_move_absolute(
&mut self,
seat: &Seat,
event: B::PointerMotionAbsoluteEvent
)

Called when a new pointer absolute movement event was received.

Arguments

seat - The Seat the event belongs to

event - The pointer absolute movement event

fn on_pointer_button(&mut self, seat: &Seat, event: B::PointerButtonEvent)

Called when a new pointer button event was received.

Arguments

seat - The Seat the event belongs to

event - The pointer button event

fn on_pointer_axis(&mut self, seat: &Seat, event: B::PointerAxisEvent)

Called when a new pointer scroll event was received.

Arguments

seat - The Seat the event belongs to

event - A upward counting variable useful for event ordering. Makes no guarantees about actual time passed between events.

fn on_touch_down(&mut self, seat: &Seat, event: B::TouchDownEvent)

Called when a new touch down event was received.

Arguments

seat - The Seat the event belongs to

event - The touch down event

fn on_touch_motion(&mut self, seat: &Seat, event: B::TouchMotionEvent)

Called when a new touch motion event was received.

Arguments

seat - The Seat the event belongs to

event - The touch motion event.

fn on_touch_up(&mut self, seat: &Seat, event: B::TouchUpEvent)

Called when a new touch up event was received.

Arguments

seat - The Seat the event belongs to

event - The touch up event.

fn on_touch_cancel(&mut self, seat: &Seat, event: B::TouchCancelEvent)

Called when a new touch cancel event was received.

Arguments

seat - The Seat the event belongs to

event - The touch cancel event.

fn on_touch_frame(&mut self, seat: &Seat, event: B::TouchFrameEvent)

Called when a new touch frame event was received.

Arguments

seat - The Seat the event belongs to

event - The touch frame event.

fn on_input_config_changed(&mut self, config: &mut B::InputConfig)

Called when the InputConfig was changed through an external event.

What kind of events can trigger this call is completely backend dependent. E.g. an input devices was attached/detached or changed it's own configuration.

Loading content...