smithay::backend::input

Trait KeyboardKeyEvent

pub trait KeyboardKeyEvent<B: InputBackend>: Event<B> {
    // Required methods
    fn key_code(&self) -> Keycode;
    fn state(&self) -> KeyState;
    fn count(&self) -> u32;
}

Trait for keyboard event

Required Methods

fn key_code(&self) -> Keycode

Returns the numerical button code of the keyboard button.

The value will correspond to one KEY_ constants from the Linux input event codes inside input-event-codes.h.

fn state(&self) -> KeyState

State of the key

fn count(&self) -> u32

Total number of keys pressed on all devices on the associated Seat