Trait smithay::backend::input::KeyboardKeyEvent

pub trait KeyboardKeyEvent<B: InputBackend>: Event<B> {
    fn key_code(&self) -> u32;

    fn state(&self) -> KeyState;

    fn count(&self) -> u32;
}

Trait for keyboard event

Required methods

fn key_code(&self) -> u32

Returns the numerical button code of the keyboard button.

The value will correspond to one KEY_ constants from the Linux input event codes inside input-event-codes.h.

fn state(&self) -> KeyState

State of the key

fn count(&self) -> u32

Total number of keys pressed on all devices on the associated Seat