Trait smithay::backend::input::PointerMotionAbsoluteEvent

pub trait PointerMotionAbsoluteEvent: Event {
    fn x(&self) -> f64;

    fn y(&self) -> f64;

    fn x_transformed(&self, width: u32) -> u32;

    fn y_transformed(&self, height: u32) -> u32;

    fn position(&self) -> (f64, f64) { ... }

    fn position_transformed(&self, coordinate_space: (u32, u32)) -> (u32, u32) { ... }
}

Trait for pointer events generated by absolute device positioning.

Required methods

fn x(&self) -> f64

Device x position in it's original coordinate space.

The format is defined by the backend implementation.

fn y(&self) -> f64

Device y position in it's original coordinate space.

The format is defined by the backend implementation.

fn x_transformed(&self, width: u32) -> u32

Device x position converted to the targets coordinate space's width. E.g. the focused output's width.

fn y_transformed(&self, height: u32) -> u32

Device y position converted to the targets coordinate space's height. E.g. the focused output's height.

Loading content...
Provided methods

fn position(&self) -> (f64, f64)

Device position in it's original coordinate space.

The format is defined by the backend implementation.

fn position_transformed(&self, coordinate_space: (u32, u32)) -> (u32, u32)

Device position converted to the targets coordinate space. E.g. the focused output's resolution.

Loading content...