Trait smithay::backend::input::PointerMotionEvent

pub trait PointerMotionEvent: Event {
    fn delta_x(&self) -> u32;

    fn delta_y(&self) -> u32;

    fn delta(&self) -> (u32, u32) { ... }
}

Trait for pointer events generated by relative device movement.

Required Methods

fn delta_x(&self) -> u32

Delta on the x axis between the last and new pointer device position interpreted as pixel movement

fn delta_y(&self) -> u32

Delta on the y axis between the last and new pointer device position interpreted as pixel movement

Provided Methods

fn delta(&self) -> (u32, u32)

Delta between the last and new pointer device position interpreted as pixel movement