Trait smithay::backend::input::PointerMotionEvent

pub trait PointerMotionEvent: Event {
    fn delta_x(&self) -> f64;

    fn delta_y(&self) -> f64;

    fn delta(&self) -> (f64, f64) { ... }
}

Trait for pointer events generated by relative device movement.

Required methods

fn delta_x(&self) -> f64

Delta on the x axis between the last and new pointer device position interpreted as pixel movement

fn delta_y(&self) -> f64

Delta on the y axis between the last and new pointer device position interpreted as pixel movement

Provided methods

fn delta(&self) -> (f64, f64)

Delta between the last and new pointer device position interpreted as pixel movement