Trait smithay::backend::input::PointerMotionEvent

pub trait PointerMotionEvent: Event {
    fn delta_x(&self) -> i32;

    fn delta_y(&self) -> i32;

    fn delta(&self) -> (i32, i32) { ... }
}

Trait for pointer events generated by relative device movement.

Required methods

fn delta_x(&self) -> i32

Delta on the x axis between the last and new pointer device position interpreted as pixel movement

fn delta_y(&self) -> i32

Delta on the y axis between the last and new pointer device position interpreted as pixel movement

Loading content...
Provided methods

fn delta(&self) -> (i32, i32)

Delta between the last and new pointer device position interpreted as pixel movement

Loading content...