Trait smithay::backend::input::TabletToolButtonEvent

pub trait TabletToolButtonEvent<B: InputBackend>: TabletToolEvent<B> + Event<B> {
    fn button(&self) -> u32;

    fn seat_button_count(&self) -> u32;

    fn button_state(&self) -> ButtonState;
}

Signals that a tool has changed a logical button state on a device with the DeviceCapability::TabletTool capability.

Required methods

fn button(&self) -> u32

Return the button that triggered this event.

fn seat_button_count(&self) -> u32

For the button of a TabletToolButtonEvent, return the total number of buttons pressed on all devices on the associated seat after the the event was triggered.

fn button_state(&self) -> ButtonState

Return the button state of the event.