Trait smithay::backend::input::TabletToolProximityEvent

pub trait TabletToolProximityEvent<B: InputBackend>: TabletToolEvent<B> + Event<B> {
    fn state(&self) -> ProximityState;
}

Trait for tablet tool proximity events.

Required methods

fn state(&self) -> ProximityState

Returns the new proximity state of a tool from a proximity event.

Used to check whether or not a tool came in or out of proximity during an TabletToolProximityEvent.

See Handling of proximity events for recommendations on proximity handling.