Trait smithay::backend::input::TouchMotionEvent

pub trait TouchMotionEvent<B: InputBackend>: Event<B> {
    fn slot(&self) -> Option<TouchSlot>;

    fn x(&self) -> f64;

    fn y(&self) -> f64;

    fn x_transformed(&self, width: i32) -> f64;

    fn y_transformed(&self, height: i32) -> f64;

    fn position(&self) -> Point<f64, Raw> { ... }

    fn position_transformed(
        &self, 
        coordinate_space: Size<i32, Logical>
    ) -> Point<f64, Logical> { ... }
}

Trait for touch events regarding movement on the screen

Required methods

fn slot(&self) -> Option<TouchSlot>

TouchSlot, if the device has multi-touch capabilities

fn x(&self) -> f64

Touch event’s x-coordinate in the device’s native coordinate space

The actual format is defined by the implementation.

fn y(&self) -> f64

Touch event’s x-coordinate in the device’s native coordinate space

The actual format is defined by the implementation.

fn x_transformed(&self, width: i32) -> f64

Touch event’s x position converted to the targets coordinate space’s width. E.g. the focused output’s width.

fn y_transformed(&self, height: i32) -> f64

Touch event’s y position converted to the targets coordinate space’s width. E.g. the focused output’s width.

Provided methods

fn position(&self) -> Point<f64, Raw>

Touch position in the device’s native coordinate space

The actual format is defined by the implementation.

fn position_transformed(
&self,
coordinate_space: Size<i32, Logical>
) -> Point<f64, Logical>

Touch position converted into the target coordinate space. E.g. the focused output’s resolution.