smithay::backend::libinput

Struct LibinputInputBackend

pub struct LibinputInputBackend { /* private fields */ }

Available on crate feature backend_libinput only.

Libinput based InputBackend.

Tracks input of all devices given manually or via a udev seat to a provided libinput context.

Implementations

impl LibinputInputBackend

pub fn new(context: Libinput) -> Self

Initialize a new LibinputInputBackend from a given already initialized libinput context.

pub fn context(&self) -> &Libinput

Returns a reference to the underlying libinput context