Struct smithay::backend::libinput::LibinputInputBackend

pub struct LibinputInputBackend { /* fields omitted */ }

Libinput based InputBackend.

Tracks input of all devices given manually or via a udev seat to a provided libinput context.

Methods

impl LibinputInputBackend

pub fn new<L>(context: Libinput, logger: L) -> Self where
L: Into<Option<Logger>>,

Initialize a new LibinputInputBackend from a given already initialized libinput context.