Module smithay::backend::renderer

Rendering functionality and abstractions

Collection of common traits and implementations to facilitate (possible hardware-accelerated) rendering.

Supported rendering apis:

Raw OpenGL ES 2

Modules

damage

Helper for effective damage tracked rendering

element

Common base for elements that can be drawn by a Renderer

gles2

Implementation of the rendering traits using OpenGL ES 2

glow

Integration for using [glow] on top of smithays OpenGL ES 2 renderer

multigpu

This module aims to make multi-gpu setups easier to handle for compositors.

utils

Utility module for helpers around drawing WlSurfaces and RenderElements with Renderers.

Enums

BufferType

Buffer type of a given wl_buffer, if managed by smithay

TextureFilter

Texture filtering methods

Traits

Bind

Abstraction for Renderers, that can render into different targets

ExportDma

Trait for renderers supporting exporting contents of framebuffers or textures as dmabufs.

ExportMem

Trait for renderers supporting exporting contents of framebuffers or textures into memory.

Frame

Helper trait for Renderer, which defines a rendering api for a currently in-progress frame during Renderer::render.

ImportAll

Common trait for renderers of any wayland buffer type

ImportDma

Trait for Renderers supporting importing dmabufs.

ImportDmaWl

Trait for Renderers supporting importing dmabuf-based wl_buffers

ImportEgl

Trait for Renderers supporting importing wl_drm-based buffers.

ImportMem

Trait for Renderers supporting importing bitmaps from memory.

ImportMemWl

Trait for Renderers supporting importing wl_buffers using shared memory.

Offscreen

Trait for renderers that support creating offscreen framebuffers to render into.

Renderer

Abstraction of commonly used rendering operations for compositors.

Texture

A two dimensional texture

TextureMapping

A downloaded texture buffer

Unbind

Functionality to unbind the current rendering target

Functions

buffer_dimensions

Returns the dimensions of a wl_buffer

buffer_has_alpha

Returns if the buffer has an alpha channel

buffer_type

Returns the type of a wl_buffer