Module smithay::backend::renderer

Rendering functionality and abstractions

Collection of common traits and implementations to facilitate (possible hardware-accelerated) rendering.

Supported rendering apis:

Raw OpenGL ES 2

Modules

gles2

Implementation of the rendering traits using OpenGL ES 2

Enums

BufferType

Buffer type of a given wl_buffer, if managed by smithay

Transform

Possible transformations to two-dimensional planes

Traits

Bind

Abstraction for Renderers, that can render into different targets

Frame

Helper trait for Renderer, which defines a rendering api for a currently in-progress frame during Renderer::render.

ImportAll

Common trait for renderers of any wayland buffer type

ImportDma

Trait for Renderers supporting importing dmabuf-based buffers.

ImportEgl

Trait for Renderers supporting importing wl_drm-based buffers.

ImportShm

Trait for Renderers supporting importing shm-based buffers.

Renderer

Abstraction of commonly used rendering operations for compositors.

Texture

A two dimensional texture

Unbind

Functionality to unbind the current rendering target

Functions

buffer_dimensions

Returns the dimensions of a wl_buffer

buffer_type

Returns the type of a wl_buffer