Module smithay::backend::renderer::element::surface

Element to render a wayland surface

Why use this implementation

The WaylandSurfaceRenderElement provides an easy way to integrate a WlSurface in the smithay rendering pipeline

How to use it

WaylandSurfaceRenderElement::from_surface allows you to obtain a WaylandSurfaceRenderElement for a single WlSurface. To retrieve WaylandSurfaceRenderElements for a whole surface tree you can use render_elements_from_surface_tree.

use smithay::{
    backend::renderer::{
        damage::DamageTrackedRenderer,
        element::surface::{render_elements_from_surface_tree, WaylandSurfaceRenderElement},
    },
    utils::{Point, Rectangle, Size, Transform},
};

// Initialize a static damage tracked renderer
let mut damage_tracked_renderer = DamageTrackedRenderer::new((800, 600), 1.0, Transform::Normal);

loop {
    // Create the render elements from the surface
    let location = Point::from((100, 100));
    let render_elements: Vec<WaylandSurfaceRenderElement> =
        render_elements_from_surface_tree(&surface, location, 1.0);

    // Render the element(s)
    damage_tracked_renderer
        .render_output(&mut renderer, 0, &*render_elements, [0.8, 0.8, 0.9, 1.0], log.clone())
        .expect("failed to render output");
}

Structs

WaylandSurfaceRenderElement

A single surface render element

Functions

render_elements_from_surface_tree

Retrieve the WaylandSurfaceRenderElements for a surface tree