smithay::backend::renderer::element

Module surface

Available on crate feature wayland_frontend only.

Element to render a wayland surface

Why use this implementation

The WaylandSurfaceRenderElement provides an easy way to integrate a WlSurface in the smithay rendering pipeline

How to use it

WaylandSurfaceRenderElement::from_surface allows you to obtain a WaylandSurfaceRenderElement for a single WlSurface. To retrieve WaylandSurfaceRenderElements for a whole surface tree you can use render_elements_from_surface_tree.

use smithay::{
    backend::renderer::{
        damage::OutputDamageTracker,
        element::{
            Kind,
            surface::{render_elements_from_surface_tree, WaylandSurfaceRenderElement},
        },
    },
    utils::{Point, Rectangle, Size, Transform},
};

// Initialize a static damage tracked renderer
let mut damage_tracker = OutputDamageTracker::new((800, 600), 1.0, Transform::Normal);

loop {
    // Create the render elements from the surface
    let location = Point::from((100, 100));
    let render_elements: Vec<WaylandSurfaceRenderElement<DummyRenderer>> =
        render_elements_from_surface_tree(&mut renderer, &surface, location, 1.0, 1.0, Kind::Unspecified);

    // Render the element(s)
    damage_tracker
        .render_output(&mut renderer, &mut framebuffer, 0, &*render_elements, [0.8, 0.8, 0.9, 1.0])
        .expect("failed to render output");
}

Structs

WaylandSurfaceRenderElement

A single surface render element

Enums

KindEvaluation

Ways to evaluation the Kind of a surface.

WaylandSurfaceTexture

Texture used for the WaylandSurfaceRenderElement

Functions

render_elements_from_surface_tree

Retrieve the WaylandSurfaceRenderElements for a surface tree