Trait smithay::backend::renderer::element::Element

pub trait Element {
    fn id(&self) -> &Id;
    fn current_commit(&self) -> CommitCounter;
    fn src(&self) -> Rectangle<f64, BufferCoords>;
    fn geometry(&self, scale: Scale<f64>) -> Rectangle<i32, Physical>;

    fn location(&self, scale: Scale<f64>) -> Point<i32, Physical> { ... }
    fn transform(&self) -> Transform { ... }
    fn damage_since(
        &self,
        scale: Scale<f64>,
        commit: Option<CommitCounter>
    ) -> Vec<Rectangle<i32, Physical>> { ... }
    fn opaque_regions(&self, _scale: Scale<f64>) -> Vec<Rectangle<i32, Physical>> { ... }
}

A single element

Required Methods

fn id(&self) -> &Id

Get the unique id of this element

fn current_commit(&self) -> CommitCounter

Get the current commit position of this element

fn src(&self) -> Rectangle<f64, BufferCoords>

Get the src of the underlying buffer

fn geometry(&self, scale: Scale<f64>) -> Rectangle<i32, Physical>

Get the geometry relative to the output

Provided Methods

fn location(&self, scale: Scale<f64>) -> Point<i32, Physical>

Get the location relative to the output

fn transform(&self) -> Transform

Get the transform of the underlying buffer

fn damage_since(
&self,
scale: Scale<f64>,
commit: Option<CommitCounter>
) -> Vec<Rectangle<i32, Physical>>

Get the damage since the provided commit relative to the element

fn opaque_regions(&self, _scale: Scale<f64>) -> Vec<Rectangle<i32, Physical>>

Get the opaque regions of the element relative to the element