Module smithay::backend::renderer::element::utils

Utilities and helpers around the Element trait.

Structs

ConstrainAlign

Defines how the elements should be aligned during constrain

CropRenderElement

A element that allows to crop another element

RelocateRenderElement

A element that allows to offset the location of an existing element

RescaleRenderElement

A element that allows to re-scale another element

Enums

ConstrainScaleBehavior

Defines the scale behavior for the constrain

Relocate

Defines how the location parameter should apply in RelocateRenderElement::from_element

Functions

constrain_as_render_elements

Convenience function to constrain something that implements AsRenderElements

constrain_render_elements

Constrain render elements on a specific location with a specific size

select_dmabuf_feedback

Select a DmabufFeedback based on the RenderElementPresentationState for a single [Element]