smithay::backend::renderer

Module gles

Available on crate feature renderer_gl only.

Implementation of the rendering traits using OpenGL ES 2

Modules

element

RenderElements specific to using a GlesRenderer

ffi

format

GL color format conversion helpers

profiler

GPU profiling helpers

Structs

GlesFrame

Handle to the currently rendered frame during GlesRenderer::render.

GlesMapping

Texture mapping of a GLES2 texture

GlesPixelProgram

Gles pixel shader

GlesRenderbuffer

Offscreen render surface

GlesRenderer

A renderer utilizing OpenGL ES

GlesTarget

A GL framebuffer

GlesTexProgram

Gles texture shader

GlesTexture

A handle to a GLES texture

Uniform

A shader uniform variable consisting out of a name and value

UniformDesc

GL location and type of a uniform shader variable

UniformName

A description of a uniform shader variable consisting out of a name and type

Enums

Capability

Capabilities of the GlesRenderer

GlesError

Error returned during rendering using GL ES

UniformType

Different value types of a shader uniform variable for the GlesRenderer.

UniformValue

Value of a uniform variable

Constants

DEBUG_FLAGS

Debug flags shader define

EXTERNAL

External texture shader define

NO_ALPHA

No alpha shader define

Functions

compile_shader⚠

Compiles a shader variant.

link_program⚠

Compiles and links a shader program.