Struct smithay::backend::renderer::gles2::Gles2Renderer

pub struct Gles2Renderer { /* fields omitted */ }

A renderer utilizing OpenGL ES 2

Implementations

impl Gles2Renderer

pub unsafe fn new<L>(
context: EGLContext,
logger: L
) -> Result<Gles2Renderer, Gles2Error> where
L: Into<Option<Logger>>,

Creates a new OpenGL ES 2 renderer from a given EGLContext.

Safety

This operation will cause undefined behavior if the given EGLContext is active in another thread.

Implementation details

Texture handles created by the resulting renderer are valid for every rendered created with an EGLContext shared with the given one (see EGLContext::new_shared) and can be used on any of these renderers.

This renderer has no default framebuffer, use Bind::bind before rendering.

Binding a new target, while another one is already bound, will replace the current target.

Shm buffers can be released after a successful import, without the texture handle becoming invalid.

impl Gles2Renderer

pub fn with_context<F, R>(&mut self, func: F) -> Result<R, Gles2Error> where
F: FnOnce(&mut Self, &Gles2) -> R,

Run custom code in the GL context owned by this renderer.

Note: Any changes to the GL state should be restored at the end of this function. Otherwise this can lead to rendering errors while using functions of this renderer. Relying on any state set by the renderer may break on any smithay update as the details about how this renderer works are considered an implementation detail.