Struct smithay::backend::renderer::multigpu::MultiRenderer

pub struct MultiRenderer<'a, 'b, R: GraphicsApi, T: GraphicsApi, Target> { /* private fields */ }

Renderer, that transparently copies rendering results to another gpu, as well as transparently importing client buffers residing on different gpus.