smithay::backend::renderer::sync

Trait Fence

pub trait Fence:
    Debug
    + Send
    + Sync
    + Downcast {
    // Required methods
    fn is_signaled(&self) -> bool;
    fn wait(&self) -> Result<(), Interrupted>;
    fn is_exportable(&self) -> bool;
    fn export(&self) -> Option<OwnedFd>;
}

A fence that will be signaled in finite time

Required Methods

fn is_signaled(&self) -> bool

Queries the state of the fence

fn wait(&self) -> Result<(), Interrupted>

Blocks the current thread until the fence is signaled

fn is_exportable(&self) -> bool

Returns whether this fence can be exported as a native fence fd

fn export(&self) -> Option<OwnedFd>

Export this fence as a native fence fd

Implementations

impl dyn Fence

pub fn is<__T: Fence>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Fence>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Fence>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Fence>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Fence>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.