Trait smithay::backend::renderer::Bind

pub trait Bind<Target>: Unbind {
    fn bind(&mut self, target: Target) -> Result<(), <Self as Renderer>::Error>;

    fn supported_formats(&self) -> Option<HashSet<Format>> { ... }
}

Abstraction for Renderers, that can render into different targets

Required methods

fn bind(&mut self, target: Target) -> Result<(), <Self as Renderer>::Error>

Bind a given rendering target, which will contain the rendering results until unbind is called.

Binding to target, while another one is already bound, is rendering defined. Some renderers might happily replace the current target, while other might drop the call or throw an error.

Provided methods

fn supported_formats(&self) -> Option<HashSet<Format>>

Supported pixel formats for given targets, if applicable.