smithay::backend::renderer

Trait Blit

pub trait Blit
where
    Self: Renderer,
{
    // Required method
    fn blit(
        &mut self,
        from: &Self::Framebuffer<'_>,
        to: &mut Self::Framebuffer<'_>,
        src: Rectangle<i32, Physical>,
        dst: Rectangle<i32, Physical>,
        filter: TextureFilter,
    ) -> Result<SyncPoint, Self::Error>;
}

Trait for renderers supporting blitting contents from one framebuffer to another.

Required Methods

fn blit( &mut self, from: &Self::Framebuffer<'_>, to: &mut Self::Framebuffer<'_>, src: Rectangle<i32, Physical>, dst: Rectangle<i32, Physical>, filter: TextureFilter, ) -> Result<SyncPoint, Self::Error>

Copies the contents of src from one provided target to dst in the other provided target, applying filter if necessary.

This operation is non destructive, the contents of the source framebuffer are kept intact as is any region not in dst for the target framebuffer.

This function may fail, if (but not limited to):

The source framebuffer is not readable

The destination framebuffer is not writable

src is out of bounds for the source framebuffer

dst is out of bounds for the destination framebuffer

src and dst sizes are different and interpolation is not supported by this renderer.

source and target framebuffer are the same

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.