Trait smithay::backend::renderer::ExportMem

pub trait ExportMem: Renderer {
    type TextureMapping: TextureMapping;

    fn copy_framebuffer(
        &mut self,
        region: Rectangle<i32, BufferCoord>
    ) -> Result<Self::TextureMapping, <Self as Renderer>::Error>;
    fn copy_texture(
        &mut self,
        texture: &Self::TextureId,
        region: Rectangle<i32, BufferCoord>
    ) -> Result<Self::TextureMapping, Self::Error>;
    fn map_texture<'a>(
        &mut self,
        texture_mapping: &'a Self::TextureMapping
    ) -> Result<&'a [u8], <Self as Renderer>::Error>;
}

Trait for renderers supporting exporting contents of framebuffers or textures into memory.

Required Associated Types

type TextureMapping: TextureMapping

Texture type representing a downloaded pixel buffer.

Required Methods

fn copy_framebuffer(
&mut self,
region: Rectangle<i32, BufferCoord>
) -> Result<Self::TextureMapping, <Self as Renderer>::Error>

Copies the contents of the currently bound framebuffer.

This operation is not destructive, the contents of the framebuffer keep being valid.

This function may fail, if (but not limited to):

The framebuffer is not readable

The region is out of bounds of the framebuffer

There is not enough space to create the mapping

fn copy_texture(
&mut self,
texture: &Self::TextureId,
region: Rectangle<i32, BufferCoord>
) -> Result<Self::TextureMapping, Self::Error>

Copies the contents of the passed texture. Note: This function may change or invalidate the current bind.

This operation is not destructive, the contents of the texture keep being valid.

This function may fail, if:

There is not enough space to create the mapping

The texture does no allow copying for implementation-specfic reasons

fn map_texture<'a>(
&mut self,
texture_mapping: &'a Self::TextureMapping
) -> Result<&'a [u8], <Self as Renderer>::Error>

Returns a read-only pointer to a previously created texture mapping.

The format of the returned slice is RGBA8.

This function may fail, if (but not limited to):

There is not enough space in memory