smithay::backend::renderer

Trait ExportMem

pub trait ExportMem: Renderer {
    type TextureMapping: TextureMapping;

    // Required methods
    fn copy_framebuffer(
        &mut self,
        target: &Self::Framebuffer<'_>,
        region: Rectangle<i32, BufferCoord>,
        format: Fourcc,
    ) -> Result<Self::TextureMapping, Self::Error>;
    fn copy_texture(
        &mut self,
        texture: &Self::TextureId,
        region: Rectangle<i32, BufferCoord>,
        format: Fourcc,
    ) -> Result<Self::TextureMapping, Self::Error>;
    fn can_read_texture(
        &mut self,
        texture: &Self::TextureId,
    ) -> Result<bool, Self::Error>;
    fn map_texture<'a>(
        &mut self,
        texture_mapping: &'a Self::TextureMapping,
    ) -> Result<&'a [u8], Self::Error>;
}

Trait for renderers supporting exporting contents of framebuffers or textures into memory.

Required Associated Types

type TextureMapping: TextureMapping

Texture type representing a downloaded pixel buffer.

Required Methods

fn copy_framebuffer( &mut self, target: &Self::Framebuffer<'_>, region: Rectangle<i32, BufferCoord>, format: Fourcc, ) -> Result<Self::TextureMapping, Self::Error>

Copies the contents of the provided target.

This operation is not destructive, the contents of the framebuffer keep being valid.

This function may fail, if (but not limited to):

The framebuffer is not readable

The region is out of bounds of the framebuffer

There is not enough space to create the mapping

It is not possible to convert the framebuffer into the provided format.

fn copy_texture( &mut self, texture: &Self::TextureId, region: Rectangle<i32, BufferCoord>, format: Fourcc, ) -> Result<Self::TextureMapping, Self::Error>

Copies the contents of the passed texture. Note: This function may change or invalidate the current bind.

Renderers are not required to support any format other than what was returned by Texture::format. This operation is not destructive, the contents of the texture keep being valid.

This function may fail, if:

There is not enough space to create the mapping

The texture does no allow copying for implementation-specific reasons

It is not possible to convert the texture into the provided format.

fn can_read_texture( &mut self, texture: &Self::TextureId, ) -> Result<bool, Self::Error>

Returns whether the renderer should be able to read-back from the given texture.

No actual copying shall be performed by this function nor is a format specified, so it is still legal for ExportMem::copy_texture to return an error, if this method returns true.

This function may fail, if:

A readability test did successfully complete (not that it returned unreadable!)

Any of the state of the renderer is irrevesibly changed

fn map_texture<'a>( &mut self, texture_mapping: &'a Self::TextureMapping, ) -> Result<&'a [u8], Self::Error>

Returns a read-only pointer to a previously created texture mapping.

The format of the returned slice is given by Texture::format of the texture mapping.

This function may fail, if (but not limited to):

There is not enough space in memory

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.