Trait smithay::backend::renderer::Frame

pub trait Frame {
    type Error: Error;
    type TextureId: Texture;

    fn clear(
        &mut self,
        color: [f32; 4],
        at: &[Rectangle<i32, Physical>]
    ) -> Result<(), Self::Error>;
    fn render_texture_from_to(
        &mut self,
        texture: &Self::TextureId,
        src: Rectangle<f64, BufferCoord>,
        dst: Rectangle<i32, Physical>,
        damage: &[Rectangle<i32, Physical>],
        src_transform: Transform,
        alpha: f32
    ) -> Result<(), Self::Error>;
    fn transformation(&self) -> Transform;

    fn render_texture_at(
        &mut self,
        texture: &Self::TextureId,
        pos: Point<i32, Physical>,
        texture_scale: i32,
        output_scale: impl Into<Scale<f64>>,
        src_transform: Transform,
        damage: &[Rectangle<i32, Physical>],
        alpha: f32
    ) -> Result<(), Self::Error> { ... }
}

Helper trait for Renderer, which defines a rendering api for a currently in-progress frame during Renderer::render.

Required Associated Types

type Error: Error

Error type returned by the rendering operations of this renderer.

type TextureId: Texture

Texture Handle type used by this renderer.

Required Methods

fn clear(
&mut self,
color: [f32; 4],
at: &[Rectangle<i32, Physical>]
) -> Result<(), Self::Error>

Clear the complete current target with a single given color.

The at parameter specifies a set of rectangles to clear in the current target. This allows partially clearing the target which may be useful for damaged rendering.

This operation is only valid in between a begin and finish-call. If called outside this operation may error-out, do nothing or modify future rendering results in any way.

fn render_texture_from_to(
&mut self,
texture: &Self::TextureId,
src: Rectangle<f64, BufferCoord>,
dst: Rectangle<i32, Physical>,
damage: &[Rectangle<i32, Physical>],
src_transform: Transform,
alpha: f32
) -> Result<(), Self::Error>

Render part of a texture as given by src to the current target into the rectangle described by dst as a flat 2d-plane after applying the inverse of the given transformation. (Meaning src_transform should match the orientation of surface being rendered).

fn transformation(&self) -> Transform

Output transformation that is applied to this frame

Provided Methods

fn render_texture_at(
&mut self,
texture: &Self::TextureId,
pos: Point<i32, Physical>,
texture_scale: i32,
output_scale: impl Into<Scale<f64>>,
src_transform: Transform,
damage: &[Rectangle<i32, Physical>],
alpha: f32
) -> Result<(), Self::Error>

Render a texture to the current target as a flat 2d-plane at a given position and applying the given transformation with the given alpha value. (Meaning src_transform should match the orientation of surface being rendered).