Trait smithay::backend::renderer::Frame

pub trait Frame {
    type Error: Error;
    type TextureId: Texture;
    fn clear(&mut self, color: [f32; 4]) -> Result<(), Self::Error>;

    fn render_texture(
        &mut self, 
        texture: &Self::TextureId, 
        matrix: Matrix3<f32>, 
        tex_coords: [Vector2<f32>; 4], 
        alpha: f32
    ) -> Result<(), Self::Error>;

    fn render_texture_at(
        &mut self, 
        texture: &Self::TextureId, 
        pos: Point<f64, Physical>, 
        texture_scale: i32, 
        output_scale: f64, 
        transform: Transform, 
        alpha: f32
    ) -> Result<(), Self::Error> { ... }

    fn render_texture_from_to(
        &mut self, 
        texture: &Self::TextureId, 
        src: Rectangle<i32, Buffer>, 
        dest: Rectangle<f64, Physical>, 
        transform: Transform, 
        alpha: f32
    ) -> Result<(), Self::Error> { ... }
}

Helper trait for Renderer, which defines a rendering api for a currently in-progress frame during Renderer::render.

Associated Types

type Error: Error

Error type returned by the rendering operations of this renderer.

type TextureId: Texture

Texture Handle type used by this renderer.

Required methods

fn clear(&mut self, color: [f32; 4]) -> Result<(), Self::Error>

Clear the complete current target with a single given color.

This operation is only valid in between a begin and finish-call. If called outside this operation may error-out, do nothing or modify future rendering results in any way.

fn render_texture(
&mut self,
texture: &Self::TextureId,
matrix: Matrix3<f32>,
tex_coords: [Vector2<f32>; 4],
alpha: f32
) -> Result<(), Self::Error>

Render a texture to the current target using given projection matrix and alpha. The given vertices are used to source the texture. This is mostly useful for cropping the texture.

Provided methods

fn render_texture_at(
&mut self,
texture: &Self::TextureId,
pos: Point<f64, Physical>,
texture_scale: i32,
output_scale: f64,
transform: Transform,
alpha: f32
) -> Result<(), Self::Error>

Render a texture to the current target as a flat 2d-plane at a given position and applying the given transformation with the given alpha value.

fn render_texture_from_to(
&mut self,
texture: &Self::TextureId,
src: Rectangle<i32, Buffer>,
dest: Rectangle<f64, Physical>,
transform: Transform,
alpha: f32
) -> Result<(), Self::Error>

Render part of a texture as given by src to the current target into the rectangle described by dest as a flat 2d-plane after applying the given transformations.