Trait smithay::backend::renderer::Renderer

pub trait Renderer {
    type Error: Error;
    type TextureId: Texture;
    type Frame: Frame<Error = Self::Error, TextureId = Self::TextureId>;

    fn id(&self) -> usize;
    fn downscale_filter(
        &mut self,
        filter: TextureFilter
    ) -> Result<(), Self::Error>;
    fn upscale_filter(
        &mut self,
        filter: TextureFilter
    ) -> Result<(), Self::Error>;
    fn render<F, R>(
        &mut self,
        output_size: Size<i32, Physical>,
        dst_transform: Transform,
        rendering: F
    ) -> Result<R, Self::Error>
    where
        F: FnOnce(&mut Self, &mut Self::Frame) -> R;
}

Abstraction of commonly used rendering operations for compositors.

Required Associated Types

type Error: Error

Error type returned by the rendering operations of this renderer.

type TextureId: Texture

Texture Handle type used by this renderer.

type Frame: Frame<Error = Self::Error, TextureId = Self::TextureId>

Type representing a currently in-progress frame during the Renderer::render-call

Required Methods

fn id(&self) -> usize

Returns an id, that is unique to all renderers, that can use TextureIds originating from any of these renderers.

fn downscale_filter(&mut self, filter: TextureFilter) -> Result<(), Self::Error>

Set the filter method to be used when rendering a texture into a smaller area than its size

fn upscale_filter(&mut self, filter: TextureFilter) -> Result<(), Self::Error>

Set the filter method to be used when rendering a texture into a larger area than its size

fn render<F, R>(
&mut self,
output_size: Size<i32, Physical>,
dst_transform: Transform,
rendering: F
) -> Result<R, Self::Error>where
F: FnOnce(&mut Self, &mut Self::Frame) -> R,

Initialize a rendering context on the current rendering target with given dimensions and transformation.

The output_size specifies the dimensions of the display before the dst_transform is applied.

This function may error, if:

The given dimensions are unsupported (too large) for this renderer

The given Transformation is not supported by the renderer (Transform::Normal is always supported).

This renderer implements Bind, no target was bound and has no default target.

(Renderers not implementing Bind always have a default target.)