smithay::backend::renderer

Trait Texture

pub trait Texture: Debug {
    // Required methods
    fn width(&self) -> u32;
    fn height(&self) -> u32;
    fn format(&self) -> Option<Fourcc>;

    // Provided method
    fn size(&self) -> Size<i32, BufferCoord> { ... }
}

A two dimensional texture

Required Methods

fn width(&self) -> u32

Width of the texture plane

fn height(&self) -> u32

Height of the texture plane

fn format(&self) -> Option<Fourcc>

Format of the texture, if available.

In case the format is hidden by the implementation, it should be assumed, that the pixel representation cannot be read.

Thus ExportMem::copy_texture, if implemented, will not succeed for this texture. Note that this does not mean every texture with a format is guaranteed to be copyable.

Provided Methods

fn size(&self) -> Size<i32, BufferCoord>

Size of the texture plane