smithay::backend::renderer::utils

Function import_surface_tree

pub fn import_surface_tree<R>(
    renderer: &mut R,
    surface: &WlSurface,
) -> Result<(), R::Error>
where
    R: Renderer + ImportAll,
    R::TextureId: 'static,

Available on crate feature wayland_frontend only.

Imports buffers of a surface and its subsurfaces using a given Renderer.

This (or import_surface) need to be called before draw_render_elements, if used later.

Note: This will do nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.