Function smithay::backend::renderer::utils::import_surface_tree

pub fn import_surface_tree<R>(
    renderer: &mut R, 
    surface: &WlSurface, 
    log: &Logger
) -> Result<(), <R as Renderer>::Error> where
    R: Renderer + ImportAll,
    <R as Renderer>::TextureId: 'static, 

Imports buffers of a surface and its subsurfaces using a given Renderer.

This can be called early as an optimization, if draw_surface_tree is used later. draw_surface_tree will also import buffers as necessary, but calling import_surface_tree already may allow buffer imports to happen before compositing takes place, depending on your event loop.

Note: This will do nothing, if you are not using crate::backend::renderer::utils::on_commit_buffer_handler to let smithay handle buffer management.