Module smithay::backend::session

Abstraction of different session APIs.

Sessions provide a way for multiple graphical systems to run in parallel by providing mechanisms to switch between and handle device access and permissions for every running instance.

They are crucial to allow unprivileged processes to use graphical or input devices.

The following mechanisms are currently provided: - direct - legacy tty / virtual terminal kernel API

Modules

auto
Implementation of the Session trait through various implementations automatically choosing the best available interface.

direct
Implementation of the Session trait through the legacy vt kernel interface.

logind
Implementation of the Session trait through the logind dbus interface.

Enums

Signal
Signals that can be generated by a session

Traits

AsErrno
Allows errors to be described by an error number

Session
General session interface.