Module smithay::backend::session::auto

Implementation of the [Session] trait through various implementations automatically choosing the best available interface.

How to use it

Initialization

To initialize a session just call AutoSession::new. A new session will be opened, if the any available interface is successful and will be closed once the AutoSessionNotifier is dropped.

Usage of the session

The session may be used to open devices manually through the [Session] interface or be passed to other objects that need it to open devices themselves. The AutoSession is clonable and may be passed to multiple devices easily.

Examples for those are e.g. the LibinputInputBackend (its context might be initialized through a [Session] via the LibinputSessionInterface).

Usage of the session notifier

The notifier might be used to pause device access, when the session gets paused (e.g. by switching the tty via AutoSession::change_vt) and to automatically enable it again, when the session becomes active again.

It is crucial to avoid errors during that state. Examples for object that might be registered for notifications are the Libinput context or the Device.

Structs

AutoId
Id's used by the [AutoSessionNotifier] internally.

Enums

AutoSession
[Session] using the best available interface

AutoSessionNotifier
[SessionNotifier] using the best available interface

BoundAutoSession
Bound session that is driven by a EventLoop.

Error
Errors related to auto sessions

Functions

auto_session_bind
Bind an [AutoSessionNotifier] to an EventLoop.