Struct smithay::backend::session::direct::DirectSession

pub struct DirectSession { /* fields omitted */ }

Session via the virtual terminal direct kernel interface

Methods

impl DirectSession

pub fn new<L>(
tty: Option<&Path>,
logger: L
) -> Result<(DirectSession, DirectSessionNotifier)> where
L: Into<Option<Logger>>,

Tries to create a new session via the legacy virtual terminal interface.

If you do not provide a tty device path, it will try to open the currently active tty if any.

pub fn vt(&self) -> i32

Get the number of the virtual terminal used by this session