Module smithay::backend::session::logind

Implementation of the Session trait through the logind dbus interface.

This requires systemd and dbus to be available and started on the system.

How to use it

Initialization

To initialize a session just call LogindSession::new. A new session will be opened, if the call is successful and will be closed once the LogindSessionNotifier is dropped.

Usage of the session

The session may be used to open devices manually through the Session interface or be passed to other objects that need it to open devices themselves. The LogindSession is clonable and may be passed to multiple devices easily.

Examples for those are e.g. the LibinputInputBackend (its context might be initialized through a Session via the LibinputSessionInterface) or the UdevBackend.

Usage of the session notifier

The notifier might be used to pause device access, when the session gets paused (e.g. by switching the tty via LogindSession::change_vt) and to automatically enable it again, when the session becomes active again.

It is crutial to avoid errors during that state. Examples for object that might be registered for notifications are the Libinput context, the UdevBackend or a DrmDevice (handled automatically by the UdevBackend, if not done manually).

Structs

BoundLogindSession
Bound logind session that is driven by the wayland_server::EventLoop.

Error
The Error type.

Id
Ids of registered SessionObservers of the LogindSessionNotifier

LogindSession
Session via the logind dbus interface

LogindSessionNotifier
SessionNotifier via the logind dbus interface

Enums

ErrorKind
The kind of an error.

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

logind_session_bind
Bind a LogindSessionNotifier to an EventLoop.

Type Definitions

Result
Convenient wrapper around std::Result.