smithay::backend::session

Trait Session

pub trait Session {
    type Error: AsErrno;

    // Required methods
    fn open(
        &mut self,
        path: &Path,
        flags: OFlags,
    ) -> Result<OwnedFd, Self::Error>;
    fn close(&mut self, fd: OwnedFd) -> Result<(), Self::Error>;
    fn change_vt(&mut self, vt: i32) -> Result<(), Self::Error>;
    fn is_active(&self) -> bool;
    fn seat(&self) -> String;
}

Available on crate feature backend_session only.

General session interface.

Provides a way to open and close devices and change the active vt.

Required Associated Types

type Error: AsErrno

Error type of the implementation

Required Methods

fn open(&mut self, path: &Path, flags: OFlags) -> Result<OwnedFd, Self::Error>

Opens a device at the given path with the given flags.

Returns a raw file descriptor

fn close(&mut self, fd: OwnedFd) -> Result<(), Self::Error>

Close a previously opened file descriptor

fn change_vt(&mut self, vt: i32) -> Result<(), Self::Error>

Change the currently active virtual terminal

fn is_active(&self) -> bool

Check if this session is currently active

fn seat(&self) -> String

Which seat this session is on