Trait smithay::backend::session::SessionNotifier

pub trait SessionNotifier {
    type Id: PartialEq + Eq;
    fn register<S: SessionObserver + 'static>(&mut self, signal: S) -> Self::Id;

    fn unregister(&mut self, signal: Self::Id);
}

Interface for registering for notifications for a given session.

Part of the session API which allows to get notified, when the given session gets paused or becomes active again. Any object implementing the [SessionObserver] trait may be registered.

Associated Types

type Id: PartialEq + Eq

Id type of registered observers

Loading content...
Required methods

fn register<S: SessionObserver + 'static>(&mut self, signal: S) -> Self::Id

Registers a given [SessionObserver].

Returns an id of the inserted observer, can be used to remove it again.

fn unregister(&mut self, signal: Self::Id)

Removes an observer by its given id from [SessionNotifier::register].

Loading content...