Module smithay::backend::udev

Provides udev related functionality for automated device scanning.

This module mainly provides the UdevBackend, which monitors available DRM devices and acts as an event source, generating events whenever these devices change.

Note: Once inserted into the event loop, the UdevBackend will only notify you about changes in the device list. To get an initial snapshot of the state during your initialization, you need to call its device_list method.

use smithay::backend::udev::{UdevBackend, UdevEvent};

let udev = UdevBackend::new("seat0", None).expect("Failed to monitor udev.");

for (dev_id, node_path) in udev.device_list() {
    // process the initial list of devices
}

// setup the event source for long-term monitoring
loop_handle.insert_source(udev, |event, _, _dispatch_data| match event {
    UdevEvent::Added { device_id, path } => {
        // a new device has been added
    },
    UdevEvent::Changed { device_id } => {
        // a device has been changed
    },
    UdevEvent::Removed { device_id } => {
        // a device has been removed
    }
}).expect("Failed to insert the udev source into the event loop");

Additionally this contains some utility functions related to scanning.

See also anvil/src/udev.rs for pure hardware backed example of a compositor utilizing this backend.

Structs

UdevBackend
Backend to monitor available drm devices.

Enums

UdevEvent
Events generated by the [UdevBackend], notifying you of changes in system devices

Functions

all_gpus
Returns the paths of all available GPU devices

primary_gpu
Returns the path of the primary GPU device if any