Module smithay::backend::udev

Provides udev related functionality for automated device scanning.

This module mainly provides the UdevBackend, which constantly monitors available DRM devices and notifies a user supplied UdevHandler of any changes.

Additionally this contains some utility functions related to scanning.

See also anvil/src/udev.rs for pure hardware backed example of a compositor utilizing this backend.

Structs

UdevBackend
Backend to monitor available drm devices.

Traits

UdevHandler
Handler for the [UdevBackend], allows to open, close and update drm devices as they change during runtime.

Functions

all_gpus
Returns the paths of all available GPU devices

primary_gpu
Returns the path of the primary GPU device if any

udev_backend_bind
Binds a [UdevBackend] to a given EventLoop.

Type Definitions

UdevSource
calloop event source associated with the Udev backend