Struct smithay::backend::udev::UdevBackend

pub struct UdevBackend { /* private fields */ }

Backend to monitor available drm devices.

Provides a way to automatically scan for available gpus and notifies the given handler of any changes. Can be used to provide hot-plug functionality for gpus and attached monitors.

Implementations

impl UdevBackend

pub fn new<S: AsRef<str>>(seat: S) -> Result<UdevBackend>

Creates a new UdevBackend

Arguments

seat - system seat which should be bound

pub fn device_list(&self) -> impl Iterator<Item = (dev_t, &Path)>

Get a list of DRM devices currently known to the backend

You should call this once before inserting the event source into your event loop, to get an initial snapshot of the device state.