Trait smithay::backend::udev::UdevHandler

pub trait UdevHandler {
    fn device_added(&mut self, device: dev_t, path: PathBuf);

    fn device_changed(&mut self, device: dev_t);

    fn device_removed(&mut self, device: dev_t);
}

Handler for the [UdevBackend], allows to open, close and update drm devices as they change during runtime.

Required methods

fn device_added(&mut self, device: dev_t, path: PathBuf)

Called when a new device is detected.

fn device_changed(&mut self, device: dev_t)

Called when an open device is changed.

This usually indicates that some connectors did become available or were unplugged. The handler should scan again for connected monitors and mode switch accordingly.

fn device_removed(&mut self, device: dev_t)

Called when a device was removed.

Loading content...