smithay::backend

Module vulkan

Available on crate feature backend_vulkan only.

Types to initialize and use Vulkan.

This module provides some thin abstractions over ash for initializing Vulkan.

This module does not provide abstractions for logical devices, rendering or memory allocation. These should instead be provided in higher level abstractions.

Smithay requires at least Vulkan 1.11.

Instance

To use Vulkan, you would first instantiate an Instance. An instance is effectively the Vulkan library and provides some information about the environment. This includes the list of supported instance extensions and the list of available physical devices.

An instance is constructed using an Instance::new or Instance::with_extensions.

Layers

The validation layers will be enabled if debug assertions are enabled and the validation layers are available on your system.

Instance extensions

Instances may be created with some enabled extensions. Note that any features gated by an extension are only available if the extension (and it’s dependencies) are enabled (you will get a validation error if you ignore that).

Some features such as window system integration are only available if their features are enabled. Available instance extensions may be obtained using Instance::enumerate_extensions.

PhysicalDevice

Once you have an instance, you may want to find a suitable device to use. A PhysicalDevice describes a Vulkan implementation that may correspond to a real or virtual device.

To get all the available devices, use PhysicalDevice::enumerate.

Physical devices are also describe the logical devices that can be created. A physical device can describe a variety of properties that may be used for device selection, including but not limited to:

Device name

Supported Vulkan API version

Type of the device

Driver information

Extensions

Features and limits

Physical devices implement Eq2, meaning two physical devices can be tested for equality.

Device extensions

Depending on the device extension (see the Vulkan specification), a device extension may indicate some physical device data is available or indicate some device feature is supported. Any features that are added using a device extension must be enabled in order to be used.

Internally Vulkan 1.1 is required because several extensions that were made part of core are used quite extensively in some of the extensions our abstractions use. The vast majority of systems using Vulkan also support at least Vulkan 1.1. If you need Vulkan 1.0 support, please open an issue and we can discuss Vulkan 1.0 support. ↩

Two physical devices are only equal if both physical devices are created from the same instance and have the same physical device handle. ↩

Modules

version

The Version type.

Structs

AppInfo

App info to be passed to the Vulkan implementation.

DriverInfo

Information about the driver providing a PhysicalDevice.

Instance

A Vulkan instance.

LoadError

Error loading the Vulkan library

PhysicalDevice

A Vulkan physical device.

Enums

InstanceError

An error that may occur when creating an Instance.

UnsupportedProperty

Error returned when a physical device property is not supported