Function smithay::backend::winit::init_from_builder_with_gl_attr

pub fn init_from_builder_with_gl_attr<L>(
    builder: WindowBuilder, 
    attributes: GlAttributes, 
    logger: L
) -> Result<(WinitGraphicsBackend, WinitEventLoop), Error> where
    L: Into<Option<Logger>>, 

Create a new WinitGraphicsBackend, which implements the Renderer trait, from a given WindowBuilder struct, as well as given GlAttributes for further customization of the rendering pipeline and a corresponding WinitEventLoop.