Module smithay::backend::x11

Implementation of the backend types using X11.

This backend provides the appropriate backend implementations to run a Wayland compositor as an X11 client.

The backend is initialized using X11Backend::new. The function will return two objects:

an X11Backend, which you will insert into an EventLoop to process events from the backend.

an X11Surface, which represents a surface that buffers are presented to for display.

Example usage

fn init_x11_backend(
   handle: calloop::LoopHandle<CompositorState>,
   logger: slog::Logger
) -> Result<(), Box<dyn Error>> {
    // Create the backend, also yielding a surface that may be used to render to the window.
    let (backend, surface) = X11Backend::new(logger)?;
    // You can get a handle to the window the backend has created for later use.
    let window = backend.window();

    // Insert the backend into the event loop to receive events.
    handle.insert_source(backend, |event, _window, state| {
        // Process events from the X server that apply to the window.
    })?;

    Ok(())
}

EGL

When using EGL, an X11Surface may be used to create an EGLDisplay.

Structs

Present

An RAII scope containing the next buffer that will be presented to the window. Presentation occurs when the Present is dropped.

Window

An X11 window.

X11Backend

Represents an active connection to the X to manage events on the Window provided by the backend.

X11Input

Marker used to define the InputBackend types for the X11 backend.

X11KeyboardInputEvent

X11-Backend internal event wrapping X11’s types into a KeyboardKeyEvent.

X11MouseInputEvent

X11-Backend internal event wrapping X11’s types into a PointerButtonEvent

X11MouseMovedEvent

X11-Backend internal event wrapping X11’s types into a PointerMotionAbsoluteEvent

X11MouseWheelEvent

X11-Backend internal event wrapping X11’s types into a PointerAxisEvent

X11Surface

An X11 surface which uses GBM to allocate and present buffers.

X11VirtualDevice

Virtual input device used by the backend to associate input events.

Enums

AllocateBuffersError

An error which may occur when allocating buffers for presentation to the window.

CreateWindowError

An error which may occur when creating an X11 window.

MissingExtensionError

An error that occurs when a required X11 extension is not present.

X11Error

An error emitted by the X11 backend during setup.

X11Event

An event emitted by the X11 backend.