smithay::desktop

Module space

Available on crate feature desktop only.

This module contains the Space helper class as well has related rendering helpers to add custom elements or different clients to a space.

Macros

space_elements

Aggregate multiple types implementing SpaceElement into a single enum type to be used with a Space.

Structs

ConstrainBehavior

Defines the behavior for constrain_space_element

Space

Represents two dimensional plane to map windows and outputs upon.

SurfaceTreewayland_frontend

A custom surface tree

Enums

ConstrainReference

Defines the reference size for the constrain behavior.

OutputError

Errors thrown by Space::elements_for_output

RenderZindex

Indicates default values for some zindexs inside smithay

SpaceRenderElements

Defines the render elements used internally by a Space

Traits

SpaceElement

Element mappable onto a Space

Functions

constrain_space_element

Constrain the render elements of a SpaceElement

render_output

Render a output

space_render_elements

Get the render elements for a specific output