Module smithay::desktop::space

This module contains the Space helper class as well has related rendering helpers to add custom elements or different clients to a space.

Structs

Space

Represents two dimensional plane to map windows and outputs upon.

SpaceOutputHash

Type to use as an owned hashable value equal to SpaceOutputTuple

SpaceOutputTuple

Newtype for (&Space, &Output) to provide a Hash implementation for damage tracking

SurfaceTree

Generic helper for drawing WlSurfaces and their subsurfaces as custom elements via RenderElement.

Enums

RenderError

Errors thrown by Space::render_output

Traits

RenderElement

Trait for custom elements to be rendered during Space::render_output.

Type Definitions

DynamicRenderElements

Elements rendered by Space::render_output in addition to windows, layers and popups.