Struct smithay::desktop::space::Space

pub struct Space<E: SpaceElement> { /* private fields */ }

Represents two dimensional plane to map windows and outputs upon.

Space is generic over the types of elements mapped onto it. The simplest usecase is a Space<Window>, but other types can be used by implementing SpaceElement. Multiple types might be quickly aggregated into an enum by using the space_elements!-macro.

Implementations

impl<E: SpaceElement + PartialEq> Space<E>

pub fn id(&self) -> usize

Gets the id of this space

pub fn map_element<P>(&mut self, element: E, location: P, activate: bool)where P: Into<Point<i32, Logical>>,

Map a SpaceElement and move it to top of the stack

This can safely be called on an already mapped window to update its location inside the space.

If activate is true it will set the new windows state to be activate and removes that state from every other mapped window.

pub fn raise_element(&mut self, element: &E, activate: bool)

Moves an already mapped SpaceElement to top of the stack

This function does nothing for unmapped windows.

If activate is true it will set the new windows state to be activate and removes that state from every other mapped window.

pub fn unmap_elem(&mut self, element: &E)

Unmap a SpaceElement from this space.

This function does nothing for already unmapped windows

pub fn elements(&self) -> impl DoubleEndedIterator<Item = &E>

Iterate elements in z-order back to front

pub fn elements_for_output<'output>( &'output self, output: &'output Output ) -> impl DoubleEndedIterator<Item = &'output E>

Iterate elements on a specific output in z-order back to front

pub fn element_under<P: Into<Point<f64, Logical>>>( &self, point: P ) -> Option<(&E, Point<i32, Logical>)>

Finds the topmost element under this point if any and returns it together with the location of this element relative to this space.

This is equivalent to iterating the elements in the space from top to bottom and testing if the point is within the elements input region and returning the first matching one.

Note that SpaceElement::is_in_input_region expects the point to be relative to the elements origin.

pub fn output_under<P: Into<Point<f64, Logical>>>( &self, point: P ) -> impl Iterator<Item = &Output>

Get a reference to the outputs under a given point

pub fn layer_for_surface( &self, surface: &WlSurface, surface_type: WindowSurfaceType ) -> Option<LayerSurface>

Returns the layer surface matching a given surface, if any

surface_type can be used to limit the types of surfaces queried for equality.

pub fn element_location(&self, elem: &E) -> Option<Point<i32, Logical>>

Returns the location of a SpaceElement inside the Space.

pub fn element_bbox(&self, elem: &E) -> Option<Rectangle<i32, Logical>>

Returns the bounding box of a SpaceElement including its relative position inside the Space.

pub fn element_geometry(&self, elem: &E) -> Option<Rectangle<i32, Logical>>

Returns the geometry of a SpaceElement including its relative position inside the Space.

This area is usually defined as the contents of the window, excluding decorations.

pub fn map_output<P: Into<Point<i32, Logical>>>( &mut self, output: &Output, location: P )

Maps an Output inside the space.

Can be safely called on an already mapped Output to update its location.

Note: Remapping an output does reset it’s damage memory.

pub fn outputs(&self) -> impl Iterator<Item = &Output>

Iterate over all mapped Outputs of this space.

pub fn unmap_output(&mut self, output: &Output)

Unmap an Output from this space.

Does nothing if the output was not previously mapped.

pub fn output_geometry(&self, o: &Output) -> Option<Rectangle<i32, Logical>>

Returns the geometry of the output including it’s relative position inside the space.

The size is matching the amount of logical pixels of the space visible on the output given is current mode and scale.

pub fn outputs_for_element(&self, elem: &E) -> Vec<Output>

Returns all Outputs a SpaceElement overlaps with.

pub fn refresh(&mut self)

Refresh some internal values and update client state, meaning this will handle output enter and leave events for mapped outputs and windows based on their position.

Needs to be called periodically, at best before every wayland socket flush.

pub fn render_elements_for_region<'a, R: Renderer, S: Into<Scale<f64>>>( &'a self, renderer: &mut R, region: &Rectangle<i32, Logical>, scale: S, alpha: f32 ) -> Vec<<E as AsRenderElements<R>>::RenderElement>where <R as Renderer>::TextureId: Texture + 'static, E: AsRenderElements<R>, <E as AsRenderElements<R>>::RenderElement: 'a,

Retrieve the render elements for a given region of the space.

Note: Because this is not rendering a specific output, this will not contain layer surfaces. Use Space::render_elements_for_output, if you care about this.

pub fn render_elements_for_output<'a, R: Renderer + ImportAll>( &'a self, renderer: &mut R, output: &Output, alpha: f32 ) -> Result<Vec<SpaceRenderElements<R, <E as AsRenderElements<R>>::RenderElement>>, OutputError>where <R as Renderer>::TextureId: Texture + 'static, E: AsRenderElements<R>, <E as AsRenderElements<R>>::RenderElement: 'a, SpaceRenderElements<R, <E as AsRenderElements<R>>::RenderElement>: From<Wrap<<E as AsRenderElements<R>>::RenderElement>>,

Retrieve the render elements for an output