Struct smithay::desktop::space::Space

pub struct Space { /* private fields */ }

Represents two dimensional plane to map windows and outputs upon.

Implementations

impl Space

pub fn new<L>(log: L) -> Space where
L: Into<Option<Logger>>,

Create a new Space

pub fn map_window<P: Into<Point<i32, Logical>>>(
&mut self,
window: &Window,
location: P,
activate: bool
)

Map a Window and move it to top of the stack

This can safely be called on an already mapped window to update its location inside the space.

If activate is true it will set the new windows state to be activate and removes that state from every other mapped window.

pub fn raise_window(&mut self, window: &Window, activate: bool)

Moves an already mapped Window to top of the stack

This function does nothing for unmapped windows.

If activate is true it will set the new windows state to be activate and removes that state from every other mapped window.

pub fn unmap_window(&mut self, window: &Window)

Unmap a Window from this space.

This function does nothing for already unmapped windows

pub fn windows(&self) -> impl DoubleEndedIterator<Item = &Window>

Iterate window in z-order back to front

pub fn window_under<P: Into<Point<f64, Logical>>>(
&self,
point: P
) -> Option<&Window>

Get a reference to the window under a given point, if any

pub fn output_under<P: Into<Point<f64, Logical>>>(
&self,
point: P
) -> impl Iterator<Item = &Output>

Get a reference to the outputs under a given point

pub fn window_for_surface(&self, surface: &WlSurface) -> Option<&Window>

Returns the window matching a given surface, if any

pub fn layer_for_surface(&self, surface: &WlSurface) -> Option<LayerSurface>

Returns the layer matching a given surface, if any

pub fn window_geometry(&self, w: &Window) -> Option<Rectangle<i32, Logical>>

Returns the geometry of a Window including its relative position inside the Space.

pub fn window_bbox(&self, w: &Window) -> Option<Rectangle<i32, Logical>>

Returns the bounding box of a Window including its relative position inside the Space.

pub fn map_output<P: Into<Point<i32, Logical>>>(
&mut self,
output: &Output,
scale: f64,
location: P
)

Maps an Output inside the space.

Can be safely called on an already mapped Output to update its scale or location.

The scale is the what is rendered for the given output and may be fractional. It is independent from the integer scale reported to clients by the output.

Note: Remapping an output does reset it’s damage memory.

pub fn outputs(&self) -> impl Iterator<Item = &Output>

Iterate over all mapped Outputs of this space.

pub fn unmap_output(&mut self, output: &Output)

Unmap an Output from this space.

Does nothing if the output was not previously mapped.

pub fn output_geometry(&self, o: &Output) -> Option<Rectangle<i32, Logical>>

Returns the geometry of the output including it’s relative position inside the space.

The size is matching the amount of logical pixels of the space visible on the output given is current mode and render scale.

pub fn output_scale(&self, o: &Output) -> Option<f64>

Returns the reder scale of a mapped output.

If the output was not previously mapped to the Space this function returns None.

pub fn outputs_for_window(&self, w: &Window) -> Vec<Output>

Returns all Outputs a Window overlaps with.

pub fn refresh(&mut self)

Refresh some internal values and update client state, meaning this will handle output enter and leave events for mapped outputs and windows based on their position.

Needs to be called periodically, at best before every wayland socket flush.

pub fn commit(&self, surface: &WlSurface)

Should be called on commit to let the space automatically call Window::refresh for the window that belongs to the given surface, if managed by this space.

pub fn render_output<R>(
&mut self,
renderer: &mut R,
output: &Output,
age: usize,
clear_color: [f32; 4],
custom_elements: &[DynamicRenderElements<R>]
) -> Result<Option<Vec<Rectangle<i32, Logical>>>, RenderError<R>> where
R: Renderer + ImportAll + 'static,
R::TextureId: 'static,
R::Error: 'static,
R::Frame: 'static,

Render a given Output using a given Renderer.

Space will render all mapped Windows, mapped LayerSurfaces of the given Output and their popups (if tracked by a PopupManager). clear_color will be used to fill all unoccupied regions.

Rendering using this function will automatically apply damage-tracking. To facilitate this you need to provide age values of the buffers bound to the given renderer. If you stop using Space temporarily for rendering or apply additional rendering operations, you need to reset the age values accordingly as Space will be unable to track your custom rendering operations to avoid rendering artifacts.

To add aditional elements without breaking damage-tracking implement the RenderElement trait and use custom_elements to provide them to this function. `custom_elements are rendered after every other element.

Returns a list of updated regions (or None if that list would be empty) in case of success.

pub fn send_frames(&self, all: bool, time: u32)

Sends the frame callback to mapped Windows and LayerSurfaces.

If all is set this will be send to all mapped surfaces. Otherwise only windows and layers previously drawn during the previous frame will be send frame events.