Trait smithay::desktop::space::RenderElement

pub trait RenderElement<R, F, E, T> where
    R: Renderer<Error = E, TextureId = T, Frame = F> + ImportAll,
    F: Frame<Error = E, TextureId = T>,
    E: Error,
    T: Texture + 'static,
    Self: Any + 'static, {
    fn id(&self) -> usize;

    fn geometry(&self) -> Rectangle<i32, Logical>;

    fn accumulated_damage(
        &self, 
        for_values: Option<SpaceOutputTuple<'_, '_>>
    ) -> Vec<Rectangle<i32, Logical>>;

    fn draw(
        &self, 
        renderer: &mut R, 
        frame: &mut F, 
        scale: f64, 
        damage: &[Rectangle<i32, Logical>], 
        log: &Logger
    ) -> Result<(), R::Error>;
}

Trait for custom elements to be rendered during Space::render_output.

Required methods

fn id(&self) -> usize

Returns an id unique to this element for the type of Self.

fn geometry(&self) -> Rectangle<i32, Logical>

Returns the bounding box of this element including its position in the space.

fn accumulated_damage(
&self,
for_values: Option<SpaceOutputTuple<'_, '_>>
) -> Vec<Rectangle<i32, Logical>>

Returns the damage of the element since it’s last update.

If you receive Some(_) for for_values you may cache that you send the damage for this Space and Output combination once and return an empty vector for subsequent calls until the contents of this element actually change again for optimization reasons.

Returning vec![Rectangle::from_loc_and_size((0, 0), (i32::MAX, i32::MAX))] is always correct, but very inefficient.

fn draw(
&self,
renderer: &mut R,
frame: &mut F,
scale: f64,
damage: &[Rectangle<i32, Logical>],
log: &Logger
) -> Result<(), R::Error>

Draws the element using the provided Frame and Renderer.

scale provides the current fractional scale value to render as

damage provides the regions you need to re-draw and may not be equivalent to the damage returned by accumulated_damage. Redrawing other parts of the element is not valid and may cause rendering artifacts.