smithay::desktop::space

Trait SpaceElement

pub trait SpaceElement: IsAlive {
    // Required methods
    fn bbox(&self) -> Rectangle<i32, Logical>;
    fn is_in_input_region(&self, point: &Point<f64, Logical>) -> bool;
    fn set_activate(&self, activated: bool);
    fn output_enter(&self, output: &Output, overlap: Rectangle<i32, Logical>);
    fn output_leave(&self, output: &Output);

    // Provided methods
    fn geometry(&self) -> Rectangle<i32, Logical> { ... }
    fn z_index(&self) -> u8 { ... }
    fn refresh(&self) { ... }
}

Available on crate feature desktop only.

Element mappable onto a Space

Required Methods

fn bbox(&self) -> Rectangle<i32, Logical>

Returns the bounding box of this element

fn is_in_input_region(&self, point: &Point<f64, Logical>) -> bool

Returns whenever a given point inside this element will be able to receive input

fn set_activate(&self, activated: bool)

Set the rendered state to activated, if applicable to this element

fn output_enter(&self, output: &Output, overlap: Rectangle<i32, Logical>)

The element is displayed on a given output.

Maybe called for an already entered output, if the overlap changes

fn output_leave(&self, output: &Output)

The element left a given output

Provided Methods

fn geometry(&self) -> Rectangle<i32, Logical>

Returns the geometry of this element.

Defaults to be equal to it’s bounding box.

fn z_index(&self) -> u8

Gets the z-index of this element

fn refresh(&self)

Periodically called to update internal state, if necessary