smithay::desktop

Struct Window

pub struct Window(/* private fields */);

Available on crate features wayland_frontend and desktop only.

Represents a single application window

Implementations

impl Window

pub fn downgrade(&self) -> WeakWindow

Create a weak reference to this window that does not prevent the inner resources from being dropped.

impl Window

pub fn new(toplevel: ToplevelSurface) -> Window

๐Deprecated

Construct a new Window from a xdg toplevel surface

This function is deprecated. Use Window::new_wayland_window instead.

pub fn new_wayland_window(toplevel: ToplevelSurface) -> Window

Construct a new Window from a xdg toplevel surface

pub fn new_x11_window(surface: X11Surface) -> Window

Available on crate feature xwayland only.

Construct a new Window from an X11 surface

pub fn is_wayland(&self) -> bool

Checks if the window is a wayland one.

pub fn is_x11(&self) -> bool

Available on crate feature xwayland only.

Checks if the window is an X11 one.

pub fn geometry(&self) -> Rectangle<i32, Logical>

Returns the geometry of this window.

pub fn bbox(&self) -> Rectangle<i32, Logical>

Returns a bounding box over this window and its children.

pub fn bbox_with_popups(&self) -> Rectangle<i32, Logical>

Returns a bounding box over this window and children including popups.

Note: You need to use a PopupManager to track popups, otherwise the bounding box will not include the popups.

pub fn set_activated(&self, active: bool) -> bool

Activate/Deactivate this window

pub fn send_frame<T, F>( &self, output: &Output, time: T, throttle: Option<Duration>, primary_scan_out_output: F, )
where T: Into<Duration>, F: FnMut(&WlSurface, &SurfaceData) -> Option<Output> + Copy,

Sends the frame callback to all the subsurfaces in this window that requested it

See send_frames_surface_tree for more information

pub fn send_dmabuf_feedback<'a, P, F>( &self, output: &Output, primary_scan_out_output: P, select_dmabuf_feedback: F, )
where P: FnMut(&WlSurface, &SurfaceData) -> Option<Output> + Copy, F: Fn(&WlSurface, &SurfaceData) -> &'a DmabufFeedback + Copy,

Sends the dmabuf feedback to all the subsurfaces in this window that requested it

See send_dmabuf_feedback_surface_tree for more information

pub fn take_presentation_feedback<F1, F2>( &self, output_feedback: &mut OutputPresentationFeedback, primary_scan_out_output: F1, presentation_feedback_flags: F2, )
where F1: FnMut(&WlSurface, &SurfaceData) -> Option<Output> + Copy, F2: FnMut(&WlSurface, &SurfaceData) -> Kind + Copy,

Takes the PresentationFeedbackCallbacks from all subsurfaces in this window

see take_presentation_feedback_surface_tree for more information

pub fn with_surfaces<F>(&self, processor: F)
where F: FnMut(&WlSurface, &SurfaceData),

Run a closure on all surfaces in this window (including itโs popups, if PopupManager is used)

pub fn on_commit(&self)

Updates internal values

Needs to be called whenever the toplevel surface or any unsynchronized subsurfaces of this window are updated to correctly update the bounding box of this window.

pub fn surface_under<P: Into<Point<f64, Logical>>>( &self, point: P, surface_type: WindowSurfaceType, ) -> Option<(WlSurface, Point<i32, Logical>)>

Finds the topmost surface under this point matching the input regions of the surface and returns it together with the location of this surface.

In case no surface input region matches the point None is returned.

point should be relative to (0,0) of the window.

pub fn toplevel(&self) -> Option<&ToplevelSurface>

Returns the underlying xdg toplevel surface

pub fn x11_surface(&self) -> Option<&X11Surface>

Available on crate feature xwayland only.

Returns the underlying X11 surface

pub fn underlying_surface(&self) -> &WindowSurface

Returns the underlying surface

pub fn override_z_index(&self, z_index: u8)

Override the z_index of this Window

pub fn user_data(&self) -> &UserDataMap

Returns a UserDataMap to allow associating arbitrary data with this window.