Module smithay::desktop::utils

Helper functions to ease dealing with surface trees

Structs

OutputPresentationFeedback

Stores the SurfacePresentationFeedback for a specific output

SurfacePresentationFeedback

Holds the presentation feedback for a surface

Functions

bbox_from_surface_tree

Returns the bounding box of a given surface and all its subsurfaces.

send_frames_surface_tree

Sends frame callbacks for a surface and its subsurfaces with the given time.

surface_presentation_feedback_flags_from_states

Retrieves the per surface [wp_presentation_feedback::Kind] flags

surface_primary_scanout_output

Retrieve a previously stored primary scan-out output from a surface

take_presentation_feedback_surface_tree

Takes the PresentationFeedbackCallbacks from the surface tree

under_from_surface_tree

Returns the topmost (sub-)surface under a given position matching the input regions of the surface.

update_surface_primary_scanout_output

Update the surface primary scan-out output from a output render report

with_surfaces_surface_tree

Run a closure on all surfaces of a surface tree