smithay

Module input

Input abstractions

This module provides some types loosely resembling instances of wayland seats, pointers, touch and keyboards. It is however not directly tied to wayland and can be used to multiplex various input operations between different handlers.

If the wayland_frontend-feature is enabled the smithay::wayland::seat-module provides additional functionality for the provided types of this module to map them to advertised wayland globals and objects.

How to use it

To start using this module you need to create a SeatState and use that to create Seats. Additionally you need to implement the SeatHandler trait.

Initialization

use smithay::input::{Seat, SeatState, SeatHandler, pointer::CursorImageStatus};

struct State {
    seat_state: SeatState<Self>,
    // ...
};

let mut seat_state = SeatState::<State>::new();

// create the seat
let seat = seat_state.new_seat(
    "seat-0",  // the name of the seat, will be advertized to clients
);

// implement the required traits
impl SeatHandler for State {
    type KeyboardFocus = Target;
    type PointerFocus = Target;
    type TouchFocus = Target;

    fn seat_state(&mut self) -> &mut SeatState<Self> {
        &mut self.seat_state
    }

    fn focus_changed(&mut self, seat: &Seat<Self>, focused: Option<&Target>) {
        // handle focus changes, if you need to ...
    }
    fn cursor_image(&mut self, seat: &Seat<Self>, image: CursorImageStatus) {
        // handle new images for the cursor ...
    }
}

Run usage

Once the seat is initialized, you can add capabilities to it.

Currently, pointer, touch and keyboard capabilities are supported by this module. tablet_manager also provides client interaction for drawing tablets.

You can add these capabilities via methods of the Seat struct: Seat::add_keyboard, Seat::add_pointer and Seat::add_touch. These methods return handles that can be cloned and sent across thread, so you can keep one around in your event-handling code to forward inputs to your clients.

Modules

dnd

Drag’n’Drop related types for smithay’s input-abstractions

keyboard

Keyboard-related types for smithay’s input abstraction

pointer

Pointer-related types for smithay’s input abstraction

touch

Touch-related types for smithay’s input abstraction

Structs

Seat

A Seat handle

SeatState

Delegate type for all Seat globals.

WeakSeat

Weak variant of an Seat

Traits

SeatHandler

Handler trait for Seats