Module smithay::input::keyboard

Keyboard-related types for smithay’s input abstraction

Modules

keysyms

Structs

GrabStartData

Data about the event that started the grab.

KeyboardHandle

An handle to a keyboard handler

KeyboardInnerHandle

This inner handle is accessed from inside a keyboard grab logic, and directly sends event to the client

KeysymHandle

Handle to the underlying keycode to allow for different conversions

ModifiersState

Represents the current state of the keyboard modifiers

XkbConfig

Configuration for xkbcommon.

Enums

Error

Errors that can be encountered when creating a keyboard handler

FilterResult

Result for key input filtering (see KeyboardHandle::input)

Traits

KeyboardGrab

A trait to implement a keyboard grab

KeyboardTarget

Trait representing object that can receive keyboard interactions

Type Definitions

Keysym

A number used to represent the symbols generated from a key on a keyboard.