smithay::input
Module keyboard

Keyboard-related types for smithay’s input abstraction

Modules

keysyms

xkb

Structs

GrabStartData

Data about the event that started the grab.

KeyboardHandle

An handle to a keyboard handler

KeyboardInnerHandle

This inner handle is accessed from inside a keyboard grab logic, and directly sends event to the client

Keycode

The keyboard code, often corresponding to a physical key.

KeymapFilewayland_frontend

Wraps an XKB keymap into a sealed file or stores as just a string for sending to WlKeyboard over an fd

Keysym

The keyboard symbol, often corresponding to a character.

KeysymHandle

Handle to the underlying keycode to allow for different conversions

Layout

Reference to the XkbLayout in the active keymap.

LedMapping

Mapping of the led of a keymap

LedState

Current state of the led when available

ModifiersState

Represents the current state of the keyboard modifiers

SerializedMods

Serialized modifier state

Xkb

An xkbcommon context, keymap, and state, that can be sent to another thread, but should not have additional ref-counts kept on one thread.

XkbConfig

Configuration for xkbcommon.

XkbContext

The currently active state of the Xkb.

Enums

Error

Errors that can be encountered when creating a keyboard handler

FilterResult

Result for key input filtering (see KeyboardHandle::input)

Traits

KeyboardGrab

A trait to implement a keyboard grab

KeyboardTarget

Trait representing object that can receive keyboard interactions