smithay::input::keyboard

Trait KeyboardGrab

pub trait KeyboardGrab<D: SeatHandler>: Downcast {
    // Required methods
    fn input(
        &mut self,
        data: &mut D,
        handle: &mut KeyboardInnerHandle<'_, D>,
        keycode: Keycode,
        state: KeyState,
        modifiers: Option<ModifiersState>,
        serial: Serial,
        time: u32,
    );
    fn set_focus(
        &mut self,
        data: &mut D,
        handle: &mut KeyboardInnerHandle<'_, D>,
        focus: Option<<D as SeatHandler>::KeyboardFocus>,
        serial: Serial,
    );
    fn start_data(&self) -> &GrabStartData<D>;
    fn unset(&mut self, data: &mut D);
}

A trait to implement a keyboard grab

In some context, it is necessary to temporarily change the behavior of the keyboard. This is typically known as a keyboard grab. A example would be, during a popup grab the keyboard focus will not be changed and stay on the grabbed popup.

This trait is the interface to intercept regular keyboard events and change them as needed, its interface mimics the KeyboardHandle interface.

If your logic decides that the grab should end, both KeyboardInnerHandle and KeyboardHandle have a method to change it.

When your grab ends (either as you requested it or if it was forcefully cancelled by the server), the struct implementing this trait will be dropped. As such you should put clean-up logic in the destructor, rather than trying to guess when the grab will end.

Required Methods

fn input( &mut self, data: &mut D, handle: &mut KeyboardInnerHandle<'_, D>, keycode: Keycode, state: KeyState, modifiers: Option<ModifiersState>, serial: Serial, time: u32, )

An input was reported.

modifiers are only passed when their state actually changes. The modifier must be sent after the key event.

fn set_focus( &mut self, data: &mut D, handle: &mut KeyboardInnerHandle<'_, D>, focus: Option<<D as SeatHandler>::KeyboardFocus>, serial: Serial, )

A focus change was requested.

fn start_data(&self) -> &GrabStartData<D>

The data about the event that started the grab.

fn unset(&mut self, data: &mut D)

The grab has been unset or replaced with another grab.

Implementations

impl<D> dyn KeyboardGrab<D>
where D: Any + 'static + SeatHandler,

pub fn is<__T: KeyboardGrab<D>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: KeyboardGrab<D>>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: KeyboardGrab<D>>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: KeyboardGrab<D>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: KeyboardGrab<D>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.