Module smithay::input::pointer

Pointer-related types for smithay’s input abstraction

Structs

AxisFrame

A frame of pointer axis events.

ButtonEvent

Pointer button event Mouse button click and release notifications. The location of the click is given by the last motion or enter event.

CursorImageAttributes

The role representing a surface set as the pointer cursor

GrabStartData

Data about the event that started the grab.

MotionEvent

Pointer motion event

PointerHandle

An handle to a pointer handler

PointerInnerHandle

This inner handle is accessed from inside a pointer grab logic, and directly sends event to the client

Enums

CursorImageStatus

Possible status of a cursor as requested by clients

Focus

Defines the focus behavior for PointerHandle::set_grab

Traits

PointerGrab

A trait to implement a pointer grab

PointerTarget

Trait representing object that can receive pointer interactions

Type Definitions

CursorImageSurfaceData

Data associated with XDG toplevel surface