smithay::input

Module pointer

Pointer-related types for smithay’s input abstraction

Structs

AxisFrame

A frame of pointer axis events. Frames of axis events should be considers as one logical action.

ButtonEvent

Pointer button event

ClickGrab

A click grab, basic grab started when an user clicks a surface to maintain it focused until the user releases the click.

CursorImageAttributes

The role representing a surface set as the pointer cursor

GestureHoldBeginEvent

Gesture hold begin event

GestureHoldEndEvent

Gesture hold end event

GesturePinchBeginEvent

Gesture pinch begin event

GesturePinchEndEvent

Gesture pinch end event

GesturePinchUpdateEvent

Gesture pinch update event

GestureSwipeBeginEvent

Gesture swipe begin event

GestureSwipeEndEvent

Gesture swipe end event

GestureSwipeUpdateEvent

Gesture swipe update event

GrabStartData

Data about the event that started the grab.

MotionEvent

Pointer motion event

PointerHandle

An handle to a pointer handler

PointerInnerHandle

This inner handle is accessed from inside a pointer grab logic, and directly sends event to the client

RelativeMotionEvent

Relative pointer motion event

Enums

CursorIcon

Describes the appearance of the (usually mouse) cursor icon.

CursorImageStatus

Possible status of a cursor as requested by clients

Focus

Defines the focus behavior for PointerHandle::set_grab

Traits

PointerGrab

A trait to implement a pointer grab

PointerTarget

Trait representing object that can receive pointer interactions

Type Aliases

CursorImageSurfaceData

Data associated with XDG toplevel surface