Struct smithay::input::pointer::PointerHandle

pub struct PointerHandle<D: SeatHandler> { /* private fields */ }

An handle to a pointer handler

It can be cloned and all clones manipulate the same internal state.

This handle gives you access to an interface to send pointer events to your clients.

When sending events using this handle, they will be intercepted by a pointer grab if any is active. See the PointerGrab trait for details.

Implementations

impl<D: SeatHandler + 'static> PointerHandle<D>

pub fn set_grab<G: PointerGrab<D> + 'static>( &self, data: &mut D, grab: G, serial: Serial, focus: Focus )

Change the current grab on this pointer to the provided grab

If focus is set to Focus::Clear any currently focused surface will be unfocused.

Overwrites any current grab.

pub fn unset_grab(&self, data: &mut D, serial: Serial, time: u32)

Remove any current grab on this pointer, resetting it to the default behavior

pub fn has_grab(&self, serial: Serial) -> bool

Check if this pointer is currently grabbed with this serial

pub fn is_grabbed(&self) -> bool

Check if this pointer is currently being grabbed

pub fn grab_start_data(&self) -> Option<GrabStartData<D>>

Returns the start data for the grab, if any.

pub fn motion( &self, data: &mut D, focus: Option<(<D as SeatHandler>::PointerFocus, Point<i32, Logical>)>, event: &MotionEvent )

Notify that the pointer moved

You provide the new location of the pointer, in the form of:

The coordinates of the pointer in the global compositor space

The surface on top of which the cursor is, and the coordinates of its origin in the global compositor space (or None of the pointer is not on top of a client surface).

This will internally take care of notifying the appropriate client objects of enter/motion/leave events.

pub fn relative_motion( &self, data: &mut D, focus: Option<(<D as SeatHandler>::PointerFocus, Point<i32, Logical>)>, event: &RelativeMotionEvent )

Notify about relative pointer motion

This will internally send the appropriate button event to the client objects matching with the currently focused surface, if the client uses the relative pointer protocol.

pub fn button(&self, data: &mut D, event: &ButtonEvent)

Notify that a button was pressed

This will internally send the appropriate button event to the client objects matching with the currently focused surface.

pub fn axis(&self, data: &mut D, details: AxisFrame)

Start an axis frame

A single frame will group multiple scroll events as if they happened in the same instance.

pub fn current_location(&self) -> Point<f64, Logical>

Access the current location of this pointer in the global space

impl<D> PointerHandle<D>where D: SeatHandler, <D as SeatHandler>::PointerFocus: Clone,

pub fn current_focus(&self) -> Option<<D as SeatHandler>::PointerFocus>

Retrieve the current pointer focus