smithay::input::pointer

Trait PointerGrab

pub trait PointerGrab<D: SeatHandler>: Send + Downcast {

    // Required methods
    fn motion(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        focus: Option<(<D as SeatHandler>::PointerFocus, Point<f64, Logical>)>,
        event: &MotionEvent,
    );
    fn relative_motion(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        focus: Option<(<D as SeatHandler>::PointerFocus, Point<f64, Logical>)>,
        event: &RelativeMotionEvent,
    );
    fn button(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &ButtonEvent,
    );
    fn axis(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        details: AxisFrame,
    );
    fn frame(&mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>);
    fn gesture_swipe_begin(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GestureSwipeBeginEvent,
    );
    fn gesture_swipe_update(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GestureSwipeUpdateEvent,
    );
    fn gesture_swipe_end(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GestureSwipeEndEvent,
    );
    fn gesture_pinch_begin(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GesturePinchBeginEvent,
    );
    fn gesture_pinch_update(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GesturePinchUpdateEvent,
    );
    fn gesture_pinch_end(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GesturePinchEndEvent,
    );
    fn gesture_hold_begin(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GestureHoldBeginEvent,
    );
    fn gesture_hold_end(
        &mut self,
        data: &mut D,
        handle: &mut PointerInnerHandle<'_, D>,
        event: &GestureHoldEndEvent,
    );
    fn start_data(&self) -> &GrabStartData<D>;
    fn unset(&mut self, data: &mut D);

}

A trait to implement a pointer grab

In some context, it is necessary to temporarily change the behavior of the pointer. This is typically known as a pointer grab. A typical example would be, during a drag’n’drop operation, the underlying surfaces will no longer receive classic pointer event, but rather special events.

This trait is the interface to intercept regular pointer events and change them as needed, its interface mimics the PointerHandle interface.

Any interactions with PointerHandle should be done using PointerInnerHandle, as handle is borrowed/locked before grab methods are called, so calling methods on PointerHandle would result in a deadlock.

If your logic decides that the grab should end, both PointerInnerHandle and PointerHandle have a method to change it.

When your grab ends (either as you requested it or if it was forcefully cancelled by the server), the struct implementing this trait will be dropped. As such you should put clean-up logic in the destructor, rather than trying to guess when the grab will end.

Required Methods

fn motion( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, focus: Option<(<D as SeatHandler>::PointerFocus, Point<f64, Logical>)>, event: &MotionEvent, )

A motion was reported

This method allows you attach additional behavior to a motion event, possibly altering it. You generally will want to invoke PointerInnerHandle::motion() as part of your processing. If you don’t, the rest of the compositor will behave as if the motion event never occurred.

Some grabs (such as drag’n’drop, shell resize and motion) unset the focus while they are active, this is achieved by just setting the focus to None when invoking PointerInnerHandle::motion().

fn relative_motion( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, focus: Option<(<D as SeatHandler>::PointerFocus, Point<f64, Logical>)>, event: &RelativeMotionEvent, )

Relative motion was reported

This method allows you attach additional behavior to a relative motion event, possibly altering it. You generally will want to invoke PointerInnerHandle::relative_motion() as part of your processing. If you don’t, the rest of the compositor will behave as if the motion event never occurred.

fn button( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &ButtonEvent, )

A button press was reported

This method allows you attach additional behavior to a button event, possibly altering it. You generally will want to invoke PointerInnerHandle::button() as part of your processing. If you don’t, the rest of the compositor will behave as if the button event never occurred.

fn axis( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, details: AxisFrame, )

An axis scroll was reported

This method allows you attach additional behavior to an axis event, possibly altering it. You generally will want to invoke PointerInnerHandle::axis() as part of your processing. If you don’t, the rest of the compositor will behave as if the axis event never occurred.

fn frame(&mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>)

End of a pointer frame

A frame groups associated events. This terminates the frame.

fn gesture_swipe_begin( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GestureSwipeBeginEvent, )

A pointer of a given seat started a swipe gesture

This method allows you attach additional behavior to a swipe gesture begin event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_swipe_begin() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_swipe_update( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GestureSwipeUpdateEvent, )

A pointer of a given seat updated a swipe gesture

This method allows you attach additional behavior to a swipe gesture update event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_swipe_update() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_swipe_end( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GestureSwipeEndEvent, )

A pointer of a given seat ended a swipe gesture

This method allows you attach additional behavior to a swipe gesture end event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_swipe_end() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_pinch_begin( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GesturePinchBeginEvent, )

A pointer of a given seat started a pinch gesture

This method allows you attach additional behavior to a pinch gesture begin event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_pinch_begin() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_pinch_update( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GesturePinchUpdateEvent, )

A pointer of a given seat updated a pinch gesture

This method allows you attach additional behavior to a pinch gesture update event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_pinch_update() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_pinch_end( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GesturePinchEndEvent, )

A pointer of a given seat ended a pinch gesture

This method allows you attach additional behavior to a pinch gesture end event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_pinch_end() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_hold_begin( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GestureHoldBeginEvent, )

A pointer of a given seat started a hold gesture

This method allows you attach additional behavior to a hold gesture begin event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_hold_begin() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn gesture_hold_end( &mut self, data: &mut D, handle: &mut PointerInnerHandle<'_, D>, event: &GestureHoldEndEvent, )

A pointer of a given seat ended a hold gesture

This method allows you attach additional behavior to a hold gesture end event, possibly altering it. You generally will want to invoke PointerInnerHandle::gesture_hold_end() as part of your processing. If you don’t, the rest of the compositor will behave as if the event never occurred.

fn start_data(&self) -> &GrabStartData<D>

The data about the event that started the grab.

fn unset(&mut self, data: &mut D)

The grab has been unset or replaced with another grab.

Implementations

impl<D> dyn PointerGrab<D>
where D: Any + 'static + SeatHandler,

pub fn is<__T: PointerGrab<D>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: PointerGrab<D>>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: PointerGrab<D>>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: PointerGrab<D>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: PointerGrab<D>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.