Trait smithay::input::pointer::PointerTarget

pub trait PointerTarget<D>: IsAlive + PartialEq + Clone + Sendwhere
    D: SeatHandler,{
    fn enter(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent);
    fn motion(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent);
    fn button(&self, seat: &Seat<D>, data: &mut D, event: &ButtonEvent);
    fn axis(&self, seat: &Seat<D>, data: &mut D, frame: AxisFrame);
    fn leave(&self, seat: &Seat<D>, data: &mut D, serial: Serial, time: u32);
}

Trait representing object that can receive pointer interactions

Required Methods

fn enter(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent)

A pointer of a given seat entered this handler

fn motion(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent)

A pointer of a given seat moved over this handler

fn button(&self, seat: &Seat<D>, data: &mut D, event: &ButtonEvent)

A pointer of a given seat clicked a button

fn axis(&self, seat: &Seat<D>, data: &mut D, frame: AxisFrame)

A pointer of a given seat scrolled on an axis

fn leave(&self, seat: &Seat<D>, data: &mut D, serial: Serial, time: u32)

A pointer of a given seat left this handler