smithay::input::pointer

Trait PointerTarget

pub trait PointerTarget<D>:
    IsAlive
    + PartialEq
    + Clone
    + Debug
    + Send
where
    D: SeatHandler,
{

    // Required methods
    fn enter(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent);
    fn motion(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent);
    fn relative_motion(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &RelativeMotionEvent,
    );
    fn button(&self, seat: &Seat<D>, data: &mut D, event: &ButtonEvent);
    fn axis(&self, seat: &Seat<D>, data: &mut D, frame: AxisFrame);
    fn frame(&self, seat: &Seat<D>, data: &mut D);
    fn gesture_swipe_begin(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GestureSwipeBeginEvent,
    );
    fn gesture_swipe_update(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GestureSwipeUpdateEvent,
    );
    fn gesture_swipe_end(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GestureSwipeEndEvent,
    );
    fn gesture_pinch_begin(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GesturePinchBeginEvent,
    );
    fn gesture_pinch_update(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GesturePinchUpdateEvent,
    );
    fn gesture_pinch_end(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GesturePinchEndEvent,
    );
    fn gesture_hold_begin(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GestureHoldBeginEvent,
    );
    fn gesture_hold_end(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &GestureHoldEndEvent,
    );
    fn leave(&self, seat: &Seat<D>, data: &mut D, serial: Serial, time: u32);

    // Provided method
    fn replace(
        &self,
        replaced: <D as SeatHandler>::PointerFocus,
        seat: &Seat<D>,
        data: &mut D,
        event: &MotionEvent,
    ) { ... }

}

Trait representing object that can receive pointer interactions

Required Methods

fn enter(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent)

A pointer of a given seat entered this handler

fn motion(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent)

A pointer of a given seat moved over this handler

fn relative_motion( &self, seat: &Seat<D>, data: &mut D, event: &RelativeMotionEvent, )

A pointer of a given seat that provides relative motion moved over this handler

fn button(&self, seat: &Seat<D>, data: &mut D, event: &ButtonEvent)

A pointer of a given seat clicked a button

fn axis(&self, seat: &Seat<D>, data: &mut D, frame: AxisFrame)

A pointer of a given seat scrolled on an axis

fn frame(&self, seat: &Seat<D>, data: &mut D)

End of a pointer frame

fn gesture_swipe_begin( &self, seat: &Seat<D>, data: &mut D, event: &GestureSwipeBeginEvent, )

A pointer of a given seat started a swipe gesture

fn gesture_swipe_update( &self, seat: &Seat<D>, data: &mut D, event: &GestureSwipeUpdateEvent, )

A pointer of a given seat updated a swipe gesture

fn gesture_swipe_end( &self, seat: &Seat<D>, data: &mut D, event: &GestureSwipeEndEvent, )

A pointer of a given seat ended a swipe gesture

fn gesture_pinch_begin( &self, seat: &Seat<D>, data: &mut D, event: &GesturePinchBeginEvent, )

A pointer of a given seat started a pinch gesture

fn gesture_pinch_update( &self, seat: &Seat<D>, data: &mut D, event: &GesturePinchUpdateEvent, )

A pointer of a given seat updated a pinch gesture

fn gesture_pinch_end( &self, seat: &Seat<D>, data: &mut D, event: &GesturePinchEndEvent, )

A pointer of a given seat ended a pinch gesture

fn gesture_hold_begin( &self, seat: &Seat<D>, data: &mut D, event: &GestureHoldBeginEvent, )

A pointer of a given seat started a hold gesture

fn gesture_hold_end( &self, seat: &Seat<D>, data: &mut D, event: &GestureHoldEndEvent, )

A pointer of a given seat ended a hold gesture

fn leave(&self, seat: &Seat<D>, data: &mut D, serial: Serial, time: u32)

A pointer of a given seat left this handler

Provided Methods

fn replace( &self, replaced: <D as SeatHandler>::PointerFocus, seat: &Seat<D>, data: &mut D, event: &MotionEvent, )

A pointer of a given seat moved from another handler to this handler

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.