smithay::input
Module touch

Touch-related types for smithay’s input abstraction

Structs

DefaultGrab

The default grab, the behavior when no particular grab is in progress

DownEvent

Pointer motion event

GrabStartData

Data about the event that started the grab.

MotionEvent

Pointer motion event

OrientationEvent

Pointer motion event

ShapeEvent

Pointer motion event

TouchDownGrab

A touch down grab, basic grab started when an user touches a surface to maintain it focused until the user releases the touch.

TouchHandle

An handle to a touch handler

TouchInnerHandle

This inner handle is accessed from inside a pointer grab logic, and directly sends event to the client

UpEvent

Pointer motion event

Traits

TouchGrab

A trait to implement a touch grab

TouchTarget

Trait representing object that can receive touch interactions