smithay::input::touch

Trait TouchGrab

pub trait TouchGrab<D: SeatHandler>: Send + Downcast {
    // Required methods
    fn down(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        focus: Option<(<D as SeatHandler>::TouchFocus, Point<f64, Logical>)>,
        event: &DownEvent,
        seq: Serial,
    );
    fn up(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        event: &UpEvent,
        seq: Serial,
    );
    fn motion(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        focus: Option<(<D as SeatHandler>::TouchFocus, Point<f64, Logical>)>,
        event: &MotionEvent,
        seq: Serial,
    );
    fn frame(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        seq: Serial,
    );
    fn cancel(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        seq: Serial,
    );
    fn shape(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        event: &ShapeEvent,
        seq: Serial,
    );
    fn orientation(
        &mut self,
        data: &mut D,
        handle: &mut TouchInnerHandle<'_, D>,
        event: &OrientationEvent,
        seq: Serial,
    );
    fn start_data(&self) -> &GrabStartData<D>;
    fn unset(&mut self, data: &mut D);
}

A trait to implement a touch grab

In some context, it is necessary to temporarily change the behavior of the touch handler. This is typically known as a touch grab. A typical example would be, during a drag’n’drop operation, the underlying surfaces will no longer receive classic pointer event, but rather special events.

This trait is the interface to intercept regular touch events and change them as needed, its interface mimics the TouchHandle interface.

Any interactions with TouchHandle should be done using TouchInnerHandle, as handle is borrowed/locked before grab methods are called, so calling methods on TouchHandle would result in a deadlock.

If your logic decides that the grab should end, both TouchInnerHandle and TouchHandle have a method to change it.

When your grab ends (either as you requested it or if it was forcefully cancelled by the server), the struct implementing this trait will be dropped. As such you should put clean-up logic in the destructor, rather than trying to guess when the grab will end.

Required Methods

fn down( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, focus: Option<(<D as SeatHandler>::TouchFocus, Point<f64, Logical>)>, event: &DownEvent, seq: Serial, )

A new touch point appeared

This method allows you attach additional behavior to a down event, possibly altering it. You generally will want to invoke TouchInnerHandle::down as part of your processing. If you don’t, the rest of the compositor will behave as if the down event never occurred.

Some grabs (such as drag’n’drop, shell resize and motion) drop down events while they are active, the default touch grab will keep the focus on the surface that started the grab.

fn up( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, event: &UpEvent, seq: Serial, )

A touch point disappeared

This method allows you attach additional behavior to a up event, possibly altering it. You generally will want to invoke TouchInnerHandle::up as part of your processing. If you don’t, the rest of the compositor will behave as if the up event never occurred.

Some grabs (such as drag’n’drop, shell resize and motion) drop up events while they are active, but will end when the touch point that initiated the grab disappeared.

fn motion( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, focus: Option<(<D as SeatHandler>::TouchFocus, Point<f64, Logical>)>, event: &MotionEvent, seq: Serial, )

A touch point has changed coordinates.

This method allows you attach additional behavior to a motion event, possibly altering it. You generally will want to invoke TouchInnerHandle::motion as part of your processing. If you don’t, the rest of the compositor will behave as if the motion event never occurred.

Note that this is not intended to update the focus of the touch point, the focus is only set on a down event. The focus provided to this function can be used to find DnD targets during touch motion.

fn frame( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, seq: Serial, )

Marks the end of a set of events that logically belong together.

This method allows you attach additional behavior to a frame event, possibly altering it. You generally will want to invoke TouchInnerHandle::frame as part of your processing. If you don’t, the rest of the compositor will behave as if the frame event never occurred.

This will to be called after one or more calls to down/motion events.

fn cancel( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, seq: Serial, )

A touch session has been cancelled.

This method allows you attach additional behavior to a cancel event, possibly altering it. You generally will want to invoke TouchInnerHandle::cancel as part of your processing. If you don’t, the rest of the compositor will behave as if the cancel event never occurred.

Usually called in case the compositor decides the touch stream is a global gesture.

fn shape( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, event: &ShapeEvent, seq: Serial, )

A touch point has changed its shape.

fn orientation( &mut self, data: &mut D, handle: &mut TouchInnerHandle<'_, D>, event: &OrientationEvent, seq: Serial, )

A touch point has changed its orientation.

fn start_data(&self) -> &GrabStartData<D>

The data about the event that started the grab.

fn unset(&mut self, data: &mut D)

The grab has been unset or replaced with another grab.

Implementations

impl<D> dyn TouchGrab<D>
where D: Any + 'static + SeatHandler,

pub fn is<__T: TouchGrab<D>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: TouchGrab<D>>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: TouchGrab<D>>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: TouchGrab<D>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: TouchGrab<D>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.