smithay::input::touch

Trait TouchTarget

pub trait TouchTarget<D>:
    IsAlive
    + Debug
    + Send
where
    D: SeatHandler,
{
    // Required methods
    fn down(&self, seat: &Seat<D>, data: &mut D, event: &DownEvent, seq: Serial);
    fn up(&self, seat: &Seat<D>, data: &mut D, event: &UpEvent, seq: Serial);
    fn motion(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &MotionEvent,
        seq: Serial,
    );
    fn frame(&self, seat: &Seat<D>, data: &mut D, seq: Serial);
    fn cancel(&self, seat: &Seat<D>, data: &mut D, seq: Serial);
    fn shape(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &ShapeEvent,
        seq: Serial,
    );
    fn orientation(
        &self,
        seat: &Seat<D>,
        data: &mut D,
        event: &OrientationEvent,
        seq: Serial,
    );
}

Trait representing object that can receive touch interactions

Required Methods

fn down(&self, seat: &Seat<D>, data: &mut D, event: &DownEvent, seq: Serial)

A new touch point has appeared on the target.

This touch point is assigned a unique ID. Future events from this touch point reference this ID. The ID ceases to be valid after a touch up event and may be reused in the future.

fn up(&self, seat: &Seat<D>, data: &mut D, event: &UpEvent, seq: Serial)

The touch point has disappeared.

No further events will be sent for this touch point and the touch point’s ID is released and may be reused in a future touch down event.

fn motion(&self, seat: &Seat<D>, data: &mut D, event: &MotionEvent, seq: Serial)

A touch point has changed coordinates.

fn frame(&self, seat: &Seat<D>, data: &mut D, seq: Serial)

Indicates the end of a set of events that logically belong together.

fn cancel(&self, seat: &Seat<D>, data: &mut D, seq: Serial)

Touch session cancelled.

Touch cancellation applies to all touch points currently active on this target. The client is responsible for finalizing the touch points, future touch points on this target may reuse the touch point ID.

fn shape(&self, seat: &Seat<D>, data: &mut D, event: &ShapeEvent, seq: Serial)

Sent when a touch point has changed its shape.

A touch point shape is approximated by an ellipse through the major and minor axis length. The major axis length describes the longer diameter of the ellipse, while the minor axis length describes the shorter diameter. Major and minor are orthogonal and both are specified in surface-local coordinates. The center of the ellipse is always at the touch point location as reported by TouchTarget::down or TouchTarget::motion.

fn orientation( &self, seat: &Seat<D>, data: &mut D, event: &OrientationEvent, seq: Serial, )

Sent when a touch point has changed its orientation.

The orientation describes the clockwise angle of a touch point’s major axis to the positive surface y-axis and is normalized to the -180 to +180 degree range. The granularity of orientation depends on the touch device, some devices only support binary rotation values between 0 and 90 degrees.