smithay::input

Trait SeatHandler

pub trait SeatHandler: Sized {
    type KeyboardFocus: KeyboardTarget<Self> + PartialEq + Clone + 'static;
    type PointerFocus: PointerTarget<Self> + PartialEq + Clone + 'static;
    type TouchFocus: TouchTarget<Self> + PartialEq + Clone + 'static;

    // Required method
    fn seat_state(&mut self) -> &mut SeatState<Self>;

    // Provided methods
    fn focus_changed(
        &mut self,
        _seat: &Seat<Self>,
        _focused: Option<&Self::KeyboardFocus>,
    ) { ... }
    fn cursor_image(&mut self, _seat: &Seat<Self>, _image: CursorImageStatus) { ... }
    fn led_state_changed(&mut self, _seat: &Seat<Self>, _led_state: LedState) { ... }
}

Handler trait for Seats

Required Associated Types

type KeyboardFocus: KeyboardTarget<Self> + PartialEq + Clone + 'static

Type used to represent the target currently holding the keyboard focus

type PointerFocus: PointerTarget<Self> + PartialEq + Clone + 'static

Type used to represent the target currently holding the pointer focus

type TouchFocus: TouchTarget<Self> + PartialEq + Clone + 'static

Type used to represent the target currently holding the touch focus

Required Methods

fn seat_state(&mut self) -> &mut SeatState<Self>

SeatState getter

Provided Methods

fn focus_changed( &mut self, _seat: &Seat<Self>, _focused: Option<&Self::KeyboardFocus>, )

Callback that will be notified whenever the focus of the seat changes.

fn cursor_image(&mut self, _seat: &Seat<Self>, _image: CursorImageStatus)

Callback that will be notified whenever a client requests to set a custom cursor image.

fn led_state_changed(&mut self, _seat: &Seat<Self>, _led_state: LedState)

Callback that will be notified whenever the keyboard led state changes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.