Module smithay::output

Output

This module provides a type helping you to abstract over various properties, that make up an Output of your compositor.

How to use it

You need to instantiate an Output.

To advertise a new output global to wayland clients take a look at crate::wayland::output, you need to have the wayland_frontend feature enabled to use it.

Additionally outputs are used to by the desktop abstractions (see crate::desktop, needs the desktop feature) to represent views into your window grid.

You can use the returned Output to change the properties of your output (if the current resolution mode changes for example). These may influence how contents will be rendered to your output, when used in conjunction with the desktop abstractions.

You can attach additional properties to your Outputs by using Output::user_data.

use smithay::output::{Output, PhysicalProperties, Scale, Mode, Subpixel};
use smithay::utils::Transform;

// Create the Output with given name and physical properties.
let output = Output::new(
    "output-0".into(), // the name of this output,
    PhysicalProperties {
        size: (200, 150).into(),        // dimensions (width, height) in mm
        subpixel: Subpixel::HorizontalRgb,  // subpixel information
        make: "Screens Inc".into(),     // make of the monitor
        model: "Monitor Ultra".into(),  // model of the monitor
    },
    None // insert a logger here
);
// Now you can configure it
output.change_current_state(
    Some(Mode { size: (1920, 1080).into(), refresh: 60000 }), // the resolution mode,
    Some(Transform::Normal), // global screen transformation
    Some(Scale::Integer(1)), // global screen scaling factor
    Some((0,0).into()) // output position
);
// set the preferred mode
output.set_preferred(Mode { size: (1920, 1080).into(), refresh: 60000 });
// add other supported modes
output.add_mode(Mode { size: (800, 600).into(), refresh: 60000 });
output.add_mode(Mode { size: (1024, 768).into(), refresh: 60000 });

Structs

Mode

An output mode

Output

An abstract output.

PhysicalProperties

The physical properties of an output

WeakOutput

Weak variant of an Output

Enums

Scale

Describes the scale advertised to clients.

Subpixel

Subpixel geometry information