smithay::output

Struct Output

pub struct Output { /* private fields */ }

An abstract output.

This handle is stored in the event loop, and allows you to notify clients about any change in the properties of this output.

Implementations

impl Output

pub fn new(name: String, physical: PhysicalProperties) -> Output

Create a new output with given name and physical properties.

pub fn set_preferred(&self, mode: Mode)

Sets the preferred mode of this output

If the provided mode was not previously known to this output, it is added to its internal list.

pub fn add_mode(&self, mode: Mode)

Adds a mode to the list of known modes to this output

pub fn current_mode(&self) -> Option<Mode>

Returns the currently advertised mode of the output

pub fn preferred_mode(&self) -> Option<Mode>

Returns the preferred mode of the output

pub fn current_transform(&self) -> Transform

Returns the currently advertised transformation of the output

pub fn current_scale(&self) -> Scale

Returns the currently set scale of the output

pub fn current_location(&self) -> Point<i32, Logical>

Returns the currently advertised location of the output

pub fn name(&self) -> String

Returns the name of the output

pub fn description(&self) -> String

Returns the description of the output, if xdg-output is initialized

pub fn physical_properties(&self) -> PhysicalProperties

Returns the physical properties of the output

pub fn modes(&self) -> Vec<Mode>

Returns the currently advertised modes of the output

pub fn delete_mode(&self, mode: Mode)

Removes a mode from the list of known modes

It will not de-advertise it from existing clients (the protocol does not allow it), but it won’t be advertised to now clients from now on.

pub fn change_current_state( &self, new_mode: Option<Mode>, new_transform: Option<Transform>, new_scale: Option<Scale>, new_location: Option<Point<i32, Logical>>, )

Change the current state of this output

You can changed the current mode, transform status, location or scale of this output. Providing None to any of these field means that the value does not change.

If the provided mode was not previously known to this output, it is added to its internal list.

By default, transform status is Normal, and scale is 1.

pub fn user_data(&self) -> &UserDataMap

Returns the user data of this output

pub fn downgrade(&self) -> WeakOutput

Create a weak reference to this output

pub fn cleanup(&self)

Cleanup some internal resource.

Needs to be called periodically, at best before every wayland socket flush.

Note: Space::refresh already calls this function internally. So in case you already use the desktop module you can skip calling this function explicitly.

impl Output

pub fn create_global<D>(&self, display: &DisplayHandle) -> GlobalId
where D: GlobalDispatch<WlOutput, WlOutputData> + 'static,

Available on crate feature wayland_frontend only.

Create a new output global.

The global is directly registered into the event loop, and this function returns the global handle in case you wish to remove this global in the future.

Calling this function multiple times without destroying the global in between, will result in multiple globals, meaning the output will be advertised to clients multiple times.

pub fn from_resource(output: &WlOutput) -> Option<Output>

Available on crate feature wayland_frontend only.

Attempt to retrieve a Output from an existing resource

pub fn owns(&self, output: &WlOutput) -> bool

Available on crate feature wayland_frontend only.

Check is given wl_output instance is managed by this Output.

pub fn client_outputs<'a>( &'a self, client: &Client, ) -> impl Iterator<Item = WlOutput> + 'a

Available on crate feature wayland_frontend only.

This function returns all managed WlOutput matching the provided Client

pub fn enter(&self, surface: &WlSurface)

Available on crate feature wayland_frontend only.

Sends wl_surface.enter for the provided surface with the matching client output

pub fn leave(&self, surface: &WlSurface)

Available on crate feature wayland_frontend only.

Sends wl_surface.leave for the provided surface with the matching client output

pub fn leave_all(&self)

Available on crate feature wayland_frontend only.

Sends wl_surface.leave for all surfaces on output