Struct smithay::utils::Point

pub struct Point<N, Kind> {
    pub x: N,
    pub y: N,
    // some fields omitted
}

A point as defined by its x and y coordinates

Fields
x: N

horizontal coordinate

y: N

vertical coordinate

Implementations

impl<N: Coordinate, Kind> Point<N, Kind>

pub fn to_size(self) -> Size<N, Kind>

Convert this Point to a Size with the same coordinates

Checks that the coordinates are positive with a debug_assert!().

pub fn to_size_abs(self) -> Size<N, Kind>

Convert this Point to a Size with the same coordinates

Ensures that the coordinates are positive by taking their absolute value

impl<N: Coordinate, Kind> Point<N, Kind>

pub fn to_f64(self) -> Point<f64, Kind>

Convert the underlying numerical type to f64 for floating point manipulations

impl<Kind> Point<f64, Kind>

pub fn to_i32_round<N: Coordinate>(self) -> Point<N, Kind>

Convert to i32 for integer-space manipulations by rounding float values

pub fn to_i32_floor<N: Coordinate>(self) -> Point<N, Kind>

Convert to i32 for integer-space manipulations by flooring float values

pub fn to_i32_ceil<N: Coordinate>(self) -> Point<N, Kind>

Convert to i32 for integer-space manipulations by ceiling float values

impl<N: Coordinate> Point<N, Logical>

pub fn to_physical(self, scale: N) -> Point<N, Physical>

Convert this logical point to physical coordinate space according to given scale factor

pub fn to_buffer(self, scale: N) -> Point<N, Buffer>

Convert this logical point to buffer coordinate space according to given scale factor

impl<N: Coordinate> Point<N, Physical>

pub fn to_logical(self, scale: N) -> Point<N, Logical>

Convert this physical point to logical coordinate space according to given scale factor

impl<N: Coordinate> Point<N, Buffer>

pub fn to_logical(self, scale: N) -> Point<N, Logical>

Convert this physical point to logical coordinate space according to given scale factor