Struct smithay::utils::Rectangle

#[repr(C)]
pub struct Rectangle<N, Kind> {
    pub loc: Point<N, Kind>,
    pub size: Size<N, Kind>,
}

A rectangle defined by its top-left corner and dimensions

Operations on rectangles are saturating.

Fields
loc: Point<N, Kind>

Location of the top-left corner of the rectangle

size: Size<N, Kind>

Size of the rectangle, as (width, height)

Implementations

impl<N: Coordinate, Kind> Rectangle<N, Kind>

pub fn to_f64(self) -> Rectangle<f64, Kind>

Convert the underlying numerical type to another

impl<N: Coordinate, Kind> Rectangle<N, Kind>

pub fn upscale(self, scale: impl Into<Scale<N>>) -> Rectangle<N, Kind>

Upscale this Rectangle by the supplied Scale

pub fn downscale(self, scale: impl Into<Scale<N>>) -> Rectangle<N, Kind>

Upscale this Rectangle by the supplied Scale

pub fn is_empty(&self) -> bool

Check if this Rectangle is empty

Returns true if either the width or the height of the Size is zero

impl<Kind> Rectangle<f64, Kind>

pub fn to_i32_round<N: Coordinate>(self) -> Rectangle<N, Kind>

Convert to i32 for integer-space manipulations by rounding float values

pub fn to_i32_down<N: Coordinate>(self) -> Rectangle<N, Kind>

Convert to i32 by returning the largest integer-space rectangle fitting into the float-based rectangle

pub fn to_i32_up<N: Coordinate>(self) -> Rectangle<N, Kind>

Convert to i32 by returning the smallest integet-space rectangle encapsulating the float-based rectangle

impl<N: Coordinate, Kind> Rectangle<N, Kind>

pub fn from_loc_and_size(
loc: impl Into<Point<N, Kind>>,
size: impl Into<Size<N, Kind>>
) -> Self

Create a new Rectangle from the coordinates of its top-left corner and its dimensions

pub fn from_extemities(
topleft: impl Into<Point<N, Kind>>,
bottomright: impl Into<Point<N, Kind>>
) -> Self

Create a new Rectangle from the coordinates of its top-left corner and its bottom-right corner

pub fn contains<P: Into<Point<N, Kind>>>(self, point: P) -> bool

Checks whether given Point is inside the rectangle

pub fn contains_rect<R: Into<Rectangle<N, Kind>>>(self, rect: R) -> bool

Checks whether given Rectangle is inside the rectangle

A rectangle is considered inside another rectangle if its location is inside the other rectangle and it does not extend outside the other rectangle. This includes rectangles with the same location and size

pub fn overlaps(self, other: impl Into<Rectangle<N, Kind>>) -> bool

Checks whether a given Rectangle overlaps with this one

pub fn intersection(self, other: impl Into<Rectangle<N, Kind>>) -> Option<Self>

Clamp rectangle to min and max corners resulting in the overlapping area of two rectangles

Returns None if the two rectangles don’t overlap

pub fn bounding_box(points: impl IntoIterator<Item = Point<N, Kind>>) -> Self

Compute the bounding box of a given set of points

pub fn merge(self, other: Self) -> Self

Merge two Rectangle by producing the smallest rectangle that contains both

pub fn subtract_rect(self, other: Self) -> Vec<Self>

Subtracts another Rectangle from this Rectangle

If the rectangles to not overlap the original rectangle will be returned. If the other rectangle contains self no rectangle will be returned, otherwise up to 4 rectangles will be returned.

impl<N: Coordinate> Rectangle<N, Logical>

pub fn to_physical(self, scale: impl Into<Scale<N>>) -> Rectangle<N, Physical>

Convert this logical rectangle to physical coordinate space according to given scale factor

pub fn to_physical_precise_round<S: Coordinate, R: Coordinate>(
self,
scale: impl Into<Scale<S>>
) -> Rectangle<R, Physical>

Convert this logical rectangle to physical coordinate space according to given scale factor and round the result

pub fn to_physical_precise_down<S: Coordinate, R: Coordinate>(
&self,
scale: impl Into<Scale<S>>
) -> Rectangle<R, Physical>

Convert this logical rectangle to physical coordinate space according to given scale factor, returning the largest N-space rectangle fitting into the N-based rectangle

This will ceil the location and floor the size after applying the scale

pub fn to_physical_precise_up<S: Coordinate, R: Coordinate>(
&self,
scale: impl Into<Scale<S>>
) -> Rectangle<R, Physical>

Convert this logical rectangle to physical coordinate space according to given scale factor, returning the smallest N-space rectangle encapsulating the N-based rectangle

This will floor the location and ceil the size after applying the scale

pub fn to_buffer(
self,
scale: impl Into<Scale<N>>,
transformation: Transform,
area: &Size<N, Logical>
) -> Rectangle<N, Buffer>

Convert this logical rectangle to buffer coordinate space according to given scale factor

impl<N: Coordinate> Rectangle<N, Physical>

pub fn to_logical(self, scale: impl Into<Scale<N>>) -> Rectangle<N, Logical>

Convert this physical rectangle to logical coordinate space according to given scale factor

impl<N: Coordinate> Rectangle<N, Buffer>

pub fn to_logical(
self,
scale: impl Into<Scale<N>>,
transformation: Transform,
area: &Size<N, Buffer>
) -> Rectangle<N, Logical>

Convert this physical rectangle to logical coordinate space according to given scale factor