Trait smithay::utils::Coordinate

pub trait Coordinate: Sized + Add<Self, Output = Self> + Sub<Self, Output = Self> + PartialOrd + Default + Copy + Debug {
    fn downscale(self, scale: Self) -> Self;

    fn upscale(self, scale: Self) -> Self;

    fn to_f64(self) -> f64;

    fn from_f64(v: f64) -> Self;

    fn non_negative(self) -> bool;

    fn abs(self) -> Self;

    fn saturating_add(self, other: Self) -> Self;

    fn saturating_sub(self, other: Self) -> Self;

    fn saturating_mul(self, other: Self) -> Self;

    fn min(self, other: Self) -> Self { ... }

    fn max(self, other: Self) -> Self { ... }
}

Trait for types serving as a coordinate for other geometry utils

Required methods

fn downscale(self, scale: Self) -> Self

Downscale the coordinate

fn upscale(self, scale: Self) -> Self

Upscale the coordinate

fn to_f64(self) -> f64

Convert the coordinate to a f64

fn from_f64(v: f64) -> Self

Convert to this coordinate from a f64

fn non_negative(self) -> bool

Test if the coordinate is not negative

fn abs(self) -> Self

Returns the absolute value of this coordinate

fn saturating_add(self, other: Self) -> Self

Saturating integer addition. Computes self + other, saturating at the numeric bounds instead of overflowing.

fn saturating_sub(self, other: Self) -> Self

Saturating integer subtraction. Computes self - other, saturating at the numeric bounds instead of overflowing.

fn saturating_mul(self, other: Self) -> Self

Saturating integer multiplication. Computes self * other, saturating at the numeric bounds instead of overflowing.

Provided methods

fn min(self, other: Self) -> Self

Compare and return the smaller one

fn max(self, other: Self) -> Self

Compare and return the larger one