Module smithay::wayland

Protocol-related utilities

This module contains several handlers to manage the Wayland protocol and the clients.

Most utilities provided in this module work in the same way:

An init function or method will take the event loop as argument and insert one or more globals into it.

If you want to remove a previously inserted global, just call the destroy() method on the associated Global. If you don’t plan to destroy the global at all, you don’t need to bother keeping the Global around.

Modules

compositor
Utilities for handling surfaces, subsurfaces and regions

data_device
Utilities for manipulating the data devices

dmabuf
Linux DMABUF protocol

explicit_synchronization
Explicit buffer synchronization per wayland surface

output
Output advertising capabilities

seat
Seat global utilities

shell
Handler utilities for the various shell protocols

shm
SHM handling helpers

Structs

Serial
A serial type, whose comparison takes into account the wrapping-around behavior of the underlying counter.

SerialCounter
A counter for generating serials, for use in the client protocol

Statics

SERIAL_COUNTER
A global SerialCounter for use in your compositor.