Module smithay::wayland::compositor

Utilities for handling surfaces, subsurfaces and regions

This module provides automatic handling of sufaces, subsurfaces and region wayland objects, by registering an implementation for for the wl_compositor and wl_subcompositor globals.

Why use this implementation

This implementation does a simple job: it stores in a coherent way the state of surface trees with subsurfaces, to provide you a direct access to the tree structure and all surface metadata.

As such, you can, given a root surface with a role requiring it to be displayed, you can iterate over the whole tree of subsurfaces to recover all the metadata you need to display the subsurface tree.

This implementation will not do anything more than present you the metadata specified by the client in a coherent and practical way. All the logic regarding to drawing itself, and the positionning of windows (surface trees) one relative to another is out of its scope.

How to use it

Initialization

To initialize this implementation, use the compositor_init method provided by this module. It'll require you to first define as few things, as shown in this example:

use smithay::wayland::compositor::compositor_init;

// Define some user data to be associated with the surfaces.
// It must implement the Default trait, which will represent the state of a surface which
// has just been created.
#[derive(Default)]
struct MyData {
    // whatever you need here
}

// Declare the roles enum
define_roles!(MyRoles);

// Call the init function:
let (token, _, _) = compositor_init::<MyData, MyRoles, _, _>(
    &mut display,
    event_loop.token(),
    |request, (surface, compositor_token)| {
        /*
          Your handling of the user requests. This closure can also
          be a struct implementing the appropriate Implementation trait.
         */
    },
    None // put a logger here
);

// this `token` is what you'll use to access the surface associated data

// You're now ready to go!

Use the surface metadata

As you can see in the previous example, in the end we are retrieving a token from the init function. This token is necessary to retrieve the metadata associated with a surface. It can be cloned. See CompositorToken for the details of what it enables you.

The surface metadata is held in the SurfaceAttributes struct. In contains double-buffered state pending from the client as defined by the protocol for wl_surface, as well as your user-defined type holding any data you need to have associated with a struct. See its documentation for details.

This CompositorToken also provides access to the metadata associated with the role of the surfaces. See the documentation of the roles submodule for a detailed explanation.

Modules

roles
Tools for handling surface roles

Structs

CompositorToken
A Compositor global token

RegionAttributes
Description of the contents of a region

SubsurfaceRole
Attributes defining the behaviour of a sub-surface relative to its parent

SurfaceAttributes
Data associated with a surface, aggreged by the handlers

Enums

Damage
Description of which part of a surface should be considered damaged and needs to be redrawn

RectangleKind
Kind of a rectangle part of a region

SurfaceEvent
User-handled events for surfaces

TraversalAction
Possible actions to do after handling a node diring tree traversal

Functions

compositor_init
Create new wl_compositor and wl_subcompositor globals.