Function smithay::wayland::compositor::compositor_init

pub fn compositor_init<R, Impl, L>(
    display: &mut Display, 
    implem: Impl, 
    logger: L
) -> (CompositorToken<R>, Global<WlCompositor>, Global<WlSubcompositor>) where
    L: Into<Option<Logger>>,
    R: Default + RoleType + Role<SubsurfaceRole> + Send + 'static,
    Impl: FnMut(SurfaceEvent, WlSurface, CompositorToken<R>) + 'static, 

Create new wl_compositor and wl_subcompositor globals.

The globals are directly registered into the event loop, and this function returns a [CompositorToken] which you'll need access the data associated to the wl_surfaces.

It also returns the two global handles, in case you wish to remove these globals from the event loop in the future.