Trait smithay::wayland::compositor::roles::Role

pub trait Role<R>: RoleType {
    fn set_with(&mut self, data: R) -> Result<(), R>;

    fn has(&self) -> bool;

    fn data(&self) -> Result<&R, WrongRole>;

    fn data_mut(&mut self) -> Result<&mut R, WrongRole>;

    fn unset(&mut self) -> Result<R, WrongRole>;

    fn set(&mut self) -> Result<(), ()>
    where
        R: Default,
    { ... }
}

A trait representing the capability of a [RoleType] to handle a given role

This trait allows to interact with the different roles a [RoleType] can handle.

This trait is meant to be used generically, for example, to retrieve the data associated with a given role with token TheRole:

let data = <MyRoles as Role<RoleData>>::data(my_roles)
                .expect("The surface does not have this role.");

The methods of this trait are mirrored on CompositorToken for easy access to the role data of the surfaces.

Note that if a role is automatically handled for you by a Handler provided by smithay, you should not set or unset it manually on a surface. Doing so would likely corrupt the internal state of these handlers, causing spurious protocol errors and unreliable behaviour overall.

Required methods

fn set_with(&mut self, data: R) -> Result<(), R>

Set the role for the associated surface with given data

Fails if the surface already has a role and returns the data

fn has(&self) -> bool

Check if the associated surface has this role

fn data(&self) -> Result<&R, WrongRole>

Access the data associated with this role if its the current one

fn data_mut(&mut self) -> Result<&mut R, WrongRole>

Mutably access the data associated with this role if its the current one

fn unset(&mut self) -> Result<R, WrongRole>

Remove this role from the associated surface

Fails if the surface does not currently have this role

Loading content...
Provided methods

fn set(&mut self) -> Result<(), ()> where
R: Default,

Set the role for the associated surface with default associated data

Fails if the surface already has a role

Loading content...