Struct smithay::wayland::compositor::SurfaceAttributes

pub struct SurfaceAttributes {
    pub buffer: Option<Option<(WlBuffer, (i32, i32))>>,
    pub buffer_scale: i32,
    pub buffer_transform: Transform,
    pub opaque_region: Option<RegionAttributes>,
    pub input_region: Option<RegionAttributes>,
    pub damage: Damage,
    pub user_data: UserDataMap,
}

Data associated with a surface, aggregated by the handlers

Most of the fields of this struct represent a double-buffered state, which should only be applied once a commit request is received from the surface.

You are responsible for setting those values as you see fit to avoid processing them two times.

Fields
buffer: Option<Option<(WlBuffer, (i32, i32))>>

Buffer defining the contents of the surface

The tuple represent the coordinates of this buffer relative to the location of the current buffer.

If set to Some(None), it means the user specifically asked for the surface to be unmapped.

You are free to set this field to None to avoid processing it several times. It'll be set to Some(...) if the user attaches a buffer (or NULL) to the surface.

buffer_scale: i32

Scale of the contents of the buffer, for higher-resolution contents.

If it matches the one of the output displaying this surface, no change is necessary.

buffer_transform: Transform

Transform under which interpret the contents of the buffer

If it matches the one of the output displaying this surface, no change is necessary.

opaque_region: Option<RegionAttributes>

Region of the surface that is guaranteed to be opaque

By default the whole surface is potentially transparent

input_region: Option<RegionAttributes>

Region of the surface that is sensitive to user input

By default the whole surface should be sensitive

damage: Damage

Damage rectangle

Hint provided by the client to suggest that only this part of the surface was changed and needs to be redrawn

user_data: UserDataMap

User-controlled data

This is your field to host whatever you need.