smithay::wayland::compositor

Trait CompositorHandler

pub trait CompositorHandler {
    // Required methods
    fn compositor_state(&mut self) -> &mut CompositorState;
    fn client_compositor_state<'a>(
        &self,
        client: &'a Client,
    ) -> &'a CompositorClientState;
    fn commit(&mut self, surface: &WlSurface);

    // Provided methods
    fn new_surface(&mut self, surface: &WlSurface) { ... }
    fn new_subsurface(&mut self, surface: &WlSurface, parent: &WlSurface) { ... }
    fn destroyed(&mut self, _surface: &WlSurface) { ... }
}

Available on crate feature wayland_frontend only.

Handler trait for compositor

Required Methods

fn compositor_state(&mut self) -> &mut CompositorState

CompositorState getter

fn client_compositor_state<'a>( &self, client: &'a Client, ) -> &'a CompositorClientState

CompositorClientState getter

The compositor implementation needs some state to be client specific. Downstream is expected to store this inside its ClientData implementation(s) to ensure automatic cleanup of the state, when the client disconnects.

fn commit(&mut self, surface: &WlSurface)

Surface commit handler

This is called when any changed state from a commit actually becomes visible. That might be some time after the actual commit has taken place, if the state changes are delayed by an added blocker (see add_blocker).

If you need to handle a commit as soon as it occurs, you might want to consider using a pre-commit hook (see add_pre_commit_hook).

Provided Methods

fn new_surface(&mut self, surface: &WlSurface)

New surface handler.

This handler can be used to setup hooks (see add_pre_commit_hook/add_post_commit_hook/add_destruction_hook), but not much else. The surface has no role or attached data at this point and cannot be rendered.

fn new_subsurface(&mut self, surface: &WlSurface, parent: &WlSurface)

New subsurface handler.

This handler can be used to run extra logic when subsurface is getting created. This is an addition to new_surface, which will be run for the subsurface surface anyway.

When your compositor knows beforehand where it’ll position subsurfaces it can send send_surface_state to them.

fn destroyed(&mut self, _surface: &WlSurface)

The surface was destroyed.

This allows the compositor to clean up any uses of the surface.

Note: Destruction might happen explicitly by the client, or implicitly when the client quits. In case of implicit destruction the order the callbacks are called in is undefined.