Function smithay::wayland::data_device::start_dnd

pub fn start_dnd<C>(
    seat: &Seat, 
    serial: u32, 
    start_data: GrabStartData, 
    metadata: SourceMetadata, 
    callback: C
) where
    C: FnMut(ServerDndEvent) + 'static, 

Start a drag'n'drop from a ressource controlled by the compositor

You'll receive events generated by the interaction of clients with your drag'n'drop in the provided callback. See ServerDndEvent for details about which events can be generated and what response is expected from you to them.