Module smithay::wayland::dmabuf

Linux DMABUF protocol

This module provides helper to handle the linux-dmabuf protocol, which allows clients to submit their contents as dmabuf file descriptors. These handlers automate the aggregation of the metadata associated with a dma buffer, and do some basic checking of the sanity of what the client sends.

How to use

To setup the dmabuf global, you will need to provide 2 things:

a list of the dmabuf formats you wish to support when creating the Global through DmabufState::create_global

an implementation of DmabufHandler to test if a dmabuf buffer can be imported by your renderer

The list of supported formats is a Vec<Format>, where you will enter all the (code, modifier) pairs you support. You can typically receive a list of supported formats for one renderer by calling ImportDma::dmabuf_formats.

Accessing a Dmabuf associated with a WlBuffer may be achieved using get_dmabuf.

use smithay::{
    delegate_dmabuf,
    backend::allocator::dmabuf::Dmabuf,
    reexports::{wayland_server::protocol::wl_buffer::WlBuffer},
    wayland::{
        buffer::BufferHandler,
        dmabuf::{DmabufGlobal, DmabufHandler, DmabufState, ImportError}
    },
};

pub struct State {
    dmabuf_state: DmabufState,
    dmabuf_global: DmabufGlobal,
}

// Smithay's "DmabufHandler" also requires the buffer management utilities, you need to implement
// "BufferHandler".
impl BufferHandler for State {
    fn buffer_destroyed(&mut self, buffer: &wayland_server::protocol::wl_buffer::WlBuffer) {
        // All renderers can handle buffer destruction at this point. Some parts of window management may
        // also use this function.
        //
        // If you need to mark a dmabuf elsewhere in your state as destroyed, you use the "get_dmabuf"
        // function defined in this module to access the dmabuf associated the "Buffer".
    }
}

impl DmabufHandler for State {
    fn dmabuf_state(&mut self) -> &mut DmabufState {
        &mut self.dmabuf_state
    }

    fn dmabuf_imported(&mut self, global: &DmabufGlobal, dmabuf: Dmabuf) -> Result<(), ImportError> {
        // Here you should import the dmabuf into your renderer.
        //
        // The return value indicates whether import was successful. If the return value is Err, then
        // the client is told dmabuf import has failed.
        Ok(())
    }
}

// Delegate dmabuf handling for State to DmabufState.
delegate_dmabuf!(State);

// First a DmabufState must be created. This type is used to create some "DmabufGlobal"s
let mut dmabuf_state = DmabufState::new();

// define your supported formats
let formats = vec![
    /* ... */
];

// And create the dmabuf global.
let dmabuf_global = dmabuf_state.create_global::<State, _>(
    &display_handle,
    formats,
    None // we don't provide a logger in this example
);

let state = State {
    dmabuf_state,
    dmabuf_global,
};

// Rest of the compositor goes here...

Structs

DmabufData

Data associated with a dmabuf global protocol object.

DmabufGlobal

A handle to a registered dmabuf global.

DmabufGlobalData

Data associated with a dmabuf global.

DmabufParamsData

Data associated with a pending Dmabuf import.

DmabufState

Delegate type for all dmabuf globals.

Enums

ImportError

Error that may occur when importing a Dmabuf.

Traits

DmabufHandler

Handler trait for Dmabuf import from the compositor.

Functions

get_dmabuf

Gets the contents of a Dmabuf backed WlBuffer.