Module smithay::wayland::dmabuf

Linux DMABUF protocol

This module provides helper to handle the linux-dmabuf protocol, which allows clients to submit their contents as dmabuf file descriptors. These handlers automate the aggregation of the metadata associated with a dma buffer, and do some basic checking of the sanity of what the client sends.

This module is only available if the backend_drm cargo feature is enabled.

How to use

To setup the dmabuf global, you will need to provide 2 things:

a list of the dmabuf formats you wish to support

an implementation of the DmabufHandler trait

The list of supported format is just a Vec<Format>, where you will enter all the (format, modifier) couples you support.

The implementation of the DmabufHandler trait will be called whenever a client has finished setting up a dma buffer. You will be handled the full details of the client's submission as a BufferInfo struct, and you need to validate it and maybe import it into your renderer. The BufferData associated type allows you to store any metadata or handle to the resource you need into the created wl_buffer, user data, to then retrieve it when it is attached to a surface to re-identify the dmabuf.

use smithay::wayland::dmabuf::{DmabufHandler, BufferInfo, init_dmabuf_global};

struct MyDmabufHandler;

struct MyBufferData {
    /* ... */
}

impl Drop for MyBufferData {
    fn drop(&mut self) {
        // This is called when all handles to this buffer have been dropped,
        // both client-side and server side.
        // You can now free the associated resources in your renderer.
    }
}

impl DmabufHandler for MyDmabufHandler {
    type BufferData = MyBufferData;
    fn validate_dmabuf(&mut self, info: BufferInfo) -> Result<Self::BufferData, ()> {
        /* validate the dmabuf and import it into your renderer state */
        Ok(MyBufferData { /* ... */ })
    }
}

// Once this is defined, you can in your setup initialize the dmabuf global:

// define your supported formats
let formats = vec![
    /* ... */
];
let dmabuf_global = init_dmabuf_global(
    &mut display,
    formats,
    MyDmabufHandler,
    None // we don't provide a logger in this example
);

Structs

BufferFlags
Possible flags for a DMA buffer

BufferInfo
The complete information provided by the client to create a dmabuf buffer

Format
Representation of a Dmabuf format, as advertized to the client

Plane
A plane send by the client

Enums

Flags

Traits

DmabufHandler
Handler trait for dmabuf validation

Functions

init_dmabuf_global
Initialize a dmabuf global.