Trait smithay::wayland::dmabuf::DmabufHandler

pub trait DmabufHandler: BufferHandler {
    fn dmabuf_state(&mut self) -> &mut DmabufState;
    fn dmabuf_imported(
        &mut self,
        global: &DmabufGlobal,
        dmabuf: Dmabuf
    ) -> Result<(), ImportError>;
}

Handler trait for Dmabuf import from the compositor.

Required Methods

fn dmabuf_state(&mut self) -> &mut DmabufState

Returns a mutable reference to the DmabufState delegate type.

fn dmabuf_imported(
&mut self,
global: &DmabufGlobal,
dmabuf: Dmabuf
) -> Result<(), ImportError>

This function is called when a client has imported a Dmabuf.

The global indicates which DmabufGlobal the buffer was imported to. You should import the dmabuf into your renderer to ensure the dmabuf may be used later when rendering.

The return value of this function indicates whether dmabuf import is successful. The renderer is responsible for determining whether the format and plane combinations are valid and should return ImportError::InvalidFormat if the format and planes are not correct.

If the import fails due to an implementation specific reason, then ImportError::Failed should be returned.