Trait smithay::wayland::dmabuf::DmabufHandler

pub trait DmabufHandler: BufferHandler {
    // Required methods
    fn dmabuf_state(&mut self) -> &mut DmabufState;
    fn dmabuf_imported(
        &mut self,
        global: &DmabufGlobal,
        dmabuf: Dmabuf
    ) -> Result<(), ImportError>;

    // Provided method
    fn new_surface_feedback(
        &mut self,
        _surface: &WlSurface,
        _global: &DmabufGlobal
    ) -> Option<DmabufFeedback> { ... }
}

Handler trait for Dmabuf import from the compositor.

Required Methods

fn dmabuf_state(&mut self) -> &mut DmabufState

Returns a mutable reference to the DmabufState delegate type.

fn dmabuf_imported( &mut self, global: &DmabufGlobal, dmabuf: Dmabuf ) -> Result<(), ImportError>

This function is called when a client has imported a Dmabuf.

The global indicates which DmabufGlobal the buffer was imported to. You should import the dmabuf into your renderer to ensure the dmabuf may be used later when rendering.

The return value of this function indicates whether dmabuf import is successful. The renderer is responsible for determining whether the format and plane combinations are valid and should return ImportError::InvalidFormat if the format and planes are not correct.

If the import fails due to an implementation specific reason, then ImportError::Failed should be returned.

Provided Methods

fn new_surface_feedback( &mut self, _surface: &WlSurface, _global: &DmabufGlobal ) -> Option<DmabufFeedback>

This function allows to override the default DmabufFeedback for a surface

Note: This will only be called if there is no alive surface feedback for the surface. Normally this will be the first time a surface requests feedback, but can also occur if all instances have been destroyed and a new surface request is sent by the client.

Returning None will use the default DmabufFeedback from the global