smithay::wayland::drm_lease

Trait DrmLeaseHandler

pub trait DrmLeaseHandler:
    GlobalDispatch<WpDrmLeaseDeviceV1, DrmLeaseDeviceGlobalData>
    + Dispatch<WpDrmLeaseConnectorV1, DrmNode, Self>
    + Dispatch<WpDrmLeaseDeviceV1, DrmNode, Self>
    + Dispatch<WpDrmLeaseRequestV1, DrmLeaseRequestData, Self>
    + Dispatch<WpDrmLeaseV1, DrmLeaseData, Self>
    + 'static {
    // Required methods
    fn drm_lease_state(&mut self, node: DrmNode) -> &mut DrmLeaseState;
    fn lease_request(
        &mut self,
        node: DrmNode,
        request: DrmLeaseRequest,
    ) -> Result<DrmLeaseBuilder, LeaseRejected>;
    fn new_active_lease(&mut self, node: DrmNode, lease: DrmLease);
    fn lease_destroyed(&mut self, node: DrmNode, lease_id: u32);
}

Available on crate features wayland_frontend and backend_drm only.

Handler trait for drm leasing from the compositor.

Required Methods

fn drm_lease_state(&mut self, node: DrmNode) -> &mut DrmLeaseState

Returns a mutable reference to the DrmLeaseState delegate type

fn lease_request( &mut self, node: DrmNode, request: DrmLeaseRequest, ) -> Result<DrmLeaseBuilder, LeaseRejected>

A client has issued a new request.

The request only contains connectors and the compositor is lease any resources it may seem fit. It is recommended however to lease the requested connectors and any resources necessary to drive them (free CRTCs per connector and at least their primary plane) or reject the request.

To reject the request return Err(LeaseRejected::default()), otherwise return a DrmLeaseBuilder with all the resources added.

fn new_active_lease(&mut self, node: DrmNode, lease: DrmLease)

A new DRM lease is active. Dropping the provided DrmLease will revoke the lease.

fn lease_destroyed(&mut self, node: DrmNode, lease_id: u32)

A DRM lease was destroyed, the previously given DrmLease should be cleaned up, if not revoked already.

Note: Destruction might happen explicitly by the client, or implicitly when the client quits. In case of implicit destruction the order the callbacks are called in is undefined.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.