Module smithay::wayland::explicit_synchronization

Explicit buffer synchronization per wayland surface

This interface allow clients to switch from a per-buffer signaling of buffer release (via the wl_buffer.release event) to a per-surface signaling using dma_fences. This is notably used for efficient synchronization of OpenGL/Vulkan clients.

At surface commit time, in addition to a buffer the client can have attached two more properties:

an acquire dma_fence file descriptor, that the compositor is required to wait on before it will try to access the contents of the associated buffer

an ExplicitBufferRelease object, that the compositor is expect to use to signal the client when it has finished using the buffer for this surface (if the same buffer is attached to multiple surfaces, the release only applies for the surface associated with this release object, not the whole buffer).

The use of these dma_fences in conjunction with the graphics stack allows for efficient synchronization between the clients and the compositor.

Usage

First, you need to initialize the global:

use smithay::wayland::explicit_synchronization::*;
init_explicit_synchronization_global(
    &mut display,
    compositor_token,
    None /* You can insert a logger here */
);

Then when handling a surface commit, you can retrieve the synchronization information for the surface data:

compositor_token.with_surface_data(&surface, |surface_attributes| {
    // While you retrieve the surface data from the commit ...
    // Check the explicit synchronization data:
    match get_explicit_synchronization_state(surface_attributes) {
        Ok(sync_state) => {
            /* This surface is explicitly synchronized, you need to handle
               the contents of sync_state
            */
        },
        Err(()) => {
            /* This surface is not explicitly synchronized, nothing more to do
            */
        }
    }
});

Structs

ExplicitBufferRelease
An object to signal end of use of a buffer

ExplicitSyncState
An explicit synchronization state

Enums

ExplicitSyncError
Possible errors you can send to an ill-behaving clients

Functions

get_explicit_synchronization_state
Retrieve the explicit synchronization state commited by the client

init_explicit_synchronization_global
Initialize the explicit synchronization global

send_explicit_synchronization_error
Send a synchronization error to a client