Module smithay::wayland::fractional_scale

Utilities for handling the wp_fractional_scale protocol

How to use it

Initialization

To initialize this implementation create the FractionalScaleManagerState, store it inside your State struct and implement the FractionScaleHandler, as shown in this example:

use smithay::delegate_fractional_scale;
use smithay::wayland::compositor;
use smithay::reexports::wayland_server::protocol::wl_surface;
use smithay::wayland::fractional_scale::{
    self,
    FractionalScaleManagerState,
    FractionScaleHandler,
};

// Create the compositor state
let fractional_scale_manager_state = FractionalScaleManagerState::new::<State, _>(
    &display.handle(),
    None // We don't add a logger in this example
);

// insert the FractionalScaleManagerState into your state
// ..

// implement the necessary traits
impl FractionScaleHandler for State {
   fn new_fractional_scale(&mut self, surface: wl_surface::WlSurface) {
       compositor::with_states(&surface, |states| {
           fractional_scale::with_fractional_scale(states, |fractional_scale| {
               // set the preferred scale for the surface
               fractional_scale.set_preferred_scale(1.25);
           });
       })
   }
}
delegate_fractional_scale!(State);

// You're now ready to go!

Use the fractional scale state

Whenever the fractional scale for a surface changes set the preferred fractional scale like shown in the example:

use smithay::wayland::compositor;
use smithay::wayland::fractional_scale;
compositor::with_states(&surface, |states| {
    fractional_scale::with_fractional_scale(states, |fractional_scale| {
        // set the preferred scale for the surface
        fractional_scale.set_preferred_scale(1.25);
    });
})

Structs

FractionalScaleManagerState

State of the wp_fractional_scale_manager_v1 Global

FractionalScaleState

State for the fractional scale

Traits

FractionScaleHandler

Fractional scale handler type

Functions

with_fractional_scale

Run a closure on the FractionalScaleState of a WlSurface

Type Definitions

FractionalScaleStateUserData

Type stored in WlSurface states data_map