Module smithay::wayland::input_method

Utilities for input method support

This module provides you with utilities to handle input methods, it must be used in conjunction with the text input module to work.

use smithay::{
    delegate_seat, delegate_tablet_manager, delegate_input_method_manager,
    delegate_text_input_manager,
};
use smithay::input::{Seat, SeatState, SeatHandler, keyboard::XkbConfig, pointer::CursorImageStatus};
use smithay::wayland::input_method::{
    InputMethodManagerState,
    InputMethodSeat
};
use smithay::wayland::text_input::TextInputManagerState;
use smithay::reexports::wayland_server::{Display, protocol::wl_surface::WlSurface};

delegate_seat!(State);
// Delegate input method handling for State to InputMethodManagerState.
delegate_input_method_manager!(State);
delegate_text_input_manager!(State);

let mut seat_state = SeatState::<State>::new();

// implement the required traits
impl SeatHandler for State {
    type KeyboardFocus = WlSurface;
    type PointerFocus = WlSurface;
    fn seat_state(&mut self) -> &mut SeatState<Self> {
        &mut self.seat_state
    }
    fn focus_changed(&mut self, seat: &Seat<Self>, focused: Option<&WlSurface>) { unimplemented!() }
    fn cursor_image(&mut self, seat: &Seat<Self>, image: CursorImageStatus) { unimplemented!() }
}

// Add the seat state to your state and create manager globals
InputMethodManagerState::new::<State>(&display_handle);
// Add text input capabilities, needed for the input method to work
TextInputManagerState::new::<State>(&display_handle);

Run usage

Once the input method and text input capabilities have been added to a seat, use the [seat.input_method().set_point] function to set the top left point of a focused surface. This is used to calculate the popup surface location.

Then use the [seat.input_method().coordinates] and [seat.input_method().with_surface] functions to draw the popup surface.

Structs

InputMethodHandle

Handle to an input method instance

InputMethodKeyboardUserData

User data of ZwpInputKeyboardGrabV2 object

InputMethodManagerState

State of wp misc input method protocol

InputMethodPopupSurfaceUserData

User data of ZwpInputPopupSurfaceV2 object

InputMethodUserData

User data of ZwpInputMethodV2 object

Traits

InputMethodSeat

Extends Seat with input method functionality