Module smithay::wayland::keyboard_shortcuts_inhibit

Available on crate feature wayland_frontend only.

Protocol for inhibiting the compositor keyboard shortcuts

This protocol specifies a way for a client to request the compositor to ignore its own keyboard shortcuts for a given seat, so that all key events from that seat get forwarded to a surface.

Structs

KeyboardShortcutsInhibitState

Delegate type for KeyboardShortcutsInhibit global.

KeyboardShortcutsInhibitor

Context object for keyboard shortcuts inhibitor

KeyboardShortcutsInhibitorUserData

User data of ZwpKeyboardShortcutsInhibitorV1 object

Traits

KeyboardShortcutsInhibitHandler

WP Keyboard shortcuts inhibit handler

KeyboardShortcutsInhibitorSeat

Seat extension used to check if shortcuts are inhibited