smithay::wayland::keyboard_shortcuts_inhibit

Trait KeyboardShortcutsInhibitHandler

pub trait KeyboardShortcutsInhibitHandler {
    // Required method
    fn keyboard_shortcuts_inhibit_state(
        &mut self,
    ) -> &mut KeyboardShortcutsInhibitState;

    // Provided methods
    fn new_inhibitor(&mut self, inhibitor: KeyboardShortcutsInhibitor) { ... }
    fn inhibitor_destroyed(&mut self, inhibitor: KeyboardShortcutsInhibitor) { ... }
}

Available on crate feature wayland_frontend only.

WP Keyboard shortcuts inhibit handler

Required Methods

fn keyboard_shortcuts_inhibit_state( &mut self, ) -> &mut KeyboardShortcutsInhibitState

KeyboardShortcutsInhibitState getter

Provided Methods

fn new_inhibitor(&mut self, inhibitor: KeyboardShortcutsInhibitor)

New keyboard shortcuts inhibitor got created by the client

In response to this event compositor can decide if inhibitor should be activated or not, usually based on user decision. You may also postpone activation based on your compositor specific policy.

fn inhibitor_destroyed(&mut self, inhibitor: KeyboardShortcutsInhibitor)

Inhibitor got destroyed