Trait smithay::wayland::keyboard_shortcuts_inhibit::KeyboardShortcutsInhibitorSeat

pub trait KeyboardShortcutsInhibitorSeat {
    fn keyboard_shortcuts_inhibited(&self) -> bool;
    fn keyboard_shortcuts_inhibitor_for_surface(
        &self,
        surface: &WlSurface
    ) -> Option<KeyboardShortcutsInhibitor>;
}

Seat extension used to check if shortcuts are inhibited

Required Methods

fn keyboard_shortcuts_inhibited(&self) -> bool

Check if keyboard_shortcuts are inhibited

fn keyboard_shortcuts_inhibitor_for_surface(
&self,
surface: &WlSurface
) -> Option<KeyboardShortcutsInhibitor>

Get inhibitors associated with given WlSurface

Can be used to check if certain surface has inhibitor on it

use smithay::input::Seat;
use smithay::wayland::keyboard_shortcuts_inhibit::KeyboardShortcutsInhibitorSeat;

let seat: Seat<State> = todo!();

if let Some(inhibitor) = seat.keyboard_shortcuts_inhibitor_for_surface(&wl_surface) {
    dbg!(inhibitor.is_active());
}