smithay::wayland

Module output

Available on crate feature wayland_frontend only.

WlOutput advertising capabilities

This module provides additional implementations for crate::output::Output to handle advertising of your compositor’s output and their capabilities to your client, as well as mapping your clients output request to your physical outputs.

How to use it

After you have instantiated an Output you need to use Output::create_global to advertise a new output global to clients. The resulting GlobalId can later be destroyed again to stop advertising it without destroying it’s state. E.g. in case the matching physical output got disabled at runtime.

If you change the properties of your output (if the current resolution mode changes for example), it’ll automatically forward any changes to the clients.

Additional protocols may piggy-back on this type. E.g. to also advertise an xdg-output for every wl-output you can use OutputManagerState::new_with_xdg_output.

use smithay::delegate_output;
use smithay::output::{Output, PhysicalProperties, Scale, Mode, Subpixel};
use smithay::utils::Transform;
use smithay::wayland::output::OutputHandler;

// Create the Output with given name and physical properties.
let output = Output::new(
    "output-0".into(), // the name of this output,
    PhysicalProperties {
        size: (200, 150).into(),        // dimensions (width, height) in mm
        subpixel: Subpixel::HorizontalRgb,  // subpixel information
        make: "Screens Inc".into(),     // make of the monitor
        model: "Monitor Ultra".into(),  // model of the monitor
        serial_number: "KL11C28J04XJ".into(), // serial number of the monitor
    },
);
// create a global, if you want to advertise it to clients
let _global = output.create_global::<State>(
    &display_handle,      // the display
); // you can drop the global, if you never intend to destroy it.
// Now you can configure it
output.change_current_state(
    Some(Mode { size: (1920, 1080).into(), refresh: 60000 }), // the resolution mode,
    Some(Transform::Normal), // global screen transformation
    Some(Scale::Integer(1)), // global screen scaling factor
    Some((0,0).into()) // output position
);
// set the preferred mode
output.set_preferred(Mode { size: (1920, 1080).into(), refresh: 60000 });
// add other supported modes
output.add_mode(Mode { size: (800, 600).into(), refresh: 60000 });
output.add_mode(Mode { size: (1024, 768).into(), refresh: 60000 });

impl OutputHandler for State {}

delegate_output!(State);

Structs

OutputManagerState

State of Smithay output manager

OutputUserData

User data for WlOutput

WlOutputData

Internal data of a wl_output global

XdgOutputUserData

User data of Xdg Output

Traits

OutputHandler

Events initiated by the clients interacting with outputs