smithay::wayland::pointer_constraints

Trait PointerConstraintsHandler

pub trait PointerConstraintsHandler: SeatHandler {
    // Required methods
    fn new_constraint(
        &mut self,
        surface: &WlSurface,
        pointer: &PointerHandle<Self>,
    );
    fn cursor_position_hint(
        &mut self,
        surface: &WlSurface,
        pointer: &PointerHandle<Self>,
        location: Point<f64, Logical>,
    );
}

Available on crate feature wayland_frontend only.

Handler for pointer constraints

Required Methods

fn new_constraint(&mut self, surface: &WlSurface, pointer: &PointerHandle<Self>)

Pointer lock or confinement constraint created for pointer on surface

Use with_pointer_constraint to access the constraint.

fn cursor_position_hint( &mut self, surface: &WlSurface, pointer: &PointerHandle<Self>, location: Point<f64, Logical>, )

The client holding a LockedPointer has committed a cursor position hint.

This is emitted upon a surface commit if the cursor position hint has been updated.

Use with_pointer_constraint to access the constraint and check if it is active.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.