Module smithay::wayland::presentation

Utilities for handling the wp_presentation protocol

How to use it

Initialization

To initialize this implementation, create PresentationState, store it in your State struct and implement the required traits, as shown in this example:

use smithay::wayland::presentation::PresentationState;
use smithay::delegate_presentation;

// Create the viewporter state:
let viewporter_state = PresentationState::new::<State>(
    &display.handle(), // the display
    1 // the id of the clock
);

// implement Dispatch for the Viewporter types
delegate_presentation!(State);

// You're now ready to go!

Use the presentation state

Before sending the frame callbacks you should drain all committed presentation feedback callbacks. After the associated frame has been presented the callbacks can be marked presented as shown in the example.

The presentation state is double-buffered and can be accessed by using the with_states function

use smithay::wayland::presentation::PresentationFeedbackCachedState;
use wayland_protocols::wp::presentation_time::server::wp_presentation_feedback;

// ... render frame ...

let presentation_feedbacks = with_states(&surface, |states| {
    std::mem::take(&mut states.cached_state.current::<PresentationFeedbackCachedState>().callbacks)
});

// ... send frame callbacks and present frame

for feedback in presentation_feedbacks {
    feedback.presented(&output, time, refresh, seq, wp_presentation_feedback::Kind::Vsync);
}

Structs

PresentationFeedbackCachedState

State of a single presentation feedback requested for a surface

PresentationFeedbackCallback

Holds a single presentation feedback

PresentationFeedbackState

State for a single presentation feedback callback

PresentationState

State of the wp_presentation global