Module smithay::wayland::seat

Seat global utilities

This module provides you with utilities for handling the seat globals and the associated input Wayland objects.

How to use it

Initialization

use smithay::delegate_seat;
use smithay::input::{Seat, SeatState, SeatHandler, pointer::CursorImageStatus};
use smithay::reexports::wayland_server::{Display, protocol::wl_surface::WlSurface};

let mut seat_state = SeatState::<State>::new();
// add the seat state to your state
// ...

// create the wl_seat
let seat = seat_state.new_wl_seat(
    &display_handle,          // the display
    "seat-0",                 // the name of the seat, will be advertized to clients
);

// implement the required traits
impl SeatHandler for State {
    type KeyboardFocus = WlSurface;
    type PointerFocus = WlSurface;
    fn seat_state(&mut self) -> &mut SeatState<Self> {
        &mut self.seat_state
    }
    fn focus_changed(&mut self, seat: &Seat<Self>, focused: Option<&WlSurface>) {
        // ...
    }
    fn cursor_image(&mut self, seat: &Seat<Self>, image: CursorImageStatus) {
        // ...
    }
}
delegate_seat!(State);

Run usage

Once the seat is initialized, you can add capabilities to it.

Currently, only pointer and keyboard capabilities are supported by smithay.

You can add these capabilities via methods of the Seat struct: Seat::add_keyboard and Seat::add_pointer. These methods return handles that can be cloned and sent across thread, so you can keep one around in your event-handling code to forward inputs to your clients.

This module further defines the "cursor_image" role, that is assigned to surfaces used by clients to change the cursor icon.

Structs

KeyboardUserData

User data for keyboard

PointerUserData

User data for pointer

SeatGlobalData

Global data of WlSeat

SeatUserData

User data for seat

TouchHandle

An handle to a touch handler.

TouchUserData

User data for keyboard

Constants

CURSOR_IMAGE_ROLE

WlSurface role of a cursor image icon

Traits

WaylandFocus

Focused objects that might have an underlying wl_surface.