Module smithay::wayland::seat

Seat global utilities

This module provides you with utilities for handling the seat globals and the associated input wayland objects.

How to use it

Initialization

use smithay::wayland::seat::Seat;

// insert the seat:
let (seat, seat_global) = Seat::new(
    &mut display, // the display
    event_loop.token(), // a LoopToken
    "seat-0".into(), // the name of the seat, will be advertize to clients
    None /* insert a logger here*/
);

Run usage

Once the seat is initialized, you can add capabilities to it.

Currently, only pointer and keyboard capabilities are supported by smithay.

You can add these capabilities via methods of the Seat struct:

let pointer_handle = seat.add_pointer();

These handles can be cloned and sent accross thread, so you can keep one around in your event-handling code to forward inputs to your clients.

Modules

keysyms

Structs

KeyboardHandle
An handle to a keyboard handler

ModifiersState
Represents the current state of the keyboard modifiers

PointerAxisHandle
A frame of pointer axis events.

PointerHandle
An handle to a keyboard handler

Seat
A Seat handle

Enums

KeyboardError
Errors that can be encountered when creating a keyboard handler

Type Definitions

Keysym
A number used to represent the symbols generated from a key on a keyboard.