Module smithay::wayland::seat

Seat global utilities

This module provides you with utilities for handling the seat globals and the associated input Wayland objects.

How to use it

Initialization

use smithay::wayland::seat::Seat;

// insert the seat:
let (seat, seat_global) = Seat::new(
    &mut display,             // the display
    "seat-0".into(),          // the name of the seat, will be advertized to clients
    None                      // insert a logger here
);

Run usage

Once the seat is initialized, you can add capabilities to it.

Currently, only pointer and keyboard capabilities are supported by smithay.

You can add these capabilities via methods of the Seat struct: Seat::add_keyboard and Seat::add_pointer. These methods return handles that can be cloned and sent across thread, so you can keep one around in your event-handling code to forward inputs to your clients.

This module further defines the "cursor_image" role, that is assigned to surfaces used by clients to change the cursor icon.

Modules

keysyms

Structs

AxisFrame

A frame of pointer axis events.

CursorImageAttributes

The role representing a surface set as the pointer cursor

GrabStartData

Data about the event that started the grab.

KeyboardHandle

An handle to a keyboard handler

ModifiersState

Represents the current state of the keyboard modifiers

PointerHandle

An handle to a pointer handler

PointerInnerHandle

This inner handle is accessed from inside a pointer grab logic, and directly sends event to the client

Seat

A Seat handle

XkbConfig

Configuration for xkbcommon.

Enums

CursorImageStatus

Possible status of a cursor as requested by clients

FilterResult

Result for key input filtering (see KeyboardHandle::input)

KeyboardError

Errors that can be encountered when creating a keyboard handler

Traits

PointerGrab

A trait to implement a pointer grab

Type Definitions

Keysym

A number used to represent the symbols generated from a key on a keyboard.