Trait smithay::wayland::seat::WaylandFocus

pub trait WaylandFocus {
    // Required method
    fn wl_surface(&self) -> Option<WlSurface>;

    // Provided method
    fn same_client_as(&self, object_id: &ObjectId) -> bool { ... }
}

Available on crate feature wayland_frontend only.

Focused objects that might have an underlying wl_surface.

Required Methods

fn wl_surface(&self) -> Option<WlSurface>

Returns the underlying wl_surface, if any.

Note: This has to return Some, if same_client_as can return true for any provided ObjectId

Provided Methods

fn same_client_as(&self, object_id: &ObjectId) -> bool

Returns true, if the underlying wayland object originates from the same client connection as the provided ObjectId.

Must return false, if there is not underlying wayland object.