smithay::wayland

Module security_context

Available on crate feature wayland_frontend only.

Utilities for handling the security context protocol

Structs

SecurityContext

A security context associated with a listener created through security context protocol, and with clients connected through that listener.

SecurityContextGlobalData

User data for a WpSecurityContextManagerV1 global

SecurityContextListenerSource

Security context listener event source.

SecurityContextState

State of the WpSecurityContextManagerV1 global

SecurityContextUserData

User data for a WpSecurityContextV1

Traits

SecurityContextHandler

Handler for security context protocol