smithay::wayland::selection

Module data_device

Available on crate feature wayland_frontend only.

Utilities for manipulating the data devices

The data device is wayland’s abstraction to represent both selection (copy/paste) and drag’n’drop actions. This module provides logic to handle this part of the protocol. Selection and drag’n’drop are per-seat notions.

This module provides the freestanding set_data_device_focus function: This function sets the data device focus for a given seat; you’d typically call it whenever the keyboard focus changes, to follow it (for example in the focus hook of your keyboards).

Using these two functions is enough for your clients to be able to interact with each other using the data devices.

The module also provides additional mechanisms allowing your compositor to see and interact with the contents of the data device:

the freestanding function set_data_device_selection allows you to set the contents of the selection for your clients

The module defines the role "dnd_icon" that is assigned to surfaces used as drag’n’drop icons.

Initialization

To initialize this implementation, create the DataDeviceState, store it inside your State struct and implement the DataDeviceHandler and SelectionHandler, as shown in this example:

use smithay::delegate_data_device;
use smithay::wayland::selection::SelectionHandler;
use smithay::wayland::selection::data_device::{WaylandDndGrabHandler, DataDeviceState, DataDeviceHandler};

// Create the data_device state
let data_device_state = DataDeviceState::new::<State>(
    &display.handle(),
);

// insert the DataDeviceState into your state
// ..

// implement the necessary traits
impl WaylandDndGrabHandler for State {
    // ... implement `dnd_requested` to handle drag&drop operations
}
impl SelectionHandler for State {
    type SelectionUserData = ();
}
impl DataDeviceHandler for State {
    fn data_device_state(&mut self) -> &mut DataDeviceState { &mut self.data_device_state }
    // ... override default implementations here to customize handling ...
}
delegate_data_device!(State);

// You're now ready to go!

Structs

DataDeviceState

State of data device

WlOfferData

Type implementing OfferData for wl_data_offer based DnD sources

Enums

SelectionRequestError

Errors happening when requesting selection contents

Constants

DND_ICON_ROLE

WlSurface role of drag and drop icon

Traits

DataDeviceHandler

Events that are generated by interactions of the clients with the data device

WaylandDndGrabHandler

Handler for wayland client initiated drag’n’drop

Functions

clear_data_device_selection

Clear the current selection for this seat

current_data_device_selection_userdata

Gets the user_data for the currently active selection, if set by the compositor

default_action_chooser

A simple action chooser for DnD negotiation

request_data_device_client_selection

Request the current data_device selection of the given seat to be written to the provided file descriptor in the given mime type.

set_data_device_focus

Set the data device focus to a certain client for a given seat

set_data_device_selection

Set a compositor-provided selection for this seat