smithay::wayland::selection::data_device

Function start_dnd

pub fn start_dnd<D>(
    dh: &DisplayHandle,
    seat: &Seat<D>,
    data: &mut D,
    serial: Serial,
    pointer_start_data: Option<GrabStartData<D>>,
    touch_start_data: Option<GrabStartData<D>>,
    metadata: SourceMetadata,
)
where
    D: SeatHandler + DataDeviceHandler + 'static,
    <D as SeatHandler>::PointerFocus: WaylandFocus,
    <D as SeatHandler>::TouchFocus: WaylandFocus,

Available on crate feature wayland_frontend only.

Start a drag’n’drop from a resource controlled by the compositor

You’ll receive events generated by the interaction of clients with your drag’n’drop in the provided callback. See ServerDndGrabHandler for details about which events can be generated and what response is expected from you to them.