smithay::wayland::selection::data_device

Trait ClientDndGrabHandler

pub trait ClientDndGrabHandler: SeatHandler + Sized {
    // Provided methods
    fn started(
        &mut self,
        source: Option<WlDataSource>,
        icon: Option<WlSurface>,
        seat: Seat<Self>,
    ) { ... }
    fn dropped(
        &mut self,
        target: Option<WlSurface>,
        validated: bool,
        seat: Seat<Self>,
    ) { ... }
}

Available on crate feature wayland_frontend only.

Events that are generated during client initiated drag’n’drop

Provided Methods

fn started( &mut self, source: Option<WlDataSource>, icon: Option<WlSurface>, seat: Seat<Self>, )

A client started a drag’n’drop as response to a user pointer action

source - The data source provided by the client. If it is None, this means the DnD is restricted to surfaces of the same client and the client will manage data transfer by itself.

icon - The icon the client requested to be used to be associated with the cursor icon during the drag’n’drop.

seat - The seat on which the DnD operation was started

fn dropped( &mut self, target: Option<WlSurface>, validated: bool, seat: Seat<Self>, )

The drag’n’drop action was finished by the user releasing the buttons

At this point, any pointer icon should be removed.

Note that this event will only be generated for client-initiated drag’n’drop session.

target - The target surface that the contents were dropped on.

validated - Whether the drop offer was negotiated and accepted. If false, the drop was cancelled or otherwise not successful.

seat - The seat on which the DnD action was finished.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.