Trait smithay::wayland::selection::data_device::ServerDndGrabHandler

pub trait ServerDndGrabHandler: SeatHandler {
    // Provided methods
    fn accept(&mut self, mime_type: Option<String>, seat: Seat<Self>) { ... }
    fn action(&mut self, action: DndAction, seat: Seat<Self>) { ... }
    fn dropped(&mut self, seat: Seat<Self>) { ... }
    fn cancelled(&mut self, seat: Seat<Self>) { ... }
    fn send(&mut self, mime_type: String, fd: OwnedFd, seat: Seat<Self>) { ... }
    fn finished(&mut self, seat: Seat<Self>) { ... }
}

Available on crate feature wayland_frontend only.

Event generated by the interactions of clients with a server initiated drag’n’drop

Provided Methods

fn accept(&mut self, mime_type: Option<String>, seat: Seat<Self>)

The client can accept the given mime type.

If mime_type is None, the client cannot accept any of the offered mime types. If the last accepted mime_type is None, the drag-and-drop operation will be cancelled. That is kept track of internally, this callback is informational only.

mime_type - The accepted mime type

seat - The seat on which the DnD action was chosen.

fn action(&mut self, action: DndAction, seat: Seat<Self>)

The client chose an action

action - The chosen action

seat - The seat on which the DnD action was chosen.

fn dropped(&mut self, seat: Seat<Self>)

The DnD resource was dropped by the user

After that, the client can still interact with your resource

seat - The seat on which the DnD was dropped.

fn cancelled(&mut self, seat: Seat<Self>)

The Dnd was cancelled

The client can no longer interact

seat - The seat on which the DnD action was cancelled.

fn send(&mut self, mime_type: String, fd: OwnedFd, seat: Seat<Self>)

The client requested for data to be sent

mime_type - The requested mime type

fd - The FD to write into

seat - The seat on which the DnD data is to be sent.

fn finished(&mut self, seat: Seat<Self>)

The client has finished interacting with the resource

This can only happen after the resource was dropped.

seat - The seat on which the DnD action was finished.

Object Safety

This trait is not object safe.