smithay::wayland::session_lock

Trait SessionLockHandler

pub trait SessionLockHandler {
    // Required methods
    fn lock_state(&mut self) -> &mut SessionLockManagerState;
    fn lock(&mut self, confirmation: SessionLocker);
    fn unlock(&mut self);
    fn new_surface(&mut self, surface: LockSurface, output: WlOutput);

    // Provided method
    fn ack_configure(
        &mut self,
        _surface: WlSurface,
        _configure: LockSurfaceConfigure,
    ) { ... }
}

Available on crate feature wayland_frontend only.

Handler trait for ext-session-lock.

Required Methods

fn lock_state(&mut self) -> &mut SessionLockManagerState

Session lock state.

fn lock(&mut self, confirmation: SessionLocker)

Handle compositor locking requests.

The SessionLocker parameter is used to confirm once the session was locked and no more client data is accessible using the SessionLocker::lock method.

If locking was not possible, dropping the SessionLocker will automatically notify the requesting client about the failure.

fn unlock(&mut self)

Handle compositor lock removal.

fn new_surface(&mut self, surface: LockSurface, output: WlOutput)

Add a new lock surface for an output.

Provided Methods

fn ack_configure( &mut self, _surface: WlSurface, _configure: LockSurfaceConfigure, )

A surface has acknowledged a configure serial.