Module smithay::wayland::shell::legacy

Utilities for handling shell surfaces with the wl_shell protocol

This module provides automatic handling of shell surfaces objects, by being registered as a global handler for wl_shell. This protocol is deprecated in favor of xdg_shell, thus this module is provided as a compatibility layer with older clients. As a consequence, you can as a compositor-writer decide to only support its functionality in a best-effort manner: as this global is part of the core protocol, you are still required to provide some support for it.

Why use this implementation

This implementation can track for you the various shell surfaces defined by the clients by handling the wl_shell protocol.

It allows you to easily access a list of all shell surfaces defined by your clients access their associated metadata and underlying wl_surfaces.

This handler only handles the protocol exchanges with the client to present you the information in a coherent and relatively easy to use manner. All the actual drawing and positioning logic of windows is out of its scope.

How to use it

Initialization

To initialize this handler, simple use the wl_shell_init function provided in this module. You need to provide a closure that will be invoked whenever some action is required from you, are represented by the ShellRequest enum.

use smithay::wayland::shell::legacy::{wl_shell_init, ShellRequest};

let (shell_state, _) = wl_shell_init(
    &mut display,
    // your implementation
    |event: ShellRequest, dispatch_data| { /* handle the shell requests here */ },
    None  // put a logger if you want
);

// You're now ready to go!

Structs

ShellState

Shell global state

ShellSurface

A handle to a shell surface

ShellSurfaceAttributes

Metadata associated with the wl_surface role

Enums

ShellRequest

A request triggered by a wl_shell_surface

ShellSurfaceKind

Possible kinds of shell surface of the wl_shell protocol

Constants

WL_SHELL_SURFACE_ROLE

The role of a wl_shell_surface.

Functions

wl_shell_init

Create a new wl_shell global