Module smithay::wayland::shell::wlr_layer

Utilities for handling shell surfaces with the wlr_layer_shell protocol

This interface should be suitable for the implementation of many desktop shell components, and a broad number of other applications that interact with the desktop.

Initialization

To initialize this handler, create the WlrLayerShellState, store it inside your State and implement the WlrLayerShellHandler, as shown in this example:

use smithay::delegate_layer_shell;
use smithay::wayland::shell::wlr_layer::{WlrLayerShellState, WlrLayerShellHandler, LayerSurface, Layer};
use smithay::reexports::wayland_server::protocol::wl_output::WlOutput;

let layer_shell_state = WlrLayerShellState::new::<State, _>(
    &display.handle(),
    None  // put a logger if you want
);

// - Put your layer_shell_state into your `State`.
// - And implement `LayerShellHandler
impl WlrLayerShellHandler for State {
    fn shell_state(&mut self) -> &mut WlrLayerShellState {
        &mut self.layer_shell_state
    }

    fn new_layer_surface(
        &mut self,
        surface: LayerSurface,
        output: Option<WlOutput>,
        layer: Layer,
        namespace: String,
    ) {
        // your implementation
    }
}
// let smithay implement wayland_server::DelegateDispatch
delegate_layer_shell!(State);

// You're now ready to go!

Structs

Anchor

Anchor bitflags, describing how the layers surface should be positioned and sized

LayerSurface

A handle to a layer surface

LayerSurfaceAttributes

Attributes for layer surface

LayerSurfaceCachedState

Represents the client pending state

LayerSurfaceConfigure

A configure message for layer surfaces

LayerSurfaceState

State of a layer surface

Margins

Describes distance from the anchor point of the output, in surface-local coordinates.

WlrLayerShellState

Shell global state

WlrLayerSurfaceUserData

User data for wlr layer surface

Enums

ExclusiveZone

Exclusive zone descriptor

KeyboardInteractivity

Types of keyboard interaction possible for a layer shell surface

Layer

Available layers for surfaces

Constants

LAYER_SURFACE_ROLE

The role of a wlr_layer_shell_surface

Traits

WlrLayerShellHandler

Handler for wlr layer shell

Type Definitions

LayerSurfaceData

Data associated with XDG popup surface