Module smithay::wayland::shell::wlr_layer

Utilities for handling shell surfaces with the wlr_layer_shell protocol

This interface should be suitable for the implementation of many desktop shell components, and a broad number of other applications that interact with the desktop.

Initialization

To initialize this handler, simple use the wlr_layer_shell_init function provided in this module. You need to provide a closure that will be invoked whenever some action is required from you, are represented by the LayerShellRequest enum.

use smithay::wayland::shell::wlr_layer::{wlr_layer_shell_init, LayerShellRequest};

let (shell_state, _) = wlr_layer_shell_init(
    &mut display,
    // your implementation
    |event: LayerShellRequest, dispatch_data| { /* handle the shell requests here */ },
    None  // put a logger if you want
);

// You're now ready to go!

Structs

Anchor

Anchor bitflags, describing how the layers surface should be positioned and sized

LayerShellState

Shell global state

LayerSurface

A handle to a layer surface

LayerSurfaceAttributes

Attributes for layer surface

LayerSurfaceCachedState

Represents the client pending state

LayerSurfaceConfigure

A configure message for layer surfaces

LayerSurfaceState

State of a layer surface

Margins

Describes distance from the anchor point of the output, in surface-local coordinates.

Enums

ExclusiveZone

Exclusive zone descriptor

KeyboardInteractivity

Types of keyboard interaction possible for a layer shell surface

Layer

Available layers for surfaces

LayerShellRequest

Events generated by layer shell surfaces

Constants

LAYER_SURFACE_ROLE

The role of a wlr_layer_shell_surface

Functions

wlr_layer_shell_init

Create a new wlr_layer_shell globals