Trait smithay::wayland::shell::wlr_layer::WlrLayerShellHandler

pub trait WlrLayerShellHandler {
    fn shell_state(&mut self) -> &mut WlrLayerShellState;
    fn new_layer_surface(
        &mut self,
        surface: LayerSurface,
        output: Option<WlOutput>,
        layer: Layer,
        namespace: String
    );

    fn new_popup(&mut self, parent: LayerSurface, popup: PopupSurface) { ... }
    fn ack_configure(
        &mut self,
        surface: WlSurface,
        configure: LayerSurfaceConfigure
    ) { ... }
    fn layer_destroyed(&mut self, surface: LayerSurface) { ... }
}

Handler for wlr layer shell

Required Methods

fn shell_state(&mut self) -> &mut WlrLayerShellState

WlrLayerShellState getter

fn new_layer_surface(
&mut self,
surface: LayerSurface,
output: Option<WlOutput>,
layer: Layer,
namespace: String
)

A new layer surface was created

You likely need to send a LayerSurfaceConfigure to the surface, to hint the client as to how its layer surface should be sized.

Provided Methods

fn new_popup(&mut self, parent: LayerSurface, popup: PopupSurface)

A new popup was assigned a layer surface as it’s parent

fn ack_configure(&mut self, surface: WlSurface, configure: LayerSurfaceConfigure)

A surface has acknowledged a configure serial.

fn layer_destroyed(&mut self, surface: LayerSurface)

A layer surface was destroyed.