Module smithay::wayland::shell::xdg

Utilities for handling shell surfaces with the xdg_shell protocol

This module provides automatic handling of shell surfaces objects, by being registered as a global handler for xdg_shell.

Why use this implementation

This implementation can track for you the various shell surfaces defined by the clients by handling the xdg_shell protocol. It also contains a compatibility layer handling its precursor, the unstable zxdg_shell_v6 protocol, which is mostly identical.

It allows you to easily access a list of all shell surfaces defined by your clients access their associated metadata and underlying wl_surfaces.

This handler only handles the protocol exchanges with the client to present you the information in a coherent and relatively easy to use manner. All the actual drawing and positioning logic of windows is out of its scope.

How to use it

Initialization

To initialize this handler, simple use the xdg_shell_init function provided in this module. You will need to provide it the CompositorToken you retrieved from an instantiation of the compositor global provided by smithay.

use smithay::wayland::compositor::roles::*;
use smithay::wayland::compositor::CompositorToken;
use smithay::wayland::shell::xdg::{xdg_shell_init, XdgToplevelSurfaceRole, XdgPopupSurfaceRole, XdgRequest};
use wayland_protocols::unstable::xdg_shell::v6::server::zxdg_shell_v6::ZxdgShellV6;

// define the roles type. You need to integrate the XdgSurface role:
define_roles!(MyRoles =>
    [XdgToplevelSurface, XdgToplevelSurfaceRole]
    [XdgPopupSurface, XdgPopupSurfaceRole]
);

// define the metadata you want associated with the shell clients
#[derive(Default)]
struct MyShellData {
    /* ... */
}

let (shell_state, _, _) = xdg_shell_init(
    &mut display,
    // token from the compositor implementation
    compositor_token,
    // your implementation
    |event: XdgRequest<_>| { /* ... */ },
    None  // put a logger if you want
);

// You're now ready to go!

Access to shell surface and clients data

There are mainly 3 kind of objects that you’ll manipulate from this implementation:

ShellClient: This is a handle representing an instantiation of a shell global you can associate client-wise metadata to it (this is the MyShellData type in the example above).

ToplevelSurface: This is a handle representing a toplevel surface, you can retrieve a list of all currently alive toplevel surface from the ShellState.

PopupSurface: This is a handle representing a popup/tooltip surface. Similarly, you can get a list of all currently alive popup surface from the ShellState.

You’ll obtain these objects though two means: either via the callback methods of the subhandler you provided, or via methods on the ShellState that you are given (in an Arc<Mutex<_>>) as return value of the init function.

Structs

PopupConfigure
A configure message for popup surface

PopupState
Represents the state of the popup

PopupSurface
A handle to a popup surface

PositionerState
The state of a positioner, as set by the client

ShellClient
A shell client

ShellState
Shell global state

ToplevelClientPending
Represents the client pending state

ToplevelConfigure
A configure message for toplevel surfaces

ToplevelState
State of a regular toplevel surface

ToplevelStateSet
Container holding the states for a XdgToplevel

ToplevelSurface
A handle to a toplevel surface

XdgPopupSurfaceRoleAttributes
Popup Surface Role

XdgToplevelSurfaceRoleAttributes
Role specific attributes for xdg_toplevel

Enums

Configure
Defines the possible configure variants for a XdgSurface that will be issued in the user_impl for notifying about a ack_configure

PendingStateError
Represents the possible errors returned from ToplevelSurface::with_pending_state and PopupSurface::with_pending_state

XdgRequest
Events generated by xdg shell surfaces

Functions

xdg_shell_init
Create a new xdg_shell globals

Type Definitions

XdgPopupSurfaceRole
A popup surface is a short-lived, temporary surface. It can be used to implement for example menus, popovers, tooltips and other similar user interface concepts.

XdgToplevelSurfaceRole
This interface defines an xdg_surface role which allows a surface to, among other things, set window-like properties such as maximize, fullscreen, and minimize, set application-specific metadata like title and id, and well as trigger user interactive operations such as interactive resize and move.