Module smithay::wayland::shell::xdg

Utilities for handling shell surfaces with the xdg_shell protocol

This module provides automatic handling of shell surfaces objects, by being registered as a global handler for xdg_shell.

Why use this implementation

This implementation can track for you the various shell surfaces defined by the clients by handling the xdg_shell protocol. It also contains a compatibility layer handling its precursor, the unstable zxdg_shell_v6 protocol, which is mostly identical.

It allows you to easily access a list of all shell surfaces defined by your clients access their associated metadata and underlying wl_surfaces.

This handler only handles the protocol exchanges with the client to present you the information in a coherent and relatively easy to use manner. All the actual drawing and positioning logic of windows is out of its scope.

How to use it

Initialization

To initialize this handler, simple use the xdg_shell_init function provided in this module. You will need to provide it the CompositorToken you retrieved from an instantiation of the compositor global provided by smithay.

use smithay::wayland::compositor::roles::*;
use smithay::wayland::compositor::CompositorToken;
use smithay::wayland::shell::xdg::{xdg_shell_init, XdgSurfaceRole, XdgRequest};
use wayland_protocols::unstable::xdg_shell::v6::server::zxdg_shell_v6::ZxdgShellV6;

// define the roles type. You need to integrate the XdgSurface role:
define_roles!(MyRoles =>
    [XdgSurface, XdgSurfaceRole]
);

// define the metadata you want associated with the shell clients
#[derive(Default)]
struct MyShellData {
    /* ... */
}

let (shell_state, _, _) = xdg_shell_init(
    &mut display,
    // token from the compositor implementation
    compositor_token,
    // your implementation
    |event: XdgRequest<_>| { /* ... */ },
    None  // put a logger if you want
);

// You're now ready to go!

Access to shell surface and clients data

There are mainly 3 kind of objects that you'll manipulate from this implementation:

ShellClient: This is a handle representing an instantiation of a shell global you can associate client-wise metadata to it (this is the MyShellData type in the example above).

ToplevelSurface: This is a handle representing a toplevel surface, you can retrieve a list of all currently alive toplevel surface from the ShellState.

PopupSurface: This is a handle representing a popup/tooltip surface. Similarly, you can get a list of all currently alive popup surface from the ShellState.

You'll obtain these objects though two means: either via the callback methods of the subhandler you provided, or via methods on the ShellState that you are given (in an Arc<Mutex<_>>) as return value of the init function.

Structs

PopupConfigure
A configure message for popup surface

PopupState
The pending state of a popup surface

PopupSurface
A handle to a popup surface

PositionerState
The state of a positioner, as set by the client

ShellClient
A shell client

ShellState
Shell global state

ToplevelConfigure
A configure message for toplevel surfaces

ToplevelState
State of a regular toplevel surface

ToplevelSurface
A handle to a toplevel surface

XdgSurfaceRole
Metadata associated with the xdg_surface role

Enums

XdgRequest
Events generated by xdg shell surfaces

XdgSurfacePendingState
Contents of the pending state of a shell surface, depending on its role

Functions

xdg_shell_init
Create a new xdg_shell globals