Module smithay::wayland::shell::xdg

Utilities for handling shell surfaces with the xdg_shell protocol

This module provides automatic handling of shell surfaces objects, by being registered as a global handler for xdg_shell.

Why use this implementation

This implementation can track for you the various shell surfaces defined by the clients by handling the xdg_shell protocol. It also contains a compatibility layer handling its precursor, the unstable zxdg_shell_v6 protocol, which is mostly identical.

It allows you to easily access a list of all shell surfaces defined by your clients access their associated metadata and underlying wl_surfaces.

This handler only handles the protocol exchanges with the client to present you the information in a coherent and relatively easy to use manner. All the actual drawing and positioning logic of windows is out of its scope.

How to use it

Initialization

To initialize this handler, simple use the xdg_shell_init function provided in this module. You need to provide a closure that will be invoked whenever some action is required from you, are represented by the XdgRequest enum.

use smithay::wayland::shell::xdg::{xdg_shell_init, XdgRequest};

let (shell_state, _) = xdg_shell_init(
    &mut display,
    // your implementation
    |event: XdgRequest, dispatch_data| { /* handle the shell requests here */ },
    None  // put a logger if you want
);

// You're now ready to go!

Access to shell surface and clients data

There are mainly 3 kind of objects that you’ll manipulate from this implementation:

ShellClient: This is a handle representing an instantiation of a shell global you can associate client-wise metadata to it through an UserDataMap.

ToplevelSurface: This is a handle representing a toplevel surface, you can retrieve a list of all currently alive toplevel surface from the ShellState.

PopupSurface: This is a handle representing a popup/tooltip surface. Similarly, you can get a list of all currently alive popup surface from the ShellState.

You’ll obtain these objects though two means: either via the callback methods of the subhandler you provided, or via methods on the ShellState that you are given (in an Arc<Mutex<_>>) as return value of the init function.

Modules

decoration

XDG Window decoration manager

Structs

PopupConfigure

A configure message for popup surface

PopupState

Represents the state of the popup

PopupSurface

A handle to a popup surface

PositionerState

The state of a positioner, as set by the client

ShellClient

A shell client

ShellState

Shell global state

SurfaceCachedState

Represents the client pending state

ToplevelConfigure

A configure message for toplevel surfaces

ToplevelState

State of a regular toplevel surface

ToplevelStateSet

Container holding the states for a XdgToplevel

ToplevelSurface

A handle to a toplevel surface

XdgPopupSurfaceRoleAttributes

Role specific attributes for xdg_popup

XdgToplevelSurfaceRoleAttributes

Role specific attributes for xdg_toplevel

Enums

Configure

Defines the possible configure variants for a XdgSurface that will be issued in the user_impl for notifying about a ack_configure

PopupConfigureError

Represents the possible errors that can be returned from PopupSurface::send_configure

XdgRequest

Events generated by xdg shell surfaces

Constants

XDG_POPUP_ROLE

The role of an XDG popup surface.

XDG_TOPLEVEL_ROLE

The role of an XDG toplevel surface.

ZXDG_POPUP_ROLE

The role of an ZXDG popup surface.

ZXDG_TOPLEVEL_ROLE

The role of an ZXDG toplevel surface.

Functions

xdg_shell_init

Create a new xdg_shell global