smithay::wayland::shell::xdg

Module decoration

Available on crate feature wayland_frontend only.

XDG Window decoration manager

This interface allows a compositor to announce support for server-side decorations.

A client can use this protocol to request being decorated by a supporting compositor.

use smithay::{delegate_xdg_decoration, delegate_xdg_shell};
use smithay::wayland::shell::xdg::{ToplevelSurface, XdgShellHandler};
use smithay::wayland::shell::xdg::decoration::{XdgDecorationState, XdgDecorationHandler};
use smithay::reexports::wayland_protocols::xdg::decoration::zv1::server::zxdg_toplevel_decoration_v1::Mode;

// Create a decoration state
let decoration_state = XdgDecorationState::new::<State>(
    &display.handle(),
);

// store that state inside your compositor state
// ...

// implement the necessary traits
impl XdgShellHandler for State {
    // ...
}
impl XdgDecorationHandler for State {
    fn new_decoration(&mut self, toplevel: ToplevelSurface) {
        toplevel.with_pending_state(|state| {
            // Advertise server side decoration
            state.decoration_mode = Some(Mode::ServerSide);
        });
        toplevel.send_configure();
    }
    fn request_mode(&mut self, toplevel: ToplevelSurface, mode: Mode) { /* ... */ }
    fn unset_mode(&mut self, toplevel: ToplevelSurface) { /* ... */ }
}

use smithay::input::{Seat, SeatState, SeatHandler, pointer::CursorImageStatus};

type Target = wl_surface::WlSurface;
impl SeatHandler for State {
    type KeyboardFocus = Target;
    type PointerFocus = Target;
    type TouchFocus = Target;

    fn seat_state(&mut self) -> &mut SeatState<Self> {
        &mut self.seat_state
    }

    fn focus_changed(&mut self, seat: &Seat<Self>, focused: Option<&Target>) {
        // handle focus changes, if you need to ...
    }
    fn cursor_image(&mut self, seat: &Seat<Self>, image: CursorImageStatus) {
        // handle new images for the cursor ...
    }
}
delegate_xdg_shell!(State);
delegate_xdg_decoration!(State);

// You are ready to go!  

Structs

XdgDecorationManagerGlobalData

Data associated with a XdgDecorationManager global.

XdgDecorationState

Delegate type for handling xdg decoration events.

Traits

XdgDecorationHandler

Handler trait for xdg decoration events.