Module smithay::wayland::shell::xdg::decoration

XDG Window decoration manager

This interface allows a compositor to announce support for server-side decorations.

A client can use this protocol to request being decorated by a supporting compositor.

use smithay::{delegate_xdg_decoration, delegate_xdg_shell};
use smithay::wayland::shell::xdg::{ToplevelSurface, XdgShellHandler};
use smithay::wayland::shell::xdg::decoration::{XdgDecorationState, XdgDecorationHandler};
use smithay::reexports::wayland_protocols::xdg::decoration::zv1::server::zxdg_toplevel_decoration_v1::Mode;

// Create a decoration state
let decoration_state = XdgDecorationState::new::<State>(
    &display.handle(),
);

// store that state inside your compositor state
// ...

// implement the necessary traits
impl XdgShellHandler for State {
    // ...
}
impl XdgDecorationHandler for State {
    fn new_decoration(&mut self, toplevel: ToplevelSurface) {
        toplevel.with_pending_state(|state| {
            // Advertise server side decoration
            state.decoration_mode = Some(Mode::ServerSide);
        });
        toplevel.send_configure();
    }
    fn request_mode(&mut self, toplevel: ToplevelSurface, mode: Mode) { /* ... */ }
    fn unset_mode(&mut self, toplevel: ToplevelSurface) { /* ... */ }
}
delegate_xdg_shell!(State);
delegate_xdg_decoration!(State);

// You are ready to go!  

Structs

XdgDecorationState

Delegate type for handling xdg decoration events.

Traits

XdgDecorationHandler

Handler trait for xdg decoration events.