Module smithay::wayland::shell::xdg::decoration

XDG Window decoration manager

This interface allows a compositor to announce support for server-side decorations.

A client can use this protocol to request being decorated by a supporting compositor.

use smithay::wayland::shell::xdg::decoration::{init_xdg_decoration_manager, XdgDecorationRequest};
use smithay::reexports::wayland_protocols::unstable::xdg_decoration::v1::server::zxdg_toplevel_decoration_v1::Mode;

init_xdg_decoration_manager(
    &mut display,
    |req, _ddata| match req {
        XdgDecorationRequest::NewToplevelDecoration { toplevel } => {
            let res = toplevel.with_pending_state(|state| {
                  // Advertise server side decoration
                state.decoration_mode = Some(Mode::ServerSide);
            });

            if res.is_ok() {
                toplevel.send_configure();
            }
        }
        XdgDecorationRequest::SetMode { .. } => {}
        XdgDecorationRequest::UnsetMode { .. } => {}
    },
    None,
);

Enums

XdgDecorationRequest

Events generated by xdg decoration manager

Functions

init_xdg_decoration_manager

Create a new XDG Decoration Manager global