smithay::wayland::shell::xdg::decoration

Trait XdgDecorationHandler

pub trait XdgDecorationHandler {
    // Required methods
    fn new_decoration(&mut self, toplevel: ToplevelSurface);
    fn request_mode(&mut self, toplevel: ToplevelSurface, mode: Mode);
    fn unset_mode(&mut self, toplevel: ToplevelSurface);
}

Available on crate feature wayland_frontend only.

Handler trait for xdg decoration events.

Required Methods

fn new_decoration(&mut self, toplevel: ToplevelSurface)

Notification the client supports server side decoration on the toplevel.

fn request_mode(&mut self, toplevel: ToplevelSurface, mode: Mode)

Notification the client prefers the provided decoration decoration mode on the toplevel.

fn unset_mode(&mut self, toplevel: ToplevelSurface)

Notification the client does not prefer a particular decoration mode on the toplevel.