Enum smithay::wayland::shell::xdg::XdgRequest

pub enum XdgRequest {

    NewClient {
        client: ShellClient,
    },
    ClientPong {
        client: ShellClient,
    },
    NewToplevel {
        surface: ToplevelSurface,
    },
    NewPopup {
        surface: PopupSurface,
        positioner: PositionerState,
    },
    Move {
        surface: ToplevelSurface,
        seat: WlSeat,
        serial: Serial,
    },
    Resize {
        surface: ToplevelSurface,
        seat: WlSeat,
        serial: Serial,
        edges: ResizeEdge,
    },
    Grab {
        surface: PopupSurface,
        seat: WlSeat,
        serial: Serial,
    },
    Maximize {
        surface: ToplevelSurface,
    },
    UnMaximize {
        surface: ToplevelSurface,
    },
    Fullscreen {
        surface: ToplevelSurface,
        output: Option<WlOutput>,
    },
    UnFullscreen {
        surface: ToplevelSurface,
    },
    Minimize {
        surface: ToplevelSurface,
    },
    ShowWindowMenu {
        surface: ToplevelSurface,
        seat: WlSeat,
        serial: Serial,
        location: Point<i32, Logical>,
    },
    AckConfigure {
        surface: WlSurface,
        configure: Configure,
    },
    RePosition {
        surface: PopupSurface,
        positioner: PositionerState,
        token: u32,
    },

}

Events generated by xdg shell surfaces

These are events that the provided implementation cannot process for you directly.

Depending on what you want to do, you might ignore some of them

Variants

NewClient

A new shell client was instantiated

Fields of NewClient

client: ShellClient

the client

ClientPong

The pong for a pending ping of this shell client was received

The ShellHandler already checked for you that the serial matches the one from the pending ping.

Fields of ClientPong

client: ShellClient

the client

NewToplevel

A new toplevel surface was created

You likely need to send a ToplevelConfigure to the surface, to hint the client as to how its window should be sized.

Fields of NewToplevel

surface: ToplevelSurface

the surface

NewPopup

A new popup surface was created

You likely need to send a PopupConfigure to the surface, to hint the client as to how its popup should be sized.

Fields of NewPopup

surface: PopupSurface

the surface

positioner: PositionerState

The state of the positioner at the time the popup was requested.

The positioner state can be used by the compositor to calculate the best placement for the popup.

For example the compositor should prevent that a popup is placed outside the visible rectangle of a output.

Move

The client requested the start of an interactive move for this surface

Fields of Move

surface: ToplevelSurface

the surface

seat: WlSeat

the seat associated to this move

serial: Serial

the grab serial

Resize

The client requested the start of an interactive resize for this surface

Fields of Resize

surface: ToplevelSurface

The surface

seat: WlSeat

The seat associated with this resize

serial: Serial

The grab serial

edges: ResizeEdge

Specification of which part of the window’s border is being dragged

Grab

This popup requests a grab of the pointer

This means it requests to be sent a popup_done event when the pointer leaves the grab area.

Fields of Grab

surface: PopupSurface

The surface

seat: WlSeat

The seat to grab

serial: Serial

The grab serial

Maximize

A toplevel surface requested to be maximized

Fields of Maximize

surface: ToplevelSurface

The surface

UnMaximize

A toplevel surface requested to stop being maximized

Fields of UnMaximize

surface: ToplevelSurface

The surface

Fullscreen

A toplevel surface requested to be set fullscreen

Fields of Fullscreen

surface: ToplevelSurface

The surface

output: Option<WlOutput>

The output (if any) on which the fullscreen is requested

UnFullscreen

A toplevel surface request to stop being fullscreen

Fields of UnFullscreen

surface: ToplevelSurface

The surface

Minimize

A toplevel surface requested to be minimized

Fields of Minimize

surface: ToplevelSurface

The surface

ShowWindowMenu

The client requests the window menu to be displayed on this surface at this location

This menu belongs to the compositor. It is typically expected to contain options for control of the window (maximize/minimize/close/move/etc…).

Fields of ShowWindowMenu

surface: ToplevelSurface

The surface

seat: WlSeat

The seat associated with this input grab

serial: Serial

the grab serial

location: Point<i32, Logical>

location of the menu request relative to the surface geometry

AckConfigure

A surface has acknowledged a configure serial.

Fields of AckConfigure

surface: WlSurface

The surface.

configure: Configure

The configure serial.

RePosition

A client requested a reposition, providing a new positioner, of a popup.

Fields of RePosition

surface: PopupSurface

The popup for which a reposition has been requested

positioner: PositionerState

The state of the positioner at the time the reposition request was made.

The positioner state can be used by the compositor to calculate the best placement for the popup.

For example the compositor should prevent that a popup is placed outside the visible rectangle of a output.

token: u32

The passed token will be sent in the corresponding xdg_popup.repositioned event. The new popup position will not take effect until the corresponding configure event is acknowledged by the client. See xdg_popup.repositioned for details. The token itself is opaque, and has no other special meaning.