Struct smithay::wayland::shell::xdg::PopupSurface

pub struct PopupSurface { /* fields omitted */ }

A handle to a popup surface

This is an unified abstraction over the popup surfaces of both wl_shell and xdg_shell.

Implementations

impl PopupSurface

pub fn alive(&self) -> bool

Is the popup surface referred by this handle still alive?

pub fn get_parent_surface(&self) -> Option<WlSurface>

Gets a reference of the parent WlSurface of this popup.

pub fn client(&self) -> Option<ShellClient>

Retrieve the shell client owning this popup surface

Returns None if the surface does actually no longer exist.

pub fn send_configure(&self) -> Result<(), PopupConfigureError>

Send a configure event to this popup surface to suggest it a new configuration

The serial of this configure will be tracked waiting for the client to ACK it.

You can manipulate the state that will be sent to the client with the with_pending_state method.

Returns [Err(PopupConfigureError)] if the initial configure has already been sent and the client protocol version disallows a re-configure or the current PositionerState is not reactive

pub fn send_repositioned(&self, token: u32)

Send a configure event, including the repositioned event to the client in response to a reposition request.

For further information see [XdgPopup::send_configure]

pub fn ensure_configured(&self) -> bool

Make sure this surface was configured

Returns true if it was, if not, returns false and raise a protocol error to the associated client. Also returns false if the surface is already destroyed.

xdg_shell mandates that a client acks a configure before committing anything.

pub fn send_popup_done(&self)

Send a popup_done event to the popup surface

It means that the use has dismissed the popup surface, or that the pointer has left the area of popup grab if there was a grab.

pub fn get_surface(&self) -> Option<&WlSurface>

Access the underlying wl_surface of this toplevel surface

Returns None if the popup surface actually no longer exists.

pub fn with_pending_state<F, T>(&self, f: F) -> Result<T, DeadResource> where
F: FnOnce(&mut PopupState) -> T,

Allows the pending state of this popup to be manipulated.

This should be used to inform the client about size and position changes, for example after a move of the parent toplevel.

The state will be sent to the client when calling send_configure.