Struct smithay::wayland::shell::xdg::ShellClient

pub struct ShellClient<R> { /* fields omitted */ }

A shell client

This represents an instantiation of a shell global (be it wl_shell or xdg_shell).

Most of the time, you can consider that a Wayland client will be a single shell client.

You can use this handle to access a storage for any client-specific data you wish to associate with it.

Methods

impl<R> ShellClient<R> where
R: Role<XdgSurfaceRole> + 'static,

pub fn alive(&self) -> bool

Is the shell client represented by this handle still connected?

pub fn equals(&self, other: &Self) -> bool

Checks if this handle and the other one actually refer to the same shell client

pub fn send_ping(&self, serial: u32) -> Result<(), ()>

Send a ping request to this shell client

You'll receive the reply as a XdgRequest::ClientPong request.

A typical use is to start a timer at the same time you send this ping request, and cancel it when you receive the pong. If the timer runs down to 0 before a pong is received, mark the client as unresponsive.

Fails if this shell client already has a pending ping or is already dead.

pub fn with_data<F, T>(&self, f: F) -> Result<T, ()> where
F: FnOnce(&mut UserDataMap) -> T,

Access the user data associated with this shell client