Struct smithay::wayland::shell::xdg::XdgToplevelSurfaceRoleAttributes

pub struct XdgToplevelSurfaceRoleAttributes {

    pub window_geometry: Option<Rectangle>,
    pub client_pending_window_geometry: Option<Rectangle>,
    pub configured: bool,
    pub configure_serial: Option<Serial>,
    pub initial_configure_sent: bool,
    pub parent: Option<WlSurface>,
    pub title: Option<String>,
    pub app_id: Option<String>,
    pub min_size: (i32, i32),
    pub max_size: (i32, i32),
    pub client_pending: Option<ToplevelClientPending>,
    pub server_pending: Option<ToplevelState>,
    pub last_acked: Option<ToplevelState>,
    pub current: ToplevelState,
    // some fields omitted

}

Role specific attributes for xdg_toplevel

Fields
window_geometry: Option<Rectangle>

Defines the current window geometry of the surface if any.

client_pending_window_geometry: Option<Rectangle>

Holds the double-buffered geometry that may be specified by xdg_surface.set_window_geometry. This should be moved to the current configuration on a commit of the underlying wl_surface.

configured: bool

Defines if the surface has received at least one xdg_surface.ack_configure from the client

configure_serial: Option<Serial>

The serial of the last acked configure

initial_configure_sent: bool

Holds the state if the surface has sent the initial configure event to the client. It is expected that during the first commit a initial configure event is sent to the client

parent: Option<WlSurface>

The parent field of a toplevel should be used by the compositor to determine which toplevel should be brought to front. If the parent is focused all of it’s child should be brought to front.

title: Option<String>

Holds the optional title the client has set for this toplevel. For example a web-browser will most likely set this to include the current uri.

This string may be used to identify the surface in a task bar, window list, or other user interface elements provided by the compositor.

app_id: Option<String>

Holds the optional app ID the client has set for this toplevel.

The app ID identifies the general class of applications to which the surface belongs. The compositor can use this to group multiple surfaces together, or to determine how to launch a new application.

For D-Bus activatable applications, the app ID is used as the D-Bus service name.

min_size: (i32, i32)

Minimum size requested for this surface

A value of 0 on an axis means this axis is not constrained

max_size: (i32, i32)

Maximum size requested for this surface

A value of 0 on an axis means this axis is not constrained

client_pending: Option<ToplevelClientPending>

Holds the pending state as set by the client.

server_pending: Option<ToplevelState>

Holds the pending state as set by the server.

last_acked: Option<ToplevelState>

Holds the last server_pending state that has been acknowledged by the client. This state should be cloned to the current during a commit.

current: ToplevelState

Holds the current state of the toplevel after a successful commit.