smithay::wayland

Module text_input

Available on crate feature wayland_frontend only.

Utilities for text input support

This module provides you with utilities to handle text input surfaces, it is usually used in conjunction with the input method module.

Text input focus is automatically set to the same surface that has keyboard focus.

use smithay::{
    delegate_seat, delegate_text_input_manager,
};
use smithay::input::{Seat, SeatState, SeatHandler, pointer::CursorImageStatus};
use smithay::wayland::text_input::TextInputManagerState;
use smithay::reexports::wayland_server::{Display, protocol::wl_surface::WlSurface};

delegate_seat!(State);
// Delegate text input handling for State to TextInputManagerState.
delegate_text_input_manager!(State);

let seat_state = SeatState::<State>::new();

// implement the required traits
impl SeatHandler for State {
    type KeyboardFocus = WlSurface;
    type PointerFocus = WlSurface;
    type TouchFocus = WlSurface;
    fn seat_state(&mut self) -> &mut SeatState<Self> {
        &mut self.seat_state
    }
    fn focus_changed(&mut self, seat: &Seat<Self>, focused: Option<&WlSurface>) { unimplemented!() }
    fn cursor_image(&mut self, seat: &Seat<Self>, image: CursorImageStatus) { unimplemented!() }
}

// Add the seat state to your state and create manager global
TextInputManagerState::new::<State>(&display_handle);

Structs

TextInputHandle

Handle to text input instances

TextInputManagerState

State of wp text input protocol

TextInputUserData

User data of ZwpTextInputV3 object

Traits

TextInputSeat

Extends Seat with text input functionality