Module smithay::wayland::viewporter

Utilities for handling the wp_viewporter protocol

How to use it

Initialization

To initialize this implementation, use the viewporter_init method provided by this module.

use smithay::wayland::viewporter::viewporter_init;

// Call the init function:
viewporter_init(&mut display);

// You're now ready to go!

Use the viewport state

The viewport state is double-buffered and can be accessed by using the with_states function

let viewport = with_states(surface, |states| {
    states.cached_state.current::<ViewportCachedState>();
});

Before accessing the state you should call ensure_viewport_valid to ensure the viewport is valid.

Note: If you already hand over buffer management to smithay by using on_commit_buffer_handler the implementation will already call ensure_viewport_valid for you.

Structs

ViewportCachedState

Represents the double-buffered viewport state of a WlSurface

Functions

ensure_viewport_valid

Ensures that the viewport, if any, is valid accordingly to the protocol specification.

viewporter_init

Create new wp_viewporter global.