smithay::wayland

Module xdg_activation

Available on crate feature wayland_frontend only.

Utilities for handling activation requests with the xdg_activation protocol

Example

use wayland_server::{protocol::wl_surface::WlSurface, DisplayHandle};
use smithay::{
    delegate_xdg_activation,
    wayland::xdg_activation::{XdgActivationHandler, XdgActivationState, XdgActivationToken, XdgActivationTokenData}
};

pub struct State {
    activation_state: XdgActivationState,
}

impl XdgActivationHandler for State {
    fn activation_state(&mut self) -> &mut XdgActivationState {
        &mut self.activation_state
    }

    fn request_activation(
        &mut self,
        token: XdgActivationToken,
        token_data: XdgActivationTokenData,
        surface: WlSurface
    ) {
        if token_data.timestamp.elapsed().as_secs() < 10 {
            // Request surface activation
        }
    }
}

// Delegate xdg activation handling for State to XdgActivationState.
delegate_xdg_activation!(State);

let state = State {
    activation_state: XdgActivationState::new::<State>(&display_handle),
};

// Rest of the compositor goes here...

Structs

ActivationTokenData

Data associated with an activation token protocol object.

XdgActivationState

Tracks the list of pending and current activation requests

XdgActivationToken

Contains the unique string token of activation request

XdgActivationTokenData

Activation data asosiated with the XdgActivationToken

Traits

XdgActivationHandler

A trait implemented to be notified of activation requests using the xdg activation protocol.