Trait smithay::wayland::xdg_activation::XdgActivationHandler

pub trait XdgActivationHandler {
    fn activation_state(&mut self) -> &mut XdgActivationState;
    fn request_activation(
        &mut self,
        token: XdgActivationToken,
        token_data: XdgActivationTokenData,
        surface: WlSurface
    );
    fn destroy_activation(
        &mut self,
        token: XdgActivationToken,
        token_data: XdgActivationTokenData,
        surface: WlSurface
    );
}

A trait implemented to be notified of activation requests using the xdg activation protocol.

Required Methods

fn activation_state(&mut self) -> &mut XdgActivationState

Returns the activation state.

fn request_activation(
&mut self,
token: XdgActivationToken,
token_data: XdgActivationTokenData,
surface: WlSurface
)

A client has requested surface activation.

The compositor may know which client requested this by checking the token data and may decide whether or not to follow through with the activation if it’s considered unwanted.

If a request is unwanted, you can discard the request using XdgActivationState::remove_request to ignore any future requests.

fn destroy_activation(
&mut self,
token: XdgActivationToken,
token_data: XdgActivationTokenData,
surface: WlSurface
)

The activation token was destroyed.

The compositor may cancel any activation requests coming from the token.

For example if your compositor blinks or highlights a window when it requests activation then the animation should stop when this function is called.