smithay::wayland

Module xdg_toplevel_icon

Available on crate feature wayland_frontend only.

XDG Toplevel Icon

This protocol allows clients to set icons for their toplevel surfaces either via the XDG icon stock (using an icon name), or from pixel data.

In order to advertise toplevel icon global call XdgToplevelIconManager::new and delegate events to it with delegate_xdg_toplevel_icon. Currently attached icon is available in double-buffered ToplevelIconCachedState

Structs

ToplevelIconCachedState

Data associated with WlSurface Represents the client pending state

XdgToplevelIconManager

Delegate type for handling xdg toplevel icon events.

XdgToplevelIconManagerUserData

User data of XdgToplevelIconManagerV1 global

XdgToplevelIconUserData

User data of XdgToplevelIconV1

Traits

XdgToplevelIconHandler

Handler trait for xdg toplevel icon events.