Module smithay::xwayland

XWayland utilities

This module contains helpers to manage XWayland from your compositor, in order to support running X11 apps.

The starting point is the XWayland struct, which represents the running XWayland instance. Dropping it will shutdown XWayland.

You need to provide an implementation of a X11 Window Manager for XWayland to function properly. You’ll need to treat XWayland (and all its X11 apps) as one special client, and play the role of an X11 Window Manager.

Smithay does not provide any helper for doing that yet, but it is planned.

Re-exports

pub use self::xwm::X11Wm;

pub use self::xwm::XwmHandler;

Modules

xwm

Xwayland Window Manager module

Structs

X11Surface

X11 window managed by an X11Wm

XWayland

The XWayland handle

XWaylandClientData

Inner ClientData-type of an xwayland client

XWaylandSource

An event source for monitoring XWayland status

Enums

XWaylandEvent

Events generated by the XWayland manager