smithay

Module xwayland

Available on crate feature xwayland only.

XWayland utilities

This module contains helpers to manage XWayland from your compositor, in order to support running X11 apps.

The starting point is the XWayland struct, which represents the running XWayland instance. Dropping it will shutdown XWayland.

You need to provide an implementation of a X11 Window Manager for XWayland to function properly. You’ll need to treat XWayland (and all its X11 apps) as one special client, and play the role of an X11 Window Manager.

Smithay does not provide any helper for doing that yet, but it is planned.

Re-exports

pub use self::xwm::X11Surface;

pub use self::xwm::X11Wm;

pub use self::xwm::XwmHandler;

Modules

xwm

Xwayland Window Manager module

Structs

XWayland

A handle to a running XWayland process. Using XWayland as an xserver for X11-based clients requires two connections: one wayland socket, where XWayland creates surfaces for its clients, and one X11 socket, where a compositor’s implementation of an X11 window manager connects and handles X11 events.

XWaylandClientData

Inner ClientData-type of an xwayland client.

Enums

XWaylandEvent

Events generated by an XWayland instance.