Module smithay::xwayland::xwm

Xwayland Window Manager module

Provides an X11Wm type, which will register itself as a window manager for a previously spawned Xwayland instances, allowing backwards-compatibility by seemlessly integrating X11 windows into a wayland compositor.

To use this functionality you must first spawn an XWayland instance to attach a X11Wm to.

struct State { /* ... */ }
impl XwmHandler for State {
    fn xwm_state(&mut self, xwm: XwmId) -> &mut X11Wm {
        // ...
    }
    fn new_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn new_override_redirect_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn map_window_request(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn mapped_override_redirect_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn unmapped_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn destroyed_window(&mut self, xwm: XwmId, window: X11Surface) { /* ... */ }
    fn configure_request(&mut self, xwm: XwmId, window: X11Surface, x: Option<i32>, y: Option<i32>, w: Option<u32>, h: Option<u32>, reorder: Option<Reorder>) { /* ... */ }
    fn configure_notify(&mut self, xwm: XwmId, window: X11Surface, geometry: Rectangle<i32, Logical>, above: Option<u32>) { /* ... */ }
    fn resize_request(&mut self, xwm: XwmId, window: X11Surface, button: u32, resize_edge: ResizeEdge) { /* ... */ }
    fn move_request(&mut self, xwm: XwmId, window: X11Surface, button: u32) { /* ... */ }
    fn send_selection(&mut self, xwm: XwmId, selection: SelectionType, mime_type: String, fd: OwnedFd) { /* ... */ }
}

let (xwayland, channel) = XWayland::new(&dh);
let ret = handle.insert_source(channel, move |event, _, data| match event {
    XWaylandEvent::Ready {
        connection,
        client,
        client_fd: _,
        display: _,
    } => {
        let wm = X11Wm::start_wm(
            handle.clone(),
            dh.clone(),
            connection,
            client,
        )
        .expect("Failed to attach X11 Window Manager");
         
        // store the WM somewhere
    }
    XWaylandEvent::Exited => {
        // cleanup your state and drop the WM again
    }
});
if let Err(e) = ret {
    tracing::error!(
        "Failed to insert the XWaylandSource into the event loop: {}", e
    );
}

Structs

Atoms

Atoms used by the XWM and X11Surface types

X11Surface

X11 window managed by an X11Wm

X11Wm

The runtime state of an reparenting XWayland window manager.

XwmId

Id of an X11 WM

Enums

Reorder

Window asks to be re-stacked

ResizeEdge

Edge values for resizing

SelectionError

Errors generated working with Xwm Selections

SelectionType

Type of a given Selection

WmWindowType

Window types of X11Surfaces

X11SurfaceError

Errors that can happen for operations on an X11Surface

Constants

X11_SURFACE_ROLE

X11 wl_surface role

Traits

X11Relatable

Trait for objects, that represent an x11 window in some shape or form and can be tested for equality.

XwmHandler

Handler trait for X11Wm interactions