Struct smithay::xwayland::xwm::X11Surface

pub struct X11Surface { /* private fields */ }

X11 window managed by an X11Wm

Implementations

impl X11Surface

pub fn new( xwm: impl Into<Option<XwmId>>, window: u32, override_redirect: bool, conn: Weak<RustConnection>, atoms: Atoms, geometry: Rectangle<i32, Logical> ) -> X11Surface

Create a new X11Surface usually handled by an [X11Wm]

Arguments

window X11 window id

override_redirect set if the X11 window has the override redirect flag set

conn Weak reference on the X11 connection

atoms Atoms struct as defined by the xwm module.

geometry Initial geometry of the window

pub fn xwm_id(&self) -> Option<XwmId>

Returns the id of the X11Wm responsible for this surface, if any

pub fn window_id(&self) -> X11Window

X11 protocol id of the underlying window

pub fn mapped_window_id(&self) -> Option<X11Window>

X11 protocol id of the reparented window, if any

pub fn set_mapped(&self, mapped: bool) -> Result<(), X11SurfaceError>

Set the X11 windows as mapped/unmapped affecting its visibility.

It is an error to call this function on override redirect windows

pub fn is_override_redirect(&self) -> bool

Returns if this window has the override redirect flag set or not

pub fn is_mapped(&self) -> bool

Returns if the window is currently mapped or not

pub fn alive(&self) -> bool

Returns if the window is still alive

pub fn configure( &self, rect: impl Into<Option<Rectangle<i32, Logical>>> ) -> Result<(), X11SurfaceError>

Send a configure to this window.

If rect is provided the new state will be send to the window. If rect is None a synthetic configure event with the existing state will be send.

pub fn wl_surface(&self) -> Option<WlSurface>

Returns the wl_surface corresponding to the underlying X11 window.

An X11 window only has a matching wl_surface if is considered visible by Xwayland, which will only happen once it is mapped.

pub fn geometry(&self) -> Rectangle<i32, Logical>

Returns the current geometry of the underlying X11 window

pub fn title(&self) -> String

Returns the current title of the underlying X11 window

pub fn class(&self) -> String

Returns the current window class of the underlying X11 window

pub fn instance(&self) -> String

Returns the current window instance of the underlying X11 window

pub fn is_popup(&self) -> bool

Returns if the window is considered to be a popup.

Corresponds to the internal _NET_WM_STATE_MODAL state of the underlying X11 window.

pub fn is_transient_for(&self) -> Option<X11Window>

Returns if the underlying window is transient to another window.

This might be used as a hint to manage windows in a group.

pub fn size_hints(&self) -> Option<WmSizeHints>

Returns the size hints for the underlying X11 window

pub fn min_size(&self) -> Option<Size<i32, Logical>>

Returns the suggested minimum size of the underlying X11 window

pub fn max_size(&self) -> Option<Size<i32, Logical>>

Returns the suggested minimum size of the underlying X11 window

pub fn base_size(&self) -> Option<Size<i32, Logical>>

Returns the suggested base size of the underlying X11 window

pub fn is_maximized(&self) -> bool

Returns if the window is in the maximized state

pub fn is_fullscreen(&self) -> bool

Returns if the window is in the fullscreen state

pub fn is_minimized(&self) -> bool

Returns if the window is in the minimized state

pub fn is_activated(&self) -> bool

Returns if the window is in the activated state

pub fn is_decorated(&self) -> bool

Returns true if the window is client-side decorated

pub fn set_maximized(&self, maximized: bool) -> Result<(), ConnectionError>

Sets the window as maximized or not.

Allows the client to reflect this state in their UI.

pub fn set_fullscreen(&self, fullscreen: bool) -> Result<(), ConnectionError>

Sets the window as fullscreen or not.

Allows the client to reflect this state in their UI.

pub fn set_minimized(&self, minimized: bool) -> Result<(), ConnectionError>

Sets the window as minimized or not.

Allows the client to e.g. stop rendering while minimized.

pub fn set_activated(&self, activated: bool) -> Result<(), ConnectionError>

Sets the window as activated or not.

Allows the client to reflect this state in their UI.

pub fn window_type(&self) -> Option<WmWindowType>

Returns the reported window type of the underlying X11 window if set.

Windows without a window type set should be considered to be of type Normal for backwards compatibility.

pub fn user_data(&self) -> &UserDataMap

Retrieve user_data associated with this X11 window

pub fn close(&self) -> Result<(), ConnectionError>

Send a close request to this window.

Will outright destroy windows that don’t support the NET_DELETE_WINDOW protocol.