Struct smithay::xwayland::xwm::X11Wm

pub struct X11Wm { /* private fields */ }

The runtime state of an reparenting XWayland window manager.

Implementations

impl X11Wm

pub fn start_wm<D>( handle: LoopHandle<'static, D>, dh: DisplayHandle, connection: UnixStream, client: Client ) -> Result<Self, Box<dyn Error>>where D: XwmHandler + 'static,

Start a new window manager for a given Xwayland connection

Arguments

handle is an eventloop handle used to queue up and handle incoming X11 events

dh is the corresponding display handle to the wayland connection of the Xwayland instance

connection is the corresponding x11 client connection of the Xwayland instance

client is the wayland client instance of the Xwayland instance

pub fn id(&self) -> XwmId

Id of this X11 WM

pub fn raise_window<'a, W: X11Relatable + 'a>( &mut self, window: &'a W ) -> Result<(), ConnectionError>

Raises a window in the internal X11 state

Needs to be called to match raising of windows inside the compositor to keep the stacking order in sync with the compositor to avoid erroneous behavior.

pub fn update_stacking_order_downwards<'a, W: X11Relatable + 'a>( &mut self, order: impl Iterator<Item = &'a W> ) -> Result<(), ConnectionError>

Updates the stacking order by matching provided windows downwards.

This function reorders provided x11 windows in such a way, that windows inside the internal X11 stack follow the provided order without moving other windows around as much as possible. The internal stack stores windows bottom -> top, so order here is also bottom -> top.

Window IDs unknown to this XWM will be ignored. The first window in order found will not be moved.

If a window is encountered in order that is stacked below the first window, it will be moved to be directly above the previous window in order. E.g. Windows C -> A -> B -> E given in order with an internal stack of D -> A -> B -> C, will be reordered as D -> C -> A -> B. First A is moved to be directly above C, then Bis moved to be directly above A.

Windows in the internal stack, that are not present in order will be skipped over in the process.

So if windows A -> C are given in order and the internal stack is A -> B -> C, no reordering will occur.

See [X11Surface::update_stacking_order_upwards] for a variant of this algorithm, which works from the bottom up or [X11Surface::raise_window] for an easier but much more limited way to reorder.

pub fn update_stacking_order_upwards<'a, W: X11Relatable + 'a>( &mut self, order: impl Iterator<Item = &'a W> ) -> Result<(), ConnectionError>

Updates the stacking order by moving provided windows upwards.

This function reorders provided x11 windows in such a way, that windows inside the internal X11 stack follow the provided order in reverse without moving other windows around as much as possible. The internal stack stores windows bottom -> top, so due to the reversal, order here is top -> bottom.

Window IDs unknown to this XWM will be ignored. The first window in order found will not be moved.

If a window is encountered in order that is stacked above the first window, it will be moved to be directly below the previous window in order. E.g. Windows C -> A -> B given in order with an internal stack of D -> A -> B -> C, will be reordered as D -> B -> A -> C. A is below C, so it isn’t moved, then B is moved to be directly below A.

Windows in the internal stack, that are not present in order will be skipped over in the process.

So if windows A -> C are given in order and the internal stack is C -> B -> A, no reordering will occur.

See [X11Surface::update_stacking_order_downwards] for a variant of this algorithm, which works from the top down or [X11Surface::raise_window] for an easier but much more limited way to reorder.

pub fn commit_hook<D: XwmHandler + 'static>(surface: &WlSurface)

This function has to be called on [CompositorState::commit] to correctly update the internal state of Xwayland WMs.

pub fn set_cursor( &mut self, pixels: &[u8], size: Size<u16, Logical>, hotspot: Point<u16, Logical> ) -> Result<(), ReplyOrIdError>

Set the default cursor used by X clients.

pixels is expected to be in rgba-format with each channel encoded as an u8.

This function will panic, if pixels is not at least size.w * size.h * 4 long.

pub fn new_selection( &mut self, selection: SelectionType, mime_types: Option<Vec<String>> ) -> Result<(), ReplyOrIdError>

Notify Xwayland of a new selection.

mime_types being None indicate there is no active selection anymore.

pub fn send_selection<D>( &mut self, selection: SelectionType, mime_type: String, fd: OwnedFd, loop_handle: LoopHandle<'_, D> ) -> Result<(), SelectionError>where D: XwmHandler + 'static,

Request to transfer the active selection for the provided mime_type to the provided file descriptor.