Trait smithay::xwayland::xwm::X11Relatable

pub trait X11Relatable {
    // Required method
    fn is_window(&self, window: &X11Surface) -> bool;
}

Available on crate feature xwayland only.

Trait for objects, that represent an x11 window in some shape or form and can be tested for equality.

Required Methods

fn is_window(&self, window: &X11Surface) -> bool

Returns if this object is considered to represent the same underlying x11 window as provided